ACAPS government measures catalogs worldwide COVID-19 government interventions

ACAPS government measures catalogs worldwide COVID-19 government interventions

acaps_government_measures_data()

Value

a data.frame

Details

The #COVID19 Government Measures Dataset puts together all the measures implemented by governments worldwide in response to the Coronavirus pandemic. Data collection includes secondary data review. The researched information available falls into five categories:

Social distancing

Movement restrictions

Public health measures

Social and economic measures

Lockdowns

Each category is broken down into several types of measures.

ACAPS government measures catalogs worldwide COVID-19 government interventionsACAPS consulted government, media, United Nations, and other organisations sources.

References

https://www.acaps.org/covid19-government-measures-dataset

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = acaps_government_measures_data()
head(res)
#> # A tibble: 6 × 18
#>      id iso   country    region admin_level_name pcode log_type category measure
#>   <dbl> <chr> <chr>      <chr>  <chr>            <lgl> <chr>    <chr>    <chr>  
#> 1  4245 AFG   Afghanist… Asia   NA               NA    Introdu… Public … Awaren…
#> 2  4246 AFG   Afghanist… Asia   NA               NA    Introdu… Public … Health…
#> 3  4247 AFG   Afghanist… Asia   NA               NA    Introdu… Movemen… Intern…
#> 4  4248 AFG   Afghanist… Asia   NA               NA    Introdu… Public … Health…
#> 5    23 AFG   Afghanist… Asia   NA               NA    Introdu… Movemen… Border…
#> 6  4249 AFG   Afghanist… Asia   Kabul            NA    Introdu… Public … Streng…
#> # … with 9 more variables: targeted_pop_group <chr>, comments <chr>,
#> #   non_compliance <chr>, date_implemented <date>, source <chr>,
#> #   source_type <chr>, link <chr>, entry_date <date>,
#> #   `alternative source` <chr>
colnames(res)
#>  [1] "id"                 "iso"                "country"           
#>  [4] "region"             "admin_level_name"   "pcode"             
#>  [7] "log_type"           "category"           "measure"           
#> [10] "targeted_pop_group" "comments"           "non_compliance"    
#> [13] "date_implemented"   "source"             "source_type"       
#> [16] "link"               "entry_date"         "alternative source"
dplyr::glimpse(res)
#> Rows: 23,923
#> Columns: 18
#> $ id                   <dbl> 4245, 4246, 4247, 4248, 23, 4249, 4250, 4252, 425…
#> $ iso                  <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", …
#> $ country              <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afg…
#> $ region               <chr> "Asia", "Asia", "Asia", "Asia", "Asia", "Asia", "…
#> $ admin_level_name     <chr> NA, NA, NA, NA, NA, "Kabul", NA, NA, NA, NA, NA, …
#> $ pcode                <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ log_type             <chr> "Introduction / extension of measures", "Introduc…
#> $ category             <chr> "Public health measures", "Public health measures…
#> $ measure              <chr> "Awareness campaigns", "Health screenings in airp…
#> $ targeted_pop_group   <chr> NA, "checked", "checked", NA, "checked", "checked…
#> $ comments             <chr> "MoPH begins announcements on their facebook to m…
#> $ non_compliance       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ date_implemented     <date> 2020-01-24, 2020-01-26, 2020-01-27, 2020-01-27, …
#> $ source               <chr> "Afghanistan MoPH", "Afghanistan MoPH", "Afghanis…
#> $ source_type          <chr> "Government", "Government", "Government", "Govern…
#> $ link                 <chr> "https://www.facebook.com/af.moph/posts/100513025…
#> $ entry_date           <date> 2020-04-07, 2020-04-07, 2020-04-07, 2020-04-07, …
#> $ `alternative source` <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…