ACAPS secondary effects of COVID-19

The objective of the dataset is to provide information for decision makers to improve their efforts in addressing the wider effects of the COVID-19 pandemic.

acaps_secondary_impact_data()

Value

a data.frame

Details

The ACAPS COVID-19 Analytical Framework, together with the data collected for the Government Measures dataset, provided the foundation for the Secondary Impacts Analytical Framework and dataset. The dataset will track secondary impacts across a wide range of relevant themes such as economy, health, migration and education.

Around 80 impact indicators, anticipated to be impacted by COVID-19, have been identified and organised across 4 pillars and 13 thematic blocks.

The data collection is ongoing and will be conducted on a country-level. Eventually, data will identify the secondary impacts of the COVID-19 pandemic in more than 190 countries. Data comes from a range of available sources including international organisations, research centres and media analysis.

The data collection is supported by a group of student volunteers from various universities worldwide. Volunteers receive training on the analytical framework and indicators to ensure coherent data. Additional guidance on analytical and data collection techniques are also provided by an ACAPS analyst team which supervises the data entered. This model has already been successfully implemented in the ACAPS Government Measures project.

References

https://www.acaps.org/secondary-impacts-covid-19

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = acaps_secondary_impact_data()

#> Warning: Expecting date in T2135 / R2135C20: got '1/12/0014'

#> Warning: Expecting date in T5921 / R5921C20: got '24/2/0020'

head(res)

#> # A tibble: 6 x 24
#>       id iso3  country `_data_type` pillar block indicator infogap
#>    <dbl> <chr> <chr>   <chr>        <chr>  <chr> <chr>     <chr>  
#> 1 1.00e6 AFG   Afghan… Quantitative Health Heal… % Most r… checked
#> 2 3.82e3 AFG   Afghan… Qualitative  Health Heal… Access t… NA     
#> 3 2.39e3 AFG   Afghan… Qualitative  Health Heal… Access t… NA     
#> 4 3.07e3 AFG   Afghan… Qualitative  Health Heal… Availabi… NA     
#> 5 4.85e2 AFG   Afghan… Qualitative  Health Heal… Mortalit… checked
#> 6 3.89e3 AFG   Afghan… Qualitative  Health Ment… Alcohol … checked
#> # … with 16 more variables: not_applicable <chr>, field_type <chr>,
#> #   values <chr>, number_value <dbl>, percentage_value <dbl>, value_date <chr>,
#> #   `drivers_of_covid-19_impact` <chr>, localised_geographic_impact <chr>,
#> #   justification <chr>, source <chr>, source_type <chr>, source_date <dttm>,
#> #   link <chr>, alternative_source <chr>, entry_date <dttm>, date <date>

dplyr::glimpse(res)

#> Rows: 6,460
#> Columns: 24
#> $ id                           <dbl> 1000020, 3820, 2390, 3069, 485, 3886, 38…
#> $ iso3                         <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG"…
#> $ country                      <chr> "Afghanistan", "Afghanistan", "Afghanist…
#> $ `_data_type`                 <chr> "Quantitative", "Qualitative", "Qualitat…
#> $ pillar                       <chr> "Health", "Health", "Health", "Health", …
#> $ block                        <chr> "Health Sector", "Health Sector", "Healt…
#> $ indicator                    <chr> "% Most recent mortality rate in 2020", …
#> $ infogap                      <chr> "checked", NA, NA, NA, "checked", "check…
#> $ not_applicable               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ field_type                   <chr> "Percent", "Qualitative", "Qualitative",…
#> $ values                       <chr> NA, "Yes", "Yes", "Yes", NA, NA, "Yes", …
#> $ number_value                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ percentage_value             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 0.063, N…
#> $ value_date                   <chr> "_", "_", "_", "_", "_", "_", "_", "_", …
#> $ `drivers_of_covid-19_impact` <chr> NA, "Government Measures", "Government M…
#> $ localised_geographic_impact  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ justification                <chr> "UN data only offers a projection and do…
#> $ source                       <chr> NA, "UNICEF", "UN Women", "WHO", NA, NA,…
#> $ source_type                  <chr> NA, "United Nations", "United Nations", …
#> $ source_date                  <dttm> NA, 2020-08-11, 2020-06-18, 2020-07-05,…
#> $ link                         <chr> NA, "https://news.un.org/en/story/2020/0…
#> $ alternative_source           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ entry_date                   <dttm> 2020-08-14, 2020-09-29, 2020-09-18, 202…
#> $ date                         <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…

dim(res)

#> [1] 6460   24