Align case tracking locations (for example) to a common baseline

When endeavoring to compare epidemic curves (cases vs date, for example), particularly when making graphical displays, it is helpful to set a "time baseline" that aligns where all the curves start.

align_to_baseline(df, filter_criteria, date_column = "date", group_vars)

Arguments

df
data.frame that includes a date column and at least one other column for filtering, typically a case count.

filter_criteria
an expression as would normally be specified directly to dplyr::filter().

date_column
character(1) column name of the column for ordering the data to define a "beginning" of the curve. It is called a "date column", but anything with a natural ordering will likely work.

group_vars
optional character() column_name(s) that specify grouping done before calculating minimum dates. Concretely, if the goal is to compare several countries, then the group_vars='country' with a column in df called country.

Value

A data.frame with a new column, index, that gives the number of time intervals (typically days) from when the baseline counts are first encountered, done by group.

Details

This function takes this basic approach:

Filter all all data using the filter_criteria, expressed as a dplyr::filter() expression.

Optionally group the dataset.

Find the minimum date left after applying the filter criteria

"Subtract" the minimum date (on a per group basis if grouping columns are used).

The result is a plot that shifts all the curves to start at the "same" starting time with respect to the "start" of the pandemic. For example, for the COVID-19 pandemic, China started much earlier than the rest of the world. To compare the time course of China versus other countries, setting the time to the point where each country had 100 cases allows direct comparison of the shapes of the countries' curves.

See also

Examples

library(dplyr)
library(ggplot2)

# use European CDC dataset
ecdc = ecdc_data()
head(ecdc)

#> # A tibble: 6 x 9
#> # Groups:   location_name, subset [6]
#>   date       location_name iso2c iso3c population_2019 continent
#>   <date>     <chr>         <chr> <chr>           <dbl> <chr>    
#> 1 2019-12-31 Afghanistan   AF    AFG          38041757 Asia     
#> 2 2019-12-31 Afghanistan   AF    AFG          38041757 Asia     
#> 3 2019-12-31 Algeria       DZ    DZA          43053054 Africa   
#> 4 2019-12-31 Algeria       DZ    DZA          43053054 Africa   
#> 5 2019-12-31 Armenia       AM    ARM           2957728 Europe   
#> 6 2019-12-31 Armenia       AM    ARM           2957728 Europe   
#> # … with 3 more variables:
#> #   `Cumulative_number_for_14_days_of_COVID-19_cases_per_100000` <dbl>,
#> #   subset <chr>, count <dbl>

dplyr::glimpse(ecdc)

#> Rows: 75,312
#> Columns: 9
#> Groups: location_name, subset [420]
#> $ date                                                         <date> 2019-12…
#> $ location_name                                                <chr> "Afghani…
#> $ iso2c                                                        <chr> "AF", "A…
#> $ iso3c                                                        <chr> "AFG", "…
#> $ population_2019                                              <dbl> 38041757…
#> $ continent                                                    <chr> "Asia", …
#> $ `Cumulative_number_for_14_days_of_COVID-19_cases_per_100000` <dbl> NA, NA, …
#> $ subset                                                       <chr> "confirm…
#> $ count                                                        <dbl> 0, 0, 0,…

# get top 10 countries by cumulative
# number of deaths
top_10 = ecdc %>%
    dplyr::group_by(location_name) %>%
    dplyr::summarize(deaths = max(deaths)) %>%
    dplyr::arrange(dplyr::desc(deaths)) %>%
    head(10)

#> `summarise()` ungrouping output (override with `.groups` argument)

top_10

#> # A tibble: 10 x 2
#>    location_name       deaths
#>    <chr>                <dbl>
#>  1 Afghanistan           3891
#>  2 Albania               3891
#>  3 Algeria               3891
#>  4 Andorra               3891
#>  5 Angola                3891
#>  6 Anguilla              3891
#>  7 Antigua_and_Barbuda   3891
#>  8 Argentina             3891
#>  9 Armenia               3891
#> 10 Aruba                 3891

# limit ecdc data to "deaths" and
# top 10 countries

ecdc_top10 = ecdc %>%
    dplyr::filter(location_name %in% top_10[['location_name']])
plot_epicurve(ecdc_top10, color='location_name', case_column='deaths')

#> Error: 

ecdc_top10_baseline = align_to_baseline(ecdc_top10, deaths>100, group_vars='location_name')

#> Error: Problem with `filter()` input `..1`.
#> ✖ Input `..1` must be of size 225 or 1, not size 72.
#> ℹ Input `..1` is `deaths > 100`.
#> ℹ The error occurred in group 1: location_name = "Afghanistan", subset = "confirmed".

plot_epicurve(ecdc_top10_baseline, date_column='index', color='location_name') +
    ggtitle('Deaths over time, aligned to date of 100 deaths per country')

#> Error in plot_epicurve(ecdc_top10_baseline, date_column = "index", color = "location_name"): object 'ecdc_top10_baseline' not found