Individual-level data contributed from around the world

Individual-level data contributed from around the world

beoutbreakprepared_data(quietly = TRUE)

Arguments

quietly
logical(1) defaults to TRUE. If FALSE, warnings generated during parsing will be displayed. These often relate to nonstandard date values that occur idiosyncratically.

Source

https://github.com/beoutbreakprepared/nCoV2019

Value

tidy data.frame of content

Note

This is individual level data, collected from diverse sources. Data may be messy and we have made limited attempts at clean up.

Citation

misckraemer2020epidemiological, author = nCoV-2019 Data Working Group, title = Epidemiological Data from the nCoV-2019 Outbreak: Early Descriptions from Publicly Available Data, howpublished = Accessed on yyyy-mm-dd from http://virological.org/t/epidemiological-data-from-the-ncov-2019-outbreak-early-descriptions-from-publicly-available-data/337, year = 2020

See also

Examples

res = beoutbreakprepared_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/beoutbreakprepared/nCoV2019/master/latest_data/latestdata.csv’
#>   local file path: ‘/Users/sdavis2/Library/Caches/sars2pack/file1475855d981f0’
#>   reason: Internal Server Error (HTTP 500).

#> Warning: bfcdownload() failed
#>   rid: BFC18
#>   file: ‘https://raw.githubusercontent.com/beoutbreakprepared/nCoV2019/master/latest_data/latestdata.csv’
#>   reason: download failed

#> Error: download failed; see warnings()

colnames(res)

#> Error in is.data.frame(x): object 'res' not found

head(res)

#> Error in head(res): object 'res' not found