US CDC Social Vulnerability Index

The CDC's Social Vulnerability Index (SVI), created and maintained by the Geospatial Research, Analysis, and Services Program (GRASP), uses US Census data to determine the social vulnerability of every county and tract. This index ranks each county and tract based upon 15 social factors including poverty, lack of vehicle access, and crowded housing, and groups them into four related themes: 1) Socioeconomic, 2) Housing Composition and Disability, 3) Minority Status and Language, and 4) Housing and Transportation.

cdc_social_vulnerability_index()

Source

https://svi.cdc.gov/

Details

Theme rankings: For each of the four themes, we summed the percentiles for the variables comprising each theme. We ordered the summed percentiles for each theme to determine theme-specific percentile rankings. The four summary theme ranking variables, detailed in the Data Dictionary below, are:

Socioeconomic - RPL_THEME1

Household Composition & Disability - RPL_THEME2

Minority Status & Language - RPL_THEME3

Housing Type & Transportation - RPL_THEME4

Overall tract rankings: We summed the sums for each theme, ordered the tracts, and then calculated overall percentile rankings. Please note; taking the sum of the sums for each theme is the same as summing individual variable rankings. The overall tract summary ranking variable is RPL_THEM

For detailed documentation, see https://svi.cdc.gov/Documents/Data/2018_SVI_Data/SVI2018Documentation.pdf

See also

Examples

res = cdc_social_vulnerability_index()
head(res)

#> # A tibble: 6 x 126
#>     fid state_fips state st_abbr county fips  location area_sqmi e_totpop
#>   <dbl> <chr>      <chr> <chr>   <chr>  <chr> <chr>        <dbl>    <dbl>
#> 1     1 00001      Alab… AL      Autau… 01001 Autauga…      594.    55200
#> 2     2 00001      Alab… AL      Baldw… 01003 Baldwin…     1590.   208107
#> 3     3 00001      Alab… AL      Barbo… 01005 Barbour…      885.    25782
#> 4     4 00001      Alab… AL      Bibb   01007 Bibb Co…      622.    22527
#> 5     5 00001      Alab… AL      Blount 01009 Blount …      645.    57645
#> 6     6 00001      Alab… AL      Bullo… 01011 Bullock…      623.    10352
#> # … with 117 more variables: m_totpop <dbl>, e_hu <dbl>, m_hu <dbl>,
#> #   e_hh <dbl>, m_hh <dbl>, e_pov <dbl>, m_pov <dbl>, e_unemp <dbl>,
#> #   m_unemp <dbl>, e_pci <dbl>, m_pci <dbl>, e_nohsdp <dbl>, m_nohsdp <dbl>,
#> #   e_age65 <dbl>, m_age65 <dbl>, e_age17 <dbl>, m_age17 <dbl>, e_disabl <dbl>,
#> #   m_disabl <dbl>, e_sngpnt <dbl>, m_sngpnt <dbl>, e_minrty <dbl>,
#> #   m_minrty <dbl>, e_limeng <dbl>, m_limeng <dbl>, e_munit <dbl>,
#> #   m_munit <dbl>, e_mobile <dbl>, m_mobile <dbl>, e_crowd <dbl>,
#> #   m_crowd <dbl>, e_noveh <dbl>, m_noveh <dbl>, e_groupq <dbl>,
#> #   m_groupq <dbl>, ep_pov <dbl>, mp_pov <dbl>, ep_unemp <dbl>, mp_unemp <dbl>,
#> #   ep_pci <dbl>, mp_pci <dbl>, ep_nohsdp <dbl>, mp_nohsdp <dbl>,
#> #   ep_age65 <dbl>, mp_age65 <dbl>, ep_age17 <dbl>, mp_age17 <dbl>,
#> #   ep_disabl <dbl>, mp_disabl <dbl>, ep_sngpnt <dbl>, mp_sngpnt <dbl>,
#> #   ep_minrty <dbl>, mp_minrty <dbl>, ep_limeng <dbl>, mp_limeng <dbl>,
#> #   ep_munit <dbl>, mp_munit <dbl>, ep_mobile <dbl>, mp_mobile <dbl>,
#> #   ep_crowd <dbl>, mp_crowd <dbl>, ep_noveh <dbl>, mp_noveh <dbl>,
#> #   ep_groupq <dbl>, mp_groupq <dbl>, epl_pov <dbl>, epl_unemp <dbl>,
#> #   epl_pci <dbl>, epl_nohsdp <dbl>, spl_theme1 <dbl>, rpl_theme1 <dbl>,
#> #   epl_age65 <dbl>, epl_age17 <dbl>, epl_disabl <dbl>, epl_sngpnt <dbl>,
#> #   spl_theme2 <dbl>, rpl_theme2 <dbl>, epl_minrty <dbl>, epl_limeng <dbl>,
#> #   spl_theme3 <dbl>, rpl_theme3 <dbl>, epl_munit <dbl>, epl_mobile <dbl>,
#> #   epl_crowd <dbl>, epl_noveh <dbl>, epl_groupq <dbl>, spl_theme4 <dbl>,
#> #   rpl_theme4 <dbl>, spl_themes <dbl>, rpl_themes <dbl>, f_pov <dbl>,
#> #   f_unemp <dbl>, f_pci <dbl>, f_nohsdp <dbl>, f_theme1 <dbl>, f_age65 <dbl>,
#> #   f_age17 <dbl>, f_disabl <dbl>, f_sngpnt <dbl>, f_theme2 <dbl>, …

# limit to index columns only 
res %>% dplyr::select(
    state_fips:e_totpop,dplyr::starts_with('rpl_'))

#> # A tibble: 3,142 x 13
#>    state_fips state st_abbr county fips  location area_sqmi e_totpop rpl_theme1
#>    <chr>      <chr> <chr>   <chr>  <chr> <chr>        <dbl>    <dbl>      <dbl>
#>  1 00001      Alab… AL      Autau… 01001 Autauga…      594.    55200      0.363
#>  2 00001      Alab… AL      Baldw… 01003 Baldwin…     1590.   208107      0.223
#>  3 00001      Alab… AL      Barbo… 01005 Barbour…      885.    25782      0.978
#>  4 00001      Alab… AL      Bibb   01007 Bibb Co…      622.    22527      0.769
#>  5 00001      Alab… AL      Blount 01009 Blount …      645.    57645      0.614
#>  6 00001      Alab… AL      Bullo… 01011 Bullock…      623.    10352      0.977
#>  7 00001      Alab… AL      Butler 01013 Butler …      777.    20025      0.846
#>  8 00001      Alab… AL      Calho… 01015 Calhoun…      606.   115098      0.787
#>  9 00001      Alab… AL      Chamb… 01017 Chamber…      597.    33826      0.690
#> 10 00001      Alab… AL      Chero… 01019 Cheroke…      554.    25853      0.688
#> # … with 3,132 more rows, and 4 more variables: rpl_theme2 <dbl>,
#> #   rpl_theme3 <dbl>, rpl_theme4 <dbl>, rpl_themes <dbl>