Time series cases, hospitalizations, and testing from Corona Data Scraper project

Corona Data Scraper pulls COVID-19 Coronavirus case data from verified sources, finds the corresponding GeoJSON features, and adds population data. All sources are cited right in the same row as the data. Covers testing, hospitalization, deaths, cases, icu. Includes population and geospatial coordinates.

coronadatascraper_data()

Source

https://coronadatascraper.com/timeseries.csv

References

https://coronadatascraper.com/#home

See also

Examples

res = coronadatascraper_data()

#> Error in st_as_sf.data.frame(res, coords = c("long", "lat"), crs = 4326,     agr = "constant"): missing values in coordinates not allowed

colnames(res)

#> Error in is.data.frame(x): object 'res' not found

head(res)

#> Error in head(res): object 'res' not found

dplyr::glimpse(res)

#> Error in dplyr::glimpse(res): object 'res' not found

length(unique(res$names))

#> Error in unique(res$names): object 'res' not found

length(unique(res$county))

#> Error in unique(res$county): object 'res' not found

length(unique(res$state))

#> Error in unique(res$state): object 'res' not found

length(unique(res$country))

#> Error in unique(res$country): object 'res' not found