CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 41
#>   record_id policy_id entry_type correct_type update_type description
#>   <chr>         <dbl> <chr>      <chr>        <chr>       <chr>      
#> 1 R_1gdyC2…   3889409 new_entry  original     <NA>        "US state …
#> 2 R_vp4xmb…   1429889 new_entry  original     <NA>        "At June 9…
#> 3 R_ZxFGXQ…   3386408 new_entry  original     <NA>        "The Parli…
#> 4 R_1KxVwu…   1846710 new_entry  original     <NA>        "Finland's…
#> 5 R_22kWz6…   1919243 new_entry  correction   <NA>        "Egypt wil…
#> 6 R_2Ytcwe…   1924743 new_entry  original     <NA>        "In India,…
#> # … with 35 more variables: date_announced <date>, date_start <date>,
#> #   date_end <date>, country <chr>, iso3c <chr>, iso2c <chr>,
#> #   init_country_level <chr>, domestic_policy <dbl>, province <chr>,
#> #   city <chr>, type <chr>, type_sub_cat <chr>, type_text <chr>,
#> #   type_quarantine_days <dbl>, type_mass_gathering <chr>, type_who_gen <chr>,
#> #   source_corr_type <chr>, target_country <chr>, target_geog_level <chr>,
#> #   target_region <chr>, target_province <chr>, target_city <chr>,
#> #   target_other <lgl>, target_who_what <chr>, target_direction <chr>,
#> #   travel_mechanism <chr>, compliance <chr>, enforcer <chr>,
#> #   index_high_est <dbl>, index_med_est <dbl>, index_low_est <dbl>,
#> #   index_country_rank <dbl>, link <chr>, date_updated <date>,
#> #   recorded_date <dttm>

colnames(res)

#>  [1] "record_id"            "policy_id"            "entry_type"          
#>  [4] "correct_type"         "update_type"          "description"         
#>  [7] "date_announced"       "date_start"           "date_end"            
#> [10] "country"              "iso3c"                "iso2c"               
#> [13] "init_country_level"   "domestic_policy"      "province"            
#> [16] "city"                 "type"                 "type_sub_cat"        
#> [19] "type_text"            "type_quarantine_days" "type_mass_gathering" 
#> [22] "type_who_gen"         "source_corr_type"     "target_country"      
#> [25] "target_geog_level"    "target_region"        "target_province"     
#> [28] "target_city"          "target_other"         "target_who_what"     
#> [31] "target_direction"     "travel_mechanism"     "compliance"          
#> [34] "enforcer"             "index_high_est"       "index_med_est"       
#> [37] "index_low_est"        "index_country_rank"   "link"                
#> [40] "date_updated"         "recorded_date"       

dplyr::glimpse(res)

#> Rows: 28,655
#> Columns: 41
#> $ record_id            <chr> "R_1gdyC25xLGlP5H1Bv", "R_vp4xmb4cOiQ3nb3NA", "R…
#> $ policy_id            <dbl> 3889409, 1429889, 3386408, 1846710, 1919243, 192…
#> $ entry_type           <chr> "new_entry", "new_entry", "new_entry", "new_entr…
#> $ correct_type         <chr> "original", "original", "original", "original", …
#> $ update_type          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ description          <chr> "US state Michigan creates new task force for ec…
#> $ date_announced       <date> 2020-05-29, 2020-06-09, 2020-03-13, 2020-05-28,…
#> $ date_start           <date> 2024-05-29, 2020-06-09, 2020-03-15, 2020-05-28,…
#> $ date_end             <date> 2024-05-29, 2023-01-01, 2022-03-13, 2021-12-31,…
#> $ country              <chr> "United States of America", "Germany", "Japan", …
#> $ iso3c                <chr> "USA", "DEU", "JPN", "FIN", "EGY", "IND", "IND",…
#> $ iso2c                <chr> "US", "DE", "JP", "FI", "EG", "IN", "IN", "DE", …
#> $ init_country_level   <chr> "Provincial", "Provincial", "National", "Nationa…
#> $ domestic_policy      <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, …
#> $ province             <chr> "Michigan", "Hesse", NA, NA, NA, "Kerala", "Kera…
#> $ city                 <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ type                 <chr> "New Task Force, Bureau or Administrative Config…
#> $ type_sub_cat         <chr> "New Task Force or Bureau (i.e. establishment of…
#> $ type_text            <chr> NA, "Economic aid by the state government of Hes…
#> $ type_quarantine_days <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ type_mass_gathering  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ type_who_gen         <chr> "No special population targeted", "No special po…
#> $ source_corr_type     <chr> NA, NA, NA, NA, "Use the old source", NA, NA, NA…
#> $ target_country       <chr> "United States of America", "Germany", "Japan", …
#> $ target_geog_level    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ target_region        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ target_province      <chr> "Michigan", "Hesse", NA, NA, NA, "Kerala", "Kera…
#> $ target_city          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ target_other         <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ target_who_what      <chr> "All Residents (Citizen Residents +Foreign Resid…
#> $ target_direction     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ travel_mechanism     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ compliance           <chr> "Mandatory (Unspecified/Implied)", "Voluntary/Re…
#> $ enforcer             <chr> "New (COVID-19 specific) Task Force", "Provincia…
#> $ index_high_est       <dbl> NA, NA, 53.30879, NA, NA, NA, NA, NA, NA, 52.558…
#> $ index_med_est        <dbl> NA, NA, 52.49780, NA, NA, NA, NA, NA, NA, 51.641…
#> $ index_low_est        <dbl> NA, NA, 51.68669, NA, NA, NA, NA, NA, NA, 50.893…
#> $ index_country_rank   <dbl> NA, NA, 158, NA, NA, NA, NA, NA, NA, 63, 63, 63,…
#> $ link                 <chr> "https://www.michigan.gov/whitmer/0,9309,7-387-9…
#> $ date_updated         <date> 2020-07-05, 2020-06-22, 2020-04-15, 2020-06-12,…
#> $ recorded_date        <dttm> 2020-07-05 06:40:29, 2020-06-22 03:20:26, 2020-…

summary(res)

#>   record_id           policy_id          entry_type        correct_type      
#>  Length:28655       Min.   :2.087e+03   Length:28655       Length:28655      
#>  Class :character   1st Qu.:2.499e+06   Class :character   Class :character  
#>  Mode  :character   Median :5.091e+06   Mode  :character   Mode  :character  
#>                     Mean   :2.189e+09                                        
#>                     3rd Qu.:7.524e+06                                        
#>                     Max.   :6.258e+13                                        
#>                                                                              
#>  update_type        description        date_announced      
#>  Length:28655       Length:28655       Min.   :2019-12-31  
#>  Class :character   Class :character   1st Qu.:2020-03-17  
#>  Mode  :character   Mode  :character   Median :2020-04-01  
#>                                        Mean   :2020-04-11  
#>                                        3rd Qu.:2020-04-30  
#>                                        Max.   :2030-02-28  
#>                                        NA's   :14          
#>    date_start            date_end            country         
#>  Min.   :2019-12-31   Min.   :2020-01-10   Length:28655      
#>  1st Qu.:2020-03-18   1st Qu.:2020-04-20   Class :character  
#>  Median :2020-04-02   Median :2020-05-13   Mode  :character  
#>  Mean   :2020-04-12   Mean   :2020-05-25                     
#>  3rd Qu.:2020-05-04   3rd Qu.:2020-06-17                     
#>  Max.   :2024-05-29   Max.   :2024-05-29                     
#>                       NA's   :17466                          
#>     iso3c              iso2c           init_country_level domestic_policy 
#>  Length:28655       Length:28655       Length:28655       Min.   :0.0000  
#>  Class :character   Class :character   Class :character   1st Qu.:1.0000  
#>  Mode  :character   Mode  :character   Mode  :character   Median :1.0000  
#>                                                           Mean   :0.9083  
#>                                                           3rd Qu.:1.0000  
#>                                                           Max.   :1.0000  
#>                                                                           
#>    province             city               type           type_sub_cat      
#>  Length:28655       Length:28655       Length:28655       Length:28655      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>   type_text         type_quarantine_days type_mass_gathering type_who_gen      
#>  Length:28655       Min.   : 1.00        Length:28655        Length:28655      
#>  Class :character   1st Qu.:10.50        Class :character    Class :character  
#>  Mode  :character   Median :19.00        Mode  :character    Mode  :character  
#>                     Mean   :19.83                                              
#>                     3rd Qu.:28.00                                              
#>                     Max.   :85.00                                              
#>                     NA's   :28456                                              
#>  source_corr_type   target_country     target_geog_level  target_region     
#>  Length:28655       Length:28655       Length:28655       Length:28655      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_province    target_city        target_other   target_who_what   
#>  Length:28655       Length:28655       Mode:logical   Length:28655      
#>  Class :character   Class :character   NA's:28655     Class :character  
#>  Mode  :character   Mode  :character                  Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_direction   travel_mechanism    compliance          enforcer        
#>  Length:28655       Length:28655       Length:28655       Length:28655      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  index_high_est  index_med_est   index_low_est   index_country_rank
#>  Min.   :48.29   Min.   :45.42   Min.   :37.03   Min.   :  1.0     
#>  1st Qu.:52.56   1st Qu.:51.43   1st Qu.:50.03   1st Qu.: 76.0     
#>  Median :53.89   Median :52.56   Median :51.30   Median :131.0     
#>  Mean   :54.16   Mean   :52.48   Mean   :50.85   Mean   :123.9     
#>  3rd Qu.:55.34   3rd Qu.:53.73   3rd Qu.:52.34   3rd Qu.:175.0     
#>  Max.   :70.74   Max.   :57.59   Max.   :56.26   Max.   :194.0     
#>  NA's   :8018    NA's   :8018    NA's   :8018    NA's   :8018      
#>      link            date_updated        recorded_date                
#>  Length:28655       Min.   :2020-03-28   Min.   :2020-03-28 05:14:33  
#>  Class :character   1st Qu.:2020-04-15   1st Qu.:2020-04-15 20:07:36  
#>  Mode  :character   Median :2020-05-27   Median :2020-05-27 10:40:33  
#>                     Mean   :2020-05-28   Mean   :2020-05-28 13:58:32  
#>                     3rd Qu.:2020-07-09   3rd Qu.:2020-07-09 14:26:57  
#>                     Max.   :2020-08-02   Max.   :2020-08-02 23:59:15  
#>