COVID GLUE datasets

COVID GLUE datasets

cov_glue_lineage_data()

References

{ \@article{VANDORP2020104351, title = "Emergence of genomic diversity and recurrent mutations in SARS-CoV-2", journal = "Infection, Genetics and Evolution", pages = "104351", year = "2020", issn = "1567-1348", doi = "https://doi.org/10.1016/j.meegid.2020.104351", url = "http://www.sciencedirect.com/science/article/pii/S1567134820301829", author = "Lucy [van Dorp] and Mislav Acman and Damien Richard and Liam P. Shaw and Charlotte E. Ford and Louise Ormond and Christopher J. Owen and Juanita Pang and Cedric C.S. Tan and Florencia A.T. Boshier and Arturo Torres Ortiz and FranÃois Balloux" } }

https://www.biorxiv.org/content/10.1101/2020.04.17.046086v1

See also

Examples

res = cov_glue_lineage_data()
head(res)

#> # A tibble: 6 x 10
#>   `GISAID ID` name  country travel_history date       epiweek lineage
#>   <chr>       <chr> <chr>   <chr>          <date>       <dbl> <chr>  
#> 1 EPI_ISL_42… USA/… USA     NA             2020-03-02      10 B.1    
#> 2 EPI_ISL_41… USA/… USA     NA             2020-02-29       9 A.1    
#> 3 EPI_ISL_42… USA/… USA     NA             2020-03-02      10 B.1    
#> 4 EPI_ISL_42… USA/… USA     NA             2020-03-01      10 B.1.5  
#> 5 EPI_ISL_41… USA/… USA     NA             2020-03-01      10 A.1    
#> 6 EPI_ISL_42… USA/… USA     NA             2020-03-01      10 B.2    
#> # … with 3 more variables: representative <dbl>, iso3c <chr>, iso2c <chr>

colnames(res)

#>  [1] "GISAID ID"      "name"           "country"        "travel_history"
#>  [5] "date"           "epiweek"        "lineage"        "representative"
#>  [9] "iso3c"          "iso2c"         

dplyr::glimpse(res)

#> Rows: 27,763
#> Columns: 10
#> $ `GISAID ID`    <chr> "EPI_ISL_420793", "EPI_ISL_417142", "EPI_ISL_420792", …
#> $ name           <chr> "USA/NY_2929/2020", "USA/WA-S89/2020", "USA/NH_0008/20…
#> $ country        <chr> "USA", "USA", "USA", "USA", "USA", "USA", "USA", "USA"…
#> $ travel_history <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ date           <date> 2020-03-02, 2020-02-29, 2020-03-02, 2020-03-01, 2020-…
#> $ epiweek        <dbl> 10, 9, 10, 10, 10, 10, 9, 9, 10, 9, 9, 9, 10, 10, 10, …
#> $ lineage        <chr> "B.1", "A.1", "B.1", "B.1.5", "A.1", "B.2", "A.1", "B.…
#> $ representative <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, …
#> $ iso3c          <chr> "USA", "USA", "USA", "USA", "USA", "USA", "USA", "USA"…
#> $ iso2c          <chr> "US", "US", "US", "US", "US", "US", "US", "US", "US", …