US State-level data including test results

Get data from the COVID tracking project https://covidtracking.com/ including daily historical data on testing results, hospitalizations, and deaths.

covidtracker_data()

Value

A tidy tbl_df

Details

From the COVID Tracking Project website:

The COVID Tracking Project collects information from 50 US states, the District of Columbia, and 5 other US territories to provide the most comprehensive testing data we can collect for the novel coronavirus, SARS-CoV-2. We attempt to report positive and negative results, pending tests, and total people tested for each state or district currently reporting that data.

See also

Examples

covidtracker_data()

#> Parsed with column specification:
#> cols(
#>   date = col_double(),
#>   state = col_character(),
#>   positive = col_double(),
#>   negative = col_double(),
#>   pending = col_double(),
#>   hospitalized = col_double(),
#>   death = col_double(),
#>   total = col_double(),
#>   hash = col_character(),
#>   dateChecked = col_datetime(format = ""),
#>   totalTestResults = col_double(),
#>   fips = col_character(),
#>   deathIncrease = col_double(),
#>   hospitalizedIncrease = col_double(),
#>   negativeIncrease = col_double(),
#>   positiveIncrease = col_double(),
#>   totalTestResultsIncrease = col_double()
#> )

#> # A tibble: 1,429 x 9
#>    date       state positive negative pending hospitalized death
#>    <date>     <chr>    <dbl>    <dbl>   <dbl>        <dbl> <dbl>
#>  1 2020-03-31 AK         119     3594      NA            7     3
#>  2 2020-03-31 AL         981     6298      NA           NA    13
#>  3 2020-03-31 AR         523     5959      NA           64     8
#>  4 2020-03-31 AS          NA        3       2           NA     0
#>  5 2020-03-31 AZ        1289    18082      NA           78    24
#>  6 2020-03-31 CA        7482    21772   57400         1617   153
#>  7 2020-03-31 CO        2627    12737      NA          414    51
#>  8 2020-03-31 CT        3128    13029      NA          608    69
#>  9 2020-03-31 DC         495     3262       2           NA     9
#> 10 2020-03-31 DE         406     3696      NA           64    10
#> # … with 1,419 more rows, and 2 more variables: dateChecked <dttm>, fips <chr>