Export all available datasets to SQL database

SQL databases are the cornerstone of many data science efforts. We provide this simple function to export all the available datasets in available_datasets() to any relational database supported by dbplyr.

datasets_to_sql(
  con,
  dataset_accessors = available_datasets()$accessor,
  overwrite = TRUE,
  ...
)

Arguments

con
a remote data source. See dplyr::copy_to()

dataset_accessors
character() vector of accessors for datasets. The accessor column of available_datasets() is the default, meaning that all data.frame-like datasets will be written to the SQL destination.

overwrite
If TRUE, will overwrite an existing table with name name. If FALSE, will throw an error if name already exists.

...
passed on to dplyr::copy_to()

Details

Create connections using RSQLite::SQLite() for RSQLite, RMariaDB::MariaDB() for RMariaDB, RPostgres::Postgres() for RPostgres, odbc::odbc() for odbc, and bigrquery::bigquery() for BigQuery.

See also

Author

Sean Davis seandavi@gmail.com

Examples

if(require(RSQLite)) {
  sql = DBI::dbConnect(RSQLite::SQLite(), ':memory:')
  datasets_to_sql(sql, dataset_accessors = "apple_mobility_data")
  DBI::dbListTables(sql)
  DBI::dbDisconnect(sql)
} else {
  print("install.packages('RSQLite') to run this example")
}

#> Loading required package: RSQLite

#> apple_mobility_data

#> writing