Daily COVID-19 deaths and confirmed cases from the European CDC

Daily COVID-19 deaths and confirmed cases from the European CDC

ecdc_data()

Source

https://www.ecdc.europa.eu/en/covid-19

Value

a data.frame

Details

From the ECDC website last accessed April 24, 2020:

Since the beginning of the coronavirus pandemic, ECDC’s Epidemic Intelligence team has been collecting the number of COVID-19 cases and deaths, based on reports from health authorities worldwide. This comprehensive and systematic process is carried out on a daily basis. To insure the accuracy and reliability of the data, this process is being constantly refined. This helps to monitor and interpret the dynamics of the COVID-19 pandemic not only in the European Union (EU), the European Economic Area (EEA), but also worldwide.

Every day between 6.00 and 10.00 CET, a team of epidemiologists screens up to 500 relevant sources to collect the latest figures. The data screening is followed by ECDC’s standard epidemic intelligence process for which every single data entry is validated and documented in an ECDC database. An extract of this database, complete with up-to-date figures and data visualisations, is then shared on the ECDC website, ensuring a maximum level of transparency.

ECDC receives regular updates from EU/EEA countries through the Early Warning and Response System (EWRS), The European Surveillance System (TESSy), the World Health Organization (WHO) and email exchanges with other international stakeholders. This information is complemented by screening up to 500 sources every day to collect COVID-19 figures from 196 countries. This includes websites of ministries of health (43% of the total number of sources), websites of public health institutes (9%), websites from other national authorities (ministries of social services and welfare, governments, prime minister cabinets, cabinets of ministries, websites on health statistics and official response teams) (6%), WHO websites and WHO situation reports (2%), and official dashboards and interactive maps from national and international institutions (10%). In addition, ECDC screens social media accounts maintained by national authorities, for example Twitter, Facebook, YouTube or Telegram accounts run by ministries of health (28%) and other official sources (e.g. official media outlets) (2%). Several media and social media sources are screened to gather additional information which can be validated with the official sources previously mentioned. Only cases and deaths reported by the national and regional competent authorities from the countries and territories listed are aggregated in our database.

See also

Author

Sean Davis seandavi@gmail.com

Examples

ecdc = ecdc_data()
colnames(ecdc)

#>  [1] "date"                                           
#>  [2] "year_week"                                      
#>  [3] "location_name"                                  
#>  [4] "iso2c"                                          
#>  [5] "iso3c"                                          
#>  [6] "population_2019"                                
#>  [7] "continent"                                      
#>  [8] "notification_rate_per_100000_population_14-days"
#>  [9] "subset"                                         
#> [10] "count"                                          

dplyr::glimpse(ecdc)

#> Rows: 18,302
#> Columns: 10
#> Groups: location_name, subset [428]
#> $ date                                              <date> 2020-01-06, 2020-0…
#> $ year_week                                         <chr> "2020-01", "2020-01…
#> $ location_name                                     <chr> "Afghanistan", "Afg…
#> $ iso2c                                             <chr> "AF", "AF", "DZ", "…
#> $ iso3c                                             <chr> "AFG", "AFG", "DZA"…
#> $ population_2019                                   <dbl> 38041757, 38041757,…
#> $ continent                                         <chr> "Asia", "Asia", "Af…
#> $ `notification_rate_per_100000_population_14-days` <dbl> NA, NA, NA, NA, NA,…
#> $ subset                                            <chr> "cases_weekly", "de…
#> $ count                                             <dbl> 0, 0, 0, 0, 0, 0, 0…