United States Employment Data from Econ Tracker and Opportunity Insights

Number of active employees, aggregating information from multiple data providers. This series is based on firm-level payroll data from Paychex and Intuit, worker-level data on employment and earnings from Earnin, and firm-level timesheet data from Kronos.

econ_tracker_employment_city_data()

econ_tracker_employment_county_data()

econ_tracker_employment_state_data()

econ_tracker_employment_national_data()

Value

a data.table with geographic details and a variable column with the following values representing the reported value:

emp_combined: Employment level for all workers.

emp_combined_inclow: Employment level for workers in the bottom quartile of the income distribution (incomes approximately under $27,000).

emp_combined_incmiddle: Employment level for workers in the middle two quartiles of the income distribution (incomes approximately $27,000 to $60,000).

emp_combined_inchigh: Employment level for workers in the top quartile of the income distribution (incomes approximately over $60,000).

emp_combined_ss40: Employment level for workers in trade, transportation and utilities (NAICS supersector 40).

emp_combined_ss60: Employment level for workers in professional and business services (NAICS supersector 60).

emp_combined_ss65: Employment level for workers in education and health services (NAICS supersector 65).

emp_combined_ss70: Employment level for workers in leisure and hospitality (NAICS supersector 70).

Details

Data Sources: Paychex, Intuit, Earnin, Kronos

Breakdowns are by:

Wages

High Income (wage greater than $60,000 per year)

Middle Income (wage between $27,000 per year and $60,000 per year)

Low Income (wage less than $27,000 per year)

Industry by NAICS supersector.

Professional and Business Services

Education and Health Services

Retail and Transportation

Leisure and Hospitality

Note

For low income workers, the change in employment is calculated using Paychex and Earnin data. For medium and high income workers, the change in employment is calculated using Paychex and Intuit data.

In order to provide closer to real time data, we forecast the most recent employment measures beyond those available in the combined Earnin, Intuit, and Paychex dataset alone. To do so, we leverage two sources of higher frequency data: Kronos timestamp data and the Paychex weekly pay cycle sample. Using this higher frequency data we forecast more recent changes in employment using a distributed lag model, constructed by regressing a given week’s employment measure on the corresponding week’s Kronos measure, as well as its current and 3 previous lagged weeks’ Paychex weekly pay cycle measure. For more details, please refer to the appendix of the accompanying paper

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = econ_tracker_employment_city_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20City%20-%20Daily.csv’
#>   local file path: ‘/tmp/RtmpyJSxab/BiocFileCache/106d4143383b_Employment%2520Combined%2520-%2520City%2520-%2520Daily.csv’
#>   reason: Not Found (HTTP 404).

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC26
#>   fpath: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20City%20-%20Daily.csv’
#>   reason: download failed

#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()

res

#> Error in eval(expr, envir, enclos): object 'res' not found

res = econ_tracker_employment_county_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20County%20-%20Daily.csv’
#>   local file path: ‘/tmp/RtmpyJSxab/BiocFileCache/106d298023d3_Employment%2520Combined%2520-%2520County%2520-%2520Daily.csv’
#>   reason: Not Found (HTTP 404).

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC27
#>   fpath: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20County%20-%20Daily.csv’
#>   reason: download failed

#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()

res

#> Error in eval(expr, envir, enclos): object 'res' not found

res = econ_tracker_employment_state_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20State%20-%20Daily.csv’
#>   local file path: ‘/tmp/RtmpyJSxab/BiocFileCache/106d4b231f65_Employment%2520Combined%2520-%2520State%2520-%2520Daily.csv’
#>   reason: Not Found (HTTP 404).

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC28
#>   fpath: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20State%20-%20Daily.csv’
#>   reason: download failed

#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()

res

#> Error in eval(expr, envir, enclos): object 'res' not found

res = econ_tracker_employment_national_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20National%20-%20Daily.csv’
#>   local file path: ‘/tmp/RtmpyJSxab/BiocFileCache/106d1c60afd_Employment%2520Combined%2520-%2520National%2520-%2520Daily.csv’
#>   reason: Not Found (HTTP 404).

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC29
#>   fpath: ‘https://raw.githubusercontent.com/OpportunityInsights/EconomicTracker/main/data/Employment%20Combined%20-%20National%20-%20Daily.csv’
#>   reason: download failed

#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()

res

#> Error in eval(expr, envir, enclos): object 'res' not found

library(ggplot2)
ggplot(res, aes(x=date,y=value,color=variable,lty=variable_type)) + 
    geom_line() + 
    ggtitle('Proportion of workforce employed in the United States', 
            subtitle='Broken down by industry and income level, normalized to January 2020') +
    ylab('Proportion of workforce') 

#> Error in ggplot(res, aes(x = date, y = value, color = variable, lty = variable_type)): object 'res' not found