Get enriched jhu dataset enriched with country-level metadata

The purpose of this function is to enrich the JHU dataset with additional country-level metadata and then get the data into a form that can be easily written to excel or used as a "wide-format" tabular dataset.

enriched_jhu_data(cols_to_remove = .cols_to_remove, warn = FALSE)

Arguments

cols_to_remove
a character vector of column names from country_metadata() to remove.

warn
logical(1) defaults to FALSE, sent to countrycode::countrycode to determine whether failed matches should be reported.

Value

A list of three data.frames named deaths, confirmed, and recovered.

Examples

res = enriched_jhu_data()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv’
#>   local file path: ‘/Users/sdavis2/Library/Caches/sars2pack/file147585cd97bb6’
#>   reason: Internal Server Error (HTTP 500).

#> Warning: bfcdownload() failed
#>   rid: BFC24
#>   file: ‘https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv’
#>   reason: download failed

#> Error: download failed; see warnings()

colnames(res)

#> Error in is.data.frame(x): object 'res' not found

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:MASS’:
#> 
#>     select

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

head(res)

#> Error in head(res): object 'res' not found

glimpse(res)

#> Error in glimpse(res): object 'res' not found