Country, region, and province level COVID-19 time series for Europe

covid19-eu-data is a dataset repository for COVID-19/SARS-CoV-2 cases in Europe. They pull data from official government websites regularly using the open-source scripts inside the repository.

eu_data_cache_data(datasets = eu_data_cache_datasets(include_ecdc = FALSE))

eu_data_cache_datasets(include_ecdc = TRUE)

Source

https://covid19-eu-data-cache.now.sh/

https://github.com/covid19-eu-zh/covid19-eu-data

Functions

eu_data_cache_datasets:

Note

Some oddities of the dataset are captured on the original site at https://github.com/covid19-eu-zh/covid19-eu-data#notes

See also

Examples

eudata = eu_data_cache_data()

#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 
#> 

colnames(eudata)

#>  [1] "country"                "nuts_2"                 "cases"                 
#>  [4] "recovered"              "deaths"                 "tests"                 
#>  [7] "hospitalized"           "intensive_care"         "datetime"              
#> [10] "date"                   "nuts_3"                 "nuts_1"                
#> [13] "quarantine"             "lau"                    "hospitalized/100k pop."

dim(eudata)

#> [1] 29829    15

head(eudata)

#> # A tibble: 6 x 15
#>   country nuts_2 cases recovered deaths tests hospitalized intensive_care
#>   <chr>   <chr>  <dbl>     <dbl>  <dbl> <dbl>        <dbl>          <dbl>
#> 1 AT      Tirol      2        NA     NA    NA           NA             NA
#> 2 AT      Salzb…     3        NA     NA    NA           NA             NA
#> 3 AT      Steie…     4        NA     NA    NA           NA             NA
#> 4 AT      Niede…     5        NA     NA    NA           NA             NA
#> 5 AT      Wien      15        NA     NA    NA           NA             NA
#> 6 AT      Tirol      2        NA     NA    NA           NA             NA
#> # … with 7 more variables: datetime <dttm>, date <date>, nuts_3 <chr>,
#> #   nuts_1 <chr>, quarantine <dbl>, lau <chr>, `hospitalized/100k pop.` <dbl>

dplyr::glimpse(eudata)

#> Rows: 29,829
#> Columns: 15
#> $ country                  <chr> "AT", "AT", "AT", "AT", "AT", "AT", "AT", "A…
#> $ nuts_2                   <chr> "Tirol", "Salzburg", "Steiermark", "Niederös…
#> $ cases                    <dbl> 2, 3, 4, 5, 15, 2, 3, 4, 12, 16, 1, 3, 4, 4,…
#> $ recovered                <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ deaths                   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ tests                    <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ hospitalized             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ intensive_care           <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ datetime                 <dttm> 2020-03-04 17:45:00, 2020-03-04 17:45:00, 2…
#> $ date                     <date> 2020-03-04, 2020-03-04, 2020-03-04, 2020-03…
#> $ nuts_3                   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ nuts_1                   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ quarantine               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ lau                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ `hospitalized/100k pop.` <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …

summary(eudata)

#>    country             nuts_2              cases           recovered     
#>  Length:29829       Length:29829       Min.   :    0.0   Min.   :   0.0  
#>  Class :character   Class :character   1st Qu.:    3.0   1st Qu.:   0.0  
#>  Mode  :character   Mode  :character   Median :   24.0   Median :   0.0  
#>                                        Mean   :  327.4   Mean   : 113.2  
#>                                        3rd Qu.:  134.0   3rd Qu.:   2.0  
#>                                        Max.   :34294.0   Max.   :8098.0  
#>                                        NA's   :5709      NA's   :29525   
#>      deaths             tests         hospitalized    intensive_care  
#>  Min.   :   0.000   Min.   :  1058   Min.   :     0   Min.   :  0.00  
#>  1st Qu.:   0.000   1st Qu.:  8958   1st Qu.:     6   1st Qu.:  6.00  
#>  Median :   0.000   Median : 16313   Median :    12   Median : 13.50  
#>  Mean   :   4.829   Mean   : 24113   Mean   :  1798   Mean   : 32.24  
#>  3rd Qu.:   0.000   3rd Qu.: 30451   3rd Qu.:    25   3rd Qu.: 32.00  
#>  Max.   :3134.000   Max.   :156801   Max.   :872779   Max.   :400.00  
#>  NA's   :10541      NA's   :29781    NA's   :20016    NA's   :29165   
#>     datetime                        date               nuts_3         
#>  Min.   :2020-02-24 18:00:00   Min.   :2020-02-24   Length:29829      
#>  1st Qu.:2020-03-17 14:00:00   1st Qu.:2020-03-17   Class :character  
#>  Median :2020-03-29 14:00:00   Median :2020-03-29   Mode  :character  
#>  Mean   :2020-03-27 02:40:59   Mean   :2020-03-26                     
#>  3rd Qu.:2020-04-07 00:00:00   3rd Qu.:2020-04-07                     
#>  Max.   :2020-04-16 00:30:00   Max.   :2020-04-16                     
#>                                                                       
#>     nuts_1            quarantine        lau            hospitalized/100k pop.
#>  Length:29829       Min.   : 59.0   Length:29829       Min.   :  0.00        
#>  Class :character   1st Qu.: 63.5   Class :character   1st Qu.: 13.38        
#>  Mode  :character   Median : 78.5   Mode  :character   Median : 21.40        
#>                     Mean   : 77.3                      Mean   : 31.48        
#>                     3rd Qu.: 86.5                      3rd Qu.: 38.20        
#>                     Max.   :100.0                      Max.   :251.00        
#>                     NA's   :29819                      NA's   :26997         

# You may wish to look at only one or a few
# of the available datasets. Here is the list
eu_data_cache_datasets()

#>     at     ch     cz     de   ecdc     fr     hu     ie     it     nl     no 
#>   "at"   "ch"   "cz"   "de" "ecdc"   "fr"   "hu"   "ie"   "it"   "nl"   "no" 
#>     pl     se     si     uk 
#>   "pl"   "se"   "si"   "uk"