Excess deaths tracking by Financial Times

From the Financial Times: "'Excess mortality' refers to the difference between deaths from all causes during the pandemic and the historic seasonal average. For many of the jurisdictions shown here, this figure is higher than the official Covid-19 fatalities that are published by national governments each day. While not all of these deaths are necessarily attributable to the disease, it does leave a number of unexplained deaths that suggests that the official figures of deaths attributed may significant undercounts of the pandemic's impact."

financial_times_excess_deaths()

Source

https://github.com/Financial-Times/coronavirus-excess-mortality-data

Value

a data.frame

Details

From the Financial Times github repository:

The data contains excess mortality data for the period covering the 2020 Covid-19 pandemic. The data has been gathered from national, regional or municipal agencies that collect death registrations and publish official mortality statistics. These original data were reshaped into a standardised format by Financial Times journalists to allow cross-national comparisons, and have been used to inform the FT’s reporting on the pandemic. The repository contains the excess mortality data for all known jurisdictions which publish all-cause mortality data meeting the following criteria:

daily, weekly or monthly level of granularity

includes equivalent historical data for at least one full year before 2020, and preferably at least five years (2015-2019) includes data up to at least April 1, 2020

Most countries publish mortality data with a longer periodicity (typically quarterly or even annually), a longer publication lag time, or both. This sort of data is not suitable for ongoing analysis during an epidemic and is therefore not included here.

License

https://creativecommons.org/licenses/by/4.0/

References

Coronavirus tracked: the latest figures as countries start to reopenFree to read, https://www.ft.com/content/a26fbf7e-48f8-11ea-aeb3-955839e06441

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = financial_times_excess_deaths()

head(res)

#> # A tibble: 6 x 11
#>   country region period  year month  week date       deaths expected_deaths
#>   <chr>   <chr>  <chr>  <dbl> <dbl> <dbl> <date>      <dbl>           <dbl>
#> 1 Russia  Russia month   2006     1    NA 2006-01-31 209058          209058
#> 2 Russia  Russia month   2006     2    NA 2006-02-28 181608          181608
#> 3 Russia  Russia month   2006     3    NA 2006-03-31 195086          195086
#> 4 Russia  Russia month   2006     4    NA 2006-04-30 172981          172981
#> 5 Russia  Russia month   2006     5    NA 2006-05-31 193808          193808
#> 6 Russia  Russia month   2006     6    NA 2006-06-30 178184          178184
#> # … with 2 more variables: excess_deaths <dbl>, excess_deaths_pct <dbl>

dplyr::glimpse(res)

#> Rows: 97,429
#> Columns: 11
#> $ country           <chr> "Russia", "Russia", "Russia", "Russia", "Russia", "…
#> $ region            <chr> "Russia", "Russia", "Russia", "Russia", "Russia", "…
#> $ period            <chr> "month", "month", "month", "month", "month", "month…
#> $ year              <dbl> 2006, 2006, 2006, 2006, 2006, 2006, 2006, 2006, 200…
#> $ month             <dbl> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1, 2, 3, 4, …
#> $ week              <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ date              <date> 2006-01-31, 2006-02-28, 2006-03-31, 2006-04-30, 20…
#> $ deaths            <dbl> 209058, 181608, 195086, 172981, 193808, 178184, 167…
#> $ expected_deaths   <dbl> 209058, 181608, 195086, 172981, 193808, 178184, 167…
#> $ excess_deaths     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ excess_deaths_pct <dbl> 17.58406, 17.58406, 17.58406, 17.58406, 17.58406, 1…