Google search trends for COVID-19 and illness-related terms

This aggregated, anonymized dataset shows trends in search patterns for symptoms and is intended to help researchers to better understand the impact of COVID-19. Public health experts indicated that trends in search patterns might be helpful in broadly understanding how COVID-19 impacts communities and even in detecting outbreaks earlier. You shouldn’t assume that the data is a recording of real-world clinical events, or use this data for medical diagnostic, prognostic, or treatment purposes.

google_search_trends_data(verbose = TRUE)

Source

https://github.com/GoogleCloudPlatform/covid-19-open-data

Value

a data.table

Details

This data reflects the volume of Google searches for a broad set of symptoms, signs and health conditions. To keep things simple in this documentation, we will refer to all of these collectively as symptoms. The data covers hundreds of symptoms such as fever, difficulty breathing, and stress—based on the following:

a symptom’s prevalence in Google’s searches

data quality and privacy considerations

For each day, we count the searches mapped to each of these symptoms and organize the data by geographic region. The resulting dataset is a daily or weekly time series for each region showing the relative frequency of searches for each symptom.

A single search query can be mapped to more than one symptom. For example, we map a search for “acid reflux and coughing up mucus” to three symptoms: Cough, Gastroesophageal reflux disease, and Heartburn.

The dataset covers the recent period and we’ll gradually expand its range as part of regular updates. Each update will bring the coverage to within three days of the day of the update.

Although we are releasing the dataset in English, we count searches in other languages. In each supported country, we include the languages needed to cover the majority of symptom search queries. For example, in the United States we support Spanish and English.

The data represents a sample of our users and might not represent the exact behavior of a wider population.

Preserving privacy

For this dataset, we use differential privacy, which adds artificial noise to our datasets while enabling high quality results without identifying any individual person.

To further protect people’s privacy, we ensure that no personal information or individual search queries are included in the dataset, and we don’t link any search-based health inferences to an individual user. More information about the privacy methods used to generate the dataset can be found in this report.

How we process the data

The data shows the relative popularity of symptoms in searches within a geographical region.

When the daily volume of the data for a given region does not meet quality or privacy thresholds, we do the following:

Try to provide a given symptom at the weekly resolution. If we cannot meet our quality or privacy thresholds at the weekly resolution, we do not provide the data for the symptom in that region. As a result, for a given region, some symptoms will appear exclusively in the daily time series, some exclusively in the weekly time series, and some might not appear at all (if the search data for those symptoms is very scarce).

To normalize and scale the daily and the weekly time series (processed separately), we do the following for each region:

First, the algorithm counts the number of searches for each symptom in that region for that day/week. Next, we divide this count by the total number of Search users in the region for that day/week to calculate relative popularity (which can be interpreted as the probability that a user in this region will search for the given symptom on that day/week). We refer to this ratio as the normalized popularity of a symptom. We then find the maximum value of the normalized popularity across the entire published time range for that region, over all symptoms using the chosen time resolution (day/week). We scale this maximum value to 100. All the other values are mapped to proportionally smaller values (linear scaling) in the range 0-100. Finally, we store the scaling factor and use it to scale values (for the same region and time resolution) in subsequent releases. In future updates, when a symptom popularity exceeds the previously-observed maximum value (found in step 3), then the new scaled value will be larger than 100. In each region, we scale all the normalized (daily / weekly) popularities using the same scaling factor. In a single region, you can compare the relative popularity of two (or more) symptoms (at the same time resolution) over any time interval. However, you should not compare the values of symptom popularity across regions or time resolutions — the region and time resolution specific scalings make these comparisons meaningless.

Note

Use of this dataset implies acceptance of Google's terms of use https://policies.google.com/terms.

References

Google LLC "Google COVID-19 Search Trends symptoms dataset". http://goo.gle/covid19symptomdataset

Author

Sean Davis seandavi@gmail.com

Examples

# \donttest{

# This example loads the dataset which is large. 
# Do not expect this example to run without sufficient
# memory.

res = google_search_trends_data()

#> This is a VERY large dataset that may take some time to load and requires about 5GB of free RAM to load

head(res[,1:10])

#>    key       date abdominal_obesity abdominal_pain  acne actinic_keratosis
#> 1:  AU 2020-01-01              3.56           5.38 10.76              0.52
#> 2:  AU 2020-01-02              3.46           5.35 11.30              0.55
#> 3:  AU 2020-01-03              3.40           5.35 11.23              0.53
#> 4:  AU 2020-01-04              3.43           5.23 10.68              0.42
#> 5:  AU 2020-01-05              3.29           5.14 10.20              0.49
#> 6:  AU 2020-01-06              3.52           5.23 10.83              0.49
#>    acute_bronchitis adrenal_crisis ageusia alcoholism
#> 1:             0.35           0.12    0.05       5.36
#> 2:             0.36           0.14    0.06       4.62
#> 3:             0.34           0.12    0.05       4.63
#> 4:             0.35           0.11    0.05       4.82
#> 5:             0.38           0.11      NA       4.26
#> 6:             0.36           0.14    0.05       4.07

colnames(res)

#>   [1] "key"                                      
#>   [2] "date"                                     
#>   [3] "abdominal_obesity"                        
#>   [4] "abdominal_pain"                           
#>   [5] "acne"                                     
#>   [6] "actinic_keratosis"                        
#>   [7] "acute_bronchitis"                         
#>   [8] "adrenal_crisis"                           
#>   [9] "ageusia"                                  
#>  [10] "alcoholism"                               
#>  [11] "allergic_conjunctivitis"                  
#>  [12] "allergy"                                  
#>  [13] "amblyopia"                                
#>  [14] "amenorrhea"                               
#>  [15] "amnesia"                                  
#>  [16] "anal_fissure"                             
#>  [17] "anaphylaxis"                              
#>  [18] "anemia"                                   
#>  [19] "angina_pectoris"                          
#>  [20] "angioedema"                               
#>  [21] "angular_cheilitis"                        
#>  [22] "anosmia"                                  
#>  [23] "anxiety"                                  
#>  [24] "aphasia"                                  
#>  [25] "aphonia"                                  
#>  [26] "apnea"                                    
#>  [27] "arthralgia"                               
#>  [28] "arthritis"                                
#>  [29] "ascites"                                  
#>  [30] "asperger_syndrome"                        
#>  [31] "asphyxia"                                 
#>  [32] "asthma"                                   
#>  [33] "astigmatism"                              
#>  [34] "ataxia"                                   
#>  [35] "atheroma"                                 
#>  [36] "attention_deficit_hyperactivity_disorder" 
#>  [37] "auditory_hallucination"                   
#>  [38] "autoimmune_disease"                       
#>  [39] "avoidant_personality_disorder"            
#>  [40] "back_pain"                                
#>  [41] "bacterial_vaginosis"                      
#>  [42] "balance_disorder"                         
#>  [43] "beaus_lines"                              
#>  [44] "bells_palsy"                              
#>  [45] "biliary_colic"                            
#>  [46] "binge_eating"                             
#>  [47] "bleeding"                                 
#>  [48] "bleeding_on_probing"                      
#>  [49] "blepharospasm"                            
#>  [50] "bloating"                                 
#>  [51] "blood_in_stool"                           
#>  [52] "blurred_vision"                           
#>  [53] "blushing"                                 
#>  [54] "boil"                                     
#>  [55] "bone_fracture"                            
#>  [56] "bone_tumor"                               
#>  [57] "bowel_obstruction"                        
#>  [58] "bradycardia"                              
#>  [59] "braxton_hicks_contractions"               
#>  [60] "breakthrough_bleeding"                    
#>  [61] "breast_pain"                              
#>  [62] "bronchitis"                               
#>  [63] "bruise"                                   
#>  [64] "bruxism"                                  
#>  [65] "bunion"                                   
#>  [66] "burn"                                     
#>  [67] "burning_chest_pain"                       
#>  [68] "burning_mouth_syndrome"                   
#>  [69] "candidiasis"                              
#>  [70] "canker_sore"                              
#>  [71] "cardiac_arrest"                           
#>  [72] "carpal_tunnel_syndrome"                   
#>  [73] "cataplexy"                                
#>  [74] "cataract"                                 
#>  [75] "chancre"                                  
#>  [76] "cheilitis"                                
#>  [77] "chest_pain"                               
#>  [78] "chills"                                   
#>  [79] "chorea"                                   
#>  [80] "chronic_pain"                             
#>  [81] "cirrhosis"                                
#>  [82] "cleft_lip_and_cleft_palate"               
#>  [83] "clouding_of_consciousness"                
#>  [84] "cluster_headache"                         
#>  [85] "colitis"                                  
#>  [86] "coma"                                     
#>  [87] "common_cold"                              
#>  [88] "compulsive_behavior"                      
#>  [89] "compulsive_hoarding"                      
#>  [90] "confusion"                                
#>  [91] "congenital_heart_defect"                  
#>  [92] "conjunctivitis"                           
#>  [93] "constipation"                             
#>  [94] "convulsion"                               
#>  [95] "cough"                                    
#>  [96] "crackles"                                 
#>  [97] "cramp"                                    
#>  [98] "crepitus"                                 
#>  [99] "croup"                                    
#> [100] "cyanosis"                                 
#> [101] "dandruff"                                 
#> [102] "delayed_onset_muscle_soreness"            
#> [103] "dementia"                                 
#> [104] "dentin_hypersensitivity"                  
#> [105] "depersonalization"                        
#> [106] "depression"                               
#> [107] "dermatitis"                               
#> [108] "desquamation"                             
#> [109] "developmental_disability"                 
#> [110] "diabetes"                                 
#> [111] "diabetic_ketoacidosis"                    
#> [112] "diarrhea"                                 
#> [113] "dizziness"                                
#> [114] "dry_eye_syndrome"                         
#> [115] "dysautonomia"                             
#> [116] "dysgeusia"                                
#> [117] "dysmenorrhea"                             
#> [118] "dyspareunia"                              
#> [119] "dysphagia"                                
#> [120] "dysphoria"                                
#> [121] "dystonia"                                 
#> [122] "dysuria"                                  
#> [123] "ear_pain"                                 
#> [124] "eczema"                                   
#> [125] "edema"                                    
#> [126] "encephalitis"                             
#> [127] "encephalopathy"                           
#> [128] "epidermoid_cyst"                          
#> [129] "epilepsy"                                 
#> [130] "epiphora"                                 
#> [131] "erectile_dysfunction"                     
#> [132] "erythema"                                 
#> [133] "erythema_chronicum_migrans"               
#> [134] "esophagitis"                              
#> [135] "excessive_daytime_sleepiness"             
#> [136] "eye_pain"                                 
#> [137] "eye_strain"                               
#> [138] "facial_nerve_paralysis"                   
#> [139] "facial_swelling"                          
#> [140] "fasciculation"                            
#> [141] "fatigue"                                  
#> [142] "fatty_liver_disease"                      
#> [143] "fecal_incontinence"                       
#> [144] "fever"                                    
#> [145] "fibrillation"                             
#> [146] "fibrocystic_breast_changes"               
#> [147] "fibromyalgia"                             
#> [148] "flatulence"                               
#> [149] "floater"                                  
#> [150] "focal_seizure"                            
#> [151] "folate_deficiency"                        
#> [152] "food_craving"                             
#> [153] "food_intolerance"                         
#> [154] "frequent_urination"                       
#> [155] "gastroesophageal_reflux_disease"          
#> [156] "gastroparesis"                            
#> [157] "generalized_anxiety_disorder"             
#> [158] "generalized_tonic–clonic_seizure"         
#> [159] "genital_wart"                             
#> [160] "gingival_recession"                       
#> [161] "gingivitis"                               
#> [162] "globus_pharyngis"                         
#> [163] "goitre"                                   
#> [164] "gout"                                     
#> [165] "grandiosity"                              
#> [166] "granuloma"                                
#> [167] "guilt"                                    
#> [168] "hair_loss"                                
#> [169] "halitosis"                                
#> [170] "hay_fever"                                
#> [171] "headache"                                 
#> [172] "heart_arrhythmia"                         
#> [173] "heart_murmur"                             
#> [174] "heartburn"                                
#> [175] "hematochezia"                             
#> [176] "hematoma"                                 
#> [177] "hematuria"                                
#> [178] "hemolysis"                                
#> [179] "hemoptysis"                               
#> [180] "hemorrhoids"                              
#> [181] "hepatic_encephalopathy"                   
#> [182] "hepatitis"                                
#> [183] "hepatotoxicity"                           
#> [184] "hiccup"                                   
#> [185] "hip_pain"                                 
#> [186] "hives"                                    
#> [187] "hot_flash"                                
#> [188] "hydrocephalus"                            
#> [189] "hypercalcaemia"                           
#> [190] "hypercapnia"                              
#> [191] "hypercholesterolemia"                     
#> [192] "hyperemesis_gravidarum"                   
#> [193] "hyperglycemia"                            
#> [194] "hyperhidrosis"                            
#> [195] "hyperkalemia"                             
#> [196] "hyperlipidemia"                           
#> [197] "hypermobility"                            
#> [198] "hyperpigmentation"                        
#> [199] "hypersomnia"                              
#> [200] "hypertension"                             
#> [201] "hyperthermia"                             
#> [202] "hyperthyroidism"                          
#> [203] "hypertriglyceridemia"                     
#> [204] "hypertrophy"                              
#> [205] "hyperventilation"                         
#> [206] "hypocalcaemia"                            
#> [207] "hypochondriasis"                          
#> [208] "hypoglycemia"                             
#> [209] "hypogonadism"                             
#> [210] "hypokalemia"                              
#> [211] "hypomania"                                
#> [212] "hyponatremia"                             
#> [213] "hypotension"                              
#> [214] "hypothyroidism"                           
#> [215] "hypoxemia"                                
#> [216] "hypoxia"                                  
#> [217] "impetigo"                                 
#> [218] "implantation_bleeding"                    
#> [219] "impulsivity"                              
#> [220] "indigestion"                              
#> [221] "infection"                                
#> [222] "inflammation"                             
#> [223] "inflammatory_bowel_disease"               
#> [224] "ingrown_hair"                             
#> [225] "insomnia"                                 
#> [226] "insulin_resistance"                       
#> [227] "intermenstrual_bleeding"                  
#> [228] "intracranial_pressure"                    
#> [229] "iron_deficiency"                          
#> [230] "irregular_menstruation"                   
#> [231] "itch"                                     
#> [232] "jaundice"                                 
#> [233] "kidney_failure"                           
#> [234] "kidney_stone"                             
#> [235] "knee_pain"                                
#> [236] "kyphosis"                                 
#> [237] "lactose_intolerance"                      
#> [238] "laryngitis"                               
#> [239] "leg_cramps"                               
#> [240] "lesion"                                   
#> [241] "leukorrhea"                               
#> [242] "lightheadedness"                          
#> [243] "low_back_pain"                            
#> [244] "low_grade_fever"                          
#> [245] "lymphedema"                               
#> [246] "major_depressive_disorder"                
#> [247] "malabsorption"                            
#> [248] "male_infertility"                         
#> [249] "manic_disorder"                           
#> [250] "melasma"                                  
#> [251] "melena"                                   
#> [252] "meningitis"                               
#> [253] "menorrhagia"                              
#> [254] "middle_back_pain"                         
#> [255] "migraine"                                 
#> [256] "milium"                                   
#> [257] "mitral_insufficiency"                     
#> [258] "mood_disorder"                            
#> [259] "mood_swing"                               
#> [260] "morning_sickness"                         
#> [261] "motion_sickness"                          
#> [262] "mouth_ulcer"                              
#> [263] "muscle_atrophy"                           
#> [264] "muscle_weakness"                          
#> [265] "myalgia"                                  
#> [266] "mydriasis"                                
#> [267] "myocardial_infarction"                    
#> [268] "myoclonus"                                
#> [269] "nasal_congestion"                         
#> [270] "nasal_polyp"                              
#> [271] "nausea"                                   
#> [272] "neck_mass"                                
#> [273] "neck_pain"                                
#> [274] "neonatal_jaundice"                        
#> [275] "nerve_injury"                             
#> [276] "neuralgia"                                
#> [277] "neutropenia"                              
#> [278] "night_sweats"                             
#> [279] "night_terror"                             
#> [280] "nocturnal_enuresis"                       
#> [281] "nodule"                                   
#> [282] "nosebleed"                                
#> [283] "nystagmus"                                
#> [284] "obesity"                                  
#> [285] "onychorrhexis"                            
#> [286] "oral_candidiasis"                         
#> [287] "orthostatic_hypotension"                  
#> [288] "osteopenia"                               
#> [289] "osteophyte"                               
#> [290] "osteoporosis"                             
#> [291] "otitis"                                   
#> [292] "otitis_externa"                           
#> [293] "otitis_media"                             
#> [294] "pain"                                     
#> [295] "palpitations"                             
#> [296] "pancreatitis"                             
#> [297] "panic_attack"                             
#> [298] "papule"                                   
#> [299] "paranoia"                                 
#> [300] "paresthesia"                              
#> [301] "pelvic_inflammatory_disease"              
#> [302] "pericarditis"                             
#> [303] "periodontal_disease"                      
#> [304] "periorbital_puffiness"                    
#> [305] "peripheral_neuropathy"                    
#> [306] "perspiration"                             
#> [307] "petechia"                                 
#> [308] "phlegm"                                   
#> [309] "photodermatitis"                          
#> [310] "photophobia"                              
#> [311] "photopsia"                                
#> [312] "pleural_effusion"                         
#> [313] "pleurisy"                                 
#> [314] "pneumonia"                                
#> [315] "podalgia"                                 
#> [316] "polycythemia"                             
#> [317] "polydipsia"                               
#> [318] "polyneuropathy"                           
#> [319] "polyuria"                                 
#> [320] "poor_posture"                             
#> [321] "post_nasal_drip"                          
#> [322] "postural_orthostatic_tachycardia_syndrome"
#> [323] "prediabetes"                              
#> [324] "proteinuria"                              
#> [325] "pruritus_ani"                             
#> [326] "psychosis"                                
#> [327] "ptosis"                                   
#> [328] "pulmonary_edema"                          
#> [329] "pulmonary_hypertension"                   
#> [330] "purpura"                                  
#> [331] "pus"                                      
#> [332] "pyelonephritis"                           
#> [333] "radiculopathy"                            
#> [334] "rectal_pain"                              
#> [335] "rectal_prolapse"                          
#> [336] "red_eye"                                  
#> [337] "renal_colic"                              
#> [338] "restless_legs_syndrome"                   
#> [339] "rheum"                                    
#> [340] "rhinitis"                                 
#> [341] "rhinorrhea"                               
#> [342] "rosacea"                                  
#> [343] "round_ligament_pain"                      
#> [344] "rumination"                               
#> [345] "scar"                                     
#> [346] "sciatica"                                 
#> [347] "scoliosis"                                
#> [348] "seborrheic_dermatitis"                    
#> [349] "self_harm"                                
#> [350] "sensitivity_to_sound"                     
#> [351] "sexual_dysfunction"                       
#> [352] "shallow_breathing"                        
#> [353] "sharp_pain"                               
#> [354] "shivering"                                
#> [355] "shortness_of_breath"                      
#> [356] "shyness"                                  
#> [357] "sinusitis"                                
#> [358] "skin_condition"                           
#> [359] "skin_rash"                                
#> [360] "skin_tag"                                 
#> [361] "skin_ulcer"                               
#> [362] "sleep_apnea"                              
#> [363] "sleep_deprivation"                        
#> [364] "sleep_disorder"                           
#> [365] "snoring"                                  
#> [366] "sore_throat"                              
#> [367] "spasticity"                               
#> [368] "splenomegaly"                             
#> [369] "sputum"                                   
#> [370] "stomach_rumble"                           
#> [371] "strabismus"                               
#> [372] "stretch_marks"                            
#> [373] "stridor"                                  
#> [374] "stroke"                                   
#> [375] "stuttering"                               
#> [376] "subdural_hematoma"                        
#> [377] "suicidal_ideation"                        
#> [378] "swelling"                                 
#> [379] "swollen_feet"                             
#> [380] "swollen_lymph_nodes"                      
#> [381] "syncope"                                  
#> [382] "tachycardia"                              
#> [383] "tachypnea"                                
#> [384] "telangiectasia"                           
#> [385] "tenderness"                               
#> [386] "testicular_pain"                          
#> [387] "throat_irritation"                        
#> [388] "thrombocytopenia"                         
#> [389] "thyroid_nodule"                           
#> [390] "tic"                                      
#> [391] "tinnitus"                                 
#> [392] "tonsillitis"                              
#> [393] "toothache"                                
#> [394] "tremor"                                   
#> [395] "trichoptilosis"                           
#> [396] "tumor"                                    
#> [397] "type_2_diabetes"                          
#> [398] "unconsciousness"                          
#> [399] "underweight"                              
#> [400] "upper_respiratory_tract_infection"        
#> [401] "urethritis"                               
#> [402] "urinary_incontinence"                     
#> [403] "urinary_tract_infection"                  
#> [404] "urinary_urgency"                          
#> [405] "uterine_contraction"                      
#> [406] "vaginal_bleeding"                         
#> [407] "vaginal_discharge"                        
#> [408] "vaginitis"                                
#> [409] "varicose_veins"                           
#> [410] "vasculitis"                               
#> [411] "ventricular_fibrillation"                 
#> [412] "ventricular_tachycardia"                  
#> [413] "vertigo"                                  
#> [414] "viral_pneumonia"                          
#> [415] "visual_acuity"                            
#> [416] "vomiting"                                 
#> [417] "wart"                                     
#> [418] "water_retention"                          
#> [419] "weakness"                                 
#> [420] "weight_gain"                              
#> [421] "wheeze"                                   
#> [422] "xeroderma"                                
#> [423] "xerostomia"                               
#> [424] "yawn"                                     
#> [425] "place_id"                                 
#> [426] "wikidata"                                 
#> [427] "datacommons"                              
#> [428] "country_code"                             
#> [429] "country_name"                             
#> [430] "subregion1_code"                          
#> [431] "subregion1_name"                          
#> [432] "subregion2_code"                          
#> [433] "subregion2_name"                          
#> [434] "locality_code"                            
#> [435] "locality_name"                            
#> [436] "3166-1-alpha-2"                           
#> [437] "3166-1-alpha-3"                           
#> [438] "aggregation_level"                        

grep('fever',colnames(res))

#> [1] 144 170 244

# }