Estimate the growth rate ratio

Calculate the growth rate ratio as defined in Badr et al.

growth_rate_ratio(v, Ws = 3, Wl = 7, smooth_window = 7)

Arguments

v

a numeric vector of new cases per day (incidence)

Ws

integer(1), the number of days to average for the shorter window

Wl

integer(1), the number of days to average to the longer window

smooth_window

integer(1), the size of the window to apply rolling mean smoothing to. The smoothing is applied before the summary process occurs

Value

a numeric vector

Details

The Growth Rate Ratio is defined as the logarithmic rate of change (number of newly reported cases) over the previous Ws days relative to the logarithmic rate of change over the longer Wl days. GR for any county j on a day t was calculated as follows: where Cjt is the number of new cases reported in county j on a day t.

References

https://doi.org/10.1016/S1473-3099(20)30553-3

Author

Sean Davis seandavi@gmail.com

Examples

if(requireNamespace("zoo")) {
nyt= nytimes_state_data() %>% dplyr::filter(fips=='00053' & subset=='confirmed') %>%
     add_incidence_column()
mgr = nyt %>% dplyr::pull(inc) %>% growth_rate_ratio()
}
#> Loading required namespace: zoo