Create +/- write excel format of enriched JHU global data

The purpose of this function is to enrich the JHU dataset with additional country-level metadata and then get the data into a form that can be easily written to excel or used as a "wide-format" tabular dataset.

jhu_data_to_excel(
  dat = jhu_data(),
  file = NA,
  cols_to_remove = c("callingCodes", "altSpellings", "latlng", "demonym", "timezones",
    "nativeName", "currencies", "languages", "translations", "flag", "regionalBlocs"),
  ...
)

Arguments

dat
a data.frame-like object with at least column CountryRegion that will be joined with country data.

file
character(1) filename to which to save excel file. If specified, use the write.xlsx() function from the openxlsx package to create (or overwrite) the file of that name. The excel file will have one tab for each of the subset records from the supplied data.frame in dat.

cols_to_remove
a character vector of column names from country_metadata() to remove.

...
passed to write.xlsx

Value

A list of three data.frames named deaths, confirmed, and recovered.

Note

This function is really a utility function to provide an excel format of data to folks more familiar with excel than with R.

Examples

xls_form = jhu_data_to_excel()

#> Warning: Some values were not matched unambiguously: MS Zaandam

#> Warning: Some values were not matched unambiguously: MS Zaandam

#> Warning: Some values were not matched unambiguously: MS Zaandam

names(xls_form)

#> [1] "confirmed" "deaths"    "recovered"

head(xls_form[[1]])

#> # A tibble: 6 x 361
#>   name  topLevelDomain alpha2Code alpha3Code capital region subregion population
#>   <chr> <list>         <chr>      <chr>      <chr>   <chr>  <chr>          <int>
#> 1 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 2 Alba… <chr [1]>      AL         ALB        Tirana  Europe Southern…    2886026
#> 3 Alge… <chr [1]>      DZ         DZA        Algiers Africa Northern…   40400000
#> 4 Ando… <chr [1]>      AD         AND        Andorr… Europe Southern…      78014
#> 5 Ango… <chr [1]>      AO         AGO        Luanda  Africa Middle A…   25868000
#> 6 Anti… <chr [1]>      AG         ATG        Saint … Ameri… Caribbean      86295
#> # … with 353 more variables: area <dbl>, gini <dbl>, borders <list>,
#> #   numericCode <chr>, cioc <chr>, ProvinceState <chr>, CountryRegion <chr>,
#> #   Lat <dbl>, Long <dbl>, `2020-01-22` <dbl>, `2020-01-23` <dbl>,
#> #   `2020-01-24` <dbl>, `2020-01-25` <dbl>, `2020-01-26` <dbl>,
#> #   `2020-01-27` <dbl>, `2020-01-28` <dbl>, `2020-01-29` <dbl>,
#> #   `2020-01-30` <dbl>, `2020-01-31` <dbl>, `2020-02-01` <dbl>,
#> #   `2020-02-02` <dbl>, `2020-02-03` <dbl>, `2020-02-04` <dbl>,
#> #   `2020-02-05` <dbl>, `2020-02-06` <dbl>, `2020-02-07` <dbl>,
#> #   `2020-02-08` <dbl>, `2020-02-09` <dbl>, `2020-02-10` <dbl>,
#> #   `2020-02-11` <dbl>, `2020-02-12` <dbl>, `2020-02-13` <dbl>,
#> #   `2020-02-14` <dbl>, `2020-02-15` <dbl>, `2020-02-16` <dbl>,
#> #   `2020-02-17` <dbl>, `2020-02-18` <dbl>, `2020-02-19` <dbl>,
#> #   `2020-02-20` <dbl>, `2020-02-21` <dbl>, `2020-02-22` <dbl>,
#> #   `2020-02-23` <dbl>, `2020-02-24` <dbl>, `2020-02-25` <dbl>,
#> #   `2020-02-26` <dbl>, `2020-02-27` <dbl>, `2020-02-28` <dbl>,
#> #   `2020-02-29` <dbl>, `2020-03-01` <dbl>, `2020-03-02` <dbl>,
#> #   `2020-03-03` <dbl>, `2020-03-04` <dbl>, `2020-03-05` <dbl>,
#> #   `2020-03-06` <dbl>, `2020-03-07` <dbl>, `2020-03-08` <dbl>,
#> #   `2020-03-09` <dbl>, `2020-03-10` <dbl>, `2020-03-11` <dbl>,
#> #   `2020-03-12` <dbl>, `2020-03-13` <dbl>, `2020-03-14` <dbl>,
#> #   `2020-03-15` <dbl>, `2020-03-16` <dbl>, `2020-03-17` <dbl>,
#> #   `2020-03-18` <dbl>, `2020-03-19` <dbl>, `2020-03-20` <dbl>,
#> #   `2020-03-21` <dbl>, `2020-03-22` <dbl>, `2020-03-23` <dbl>,
#> #   `2020-03-24` <dbl>, `2020-03-25` <dbl>, `2020-03-26` <dbl>,
#> #   `2020-03-27` <dbl>, `2020-03-28` <dbl>, `2020-03-29` <dbl>,
#> #   `2020-03-30` <dbl>, `2020-03-31` <dbl>, `2020-04-01` <dbl>,
#> #   `2020-04-02` <dbl>, `2020-04-03` <dbl>, `2020-04-04` <dbl>,
#> #   `2020-04-05` <dbl>, `2020-04-06` <dbl>, `2020-04-07` <dbl>,
#> #   `2020-04-08` <dbl>, `2020-04-09` <dbl>, `2020-04-10` <dbl>,
#> #   `2020-04-11` <dbl>, `2020-04-12` <dbl>, `2020-04-13` <dbl>,
#> #   `2020-04-14` <dbl>, `2020-04-15` <dbl>, `2020-04-16` <dbl>,
#> #   `2020-04-17` <dbl>, `2020-04-18` <dbl>, `2020-04-19` <dbl>,
#> #   `2020-04-20` <dbl>, `2020-04-21` <dbl>, …