Nextstrain/GISAID metadata table (raw input)

This function gathers the raw input data associated with https://nextstrain.org/ for COVID-19 as gathered from GISAID.

nextstrain_case_data()

Source

https://nextstrain.org/

https://raw.githubusercontent.com/nextstrain/ncov/master/data/metadata.tsv

Details

Nextstrain is an open-source project to harness the scientific and public health potential of pathogen genome data. They provide a continually-updated view of publicly available data alongside powerful analytic and visualization tools for use by the community. Their goal is to aid epidemiological understanding and improve outbreak response.

Nextstrain overview

From the Nextstrain website:

In the course of an infection and over an epidemic, pathogens naturally accumulate random mutations to their genomes. This is an inevitable consequence of error-prone genome replication. Since different genomes typically pick up different mutations, mutations can be used as a marker of transmission in which closely related genomes indicate closely related infections. By reconstructing a phylogeny we can learn about important epidemiological phenomena such as spatial spread, introduction timings and epidemic growth rate.

However, if pathogen genome sequences are going to inform public health interventions, then analyses have to be rapidly conducted and results widely disseminated. Current scientific publishing practices hinder the rapid dissemination of epidemiologically relevant results. We thought an open online system that implements robust bioinformatic pipelines to synthesize data from across research groups has the best capacity to make epidemiologically actionable inferences.

Workflow for Nextstrain processing:

https://github.com/nextstrain/ncov/raw/master/docs/images/basic_snakemake_build.png

References

James Hadfield, Colin Megill, Sidney M Bell, John Huddleston, Barney Potter, Charlton Callender, Pavel Sagulenko, Trevor Bedford, Richard A Neher, Nextstrain: real-time tracking of pathogen evolution, Bioinformatics, Volume 34, Issue 23, 01 December 2018, Pages 4121–4123, https://doi.org/10.1093/bioinformatics/bty407

See also

Examples

res = nextstrain_case_data()
colnames(res)

#>  [1] "strain"            "virus"             "gisaid_epi_isl"   
#>  [4] "genbank_accession" "date"              "region"           
#>  [7] "country"           "division"          "location"         
#> [10] "region_exposure"   "country_exposure"  "division_exposure"
#> [13] "segment"           "length"            "host"             
#> [16] "age"               "sex"               "originating_lab"  
#> [19] "submitting_lab"    "authors"           "url"              
#> [22] "title"             "date_submitted"   

head(res)

#> # A tibble: 6 x 23
#>   strain virus gisaid_epi_isl genbank_accessi… date  region country division
#>   <chr>  <chr> <chr>          <chr>            <chr> <chr>  <chr>   <chr>   
#> 1 Alger… ncov  EPI_ISL_418241 ?                2020… Africa Algeria Boufarik
#> 2 Alger… ncov  EPI_ISL_418242 ?                2020… Africa Algeria Blida   
#> 3 Alger… ncov  EPI_ISL_420037 ?                2020… Africa Algeria Boufarik
#> 4 Anhui… ncov  EPI_ISL_413485 ?                2020… Asia   China   Anhui   
#> 5 Argen… ncov  EPI_ISL_420600 ?                2020… South… Argent… Argenti…
#> 6 Argen… ncov  EPI_ISL_420599 ?                2020… South… Argent… Argenti…
#> # … with 15 more variables: location <chr>, region_exposure <chr>,
#> #   country_exposure <chr>, division_exposure <chr>, segment <chr>,
#> #   length <dbl>, host <chr>, age <chr>, sex <chr>, originating_lab <chr>,
#> #   submitting_lab <chr>, authors <chr>, url <chr>, title <chr>,
#> #   date_submitted <date>

library(dplyr)
dplyr::glimpse(res)

#> Rows: 15,908
#> Columns: 23
#> $ strain            <chr> "Algeria/G0638_2264/2020", "Algeria/G0640_2265/2020…
#> $ virus             <chr> "ncov", "ncov", "ncov", "ncov", "ncov", "ncov", "nc…
#> $ gisaid_epi_isl    <chr> "EPI_ISL_418241", "EPI_ISL_418242", "EPI_ISL_420037…
#> $ genbank_accession <chr> "?", "?", "?", "?", "?", "?", "?", "?", "?", "?", "…
#> $ date              <chr> "2020-03-02", "2020-03-08", "2020-03-02", "2020-01-…
#> $ region            <chr> "Africa", "Africa", "Africa", "Asia", "South Americ…
#> $ country           <chr> "Algeria", "Algeria", "Algeria", "China", "Argentin…
#> $ division          <chr> "Boufarik", "Blida", "Boufarik", "Anhui", "Argentin…
#> $ location          <chr> NA, NA, NA, "Suzhou", NA, NA, NA, NA, NA, NA, NA, N…
#> $ region_exposure   <chr> "Africa", "Africa", "Africa", "Asia", "South Americ…
#> $ country_exposure  <chr> "Algeria", "Algeria", "Algeria", "China", "Argentin…
#> $ division_exposure <chr> "Boufarik", "Blida", "Boufarik", "Anhui", NA, NA, N…
#> $ segment           <chr> "genome", "genome", "genome", "genome", "genome", "…
#> $ length            <dbl> 29862, 29867, 29862, 29860, 29903, 29903, 29903, 29…
#> $ host              <chr> "Human", "Human", "Human", "Human", "Human", "Human…
#> $ age               <chr> "28", "87", "41", "58", "51", "47", "73", "40", "40…
#> $ sex               <chr> "Female", "Male", "Male", "Male", "Male", "Male", "…
#> $ originating_lab   <chr> "NIC Viral Respiratory Unit - Institut Pasteur of A…
#> $ submitting_lab    <chr> "National Reference Center for Viruses of Respirato…
#> $ authors           <chr> "Albert et al", "Albert et al", "Albert et al", "Li…
#> $ url               <chr> "https://www.gisaid.org", "https://www.gisaid.org",…
#> $ title             <chr> "?", "?", "?", "?", "?", "?", "?", "?", "?", "?", "…
#> $ date_submitted    <date> 2020-03-29, 2020-03-29, 2020-04-04, 2020-03-05, 20…