Published parameter estimates from the literature

Published estimates of epidemiological characteristics from both peer-reviewed and non-peer-reviewed resources, encoded by community members and approved by authors. For complete information on each estimate, see the actual dataset items.

param_estimates_published()

Source

https://github.com/midas-network/COVID-19/blob/master/parameter_estimates/2019_novel_coronavirus/estimates.csv

Value

a data.frame

References

https://midasnetwork.us/covid-19/

See also

Examples

library(dplyr)
params = param_estimates_published()
colnames(params)
#>  [1] "id"                   "peer_review"          "peer_review_location"
#>  [4] "name"                 "abbreviation"         "units"               
#>  [7] "country"              "location_name"        "location_type"       
#> [10] "start_date"           "end_date"             "value_type"          
#> [13] "value"                "uncertainty_type"     "lower_bound"         
#> [16] "upper_bound"          "population"           "age_range"           
#> [19] "method_description"   "data_description"     "data_URL"            
#> [22] "date_publication"     "title_publication"    "type_publication"    
#> [25] "authors"              "publication_URL"      "entry_date"          
#> [28] "entry_person"         "entry_email"         
head(params)
#> # A tibble: 6 × 29
#>   id     peer_review peer_review_location name        abbreviation units country
#>   <chr>  <chr>       <chr>                <chr>       <chr>        <chr> <chr>  
#> 1 covid1 NA          NA                   cumulative… NA           pers… China  
#> 2 covid2 NA          MedRxiv              cumulative… NA           pers… China  
#> 3 covid3 NA          NA                   date of em… NA           NA    China  
#> 4 covid4 positive    Eurosurveillance     basic repr… R0           Nb    China  
#> 5 covid5 NA          BioRxiv              basic repr… R0           Nb    China  
#> 6 covid6 NA          NA                   basic repr… R0           Nb    China  
#> # … with 22 more variables: location_name <chr>, location_type <chr>,
#> #   start_date <date>, end_date <date>, value_type <chr>, value <chr>,
#> #   uncertainty_type <chr>, lower_bound <chr>, upper_bound <chr>,
#> #   population <chr>, age_range <chr>, method_description <chr>,
#> #   data_description <chr>, data_URL <chr>, date_publication <date>,
#> #   title_publication <chr>, type_publication <chr>, authors <chr>,
#> #   publication_URL <chr>, entry_date <date>, entry_person <chr>, …
dplyr::glimpse(params)
#> Rows: 845
#> Columns: 29
#> $ id                   <chr> "covid1", "covid2", "covid3", "covid4", "covid5",…
#> $ peer_review          <chr> NA, NA, NA, "positive", NA, NA, "positive", NA, N…
#> $ peer_review_location <chr> NA, "MedRxiv", NA, "Eurosurveillance", "BioRxiv",…
#> $ name                 <chr> "cumulative case count", "cumulative case count",…
#> $ abbreviation         <chr> NA, NA, NA, "R0", "R0", "R0", "R0", "R0", "k", NA…
#> $ units                <chr> "persons", "persons", NA, "Nb", "Nb", "Nb", "Nb",…
#> $ country              <chr> "China", "China", "China", "China", "China", "Chi…
#> $ location_name        <chr> "Wuhan City", "Wuhan City", "Wuhan City", "Wuhan …
#> $ location_type        <chr> "city", "metropolitan area", "city", "city", "cit…
#> $ start_date           <date> 2020-01-18, 2019-12-31, NA, 2020-01-09, 2019-11-…
#> $ end_date             <date> 2020-01-18, 2020-01-29, NA, 2020-01-11, 2020-01-…
#> $ value_type           <chr> "central estimate, unspecified", "median", NA, NA…
#> $ value                <chr> "4000", "31200", NA, "0.3", "2.2", "2.6", "2.24",…
#> $ uncertainty_type     <chr> "confidence interval 95%", "confidence interval 9…
#> $ lower_bound          <chr> "1000", "23400", "2020-11-19", "0.17", "1.4", "1.…
#> $ upper_bound          <chr> "9700", "40400", "2020-12-19", "0.44", "3.8", "3.…
#> $ population           <chr> "general population", "general population", "gene…
#> $ age_range            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ method_description   <chr> NA, "mobility model assuming travelers are expose…
#> $ data_description     <chr> "7 imported cases internationally as of  9:00 GMT…
#> $ data_URL             <chr> NA, NA, NA, NA, "https://github.com/jriou/wcov", …
#> $ date_publication     <date> 2020-01-22, 2020-01-29, 2020-01-23, 2020-01-12, …
#> $ title_publication    <chr> "Report 3: Transmissibility of 2019-nCoV", "Early…
#> $ type_publication     <chr> "report", "report", "executive summary", "report"…
#> $ authors              <chr> "Imai N., Dorigatti I., Cori A., Donnelly C., Ril…
#> $ publication_URL      <chr> "https://www.imperial.ac.uk/mrc-global-infectious…
#> $ entry_date           <date> 2020-01-24, 2020-01-24, 2020-01-24, 2020-01-24, …
#> $ entry_person         <chr> "Cecile Viboud", "Cecile Viboud", "Cecile Viboud"…
#> $ entry_email          <chr> "viboudc@mail.nih.gov", "viboudc@mail.nih.gov", "…

summary(params)
#>       id            peer_review        peer_review_location     name          
#>  Length:845         Length:845         Length:845           Length:845        
#>  Class :character   Class :character   Class :character     Class :character  
#>  Mode  :character   Mode  :character   Mode  :character     Mode  :character  
#>                                                                               
#>                                                                               
#>                                                                               
#>                                                                               
#>  abbreviation          units             country          location_name     
#>  Length:845         Length:845         Length:845         Length:845        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  location_type        start_date            end_date         
#>  Length:845         Min.   :2019-11-01   Min.   :2019-02-04  
#>  Class :character   1st Qu.:2019-12-31   1st Qu.:2020-02-13  
#>  Mode  :character   Median :2020-01-14   Median :2020-04-16  
#>                     Mean   :2020-01-21   Mean   :2020-03-26  
#>                     3rd Qu.:2020-01-23   3rd Qu.:2020-04-24  
#>                     Max.   :2020-12-24   Max.   :2020-07-25  
#>                     NA's   :324          NA's   :131         
#>   value_type           value           uncertainty_type   lower_bound       
#>  Length:845         Length:845         Length:845         Length:845        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  upper_bound         population         age_range         method_description
#>  Length:845         Length:845         Length:845         Length:845        
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  data_description     data_URL         date_publication     title_publication 
#>  Length:845         Length:845         Min.   :2020-01-12   Length:845        
#>  Class :character   Class :character   1st Qu.:2020-03-19   Class :character  
#>  Mode  :character   Mode  :character   Median :2020-04-01   Mode  :character  
#>                                        Mean   :2020-04-04                     
#>                                        3rd Qu.:2020-04-26                     
#>                                        Max.   :2020-06-08                     
#>                                                                               
#>  type_publication     authors          publication_URL      entry_date        
#>  Length:845         Length:845         Length:845         Min.   :2020-01-24  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-24  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-04-02  
#>                                                           Mean   :2020-04-09  
#>                                                           3rd Qu.:2020-04-29  
#>                                                           Max.   :2020-06-19  
#>                                                                               
#>  entry_person       entry_email       
#>  Length:845         Length:845        
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>                                       

if(require(DT) & interactive()) {
    datatable(params)
}
#> Loading required package: DT