plot the disease course and death projections

Using data from https://covid19.healthdata.org/projections, either render a plot or return a ggplot object for further manipulation.

plot_projection(
  location_name,
  projections = healthdata_projections_data(),
  metrics = unique(projections$metric)
)

Arguments

location_name
character(1), chosen from unique(projections$location_name) that is used to filter the projections data.frame to one location.

projections
supply a data.frame, typically from a call to healthdata_projections_data(). This data.frame must have columns that include date, location_name, metric, mean, upper, and lower.

metrics
character vector of metrics to include on plot. Get a list of these from the metric column.

Source

https://covid19.healthdata.org/projections

Details

This function is a convenience function to return a ggplot2 object that displays projections for a given location. Projections are typically coming from a call to healthdata_projections_data(). If assigned to a variable, the results can be further manipulated or passed into any function that accepts a ggplot2 object (like plotly, cowplot, etc.). See examples for details.

See also

Examples

library(sars2pack)

hdp = healthdata_projections_data()

library(ggplot2)
# possible locations
sort(unique(hdp$location_name))

#>   [1] "Abruzzo"                            "Acre"                              
#>   [3] "Afghanistan"                        "Aguascalientes"                    
#>   [5] "Alabama"                            "Alagoas"                           
#>   [7] "Alaska"                             "Albania"                           
#>   [9] "Alberta"                            "Algeria"                           
#>  [11] "Amapá"                              "Amazonas"                          
#>  [13] "Andalucia"                          "Andorra"                           
#>  [15] "Angola"                             "Antigua and Barbuda"               
#>  [17] "Aragon"                             "Argentina"                         
#>  [19] "Arizona"                            "Arkansas"                          
#>  [21] "Armenia"                            "Asturias"                          
#>  [23] "Australia"                          "Austria"                           
#>  [25] "Azad Jammu & Kashmir"               "Azerbaijan"                        
#>  [27] "Baden-Württemberg"                  "Bahamas"                           
#>  [29] "Bahia"                              "Bahrain"                           
#>  [31] "Baja California"                    "Baja California Sur"               
#>  [33] "Balearic Islands"                   "Balochistan"                       
#>  [35] "Bangladesh"                         "Barbados"                          
#>  [37] "Basilicata"                         "Basque Country"                    
#>  [39] "Bavaria"                            "Belarus"                           
#>  [41] "Belgium"                            "Belize"                            
#>  [43] "Benin"                              "Berlin"                            
#>  [45] "Bolivia (Plurinational State of)"   "Bosnia and Herzegovina"            
#>  [47] "Botswana"                           "Brandenburg"                       
#>  [49] "Brazil"                             "Bremen"                            
#>  [51] "British Columbia"                   "Brunei Darussalam"                 
#>  [53] "Bulgaria"                           "Burkina Faso"                      
#>  [55] "Cabo Verde"                         "Calabria"                          
#>  [57] "California"                         "Cameroon"                          
#>  [59] "Campania"                           "Campeche"                          
#>  [61] "Canada"                             "Canary Islands"                    
#>  [63] "Cantabria"                          "Castile and León"                  
#>  [65] "Castilla-La Mancha"                 "Catalonia"                         
#>  [67] "Ceará"                              "Central African Republic"          
#>  [69] "Ceuta"                              "Chad"                              
#>  [71] "Chiapas"                            "Chihuahua"                         
#>  [73] "Chile"                              "Coahuila"                          
#>  [75] "Colima"                             "Colombia"                          
#>  [77] "Colorado"                           "Community of Madrid"               
#>  [79] "Comoros"                            "Congo"                             
#>  [81] "Connecticut"                        "Costa Rica"                        
#>  [83] "Côte d'Ivoire"                      "Croatia"                           
#>  [85] "Cuba"                               "Cyprus"                            
#>  [87] "Czechia"                            "Delaware"                          
#>  [89] "Democratic Republic of the Congo"   "Denmark"                           
#>  [91] "District of Columbia"               "Distrito Federal"                  
#>  [93] "Djibouti"                           "Dominican Republic"                
#>  [95] "Durango"                            "Ecuador"                           
#>  [97] "Egypt"                              "El Salvador"                       
#>  [99] "Emilia-Romagna"                     "Equatorial Guinea"                 
#> [101] "Eritrea"                            "Espírito Santo"                    
#> [103] "Estonia"                            "Eswatini"                          
#> [105] "Extremadura"                        "Finland"                           
#> [107] "Florida"                            "France"                            
#> [109] "Friuli-Venezia Giulia"              "Gabon"                             
#> [111] "Galicia"                            "Gambia"                            
#> [113] "Georgia"                            "Germany"                           
#> [115] "Ghana"                              "Gilgit-Baltistan"                  
#> [117] "Goiás"                              "Greece"                            
#> [119] "Guam"                               "Guanajuato"                        
#> [121] "Guatemala"                          "Guerrero"                          
#> [123] "Guinea"                             "Guinea-Bissau"                     
#> [125] "Guyana"                             "Haiti"                             
#> [127] "Hamburg"                            "Hawaii"                            
#> [129] "Hesse"                              "Hidalgo"                           
#> [131] "Honduras"                           "Hungary"                           
#> [133] "Iceland"                            "Idaho"                             
#> [135] "Illinois"                           "Indiana"                           
#> [137] "Iowa"                               "Iran (Islamic Republic of)"        
#> [139] "Iraq"                               "Ireland"                           
#> [141] "Islamabad Capital Territory"        "Israel"                            
#> [143] "Italy"                              "Jalisco"                           
#> [145] "Jamaica"                            "Japan"                             
#> [147] "Jordan"                             "Kansas"                            
#> [149] "Kazakhstan"                         "Kentucky"                          
#> [151] "Kenya"                              "Khyber Pakhtunkhwa"                
#> [153] "Kuwait"                             "Kyrgyzstan"                        
#> [155] "La Rioja"                           "Latvia"                            
#> [157] "Lazio"                              "Lebanon"                           
#> [159] "Liberia"                            "Libya"                             
#> [161] "Liguria"                            "Lithuania"                         
#> [163] "Lombardia"                          "Louisiana"                         
#> [165] "Lower Saxony"                       "Luxembourg"                        
#> [167] "Madagascar"                         "Maine"                             
#> [169] "Malawi"                             "Malaysia"                          
#> [171] "Maldives"                           "Mali"                              
#> [173] "Malta"                              "Manitoba"                          
#> [175] "Maranhão"                           "Marche"                            
#> [177] "Maryland"                           "Massachusetts"                     
#> [179] "Mato Grosso"                        "Mato Grosso do Sul"                
#> [181] "Mauritania"                         "Mauritius"                         
#> [183] "Mecklenburg-Vorpommern"             "Melilla"                           
#> [185] "Mexico"                             "México"                            
#> [187] "Mexico City"                        "Michigan"                          
#> [189] "Michoacán de Ocampo"                "Minas Gerais"                      
#> [191] "Minnesota"                          "Mississippi"                       
#> [193] "Missouri"                           "Molise"                            
#> [195] "Monaco"                             "Montana"                           
#> [197] "Montenegro"                         "Morelos"                           
#> [199] "Morocco"                            "Mozambique"                        
#> [201] "Murcia"                             "Myanmar"                           
#> [203] "Navarre"                            "Nayarit"                           
#> [205] "Nebraska"                           "Nepal"                             
#> [207] "Netherlands"                        "Nevada"                            
#> [209] "New Brunswick"                      "New Hampshire"                     
#> [211] "New Jersey"                         "New Mexico"                        
#> [213] "New York"                           "New Zealand"                       
#> [215] "Newfoundland and Labrador"          "Nicaragua"                         
#> [217] "Niger"                              "North Carolina"                    
#> [219] "North Dakota"                       "North Macedonia"                   
#> [221] "North Rhine-Westphalia"             "Northwest Territories"             
#> [223] "Norway"                             "Nova Scotia"                       
#> [225] "Nuevo León"                         "Nunavut"                           
#> [227] "Oaxaca"                             "Ohio"                              
#> [229] "Oklahoma"                           "Oman"                              
#> [231] "Ontario"                            "Oregon"                            
#> [233] "Pakistan"                           "Palestine"                         
#> [235] "Panama"                             "Papua New Guinea"                  
#> [237] "Pará"                               "Paraguay"                          
#> [239] "Paraíba"                            "Paraná"                            
#> [241] "Pennsylvania"                       "Pernambuco"                        
#> [243] "Peru"                               "Philippines"                       
#> [245] "Piaui"                              "Piemonte"                          
#> [247] "Poland"                             "Portugal"                          
#> [249] "Prince Edward Island"               "Provincia autonoma di Bolzano"     
#> [251] "Provincia autonoma di Trento"       "Puebla"                            
#> [253] "Puerto Rico"                        "Puglia"                            
#> [255] "Punjab"                             "Qatar"                             
#> [257] "Quebec"                             "Querétaro"                         
#> [259] "Quintana Roo"                       "Republic of Korea"                 
#> [261] "Republic of Moldova"                "Rhineland-Palatinate"              
#> [263] "Rhode Island"                       "Rio de Janeiro"                    
#> [265] "Rio Grande do Norte"                "Rio Grande do Sul"                 
#> [267] "Romania"                            "Rondônia"                          
#> [269] "Roraima"                            "Russian Federation"                
#> [271] "Rwanda"                             "Saarland"                          
#> [273] "San Luis Potosí"                    "San Marino"                        
#> [275] "Santa Catarina"                     "São Paulo"                         
#> [277] "Sao Tome and Principe"              "Sardegna"                          
#> [279] "Saskatchewan"                       "Saudi Arabia"                      
#> [281] "Saxony"                             "Saxony-Anhalt"                     
#> [283] "Schleswig-Holstein"                 "Senegal"                           
#> [285] "Serbia"                             "Sergipe"                           
#> [287] "Sicilia"                            "Sierra Leone"                      
#> [289] "Sinaloa"                            "Sindh"                             
#> [291] "Singapore"                          "Slovakia"                          
#> [293] "Slovenia"                           "Somalia"                           
#> [295] "Sonora"                             "South Carolina"                    
#> [297] "South Dakota"                       "South Sudan"                       
#> [299] "Spain"                              "Sri Lanka"                         
#> [301] "Sudan"                              "Suriname"                          
#> [303] "Sweden"                             "Switzerland"                       
#> [305] "Syrian Arab Republic"               "Tabasco"                           
#> [307] "Tajikistan"                         "Tamaulipas"                        
#> [309] "Tennessee"                          "Texas"                             
#> [311] "Thailand"                           "Thuringia"                         
#> [313] "Tlaxcala"                           "Tocantins"                         
#> [315] "Togo"                               "Toscana"                           
#> [317] "Trinidad and Tobago"                "Turkey"                            
#> [319] "Uganda"                             "Ukraine"                           
#> [321] "Umbria"                             "United Arab Emirates"              
#> [323] "United Kingdom"                     "United States of America"          
#> [325] "United States Virgin Islands"       "Uruguay"                           
#> [327] "Utah"                               "Uzbekistan"                        
#> [329] "Valencian Community"                "Valle d'Aosta"                     
#> [331] "Veneto"                             "Venezuela (Bolivarian Republic of)"
#> [333] "Veracruz de Ignacio de la Llave"    "Vermont"                           
#> [335] "Virginia"                           "Washington"                        
#> [337] "West Virginia"                      "Wisconsin"                         
#> [339] "Wyoming"                            "Yemen"                             
#> [341] "Yucatán"                            "Yukon"                             
#> [343] "Zacatecas"                          "Zambia"                            
#> [345] "Zimbabwe"                          

# create a ggplot2 object
nyp = plot_projection(projections = hdp,
    location_name='New York')

#> Warning: Unknown or uninitialised column: `metric`.

#> Error: 

# and do the plot
print(nyp)

#> Error in print(nyp): object 'nyp' not found

# Zoomable html version using plotly
if(require(plotly)) {
  ggplotly(nyp)
}

#> Error in ggplotly(nyp): object 'nyp' not found

# California
calip = plot_projection(projections = hdp,
    location_name='California')

#> Warning: Unknown or uninitialised column: `metric`.

#> Error: 

print(calip)

#> Error in print(calip): object 'calip' not found

# entire US
plot_projection(location_name = "United States of America", projections = hdp)

#> Warning: Unknown or uninitialised column: `metric`.

#> Error: 

# Compare two countries
spain_proj = plot_projection(location_name = "Spain", projections = hdp)

#> Warning: Unknown or uninitialised column: `metric`.

#> Error: 

portugal_proj = plot_projection(location_name = "Portugal", projections = hdp)

#> Warning: Unknown or uninitialised column: `metric`.

#> Error: 

#'
print(spain_proj)

#> Error in print(spain_proj): object 'spain_proj' not found

print(portugal_proj)

#> Error in print(portugal_proj): object 'portugal_proj' not found

# plot both on same plot
if(require(cowplot)) {
  plot_grid(spain_proj, portugal_proj)
}

#> Loading required package: cowplot

#> 
#> ********************************************************

#> Note: As of version 1.0.0, cowplot does not change the

#>   default ggplot2 theme anymore. To recover the previous

#>   behavior, execute:
#>   theme_set(theme_cowplot())

#> ********************************************************

#> Error in plot_grid(spain_proj, portugal_proj): object 'spain_proj' not found