United States county health rankings

The County Health Rankings, a collaboration between the Robert Wood Johnson Foundation and the University of Wisconsin Population Health Institute, measure the health of nearly all counties in the nation and rank them within states. The Rankings are compiled using county-level measures from a variety of national and state data sources.

us_county_health_rankings()

Details

The County Health Rankings gather data from around the nation that is comparable between counties, within states. For most of their measures, county data is also comparable across state lines. However, for a few of their measures, some caution must be exercised when making comparisons between counties in different states. See references below for details.

References

https://www.countyhealthrankings.org/sites/default/files/media/document/2020%20Analytic%20Documentation_0.pdf

https://www.countyhealthrankings.org/sites/default/files/media/document/2020%20Comparability%20across%20states.pdf

See also

Examples

ret = us_county_health_rankings()
head(ret)

#> # A tibble: 6 x 9
#>   statecode countycode fips  state county   year county_ranked variable    value
#>   <chr>     <chr>      <chr> <chr> <chr>   <dbl> <lgl>         <chr>       <dbl>
#> 1 00        000        00000 US    United…  2020 NA            Premature… 6.94e3
#> 2 00        000        00000 US    United…  2020 NA            Premature… 3.81e6
#> 3 00        000        00000 US    United…  2020 NA            Premature… 9.12e8
#> 4 00        000        00000 US    United…  2020 NA            Premature… 6.93e3
#> 5 00        000        00000 US    United…  2020 NA            Premature… 6.95e3
#> 6 00        000        00000 US    United…  2020 NA            Premature… 0.    

colnames(ret)

#> [1] "statecode"     "countycode"    "fips"          "state"        
#> [5] "county"        "year"          "county_ranked" "variable"     
#> [9] "value"