US hospital dataset

Includes details of beds, NAICS codes, and geolocation data for US states and territories.

us_hospital_details()

Source

https://hub.arcgis.com/datasets/geoplatform::hospitals

Value

a data.frame

Details

This dataset contains locations of Hospitals for 50 US states, Washington D.C., US territories of Puerto Rico, Guam, American Samoa, Northern Mariana Islands, Palau, and Virgin Islands. The dataset only includes hospital facilities based on data acquired from various state departments or federal sources which has been referenced in the SOURCE field. Hospital facilities which do not occur in these sources will be not present in the database. The source data was available in a variety of formats (pdfs, tables, webpages, etc.) which was cleaned and geocoded and then converted into a spatial database. The database does not contain nursing homes or health centers. Hospitals have been categorized into children, chronic disease, critical access, general acute care, long term care, military, psychiatric, rehabilitation, special, and women based on the range of the available values from the various sources after removing similarities. In this update the TRAUMA field was populated for 172 additional hospitals and helipad presence were verified for all hospitals.

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_hospital_details()
head(res)
#> Simple feature collection with 6 features and 33 fields
#> Attribute-geometry relationship: 33 constant, 0 aggregate, 0 identity
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -119.6457 ymin: 33.85971 xmax: -117.9674 ymax: 36.33616
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 34
#>       x     y objectid id        name  address city  state zip   zip4  telephone
#>   <dbl> <dbl>    <dbl> <chr>     <chr> <chr>   <chr> <chr> <chr> <chr> <chr>    
#> 1 -120.  36.3        1 00057932… CENT… 1025 N… HANF… CA    93230 NOT … NOT AVAI…
#> 2 -119.  34.2        2 00533913… LOS … 150 VI… WEST… CA    91362 NOT … NOT AVAI…
#> 3 -118.  34.0        3 00111900… EAST… 4060 W… LOS … CA    90023 NOT … NOT AVAI…
#> 4 -118.  34.1        4 00170900… SOUT… 6245 D… HOLL… CA    90028 NOT … (323) 46…
#> 5 -118.  34.1        5 00236917… KIND… 14148 … BALD… CA    91706 NOT … NOT AVAI…
#> 6 -118.  33.9        6 00251907… LAKE… 3700 E… LAKE… CA    90712 NOT … NOT AVAI…
#> # … with 23 more variables: type <chr>, status <chr>, population <dbl>,
#> #   county <chr>, countyfips <chr>, country <chr>, naics_code <dbl>,
#> #   naics_desc <chr>, source <chr>, sourcedate <chr>, val_method <chr>,
#> #   val_date <chr>, website <chr>, state_id <chr>, alt_name <chr>,
#> #   st_fips <chr>, owner <chr>, ttl_staff <dbl>, beds <dbl>, trauma <chr>,
#> #   helipad <chr>, globalid <chr>, geometry <POINT [°]>
colnames(res)
#>  [1] "x"          "y"          "objectid"   "id"         "name"      
#>  [6] "address"    "city"       "state"      "zip"        "zip4"      
#> [11] "telephone"  "type"       "status"     "population" "county"    
#> [16] "countyfips" "country"    "naics_code" "naics_desc" "source"    
#> [21] "sourcedate" "val_method" "val_date"   "website"    "state_id"  
#> [26] "alt_name"   "st_fips"    "owner"      "ttl_staff"  "beds"      
#> [31] "trauma"     "helipad"    "globalid"   "geometry"  
summary(res)
#>        x                 y             objectid          id           
#>  Min.   :-176.64   Min.   :-14.29   Min.   :    1   Length:76398      
#>  1st Qu.: -98.23   1st Qu.: 33.45   1st Qu.:19115   Class :character  
#>  Median : -90.14   Median : 37.98   Median :38214   Mode  :character  
#>  Mean   : -92.45   Mean   : 37.33   Mean   :38214                     
#>  3rd Qu.: -81.80   3rd Qu.: 41.35   3rd Qu.:57314                     
#>  Max.   : 145.72   Max.   : 71.30   Max.   :76413                     
#>      name             address              city              state          
#>  Length:76398       Length:76398       Length:76398       Length:76398      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>      zip                zip4            telephone             type          
#>  Length:76398       Length:76398       Length:76398       Length:76398      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>     status            population         county           countyfips       
#>  Length:76398       Min.   :-999.00   Length:76398       Length:76398      
#>  Class :character   1st Qu.:  25.00   Class :character   Class :character  
#>  Mode  :character   Median :  70.00   Mode  :character   Mode  :character  
#>                     Mean   :  90.35                                        
#>                     3rd Qu.: 184.00                                        
#>                     Max.   :1592.00                                        
#>    country            naics_code      naics_desc           source         
#>  Length:76398       Min.   :622110   Length:76398       Length:76398      
#>  Class :character   1st Qu.:622110   Class :character   Class :character  
#>  Mode  :character   Median :622110   Mode  :character   Mode  :character  
#>                     Mean   :622142                                        
#>                     3rd Qu.:622110                                        
#>                     Max.   :622310                                        
#>   sourcedate         val_method          val_date           website         
#>  Length:76398       Length:76398       Length:76398       Length:76398      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>    state_id           alt_name           st_fips             owner          
#>  Length:76398       Length:76398       Length:76398       Length:76398      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>    ttl_staff         beds            trauma            helipad         
#>  Min.   :-999   Min.   :-999.00   Length:76398       Length:76398      
#>  1st Qu.:-999   1st Qu.:  25.00   Class :character   Class :character  
#>  Median :-999   Median :  70.00   Mode  :character   Mode  :character  
#>  Mean   :-999   Mean   :  90.35                                        
#>  3rd Qu.:-999   3rd Qu.: 184.00                                        
#>  Max.   :-999   Max.   :1592.00                                        
#>    globalid                  geometry    
#>  Length:76398       POINT        :76398  
#>  Class :character   epsg:4326    :    0  
#>  Mode  :character   +proj=long...:    0  
#>                                          
#>                                          
#>