United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Examples

res = us_state_distancing_policy()
colnames(res)

#>  [1] "PID"                 "location_id"         "StateFIPS"          
#>  [4] "iso2c"               "state"               "StatePolicy"        
#>  [7] "Mandate"             "StateWide"           "SWGeo"              
#> [10] "SWPop"               "Encourages"          "InGathLim"          
#> [13] "OutGathLim"          "InGathLimReligious"  "OutGathLimReligious"
#> [16] "PublicMaskLevel"     "DateIssued"          "DateEnacted"        
#> [19] "DateExpiry"          "DateEnded"           "Extends"            
#> [22] "Expands"             "Eases"               "Ends"               
#> [25] "Depends"             "Joins"               "PolicyCodingNotes"  
#> [28] "PolicySource"        "PolicySourceID"      "LastUpdated"        
#> [31] "LastUpdatedNotes"    "UpdatingSource1"     "UpdatingSource2"    

library(dplyr)
dplyr::glimpse(res)

#> Rows: 5,130
#> Columns: 33
#> $ PID                 <chr> "AL0001", "AL0002", "AL0080", "AL0013", "AL0015",…
#> $ location_id         <dbl> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523,…
#> $ StateFIPS           <chr> "00001", "00001", "00001", "00001", "00001", "000…
#> $ iso2c               <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", "…
#> $ state               <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Alab…
#> $ StatePolicy         <chr> "EmergDec", "SchoolClose", "BarRestrict", "GathRe…
#> $ Mandate             <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, T…
#> $ StateWide           <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, TRUE, TRU…
#> $ SWGeo               <dbl> 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ SWPop               <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ Encourages          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ InGathLim           <dbl> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA, 1…
#> $ OutGathLim          <dbl> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA, 1…
#> $ InGathLimReligious  <dbl> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA, 1…
#> $ OutGathLimReligious <dbl> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA, 1…
#> $ PublicMaskLevel     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ DateIssued          <date> 2020-03-13, 2020-03-13, 2020-03-17, 2020-03-17, …
#> $ DateEnacted         <date> 2020-03-13, 2020-03-19, 2020-03-17, 2020-03-17, …
#> $ DateExpiry          <date> NA, 2020-04-06, 2020-03-24, NA, 2020-03-24, 2020…
#> $ DateEnded           <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ Extends             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ Expands             <chr> NA, NA, NA, NA, NA, NA, "AL0080", "AL0013", NA, "…
#> $ Eases               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ Ends                <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ Depends             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ Joins               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ PolicyCodingNotes   <chr> "Public Health Emergency declared. No clear expir…
#> $ PolicySource        <chr> "https://governor.alabama.gov/assets/2020/03/2020…
#> $ PolicySourceID      <chr> "Alabama0001", "Alabama0002", "Alabama0004", "Ala…
#> $ LastUpdated         <date> 2020-03-24, 2020-04-30, 2020-07-03, 2020-07-24, …
#> $ LastUpdatedNotes    <chr> NA, "Updated DateIssued to 20200313, DateEnacted …
#> $ UpdatingSource1     <chr> "https://covid19.alaska.gov/health-mandates/", "h…
#> $ UpdatingSource2     <chr> "https://www.alabamapublichealth.gov/legal/orders…

summary(res)

#>      PID             location_id     StateFIPS            iso2c          
#>  Length:5130        Min.   :523.0   Length:5130        Length:5130       
#>  Class :character   1st Qu.:536.0   Class :character   Class :character  
#>  Mode  :character   Median :550.0   Mode  :character   Mode  :character  
#>                     Mean   :549.1                                        
#>                     3rd Qu.:561.0                                        
#>                     Max.   :575.0                                        
#>                                                                          
#>     state           StatePolicy         Mandate        StateWide      
#>  Length:5130        Length:5130        Mode :logical   Mode :logical  
#>  Class :character   Class :character   FALSE:521       FALSE:1352     
#>  Mode  :character   Mode  :character   TRUE :4609      TRUE :3778     
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>      SWGeo            SWPop          Encourages     InGathLim      
#>  Min.   :0.0000   Min.   :0.0000   Min.   :1      Min.   :    0.0  
#>  1st Qu.:1.0000   1st Qu.:1.0000   1st Qu.:1      1st Qu.:   10.0  
#>  Median :1.0000   Median :1.0000   Median :1      Median :   10.0  
#>  Mean   :0.7573   Mean   :0.9774   Mean   :1      Mean   :  113.8  
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:1      3rd Qu.:   50.0  
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1      Max.   :10000.0  
#>                                    NA's   :5074   NA's   :4572     
#>    OutGathLim    InGathLimReligious OutGathLimReligious PublicMaskLevel   
#>  Min.   :    0   Min.   :   0.00    Min.   :   0.0      Length:5130       
#>  1st Qu.:   10   1st Qu.:  10.00    1st Qu.:  10.0      Class :character  
#>  Median :   10   Median :  15.00    Median :  10.0      Mode  :character  
#>  Mean   :  124   Mean   :  60.13    Mean   :  54.3                        
#>  3rd Qu.:   50   3rd Qu.:  50.00    3rd Qu.:  50.0                        
#>  Max.   :10000   Max.   :1000.00    Max.   :1000.0                        
#>  NA's   :4573    NA's   :4728       NA's   :4757                          
#>    DateIssued          DateEnacted           DateExpiry        
#>  Min.   :2020-02-29   Min.   :2020-02-29   Min.   :2020-03-21  
#>  1st Qu.:2020-04-28   1st Qu.:2020-04-30   1st Qu.:2020-05-15  
#>  Median :2020-05-28   Median :2020-05-30   Median :2020-06-15  
#>  Mean   :2020-05-26   Mean   :2020-05-27   Mean   :2020-06-18  
#>  3rd Qu.:2020-06-29   3rd Qu.:2020-06-30   3rd Qu.:2020-07-22  
#>  Max.   :2020-08-14   Max.   :2020-09-14   Max.   :2021-06-30  
#>                                            NA's   :1759        
#>    DateEnded            Extends            Expands             Eases          
#>  Min.   :2020-04-26   Length:5130        Length:5130        Length:5130       
#>  1st Qu.:2020-05-26   Class :character   Class :character   Class :character  
#>  Median :2020-06-10   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :2020-06-15                                                           
#>  3rd Qu.:2020-07-06                                                           
#>  Max.   :2020-09-14                                                           
#>  NA's   :4771                                                                 
#>      Ends             Depends             Joins           PolicyCodingNotes 
#>  Length:5130        Length:5130        Length:5130        Length:5130       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  PolicySource       PolicySourceID      LastUpdated         LastUpdatedNotes  
#>  Length:5130        Length:5130        Min.   :2020-03-24   Length:5130       
#>  Class :character   Class :character   1st Qu.:2020-06-01   Class :character  
#>  Mode  :character   Mode  :character   Median :2020-07-05   Mode  :character  
#>                                        Mean   :2020-06-27                     
#>                                        3rd Qu.:2020-07-24                     
#>                                        Max.   :2020-08-16                     
#>                                                                               
#>  UpdatingSource1    UpdatingSource2   
#>  Length:5130        Length:5130       
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>