retrieve time series dataset from USA Facts

This always does a web call. The assumption here is that the user is online and that github is up and that URLs are stable.

usa_facts_data()

Value

an object of class s2p_long_df that inherits from tbl_df

Note

Uses https://usafacts.org/visualizations/coronavirus-covid-19-spread-map/ as data source, then modifies column names and munges to long form table.

See also

Examples

res = usa_facts_data()

#> Warning: Missing column names filled in: 'X70' [70], 'X71' [71], 'X72' [72], 'X73' [73], 'X74' [74], 'X75' [75], 'X76' [76], 'X77' [77], 'X78' [78], 'X79' [79], 'X80' [80], 'X81' [81], 'X82' [82], 'X83' [83], 'X84' [84], 'X85' [85], 'X86' [86], 'X87' [87], 'X88' [88], 'X89' [89], 'X90' [90], 'X91' [91], 'X92' [92], 'X93' [93], 'X94' [94], 'X95' [95], 'X96' [96], 'X97' [97], 'X98' [98], 'X99' [99]

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   `County Name` = col_character(),
#>   State = col_character(),
#>   X70 = col_logical(),
#>   X71 = col_logical(),
#>   X72 = col_logical(),
#>   X73 = col_logical(),
#>   X74 = col_logical(),
#>   X75 = col_logical(),
#>   X76 = col_logical(),
#>   X77 = col_logical(),
#>   X78 = col_logical(),
#>   X79 = col_logical(),
#>   X80 = col_logical(),
#>   X81 = col_logical(),
#>   X82 = col_logical(),
#>   X83 = col_logical(),
#>   X84 = col_logical(),
#>   X85 = col_logical(),
#>   X86 = col_logical(),
#>   X87 = col_logical()
#>   # ... with 12 more columns
#> )

#> See spec(...) for full column specifications.

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   `County Name` = col_character(),
#>   State = col_character()
#> )

#> See spec(...) for full column specifications.

#> Warning:  100200 failed to parse.

colnames(res)

#> [1] "countyFIPS" "County"     "State"      "stateFIPS"  "subset"    
#> [6] "date"       "count"     

head(res)

#> # A tibble: 6 x 7
#>   countyFIPS County                State stateFIPS subset    date       count
#>        <dbl> <chr>                 <chr>     <dbl> <chr>     <date>     <dbl>
#> 1          0 Statewide Unallocated AL            1 confirmed 2020-01-22     0
#> 2          0 Statewide Unallocated AL            1 confirmed 2020-01-23     0
#> 3          0 Statewide Unallocated AL            1 confirmed 2020-01-24     0
#> 4          0 Statewide Unallocated AL            1 confirmed 2020-01-25     0
#> 5          0 Statewide Unallocated AL            1 confirmed 2020-01-26     0
#> 6          0 Statewide Unallocated AL            1 confirmed 2020-01-27     0