Access Apple mobility data

From Apple's website: "Learn about COVID‑19 mobility trends in countries/regions and cities. Reports are published daily and reflect requests for directions in Apple Maps. Privacy is one of our core values, so Maps doesn’t associate your data with your Apple ID, and Apple doesn’t keep a history of where you’ve been."

apple_mobility_data()

Details

The CSV file and charts on this site show a relative volume of directions requests per country/region or city compared to a baseline volume on January 13th, 2020.

We define our day as midnight-to-midnight, Pacific time. Cities represent usage in greater metropolitan areas and are stably defined during this period. In many countries/regions and cities, relative volume has increased since January 13th, consistent with normal, seasonal usage of Apple Maps. Day of week effects are important to normalize as you use this data.

Data that is sent from users’ devices to the Maps service is associated with random, rotating identifiers so Apple doesn’t have a profile of your movements and searches. Apple Maps has no demographic information about our users, so we can’t make any statements about the representativeness of our usage against the overall population.

These data are available from a URL that changes daily. The parent page is the place to check to see what is going on if there are problems.

Note

Apple requires that all users agree to their terms of use. See https://www.apple.com/covid19/mobility.

References

https://www.apple.com/covid19/mobility

.

See also

Examples

dat = apple_mobility_data()
colnames(res)

#> Error in is.data.frame(x): object 'res' not found

head(res)

#> Error in head(res): object 'res' not found

table(dat$transportation_type)

#> 
#> driving transit walking 
#>   15200    9100   15200