list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 36 x 8
#>    name    accessor  data_type geographical geospatial region resolution url    
#>    <chr>   <chr>     <chr>     <lgl>        <lgl>      <chr>  <chr>      <chr>  
#>  1 Corona… coronane… public p… TRUE         FALSE      World  admin0     https:…
#>  2 Corona… coronane… public p… TRUE         FALSE      World  admin1     https:…
#>  3 Our Wo… owid_data time ser… TRUE         FALSE      World  admin0     https:…
#>  4 Our Wo… owid_data cumulati… TRUE         FALSE      World  admin0     https:…
#>  5 Our Wo… owid_data deaths    TRUE         FALSE      World  admin0     https:…
#>  6 Our Wo… owid_data tests     TRUE         FALSE      World  admin0     https:…
#>  7 GISAID… cov_glue… individu… TRUE         FALSE      World  multiple   https:…
#>  8 Newick… cov_glue… phylogen… FALSE        FALSE      World  multiple   https:…
#>  9 US CDC… cdc_exce… miscella… TRUE         FALSE      US     admin1     https:…
#> 10 Publis… param_es… miscella… FALSE        FALSE      NA     NA         https:…
#> # … with 26 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 13,666 x 40
#>    record_id policy_id recorded_date       date_updated date_announced
#>    <chr>         <dbl> <dttm>              <date>       <date>        
#>  1 1000775Cn   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14    
#>  2 1000775Cf   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14    
#>  3 1003875NA   1003875 2020-04-15 11:04:50 2020-04-15   2020-04-07    
#>  4 1004681NA   2575122 2020-04-03 08:33:25 2020-04-10   2020-03-22    
#>  5 4866080NA   4866080 2020-03-30 05:21:44 2020-05-04   2020-03-20    
#>  6 1011795NA   4951252 2020-04-03 03:29:36 2020-04-03   2020-03-08    
#>  7 1012765NA   1012765 2020-04-05 07:35:13 2020-04-10   2020-04-04    
#>  8 1014171NA   1014171 2020-04-01 07:31:20 2020-04-01   2020-02-10    
#>  9 1014978NA   1014978 2020-04-05 09:34:00 2020-04-05   2020-03-30    
#> 10 1017233NA   1017233 2020-04-04 06:44:27 2020-04-10   2020-03-29    
#> # … with 13,656 more rows, and 35 more variables: date_start <date>,
#> #   date_end <date>, entry_type <chr>, update_type <chr>,
#> #   event_description <chr>, domestic_policy <dbl>, type <chr>,
#> #   type_sub_cat <chr>, type_text <dbl>, index_high_est <dbl>,
#> #   index_med_est <dbl>, index_low_est <dbl>, index_country_rank <dbl>,
#> #   country <chr>, init_country_level <chr>, province <chr>, city <chr>,
#> #   corr_entry_type <chr>, type_new_admin_coop <lgl>, source_file_Type <chr>,
#> #   source_file_2_Type <lgl>, target_country <chr>, target_geog_level <chr>,
#> #   target_region <chr>, target_province <chr>, target_city <chr>,
#> #   target_other <lgl>, target_who_what <chr>, target_direction <chr>,
#> #   travel_mechanism <chr>, compliance <chr>, enforcer <chr>, link <chr>,
#> #   iso3c <chr>, iso2c <chr>

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 25
#>   ..$ coronanet_government_response_data:List of 2
#>   .. ..$ columns   :List of 40
#>   .. .. ..$ record_id          : chr "character"
#>   .. .. ..$ policy_id          : chr "numeric"
#>   .. .. ..$ recorded_date      : chr [1:2] "POSIXct" "POSIXt"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 13666
#>   .. .. ..$ ncol: int 40
#>   ..$ owid_data                         :List of 3
#>   .. ..$ columns   :List of 7
#>   .. .. ..$ iso3c      : chr "character"
#>   .. .. ..$ country    : chr "character"
#>   .. .. ..$ date       : chr "Date"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 16592
#>   .. .. ..$ ncol: int 7
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2019-12-31"
#>   .. .. ..$ max: chr "2020-05-11"
#>   ..$ cov_glue_lineage_data             :List of 3
#>   .. ..$ columns   :List of 9
#>   .. .. ..$ name          : chr "character"
#>   .. .. ..$ country       : chr "character"
#>   .. .. ..$ travel_history: chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 15194
#>   .. .. ..$ ncol: int 9
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2019-12-24"
#>   .. .. ..$ max: chr "2020-04-23"
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-05-11"

names(dd$datasets)

#>  [1] "coronanet_government_response_data" "owid_data"                         
#>  [3] "cov_glue_lineage_data"              "cov_glue_newick_data"              
#>  [5] "cdc_excess_deaths"                  "param_estimates_published"         
#>  [7] "cdc_aggregated_projections"         "google_mobility_data"              
#>  [9] "ecdc_data"                          "beoutbreakprepared_data"           
#> [11] "eu_data_cache_data"                 "descartes_mobility_data"           
#> [13] "apple_mobility_data"                "usa_facts_data"                    
#> [15] "jhu_data"                           "jhu_us_data"                       
#> [17] "nytimes_county_data"                "nytimes_state_data"                
#> [19] "kff_icu_beds"                       "us_county_geo_details"             
#> [21] "covidtracker_data"                  "us_healthcare_capacity"            
#> [23] "government_policy_timeline"         "healthdata_projections_data"       
#> [25] "us_state_distancing_policy"        

# evaluated
dd$eval_date

#> [1] "2020-05-11"