list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 34 x 8
#>    name    accessor  data_type geographical geospatial resolution region url    
#>    <chr>   <chr>     <chr>     <lgl>        <lgl>      <chr>      <chr>  <chr>  
#>  1 Our Wo… owid_data time ser… TRUE         FALSE      admin0     World  https:…
#>  2 Our Wo… owid_data cumulati… TRUE         FALSE      admin0     World  https:…
#>  3 Our Wo… owid_data deaths    TRUE         FALSE      admin0     World  https:…
#>  4 Our Wo… owid_data tests     TRUE         FALSE      admin0     World  https:…
#>  5 GISAID… cov_glue… individu… TRUE         FALSE      multiple   World  https:…
#>  6 Newick… cov_glue… phylogen… FALSE        FALSE      multiple   World  https:…
#>  7 US CDC… cdc_exce… miscella… TRUE         FALSE      admin1     US     https:…
#>  8 Publis… param_es… miscella… FALSE        FALSE      NA         NA     https:…
#>  9 Aggreg… cdc_aggr… projecti… TRUE         FALSE      admin0     NA     https:…
#> 10 Aggreg… cdc_aggr… projecti… TRUE         FALSE      admin1     NA     https:…
#> # … with 24 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 16,592 x 7
#>    iso3c country date       confirmed deaths tests tests_units
#>    <chr> <chr>   <date>         <dbl>  <dbl> <dbl> <chr>      
#>  1 ABW   Aruba   2020-03-13         2      0    NA NA         
#>  2 ABW   Aruba   2020-03-20         4      0    NA NA         
#>  3 ABW   Aruba   2020-03-24        12      0    NA NA         
#>  4 ABW   Aruba   2020-03-25        17      0    NA NA         
#>  5 ABW   Aruba   2020-03-26        19      0    NA NA         
#>  6 ABW   Aruba   2020-03-27        28      0    NA NA         
#>  7 ABW   Aruba   2020-03-28        28      0    NA NA         
#>  8 ABW   Aruba   2020-03-29        28      0    NA NA         
#>  9 ABW   Aruba   2020-03-30        50      0    NA NA         
#> 10 ABW   Aruba   2020-04-01        55      0    NA NA         
#> # … with 16,582 more rows

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 23
#>   ..$ cov_glue_lineage_data      :List of 3
#>   .. ..$ columns   :List of 9
#>   .. .. ..$ name          : chr "character"
#>   .. .. ..$ country       : chr "character"
#>   .. .. ..$ travel_history: chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 15194
#>   .. .. ..$ ncol: int 9
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2019-12-24"
#>   .. .. ..$ max: chr "2020-04-23"
#>   ..$ cov_glue_newick_data       :List of 2
#>   .. ..$ columns   :List of 5
#>   .. .. ..$ edge       : chr "matrix"
#>   .. .. ..$ edge.length: chr "numeric"
#>   .. .. ..$ Nnode      : chr "integer"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: NULL
#>   .. .. ..$ ncol: NULL
#>   ..$ cdc_excess_deaths          :List of 3
#>   .. ..$ columns   :List of 8
#>   .. .. ..$ date        : chr "Date"
#>   .. .. ..$ state       : chr "character"
#>   .. .. ..$ deaths      : chr "numeric"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 36464
#>   .. .. ..$ ncol: int 8
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2017-01-14"
#>   .. .. ..$ max: chr "2020-04-25"
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-05-08"

names(dd$datasets)

#>  [1] "cov_glue_lineage_data"       "cov_glue_newick_data"       
#>  [3] "cdc_excess_deaths"           "param_estimates_published"  
#>  [5] "cdc_aggregated_projections"  "google_mobility_data"       
#>  [7] "ecdc_data"                   "beoutbreakprepared_data"    
#>  [9] "eu_data_cache_data"          "descartes_mobility_data"    
#> [11] "apple_mobility_data"         "usa_facts_data"             
#> [13] "jhu_data"                    "jhu_us_data"                
#> [15] "nytimes_county_data"         "nytimes_state_data"         
#> [17] "kff_icu_beds"                "us_county_geo_details"      
#> [19] "covidtracker_data"           "us_healthcare_capacity"     
#> [21] "government_policy_timeline"  "healthdata_projections_data"
#> [23] "us_state_distancing_policy" 

# evaluated
dd$eval_date

#> [1] "2020-05-08"