list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 30 x 8
#>    name    accessor  data_type geographical geospatial region resolution url    
#>    <chr>   <chr>     <list>    <lgl>        <lgl>      <list> <list>     <chr>  
#>  1 Extens… us_hospi… <chr [1]> TRUE         TRUE       <chr … <chr [1]>  https:…
#>  2 The Ec… economis… <chr [3]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  3 US cou… us_count… <chr [1]> TRUE         FALSE      <chr … <chr [3]>  https:…
#>  4 Corona… coronane… <chr [1]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  5 Our Wo… owid_data <chr [4]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  6 GISAID… cov_glue… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  7 Newick… cov_glue… <chr [1]> FALSE        FALSE      <chr … <chr [1]>  https:…
#>  8 US CDC… cdc_exce… <chr [3]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  9 Publis… param_es… <chr [1]> FALSE        FALSE      <NULL> <NULL>     https:…
#> 10 Aggreg… cdc_aggr… <chr [1]> TRUE         FALSE      <NULL> <chr [2]>  https:…
#> # … with 20 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> Simple feature collection with 7596 features and 32 fields
#> Attribute-geometry relationship: 32 constant, 0 aggregate, 0 identity
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -176.6403 ymin: -14.29024 xmax: 145.7245 ymax: 71.29773
#> CRS:            EPSG:4326
#> # A tibble: 7,596 x 33
#>        x     y   fid id    name  address city  state zip   zip4  telephone type 
#>    <dbl> <dbl> <dbl> <chr> <chr> <chr>   <chr> <chr> <chr> <chr> <chr>     <chr>
#>  1 -120.  36.3     1 0005… CENT… 1025 N… HANF… CA    93230 NOT … NOT AVAI… GENE…
#>  2 -119.  34.2     2 0053… LOS … 150 VI… WEST… CA    91362 NOT … NOT AVAI… GENE…
#>  3 -118.  34.0     3 0011… EAST… 4060 W… LOS … CA    90023 NOT … NOT AVAI… GENE…
#>  4 -118.  34.1     4 0017… SOUT… 6245 D… HOLL… CA    90028 NOT … (323) 46… GENE…
#>  5 -118.  34.1     5 0023… KIND… 14148 … BALD… CA    91706 NOT … NOT AVAI… GENE…
#>  6 -118.  33.9     6 0025… LAKE… 3700 E… LAKE… CA    90712 NOT … NOT AVAI… GENE…
#>  7 -118.  34.1     7 0030… MONR… 323 SO… MONR… CA    91016 NOT … NOT AVAI… GENE…
#>  8 -118.  34.0     8 0030… MONT… 900 SO… MONT… CA    91754 NOT … NOT AVAI… GENE…
#>  9 -118.  33.8     9 0047… STAR… 4025 W… TORR… CA    90505 NOT … NOT AVAI… PSYC…
#> 10 -118.  33.9    10 0009… CORO… 730 MA… CORO… CA    92879 NOT … NOT AVAI… GENE…
#> # … with 7,586 more rows, and 21 more variables: status <chr>,
#> #   population <dbl>, county <chr>, countyfips <chr>, country <chr>,
#> #   naics_code <dbl>, naics_desc <chr>, source <chr>, sourcedate <chr>,
#> #   val_method <chr>, val_date <chr>, website <chr>, state_id <chr>,
#> #   alt_name <chr>, st_fips <chr>, owner <chr>, ttl_staff <dbl>, beds <dbl>,
#> #   trauma <chr>, helipad <chr>, geometry <POINT [°]>

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 30
#>   ..$ us_hospital_details               :List of 2
#>   .. ..$ columns   :List of 33
#>   .. .. ..$ x         : chr "numeric"
#>   .. .. ..$ y         : chr "numeric"
#>   .. .. ..$ fid       : chr "numeric"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 7596
#>   .. .. ..$ ncol: int 33
#>   ..$ economist_excess_deaths           :List of 2
#>   .. ..$ columns   :List of 12
#>   .. .. ..$ country        : chr "character"
#>   .. .. ..$ region         : chr "character"
#>   .. .. ..$ region_code    : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 35900
#>   .. .. ..$ ncol: int 12
#>   ..$ us_county_health_rankings         :List of 2
#>   .. ..$ columns   :List of 9
#>   .. .. ..$ statecode    : chr "character"
#>   .. .. ..$ countycode   : chr "character"
#>   .. .. ..$ fips         : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 2488126
#>   .. .. ..$ ncol: int 9
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-05-31"

names(dd$datasets)

#>  [1] "us_hospital_details"                "economist_excess_deaths"           
#>  [3] "us_county_health_rankings"          "coronanet_government_response_data"
#>  [5] "owid_data"                          "cov_glue_lineage_data"             
#>  [7] "cov_glue_newick_data"               "cdc_excess_deaths"                 
#>  [9] "param_estimates_published"          "cdc_aggregated_projections"        
#> [11] "google_mobility_data"               "ecdc_data"                         
#> [13] "beoutbreakprepared_data"            "eu_data_cache_data"                
#> [15] "descartes_mobility_data"            "apple_mobility_data"               
#> [17] "usa_facts_data"                     "jhu_data"                          
#> [19] "jhu_us_data"                        "nytimes_county_data"               
#> [21] "nytimes_state_data"                 "kff_icu_beds"                      
#> [23] "us_county_geo_details"              "covidtracker_data"                 
#> [25] "us_healthcare_capacity"             "government_policy_timeline"        
#> [27] "healthdata_testing_data"            "healthdata_mobility_data"          
#> [29] "healthdata_projections_data"        "us_state_distancing_policy"        

# evaluated
dd$eval_date

#> [1] "2020-05-31"