list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 22 x 8
#>    name   accessor  data_type  region resolution geographical geospatial url    
#>    <chr>  <chr>     <chr>      <chr>  <chr>      <lgl>        <lgl>      <chr>  
#>  1 Googl… google_m… mobility   World  admin0     TRUE         FALSE      https:…
#>  2 Googl… google_m… mobility   World  admin1     TRUE         FALSE      https:…
#>  3 Googl… google_m… mobility   World  admin2     TRUE         FALSE      https:…
#>  4 Europ… ecdc_data time seri… World  country    TRUE         FALSE      https:…
#>  5 beout… beoutbre… individua… World  patient    TRUE         FALSE      https:…
#>  6 EU da… eu_data_… time seri… Europe country a… TRUE         FALSE      https:…
#>  7 Desca… descarte… mobility   Unite… county     TRUE         FALSE      https:…
#>  8 Apple… apple_mo… mobility   World  country a… TRUE         FALSE      https:…
#>  9 USA F… usa_fact… time seri… Unite… county     TRUE         FALSE      https:…
#> 10 Johns… jhu_data  time seri… World  country    TRUE         FALSE      https:…
#> # … with 12 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 1,384,968 x 8
#>    iso2c country_region admin1 admin2 date       admin_level places_category
#>    <chr> <chr>          <chr>  <chr>  <date>           <dbl> <chr>          
#>  1 AE    United Arab E… NA     NA     2020-02-15           0 retail_and_rec…
#>  2 AE    United Arab E… NA     NA     2020-02-15           0 grocery_and_ph…
#>  3 AE    United Arab E… NA     NA     2020-02-15           0 parks          
#>  4 AE    United Arab E… NA     NA     2020-02-15           0 transit_statio…
#>  5 AE    United Arab E… NA     NA     2020-02-15           0 workplaces     
#>  6 AE    United Arab E… NA     NA     2020-02-15           0 residential    
#>  7 AE    United Arab E… NA     NA     2020-02-16           0 retail_and_rec…
#>  8 AE    United Arab E… NA     NA     2020-02-16           0 grocery_and_ph…
#>  9 AE    United Arab E… NA     NA     2020-02-16           0 parks          
#> 10 AE    United Arab E… NA     NA     2020-02-16           0 transit_statio…
#> # … with 1,384,958 more rows, and 1 more variable:
#> #   percent_change_from_baseline <dbl>

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 18
#>   ..$ google_mobility_data       :List of 3
#>   .. ..$ columns   :List of 8
#>   .. .. ..$ iso2c                       : chr "character"
#>   .. .. ..$ country_region              : chr "character"
#>   .. .. ..$ admin1                      : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 1384968
#>   .. .. ..$ ncol: int 8
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2020-02-15"
#>   .. .. ..$ max: chr "2020-04-17"
#>   ..$ ecdc_data                  :List of 3
#>   .. ..$ columns   :List of 8
#>   .. .. ..$ date           : chr "Date"
#>   .. .. ..$ location_name  : chr "character"
#>   .. .. ..$ iso2c          : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 26836
#>   .. .. ..$ ncol: int 8
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2019-12-31"
#>   .. .. ..$ max: chr "2020-04-27"
#>   ..$ beoutbreakprepared_data    :List of 2
#>   .. ..$ columns   :List of 33
#>   .. .. ..$ ID                      : chr "character"
#>   .. .. ..$ age                     : chr "character"
#>   .. .. ..$ sex                     : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 442315
#>   .. .. ..$ ncol: int 33
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-04-27"

names(dd$datasets)

#>  [1] "google_mobility_data"        "ecdc_data"                  
#>  [3] "beoutbreakprepared_data"     "eu_data_cache_data"         
#>  [5] "descartes_mobility_data"     "apple_mobility_data"        
#>  [7] "usa_facts_data"              "jhu_data"                   
#>  [9] "jhu_us_data"                 "nytimes_county_data"        
#> [11] "nytimes_state_data"          "kff_icu_beds"               
#> [13] "us_county_geo_details"       "covidtracker_data"          
#> [15] "us_healthcare_capacity"      "government_policy_timeline" 
#> [17] "healthdata_projections_data" "us_state_distancing_policy" 

# evaluated
dd$eval_date

#> [1] "2020-04-27"