list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Examples

res = available_datasets()

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

res

#>                                                             name
#> 1                                             beoutbreakprepared
#> 2                                   Descartes Labs Mobility Data
#> 3                                                      USA Facts
#> 4                                          Johns Hopkins dataset
#> 5                                  Johns Hopkins US-centric data
#> 6                               New York Times county level data
#> 7                                New York Times state level data
#> 8                          Kaiser Family Foundation ICU bed data
#> 9                  United States county-level geographic details
#> 10                 United States county-level geographic details
#> 11                                             CovidTracker data
#> 12                   CovidCare United States Healthcare Capacity
#> 13             Oxford Government Policy Intervention time series
#> 14                  Johns Hopkins United States specific dataset
#> 15 Healthdata.org projections of hospital utilization and deaths
#> 16 Healthdata.org projections of hospital utilization and deaths
#>                       accessor                     data_type        region
#> 1      beoutbreakprepared_data individual patient level data         World
#> 2      descartes_mobility_data                      mobility United States
#> 3               usa_facts_data  time series cumulative cases United States
#> 4                     jhu_data  time series cumulative cases         World
#> 5                  jhu_us_data  time series cumulative cases United States
#> 6          nytimes_county_data  time series cumulative cases United States
#> 7           nytimes_state_data  time series cumulative cases United States
#> 8                 kff_icu_beds           healthcare capacity United States
#> 9        us_county_geo_details                   demographic United States
#> 10       us_county_geo_details                    geographic United States
#> 11           covidtracker_data           healthcare capacity United States
#> 12      us_healthcare_capacity           healthcare capacity United States
#> 13  government_policy_timeline                 public policy United States
#> 14                 jhu_us_data  time series cumulative cases United States
#> 15 healthdata_projections_data       time series projections United States
#> 16 healthdata_projections_data       time series projections International
#>             resolution geographical geospatial
#> 1              patient         TRUE      FALSE
#> 2               county         TRUE      FALSE
#> 3               county         TRUE      FALSE
#> 4              country         TRUE      FALSE
#> 5               county         TRUE      FALSE
#> 6               county         TRUE      FALSE
#> 7                state         TRUE      FALSE
#> 8  Individual hospital         TRUE       TRUE
#> 9               county         TRUE       TRUE
#> 10              county         TRUE       TRUE
#> 11 Individual hospital         TRUE      FALSE
#> 12 Individual hospital         TRUE       TRUE
#> 13 Individual hospital         TRUE      FALSE
#> 14              county         TRUE      FALSE
#> 15               state         TRUE      FALSE
#> 16             country         TRUE      FALSE

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 261,558 x 34
#>    ID    age   sex   city  province country `wuhan(0)_not_w… latitude longitude
#>    <chr> <chr> <chr> <chr> <chr>    <chr>              <dbl>    <dbl>     <dbl>
#>  1 000-… 30-39 fema… Snoh… Washing… United…                1     48.0  -122.   
#>  2 000-… NA    NA    NA    Khuzest… Iran                   1     31.5    49.0  
#>  3 000-… NA    NA    NA    NA       NA                    NA     NA      NA    
#>  4 000-… 50-59 male  Snoh… Washing… United…                1     48.0  -122.   
#>  5 000-… NA    NA    NA    Pays de… France                 1     47.5    -0.811
#>  6 000-… NA    NA    NA    NA       NA                    NA     NA      NA    
#>  7 000-… NA    NA    NA    NA       NA                    NA     NA      NA    
#>  8 000-… NA    NA    NA    NA       NA                    NA     NA      NA    
#>  9 000-… 30-39 fema… Snoh… Washing… United…                1     48.0  -122.   
#> 10 000-… 50-59 male  Snoh… Washing… United…                1     48.0  -122.   
#> # … with 261,548 more rows, and 25 more variables: geo_resolution <chr>,
#> #   date_onset_symptoms <date>, date_admission_hospital <date>,
#> #   date_confirmation <date>, symptoms <chr>, lives_in_Wuhan <chr>,
#> #   travel_history_dates <date>, travel_history_location <chr>,
#> #   reported_market_exposure <chr>, additional_information <chr>,
#> #   chronic_disease_binary <chr>, chronic_disease <chr>, source <chr>,
#> #   sequence_available <chr>, outcome <chr>, date_death_or_discharge <date>,
#> #   notes_for_discussion <chr>, location <chr>, admin3 <chr>, admin2 <chr>,
#> #   admin1 <chr>, country_new <chr>, admin_id <chr>,
#> #   data_moderator_initials <chr>, travel_history_binary <lgl>