list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 25 x 8
#>    name    accessor  data_type geographical geospatial url     resolution region
#>    <chr>   <chr>     <chr>     <lgl>        <lgl>      <chr>   <chr>      <chr> 
#>  1 Publis… param_es… miscella… FALSE        FALSE      https:… NA         NA    
#>  2 Aggreg… cdc_aggr… projecti… TRUE         FALSE      https:… country    NA    
#>  3 Aggreg… cdc_aggr… projecti… TRUE         FALSE      https:… admin1     NA    
#>  4 Google… google_m… mobility  TRUE         FALSE      https:… admin0     World 
#>  5 Google… google_m… mobility  TRUE         FALSE      https:… admin1     World 
#>  6 Google… google_m… mobility  TRUE         FALSE      https:… admin2     World 
#>  7 Europe… ecdc_data time ser… TRUE         FALSE      https:… country    World 
#>  8 beoutb… beoutbre… individu… TRUE         FALSE      https:… patient    World 
#>  9 EU dat… eu_data_… time ser… TRUE         FALSE      https:… country a… Europe
#> 10 Descar… descarte… mobility  TRUE         FALSE      https:… county     Unite…
#> # … with 15 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> Warning: 1 parsing failure.
#> row      col   expected     actual                                                                 file
#> 141 end_date valid date 2020-28-02 '/Users/sdavis2/Library/Caches/sars2pack/16733a743e7a_estimates.csv'

#> # A tibble: 537 x 29
#>    id    peer_review peer_review_loc… name  abbreviation units country
#>    <chr> <chr>       <chr>            <chr> <chr>        <chr> <chr>  
#>  1 covi… NA          NA               cumu… NA           pers… China  
#>  2 covi… NA          MedRxiv          cumu… NA           pers… China  
#>  3 covi… NA          NA               date… NA           NA    China  
#>  4 covi… positive    Eurosurveillance basi… R0           Nb    China  
#>  5 covi… NA          BioRxiv          basi… R0           Nb    China  
#>  6 covi… NA          NA               basi… R0           Nb    China  
#>  7 covi… positive    International J… basi… R0           Nb    China  
#>  8 covi… NA          NA               basi… R0           Nb    China  
#>  9 covi… NA          BioRxiv          disp… k            Nb    China  
#> 10 covi… positive    Eurosurveillance prop… NA           prop… China  
#> # … with 527 more rows, and 22 more variables: location_name <chr>,
#> #   location_type <chr>, start_date <date>, end_date <date>, value_type <chr>,
#> #   value <chr>, uncertainty_type <chr>, lower_bound <chr>, upper_bound <chr>,
#> #   population <chr>, age_range <chr>, method_description <chr>,
#> #   data_description <chr>, data_URL <chr>, date_publication <date>,
#> #   title_publication <chr>, type_publication <chr>, authors <chr>,
#> #   publication_URL <chr>, entry_date <date>, entry_person <chr>,
#> #   entry_email <chr>

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 20
#>   ..$ param_estimates_published  :List of 2
#>   .. ..$ columns   :List of 29
#>   .. .. ..$ id                  : chr "character"
#>   .. .. ..$ peer_review         : chr "character"
#>   .. .. ..$ peer_review_location: chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 537
#>   .. .. ..$ ncol: int 29
#>   ..$ cdc_aggregated_projections :List of 2
#>   .. ..$ columns   :List of 8
#>   .. .. ..$ model               : chr "character"
#>   .. .. ..$ forecast_date       : chr "Date"
#>   .. .. ..$ target              : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 5209
#>   .. .. ..$ ncol: int 8
#>   ..$ google_mobility_data       :List of 3
#>   .. ..$ columns   :List of 8
#>   .. .. ..$ iso2c                       : chr "character"
#>   .. .. ..$ country_region              : chr "character"
#>   .. .. ..$ admin1                      : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 1706934
#>   .. .. ..$ ncol: int 8
#>   .. ..$ dates     :List of 2
#>   .. .. ..$ min: chr "2020-02-15"
#>   .. .. ..$ max: chr "2020-04-26"
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-05-01"

names(dd$datasets)

#>  [1] "param_estimates_published"   "cdc_aggregated_projections" 
#>  [3] "google_mobility_data"        "ecdc_data"                  
#>  [5] "beoutbreakprepared_data"     "eu_data_cache_data"         
#>  [7] "descartes_mobility_data"     "apple_mobility_data"        
#>  [9] "usa_facts_data"              "jhu_data"                   
#> [11] "jhu_us_data"                 "nytimes_county_data"        
#> [13] "nytimes_state_data"          "kff_icu_beds"               
#> [15] "us_county_geo_details"       "covidtracker_data"          
#> [17] "us_healthcare_capacity"      "government_policy_timeline" 
#> [19] "healthdata_projections_data" "us_state_distancing_policy" 

# evaluated
dd$eval_date

#> [1] "2020-05-01"