list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Examples

res = available_datasets()

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

res

#>                                                             name
#> 1                                             beoutbreakprepared
#> 2                                   Descartes Labs Mobility Data
#> 3                                                      USA Facts
#> 4                                          Johns Hopkins dataset
#> 5                                  Johns Hopkins US-centric data
#> 6                               New York Times county level data
#> 7                                New York Times state level data
#> 8                          Kaiser Family Foundation ICU bed data
#> 9                  United States county-level geographic details
#> 10                 United States county-level geographic details
#> 11                                             CovidTracker data
#> 12                   CovidCare United States Healthcare Capacity
#> 13             Oxford Government Policy Intervention time series
#> 14                  Johns Hopkins United States specific dataset
#> 15 Healthdata.org projections of hospital utilization and deaths
#> 16 Healthdata.org projections of hospital utilization and deaths
#>                       accessor                     data_type        region
#> 1      beoutbreakprepared_data individual patient level data         World
#> 2      descartes_mobility_data                      mobility United States
#> 3               usa_facts_data  time series cumulative cases United States
#> 4                     jhu_data  time series cumulative cases         World
#> 5                  jhu_us_data  time series cumulative cases United States
#> 6          nytimes_county_data  time series cumulative cases United States
#> 7           nytimes_state_data  time series cumulative cases United States
#> 8                 kff_icu_beds           healthcare capacity United States
#> 9        us_county_geo_details                   demographic United States
#> 10       us_county_geo_details                    geographic United States
#> 11           covidtracker_data           healthcare capacity United States
#> 12      us_healthcare_capacity           healthcare capacity United States
#> 13  government_policy_timeline                 public policy United States
#> 14                 jhu_us_data  time series cumulative cases United States
#> 15 healthdata_projections_data       time series projections United States
#> 16 healthdata_projections_data       time series projections International
#>             resolution geographical geospatial
#> 1              patient         TRUE      FALSE
#> 2               county         TRUE      FALSE
#> 3               county         TRUE      FALSE
#> 4              country         TRUE      FALSE
#> 5               county         TRUE      FALSE
#> 6               county         TRUE      FALSE
#> 7                state         TRUE      FALSE
#> 8  Individual hospital         TRUE       TRUE
#> 9               county         TRUE       TRUE
#> 10              county         TRUE       TRUE
#> 11 Individual hospital         TRUE      FALSE
#> 12 Individual hospital         TRUE       TRUE
#> 13 Individual hospital         TRUE      FALSE
#> 14              county         TRUE      FALSE
#> 15               state         TRUE      FALSE
#> 16             country         TRUE      FALSE

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> using temporary cache /tmp/RtmpxN4CiE/BiocFileCache

#> # A tibble: 267,492 x 33
#>    ID    age   sex   city  province country latitude longitude geo_resolution
#>    <chr> <chr> <chr> <chr> <chr>    <chr>      <dbl>     <dbl> <chr>         
#>  1 000-… 60-69 fema… What… Washing… United…     48.8   -122.   admin2        
#>  2 000-… 30    male  Chao… Anhui    China       31.6    118.   admin3        
#>  3 000-… 30    male  Hefe… Anhui    China       31.8    117.   admin2        
#>  4 000-… 60-69 fema… Pitk… Colorado United…     39.2   -107.   admin2        
#>  5 000-… 44    male  <NA>  Zurich   Switze…     47.4      8.66 admin1        
#>  6 000-… 65    fema… Shen… Guangdo… China       22.7    114.   admin2        
#>  7 000-… <NA>  <NA>  Nant… Jiangsu  China       32.2    121.   admin2        
#>  8 000-… <NA>  <NA>  <NA>  Hamburg  Germany     53.5     10.0  admin1        
#>  9 000-… 8-68  <NA>  Xuzh… Jiangsu  China       34.4    118.   admin2        
#> 10 000-… 8-68  <NA>  Chan… Jiangsu  China       31.6    120.   admin2        
#> # … with 267,482 more rows, and 24 more variables: date_onset_symptoms <date>,
#> #   date_admission_hospital <date>, date_confirmation <date>, symptoms <chr>,
#> #   lives_in_Wuhan <chr>, travel_history_dates <date>,
#> #   travel_history_location <chr>, reported_market_exposure <chr>,
#> #   additional_information <chr>, chronic_disease_binary <chr>,
#> #   chronic_disease <chr>, source <chr>, sequence_available <chr>,
#> #   outcome <chr>, date_death_or_discharge <date>, notes_for_discussion <chr>,
#> #   location <chr>, admin3 <chr>, admin2 <chr>, admin1 <chr>,
#> #   country_new <chr>, admin_id <dbl>, data_moderator_initials <chr>,
#> #   travel_history_binary <lgl>