list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#>                                                                     name
#> 1                           United States CDC Social Vulnerability Index
#> 2                          Extensive United States hospital capabilities
#> 3                     The Economist: Excess deaths during COVID pandemic
#> 4                              The : Excess deaths during COVID pandemic
#> 5  US county health rankings from 'https://www.countyhealthrankings.org'
#> 6                                 CoronaNet government response database
#> 7                                 Country metadata from restcountries.eu
#> 8                          Our World In Data testing and cases reporting
#> 9                 GISAID metadata from thousands of SARS-CoV-2 sequences
#> 10                    Newick tree from thousands of SARS-CoV-2 sequences
#> 11                                          US CDC excess deaths dataset
#> 12                            Published epidemic parameters for COVID-19
#> 13                                    Aggregated projections from US CDC
#> 14                                                  Google mobility data
#> 15                         World Health Organization (WHO) case tracking
#> 16                                           European CDC world tracking
#> 17                                                    beoutbreakprepared
#> 18                                             EU data Github aggregator
#> 19                                          Descartes Labs Mobility Data
#> 20                                         Apple mobility data from maps
#> 21                                                             USA Facts
#> 22                                                 Johns Hopkins dataset
#> 23                                         Johns Hopkins US-centric data
#> 24                                      New York Times county level data
#> 25                                       New York Times state level data
#> 26                                 Kaiser Family Foundation ICU bed data
#> 27              United Nations World population by gender, age, and year
#> 28                         United States county-level geographic details
#> 29                                                     CovidTracker data
#> 30                           CovidCare United States Healthcare Capacity
#> 31                     Oxford Government Policy Intervention time series
#> 32                                          Healthdata.org mobility data
#> 33         Healthdata.org projections of hospital utilization and deaths
#> 34                              United States social distancing policies
#> 35                                  OECD International Unemployment Data
#> 36         healthdata.org COVID-19 Mobility Observations and Projections
#> 37          healthdata.org COVID-19 Testing Observations and Projections
#>                              accessor                               data_type
#> 1      cdc_social_vulnerability_index                            demographics
#> 2                 us_hospital_details                     healthcare capacity
#> 3             economist_excess_deaths      time series, deaths, excess deaths
#> 4       financial_times_excess_deaths      time series, deaths, excess deaths
#> 5           us_county_health_rankings                            demographics
#> 6  coronanet_government_response_data                           public policy
#> 7                    country_metadata                            demographics
#> 8                           owid_data     time series, cases, deaths, testing
#> 9               cov_glue_lineage_data                               line list
#> 10               cov_glue_newick_data                            phylogenetic
#> 11                  cdc_excess_deaths      time series, deaths, excess deaths
#> 12          param_estimates_published                           miscellaneous
#> 13         cdc_aggregated_projections                             projections
#> 14               google_mobility_data                                mobility
#> 15                          ecdc_data              time series, cases, deaths
#> 16                          ecdc_data              time series, cases, deaths
#> 17            beoutbreakprepared_data                               line list
#> 18                 eu_data_cache_data              time series, cases, deaths
#> 19            descartes_mobility_data                   time series, mobility
#> 20                apple_mobility_data                   time series, mobility
#> 21                     usa_facts_data              time series, cases, deaths
#> 22                           jhu_data              time series, cases, deaths
#> 23                        jhu_us_data              time series, cases, deaths
#> 24                nytimes_county_data              time series, cases, deaths
#> 25                 nytimes_state_data              time series, cases, deaths
#> 26                       kff_icu_beds                     healthcare capacity
#> 27              world_population_data                demographics, geographic
#> 28              us_county_geo_details                demographics, geographic
#> 29                  covidtracker_data     time series, cases, deaths, testing
#> 30             us_healthcare_capacity                     healthcare capacity
#> 31         government_policy_timeline                           public policy
#> 32           healthdata_mobility_data      time series, projections, mobility
#> 33        healthdata_projections_data time series, projections, cases, deaths
#> 34         us_state_distancing_policy                           public policy
#> 35             oecd_unemployment_data                  economics, time series
#> 36           healthdata_mobility_data      mobility, time series, projections
#> 37            healthdata_testing_data       testing, time series, projections
#>    geographical geospatial               region                     resolution
#> 1          TRUE      FALSE        United States                         admin2
#> 2          TRUE       TRUE        United States            individual hospital
#> 3          TRUE      FALSE        International                 admin0, admin1
#> 4          TRUE      FALSE        International                 admin0, admin1
#> 5          TRUE      FALSE        United States         admin0, admin1, admin2
#> 6          TRUE      FALSE                World                 admin0, admin1
#> 7          TRUE      FALSE                World                         admin0
#> 8          TRUE      FALSE                World                         admin0
#> 9          TRUE      FALSE                World                       multiple
#> 10        FALSE      FALSE                World                       multiple
#> 11         TRUE      FALSE        United States                         admin1
#> 12        FALSE      FALSE                 NULL                           NULL
#> 13         TRUE      FALSE                 NULL                 admin0, admin1
#> 14         TRUE      FALSE                World         admin0, admin1, admin2
#> 15         TRUE      FALSE                World                         admin0
#> 16         TRUE      FALSE                World                         admin0
#> 17         TRUE      FALSE                World                        patient
#> 18         TRUE      FALSE               Europe                 admin0, admin1
#> 19         TRUE      FALSE        United States                         admin1
#> 20         TRUE      FALSE                World admin0, admin1, admin2, admin3
#> 21         TRUE      FALSE        United States                         admin1
#> 22         TRUE      FALSE                World                         admin0
#> 23         TRUE      FALSE        United States                 admin1, admin2
#> 24         TRUE      FALSE        United States                         admin2
#> 25         TRUE      FALSE        United States                         admin1
#> 26         TRUE       TRUE        United States            Individual hospital
#> 27         TRUE       TRUE                World                         admon1
#> 28         TRUE       TRUE        United States                         admin2
#> 29         TRUE      FALSE        United States                         admin1
#> 30         TRUE       TRUE        United States            Individual hospital
#> 31         TRUE      FALSE                World                         admin0
#> 32         TRUE      FALSE United States, World                 admin1, admin2
#> 33         TRUE      FALSE United States, World                 admin1, admin2
#> 34         TRUE      FALSE        United States                         admin1
#> 35         TRUE      FALSE                World                         admin0
#> 36         TRUE      FALSE        International                 admin0, admin1
#> 37         TRUE      FALSE        International                 admin0, admin1
#>                                                                                                                 url
#> 1                                                                                              https://svi.cdc.gov/
#> 2                                                            https://hub.arcgis.com/datasets/geoplatform::hospitals
#> 3                                                    https://github.com/TheEconomist/covid-19-excess-deaths-tracker
#> 4                                              https://github.com/Financial-Times/coronavirus-excess-mortality-data
#> 5                                                                              https://www.countyhealthrankings.org
#> 6                                                                          https://coronanet-project.org/index.html
#> 7                                                                                          https://restcountries.eu
#> 8                                                                            https://ourworldindata.org/coronavirus
#> 9                                                                             https://github.com/hCoV-2019/lineages
#> 10                                                                            https://github.com/hCoV-2019/lineages
#> 11                                                    https://www.cdc.gov/nchs/nvss/vsrr/covid19/excess_deaths.html
#> 12   https://github.com/midas-network/COVID-19/blob/master/parameter_estimates/2019_novel_coronavirus/estimates.csv
#> 13                                         https://www.cdc.gov/coronavirus/2019-ncov/covid-data/forecasting-us.html
#> 14                                                                         https://www.google.com/covid19/mobility/
#> 15                                                                                         https://covid19.who.int/
#> 16                                                                           https://www.ecdc.europa.eu/en/covid-19
#> 17                                                                   https://github.com/beoutbreakprepared/nCoV2019
#> 18                                                                 https://github.com/covid19-eu-zh/covid19-eu-data
#> 19                                                      https://raw.githubusercontent.com/descarteslabs/DL-COVID-19
#> 20                                                                           https://www.apple.com/covid19/mobility
#> 21                                             https://usafacts.org/visualizations/coronavirus-covid-19-spread-map/
#> 22                                                                       https://github.com/CSSEGISandData/COVID-19
#> 23                                                                       https://github.com/CSSEGISandData/COVID-19
#> 24                                                          https://raw.githubusercontent.com/nytimes/covid-19-data
#> 25                                                          https://raw.githubusercontent.com/nytimes/covid-19-data
#> 26 https://khn.org/news/as-coronavirus-spreads-widely-millions-of-older-americans-live-in-counties-with-no-icu-beds
#> 27                                                      https://population.un.org/wpp/Download/Standard/Population/
#> 28                                                                     https://github.com/josh-byster/fips_lat_long
#> 29                                                                                       https://covidtracking.com/
#> 30                                                     https://github.com/covidcaremap/covid19-healthsystemcapacity
#> 31                  https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker
#> 32                                                                                  http://www.healthdata.org/covid
#> 33                                                                                  http://www.healthdata.org/covid
#> 34                                                          https://github.com/COVID19StatePolicy/SocialDistancing/
#> 35                                                                                                 https://oecd.org
#> 36                                                                       https://covid19.healthdata.org/projections
#> 37                                                                       https://covid19.healthdata.org/projections

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 3,142 x 126
#>      fid state_fips state st_abbr county fips  location area_sqmi e_totpop
#>    <dbl> <chr>      <chr> <chr>   <chr>  <chr> <chr>        <dbl>    <dbl>
#>  1     1 00001      Alab… AL      Autau… 01001 Autauga…      594.    55200
#>  2     2 00001      Alab… AL      Baldw… 01003 Baldwin…     1590.   208107
#>  3     3 00001      Alab… AL      Barbo… 01005 Barbour…      885.    25782
#>  4     4 00001      Alab… AL      Bibb   01007 Bibb Co…      622.    22527
#>  5     5 00001      Alab… AL      Blount 01009 Blount …      645.    57645
#>  6     6 00001      Alab… AL      Bullo… 01011 Bullock…      623.    10352
#>  7     7 00001      Alab… AL      Butler 01013 Butler …      777.    20025
#>  8     8 00001      Alab… AL      Calho… 01015 Calhoun…      606.   115098
#>  9     9 00001      Alab… AL      Chamb… 01017 Chamber…      597.    33826
#> 10    10 00001      Alab… AL      Chero… 01019 Cheroke…      554.    25853
#> # … with 3,132 more rows, and 117 more variables: m_totpop <dbl>, e_hu <dbl>,
#> #   m_hu <dbl>, e_hh <dbl>, m_hh <dbl>, e_pov <dbl>, m_pov <dbl>,
#> #   e_unemp <dbl>, m_unemp <dbl>, e_pci <dbl>, m_pci <dbl>, e_nohsdp <dbl>,
#> #   m_nohsdp <dbl>, e_age65 <dbl>, m_age65 <dbl>, e_age17 <dbl>, m_age17 <dbl>,
#> #   e_disabl <dbl>, m_disabl <dbl>, e_sngpnt <dbl>, m_sngpnt <dbl>,
#> #   e_minrty <dbl>, m_minrty <dbl>, e_limeng <dbl>, m_limeng <dbl>,
#> #   e_munit <dbl>, m_munit <dbl>, e_mobile <dbl>, m_mobile <dbl>,
#> #   e_crowd <dbl>, m_crowd <dbl>, e_noveh <dbl>, m_noveh <dbl>, e_groupq <dbl>,
#> #   m_groupq <dbl>, ep_pov <dbl>, mp_pov <dbl>, ep_unemp <dbl>, mp_unemp <dbl>,
#> #   ep_pci <dbl>, mp_pci <dbl>, ep_nohsdp <dbl>, mp_nohsdp <dbl>,
#> #   ep_age65 <dbl>, mp_age65 <dbl>, ep_age17 <dbl>, mp_age17 <dbl>,
#> #   ep_disabl <dbl>, mp_disabl <dbl>, ep_sngpnt <dbl>, mp_sngpnt <dbl>,
#> #   ep_minrty <dbl>, mp_minrty <dbl>, ep_limeng <dbl>, mp_limeng <dbl>,
#> #   ep_munit <dbl>, mp_munit <dbl>, ep_mobile <dbl>, mp_mobile <dbl>,
#> #   ep_crowd <dbl>, mp_crowd <dbl>, ep_noveh <dbl>, mp_noveh <dbl>,
#> #   ep_groupq <dbl>, mp_groupq <dbl>, epl_pov <dbl>, epl_unemp <dbl>,
#> #   epl_pci <dbl>, epl_nohsdp <dbl>, spl_theme1 <dbl>, rpl_theme1 <dbl>,
#> #   epl_age65 <dbl>, epl_age17 <dbl>, epl_disabl <dbl>, epl_sngpnt <dbl>,
#> #   spl_theme2 <dbl>, rpl_theme2 <dbl>, epl_minrty <dbl>, epl_limeng <dbl>,
#> #   spl_theme3 <dbl>, rpl_theme3 <dbl>, epl_munit <dbl>, epl_mobile <dbl>,
#> #   epl_crowd <dbl>, epl_noveh <dbl>, epl_groupq <dbl>, spl_theme4 <dbl>,
#> #   rpl_theme4 <dbl>, spl_themes <dbl>, rpl_themes <dbl>, f_pov <dbl>,
#> #   f_unemp <dbl>, f_pci <dbl>, f_nohsdp <dbl>, f_theme1 <dbl>, f_age65 <dbl>,
#> #   f_age17 <dbl>, f_disabl <dbl>, f_sngpnt <dbl>, f_theme2 <dbl>, …

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 35
#>   ..$ cdc_social_vulnerability_index    :List of 2
#>   .. ..$ columns   :List of 126
#>   .. .. ..$ fid         : chr "numeric"
#>   .. .. ..$ state_fips  : chr "character"
#>   .. .. ..$ state       : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 3142
#>   .. .. ..$ ncol: int 126
#>   ..$ us_hospital_details               :List of 2
#>   .. ..$ columns   :List of 33
#>   .. .. ..$ x         : chr "numeric"
#>   .. .. ..$ y         : chr "numeric"
#>   .. .. ..$ fid       : chr "numeric"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 7596
#>   .. .. ..$ ncol: int 33
#>   ..$ economist_excess_deaths           :List of 2
#>   .. ..$ columns   :List of 12
#>   .. .. ..$ country        : chr "character"
#>   .. .. ..$ region         : chr "character"
#>   .. .. ..$ region_code    : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 8658
#>   .. .. ..$ ncol: int 12
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-08-19"

names(dd$datasets)

#>  [1] "cdc_social_vulnerability_index"     "us_hospital_details"               
#>  [3] "economist_excess_deaths"            "financial_times_excess_deaths"     
#>  [5] "us_county_health_rankings"          "coronanet_government_response_data"
#>  [7] "country_metadata"                   "owid_data"                         
#>  [9] "cov_glue_lineage_data"              "cov_glue_newick_data"              
#> [11] "cdc_excess_deaths"                  "param_estimates_published"         
#> [13] "cdc_aggregated_projections"         "google_mobility_data"              
#> [15] "ecdc_data"                          "beoutbreakprepared_data"           
#> [17] "eu_data_cache_data"                 "descartes_mobility_data"           
#> [19] "apple_mobility_data"                "usa_facts_data"                    
#> [21] "jhu_data"                           "jhu_us_data"                       
#> [23] "nytimes_county_data"                "nytimes_state_data"                
#> [25] "kff_icu_beds"                       "world_population_data"             
#> [27] "us_county_geo_details"              "covidtracker_data"                 
#> [29] "us_healthcare_capacity"             "government_policy_timeline"        
#> [31] "healthdata_mobility_data"           "healthdata_projections_data"       
#> [33] "us_state_distancing_policy"         "oecd_unemployment_data"            
#> [35] "healthdata_testing_data"           

# evaluated
dd$eval_date

#> [1] "2020-08-19"