list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

a list, with each dataset as an item

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 50 x 8
#>    name     accessor data_type geographical geospatial region resolution url    
#>    <chr>    <chr>    <list>    <lgl>        <lgl>      <list> <list>     <chr>  
#>  1 United … cdc_soc… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  2 Extensi… us_hosp… <chr [1]> TRUE         TRUE       <chr … <chr [1]>  https:…
#>  3 The Eco… economi… <chr [3]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  4 The : E… financi… <chr [3]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  5 US coun… us_coun… <chr [1]> TRUE         FALSE      <chr … <chr [3]>  https:…
#>  6 CoronaN… coronan… <chr [1]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  7 Country… country… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  8 Our Wor… owid_da… <chr [4]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  9 GISAID … cov_glu… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#> 10 Newick … cov_glu… <chr [1]> FALSE        FALSE      <chr … <chr [1]>  https:…
#> # … with 40 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 3,142 x 126
#>    grasp_id state_fips state  st_abbr county fips  location   area_sqmi e_totpop
#>       <dbl> <chr>      <chr>  <chr>   <chr>  <chr> <chr>          <dbl>    <dbl>
#>  1        1 00001      Alaba… AL      Autau… 01001 Autauga C…      594.    55200
#>  2        2 00001      Alaba… AL      Baldw… 01003 Baldwin C…     1590.   208107
#>  3        3 00001      Alaba… AL      Barbo… 01005 Barbour C…      885.    25782
#>  4        4 00001      Alaba… AL      Bibb   01007 Bibb Coun…      622.    22527
#>  5        5 00001      Alaba… AL      Blount 01009 Blount Co…      645.    57645
#>  6        6 00001      Alaba… AL      Bullo… 01011 Bullock C…      623.    10352
#>  7        7 00001      Alaba… AL      Butler 01013 Butler Co…      777.    20025
#>  8        8 00001      Alaba… AL      Calho… 01015 Calhoun C…      606.   115098
#>  9        9 00001      Alaba… AL      Chamb… 01017 Chambers …      597.    33826
#> 10       10 00001      Alaba… AL      Chero… 01019 Cherokee …      554.    25853
#> # … with 3,132 more rows, and 117 more variables: m_totpop <dbl>, e_hu <dbl>,
#> #   m_hu <dbl>, e_hh <dbl>, m_hh <dbl>, e_pov <dbl>, m_pov <dbl>,
#> #   e_unemp <dbl>, m_unemp <dbl>, e_pci <dbl>, m_pci <dbl>, e_nohsdp <dbl>,
#> #   m_nohsdp <dbl>, e_age65 <dbl>, m_age65 <dbl>, e_age17 <dbl>, m_age17 <dbl>,
#> #   e_disabl <dbl>, m_disabl <dbl>, e_sngpnt <dbl>, m_sngpnt <dbl>,
#> #   e_minrty <dbl>, m_minrty <dbl>, e_limeng <dbl>, m_limeng <dbl>,
#> #   e_munit <dbl>, m_munit <dbl>, e_mobile <dbl>, m_mobile <dbl>,
#> #   e_crowd <dbl>, m_crowd <dbl>, e_noveh <dbl>, m_noveh <dbl>, e_groupq <dbl>,
#> #   m_groupq <dbl>, ep_pov <dbl>, mp_pov <dbl>, ep_unemp <dbl>, mp_unemp <dbl>,
#> #   ep_pci <dbl>, mp_pci <dbl>, ep_nohsdp <dbl>, mp_nohsdp <dbl>,
#> #   ep_age65 <dbl>, mp_age65 <dbl>, ep_age17 <dbl>, mp_age17 <dbl>,
#> #   ep_disabl <dbl>, mp_disabl <dbl>, ep_sngpnt <dbl>, mp_sngpnt <dbl>,
#> #   ep_minrty <dbl>, mp_minrty <dbl>, ep_limeng <dbl>, mp_limeng <dbl>,
#> #   ep_munit <dbl>, mp_munit <dbl>, ep_mobile <dbl>, mp_mobile <dbl>,
#> #   ep_crowd <dbl>, mp_crowd <dbl>, ep_noveh <dbl>, mp_noveh <dbl>,
#> #   ep_groupq <dbl>, mp_groupq <dbl>, epl_pov <dbl>, epl_unemp <dbl>,
#> #   epl_pci <dbl>, epl_nohsdp <dbl>, spl_theme1 <dbl>, rpl_theme1 <dbl>,
#> #   epl_age65 <dbl>, epl_age17 <dbl>, epl_disabl <dbl>, epl_sngpnt <dbl>,
#> #   spl_theme2 <dbl>, rpl_theme2 <dbl>, epl_minrty <dbl>, epl_limeng <dbl>,
#> #   spl_theme3 <dbl>, rpl_theme3 <dbl>, epl_munit <dbl>, epl_mobile <dbl>,
#> #   epl_crowd <dbl>, epl_noveh <dbl>, epl_groupq <dbl>, spl_theme4 <dbl>,
#> #   rpl_theme4 <dbl>, spl_themes <dbl>, rpl_themes <dbl>, f_pov <dbl>,
#> #   f_unemp <dbl>, f_pci <dbl>, f_nohsdp <dbl>, f_theme1 <dbl>, f_age65 <dbl>,
#> #   f_age17 <dbl>, f_disabl <dbl>, f_sngpnt <dbl>, f_theme2 <dbl>, …

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 50
#>   ..$ cdc_social_vulnerability_index              :List of 2
#>   .. ..$ columns   :List of 126
#>   .. .. ..$ grasp_id    : chr "numeric"
#>   .. .. ..$ state_fips  : chr "character"
#>   .. .. ..$ state       : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 3142
#>   .. .. ..$ ncol: int 126
#>   ..$ us_hospital_details                         :List of 2
#>   .. ..$ columns   :List of 33
#>   .. .. ..$ x         : chr "numeric"
#>   .. .. ..$ y         : chr "numeric"
#>   .. .. ..$ objectid  : chr "numeric"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 7596
#>   .. .. ..$ ncol: int 33
#>   ..$ economist_excess_deaths                     :List of 2
#>   .. ..$ columns   :List of 14
#>   .. .. ..$ country        : chr "character"
#>   .. .. ..$ region         : chr "character"
#>   .. .. ..$ region_code    : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 25086
#>   .. .. ..$ ncol: int 14
#>   .. [list output truncated]
#>  $ eval_date: chr "2021-05-13"

names(dd$datasets)

#>  [1] "cdc_social_vulnerability_index"              
#>  [2] "us_hospital_details"                         
#>  [3] "economist_excess_deaths"                     
#>  [4] "financial_times_excess_deaths"               
#>  [5] "us_county_health_rankings"                   
#>  [6] "coronanet_government_response_data"          
#>  [7] "country_metadata"                            
#>  [8] "owid_data"                                   
#>  [9] "cov_glue_lineage_data"                       
#> [10] "cov_glue_newick_data"                        
#> [11] "cdc_excess_deaths"                           
#> [12] "param_estimates_published"                   
#> [13] "cdc_aggregated_projections"                  
#> [14] "google_mobility_data"                        
#> [15] "google_search_trends_data"                   
#> [16] "who_cases"                                   
#> [17] "ecdc_data"                                   
#> [18] "beoutbreakprepared_data"                     
#> [19] "descartes_mobility_data"                     
#> [20] "apple_mobility_data"                         
#> [21] "usa_facts_data"                              
#> [22] "jhu_data"                                    
#> [23] "jhu_us_data"                                 
#> [24] "nytimes_county_data"                         
#> [25] "nytimes_state_data"                          
#> [26] "kff_icu_beds"                                
#> [27] "world_population_data"                       
#> [28] "us_county_geo_details"                       
#> [29] "covidtracker_data"                           
#> [30] "us_healthcare_capacity"                      
#> [31] "government_policy_timeline"                  
#> [32] "us_state_distancing_policy"                  
#> [33] "oecd_unemployment_data"                      
#> [34] "acaps_secondary_impact_data"                 
#> [35] "cci_us_vaccine_data"                         
#> [36] "coronadatascraper_data"                      
#> [37] "test_and_trace_data"                         
#> [38] "cdc_us_linelist_data"                        
#> [39] "econ_tracker_unemp_national_data"            
#> [40] "econ_tracker_unemp_state_data"               
#> [41] "econ_tracker_unemp_city_data"                
#> [42] "econ_tracker_unemp_county_data"              
#> [43] "econ_tracker_consumer_spending_city_data"    
#> [44] "econ_tracker_consumer_spending_county_data"  
#> [45] "econ_tracker_consumer_spending_state_data"   
#> [46] "econ_tracker_consumer_spending_national_data"
#> [47] "econ_tracker_employment_national_data"       
#> [48] "econ_tracker_employment_city_data"           
#> [49] "econ_tracker_employment_state_data"          
#> [50] "econ_tracker_employment_county_data"         

# evaluated
dd$eval_date

#> [1] "2021-05-13"