list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

a list, with each dataset as an item

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 35 x 8
#>    name     accessor data_type geographical geospatial region resolution url    
#>    <chr>    <chr>    <list>    <lgl>        <lgl>      <list> <list>     <chr>  
#>  1 United … cdc_soc… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  2 Extensi… us_hosp… <chr [1]> TRUE         TRUE       <chr … <chr [1]>  https:…
#>  3 The Eco… economi… <chr [3]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  4 The : E… financi… <chr [3]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  5 US coun… us_coun… <chr [1]> TRUE         FALSE      <chr … <chr [3]>  https:…
#>  6 CoronaN… coronan… <chr [1]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  7 Country… country… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  8 Our Wor… owid_da… <chr [4]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  9 GISAID … cov_glu… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#> 10 Newick … cov_glu… <chr [1]> FALSE        FALSE      <chr … <chr [1]>  https:…
#> # … with 25 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 3,142 x 126
#>    grasp_id state_fips state st_abbr county fips  location area_sqmi e_totpop
#>       <dbl> <chr>      <chr> <chr>   <chr>  <chr> <chr>        <dbl>    <dbl>
#>  1        1 00001      Alab… AL      Autau… 01001 Autauga…      594.    55200
#>  2        2 00001      Alab… AL      Baldw… 01003 Baldwin…     1590.   208107
#>  3        3 00001      Alab… AL      Barbo… 01005 Barbour…      885.    25782
#>  4        4 00001      Alab… AL      Bibb   01007 Bibb Co…      622.    22527
#>  5        5 00001      Alab… AL      Blount 01009 Blount …      645.    57645
#>  6        6 00001      Alab… AL      Bullo… 01011 Bullock…      623.    10352
#>  7        7 00001      Alab… AL      Butler 01013 Butler …      777.    20025
#>  8        8 00001      Alab… AL      Calho… 01015 Calhoun…      606.   115098
#>  9        9 00001      Alab… AL      Chamb… 01017 Chamber…      597.    33826
#> 10       10 00001      Alab… AL      Chero… 01019 Cheroke…      554.    25853
#> # … with 3,132 more rows, and 117 more variables: m_totpop <dbl>, e_hu <dbl>,
#> #   m_hu <dbl>, e_hh <dbl>, m_hh <dbl>, e_pov <dbl>, m_pov <dbl>,
#> #   e_unemp <dbl>, m_unemp <dbl>, e_pci <dbl>, m_pci <dbl>, e_nohsdp <dbl>,
#> #   m_nohsdp <dbl>, e_age65 <dbl>, m_age65 <dbl>, e_age17 <dbl>, m_age17 <dbl>,
#> #   e_disabl <dbl>, m_disabl <dbl>, e_sngpnt <dbl>, m_sngpnt <dbl>,
#> #   e_minrty <dbl>, m_minrty <dbl>, e_limeng <dbl>, m_limeng <dbl>,
#> #   e_munit <dbl>, m_munit <dbl>, e_mobile <dbl>, m_mobile <dbl>,
#> #   e_crowd <dbl>, m_crowd <dbl>, e_noveh <dbl>, m_noveh <dbl>, e_groupq <dbl>,
#> #   m_groupq <dbl>, ep_pov <dbl>, mp_pov <dbl>, ep_unemp <dbl>, mp_unemp <dbl>,
#> #   ep_pci <dbl>, mp_pci <dbl>, ep_nohsdp <dbl>, mp_nohsdp <dbl>,
#> #   ep_age65 <dbl>, mp_age65 <dbl>, ep_age17 <dbl>, mp_age17 <dbl>,
#> #   ep_disabl <dbl>, mp_disabl <dbl>, ep_sngpnt <dbl>, mp_sngpnt <dbl>,
#> #   ep_minrty <dbl>, mp_minrty <dbl>, ep_limeng <dbl>, mp_limeng <dbl>,
#> #   ep_munit <dbl>, mp_munit <dbl>, ep_mobile <dbl>, mp_mobile <dbl>,
#> #   ep_crowd <dbl>, mp_crowd <dbl>, ep_noveh <dbl>, mp_noveh <dbl>,
#> #   ep_groupq <dbl>, mp_groupq <dbl>, epl_pov <dbl>, epl_unemp <dbl>,
#> #   epl_pci <dbl>, epl_nohsdp <dbl>, spl_theme1 <dbl>, rpl_theme1 <dbl>,
#> #   epl_age65 <dbl>, epl_age17 <dbl>, epl_disabl <dbl>, epl_sngpnt <dbl>,
#> #   spl_theme2 <dbl>, rpl_theme2 <dbl>, epl_minrty <dbl>, epl_limeng <dbl>,
#> #   spl_theme3 <dbl>, rpl_theme3 <dbl>, epl_munit <dbl>, epl_mobile <dbl>,
#> #   epl_crowd <dbl>, epl_noveh <dbl>, epl_groupq <dbl>, spl_theme4 <dbl>,
#> #   rpl_theme4 <dbl>, spl_themes <dbl>, rpl_themes <dbl>, f_pov <dbl>,
#> #   f_unemp <dbl>, f_pci <dbl>, f_nohsdp <dbl>, f_theme1 <dbl>, f_age65 <dbl>,
#> #   f_age17 <dbl>, f_disabl <dbl>, f_sngpnt <dbl>, f_theme2 <dbl>, …

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 35
#>   ..$ cdc_social_vulnerability_index    :List of 2
#>   .. ..$ columns   :List of 126
#>   .. .. ..$ grasp_id    : chr "numeric"
#>   .. .. ..$ state_fips  : chr "character"
#>   .. .. ..$ state       : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 3142
#>   .. .. ..$ ncol: int 126
#>   ..$ us_hospital_details               :List of 2
#>   .. ..$ columns   :List of 33
#>   .. .. ..$ x         : chr "numeric"
#>   .. .. ..$ y         : chr "numeric"
#>   .. .. ..$ objectid  : chr "numeric"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 7596
#>   .. .. ..$ ncol: int 33
#>   ..$ economist_excess_deaths           :List of 2
#>   .. ..$ columns   :List of 12
#>   .. .. ..$ country        : chr "character"
#>   .. .. ..$ region         : chr "character"
#>   .. .. ..$ region_code    : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 13592
#>   .. .. ..$ ncol: int 12
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-12-22"

names(dd$datasets)

#>  [1] "cdc_social_vulnerability_index"     "us_hospital_details"               
#>  [3] "economist_excess_deaths"            "financial_times_excess_deaths"     
#>  [5] "us_county_health_rankings"          "coronanet_government_response_data"
#>  [7] "country_metadata"                   "owid_data"                         
#>  [9] "cov_glue_lineage_data"              "cov_glue_newick_data"              
#> [11] "cdc_excess_deaths"                  "param_estimates_published"         
#> [13] "cdc_aggregated_projections"         "google_mobility_data"              
#> [15] "google_search_trends_data"          "who_cases"                         
#> [17] "ecdc_data"                          "beoutbreakprepared_data"           
#> [19] "descartes_mobility_data"            "apple_mobility_data"               
#> [21] "usa_facts_data"                     "jhu_data"                          
#> [23] "jhu_us_data"                        "nytimes_county_data"               
#> [25] "nytimes_state_data"                 "kff_icu_beds"                      
#> [27] "world_population_data"              "us_county_geo_details"             
#> [29] "covidtracker_data"                  "us_healthcare_capacity"            
#> [31] "government_policy_timeline"         "us_state_distancing_policy"        
#> [33] "oecd_unemployment_data"             "yougov_behavioral_data"            
#> [35] "acaps_secondary_impact_data"       

# evaluated
dd$eval_date

#> [1] "2020-12-22"