list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 19 x 8
#>    name    accessor  data_type  region resolution geographical geospatial url   
#>    <chr>   <chr>     <chr>      <chr>  <chr>      <lgl>        <lgl>      <chr> 
#>  1 Europe… ecdc_data time seri… World  country    TRUE         FALSE      https…
#>  2 beoutb… beoutbre… individua… World  patient    TRUE         FALSE      https…
#>  3 EU dat… eu_data_… time seri… Europe country a… TRUE         FALSE      https…
#>  4 Descar… descarte… mobility   Unite… county     TRUE         FALSE      https…
#>  5 Apple … apple_mo… mobility   World  country a… TRUE         FALSE      https…
#>  6 USA Fa… usa_fact… time seri… Unite… county     TRUE         FALSE      https…
#>  7 Johns … jhu_data  time seri… World  country    TRUE         FALSE      https…
#>  8 Johns … jhu_us_d… time seri… Unite… county     TRUE         FALSE      https…
#>  9 New Yo… nytimes_… time seri… Unite… county     TRUE         FALSE      https…
#> 10 New Yo… nytimes_… time seri… Unite… state      TRUE         FALSE      https…
#> 11 Kaiser… kff_icu_… healthcar… Unite… Individua… TRUE         TRUE       https…
#> 12 United… us_count… demograph… Unite… county     TRUE         TRUE       https…
#> 13 United… us_count… geographic Unite… county     TRUE         TRUE       https…
#> 14 CovidT… covidtra… healthcar… Unite… Individua… TRUE         FALSE      https…
#> 15 CovidC… us_healt… healthcar… Unite… Individua… TRUE         TRUE       https…
#> 16 Oxford… governme… public po… Inter… country    TRUE         FALSE      https…
#> 17 Health… healthda… time seri… Unite… state      TRUE         FALSE      http:…
#> 18 Health… healthda… time seri… Inter… country    TRUE         FALSE      http:…
#> 19 United… us_state… public po… Unite… state      TRUE         FALSE      https…

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> # A tibble: 26,836 x 8
#> # Groups:   location_name [207]
#>    date       location_name iso2c iso3c population_2018 continent subset   count
#>    <date>     <chr>         <chr> <chr>           <dbl> <chr>     <chr>    <dbl>
#>  1 2019-12-31 Afghanistan   AF    AFG          37172386 Asia      deaths       0
#>  2 2019-12-31 Afghanistan   AF    AFG          37172386 Asia      confirm…     0
#>  3 2020-01-01 Afghanistan   AF    AFG          37172386 Asia      deaths       0
#>  4 2020-01-01 Afghanistan   AF    AFG          37172386 Asia      confirm…     0
#>  5 2020-01-02 Afghanistan   AF    AFG          37172386 Asia      deaths       0
#>  6 2020-01-02 Afghanistan   AF    AFG          37172386 Asia      confirm…     0
#>  7 2020-01-03 Afghanistan   AF    AFG          37172386 Asia      deaths       0
#>  8 2020-01-03 Afghanistan   AF    AFG          37172386 Asia      confirm…     0
#>  9 2020-01-04 Afghanistan   AF    AFG          37172386 Asia      deaths       0
#> 10 2020-01-04 Afghanistan   AF    AFG          37172386 Asia      confirm…     0
#> # … with 26,826 more rows

dd = dataset_details()
str(dd,list.len=3)

#> List of 18
#>  $ ecdc_data                  :List of 3
#>   ..$ columns   :List of 8
#>   .. ..$ date           : chr "Date"
#>   .. ..$ location_name  : chr "character"
#>   .. ..$ iso2c          : chr "character"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 25602
#>   .. ..$ ncol: int 8
#>   ..$ dates     :List of 2
#>   .. ..$ min: chr "2019-12-31"
#>   .. ..$ max: chr "2020-04-24"
#>  $ beoutbreakprepared_data    :List of 2
#>   ..$ columns   :List of 33
#>   .. ..$ ID                      : chr "character"
#>   .. ..$ age                     : chr "character"
#>   .. ..$ sex                     : chr "character"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 441082
#>   .. ..$ ncol: int 33
#>  $ eu_data_cache_data         :List of 3
#>   ..$ columns   :List of 15
#>   .. ..$ country               : chr "character"
#>   .. ..$ nuts_2                : chr "character"
#>   .. ..$ cases                 : chr "numeric"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 29829
#>   .. ..$ ncol: int 15
#>   ..$ dates     :List of 2
#>   .. ..$ min: chr "2020-02-24"
#>   .. ..$ max: chr "2020-04-16"
#>   [list output truncated]

names(dd)

#>  [1] "ecdc_data"                   "beoutbreakprepared_data"    
#>  [3] "eu_data_cache_data"          "descartes_mobility_data"    
#>  [5] "apple_mobility_data"         "usa_facts_data"             
#>  [7] "jhu_data"                    "jhu_us_data"                
#>  [9] "nytimes_county_data"         "nytimes_state_data"         
#> [11] "kff_icu_beds"                "us_county_geo_details"      
#> [13] "covidtracker_data"           "us_healthcare_capacity"     
#> [15] "government_policy_timeline"  "healthdata_projections_data"
#> [17] "us_state_distancing_policy"  "eval_date"