list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 15 x 8
#>    name    accessor  data_type  region resolution geographical geospatial url   
#>    <chr>   <chr>     <chr>      <chr>  <chr>      <lgl>        <lgl>      <chr> 
#>  1 beoutb… beoutbre… individua… World  patient    TRUE         FALSE      https…
#>  2 Descar… descarte… mobility   Unite… county     TRUE         FALSE      https…
#>  3 USA Fa… usa_fact… time seri… Unite… county     TRUE         FALSE      https…
#>  4 Johns … jhu_data  time seri… World  country    TRUE         FALSE      https…
#>  5 Johns … jhu_us_d… time seri… Unite… county     TRUE         FALSE      https…
#>  6 New Yo… nytimes_… time seri… Unite… county     TRUE         FALSE      https…
#>  7 New Yo… nytimes_… time seri… Unite… state      TRUE         FALSE      https…
#>  8 Kaiser… kff_icu_… healthcar… Unite… Individua… TRUE         TRUE       https…
#>  9 United… us_count… demograph… Unite… county     TRUE         TRUE       https…
#> 10 United… us_count… geographic Unite… county     TRUE         TRUE       https…
#> 11 CovidT… covidtra… healthcar… Unite… Individua… TRUE         FALSE      https…
#> 12 CovidC… us_healt… healthcar… Unite… Individua… TRUE         TRUE       https…
#> 13 Oxford… governme… public po… Unite… Individua… TRUE         FALSE      https…
#> 14 Health… healthda… time seri… Unite… state      TRUE         FALSE      http:…
#> 15 Health… healthda… time seri… Inter… country    TRUE         FALSE      http:…

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> using temporary cache /tmp/Rtmp7hsYhy/BiocFileCache

#> # A tibble: 419,904 x 33
#>    ID    age   sex   city  province country latitude longitude geo_resolution
#>    <chr> <chr> <chr> <chr> <chr>    <chr>      <dbl>     <dbl> <chr>         
#>  1 000-… <NA>  male  Shek… Hong Ko… China      22.4      114.  point         
#>  2 000-… 78    male  Vo E… Veneto   Italy      45.3       11.7 point         
#>  3 000-… 61    fema… <NA>  <NA>     Singap…     1.35     104.  admin0        
#>  4 000-… <NA>  <NA>  Zhen… Henan    China      34.6      113.  admin2        
#>  5 000-… <NA>  <NA>  Ping… Jiangxi  China      27.5      114.  admin2        
#>  6 000-… <NA>  <NA>  Yich… Jiangxi  China      28.3      115.  admin2        
#>  7 000-… <NA>  <NA>  Shan… Jiangxi  China      28.8      117.  admin2        
#>  8 000-… <NA>  <NA>  Fuzh… Jiangxi  China      27.5      116.  admin2        
#>  9 000-… <NA>  <NA>  Nanc… Jiangxi  China      28.7      116.  admin2        
#> 10 000-… <NA>  <NA>  Nanc… Jiangxi  China      28.7      116.  admin2        
#> # … with 419,894 more rows, and 24 more variables: date_onset_symptoms <date>,
#> #   date_admission_hospital <date>, date_confirmation <date>, symptoms <chr>,
#> #   lives_in_Wuhan <chr>, travel_history_dates <date>,
#> #   travel_history_location <chr>, reported_market_exposure <chr>,
#> #   additional_information <chr>, chronic_disease_binary <lgl>,
#> #   chronic_disease <chr>, source <chr>, sequence_available <chr>,
#> #   outcome <chr>, date_death_or_discharge <date>, notes_for_discussion <chr>,
#> #   location <chr>, admin3 <chr>, admin2 <chr>, admin1 <chr>,
#> #   country_new <chr>, admin_id <dbl>, data_moderator_initials <chr>,
#> #   travel_history_binary <lgl>

dd = dataset_details()
str(dd,list.len=3)

#> List of 14
#>  $ beoutbreakprepared_data    :List of 2
#>   ..$ columns   :List of 33
#>   .. ..$ ID                      : chr "character"
#>   .. ..$ age                     : chr "character"
#>   .. ..$ sex                     : chr "character"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 419904
#>   .. ..$ ncol: int 33
#>  $ descartes_mobility_data    :List of 3
#>   ..$ columns   :List of 9
#>   .. ..$ date        : chr "Date"
#>   .. ..$ country_code: chr "character"
#>   .. ..$ admin_level : chr "numeric"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 119979
#>   .. ..$ ncol: int 9
#>   ..$ dates     :List of 2
#>   .. ..$ min: chr NA
#>   .. ..$ max: chr NA
#>  $ usa_facts_data             :List of 3
#>   ..$ columns   :List of 6
#>   .. ..$ fips  : chr "character"
#>   .. ..$ county: chr "character"
#>   .. ..$ state : chr "character"
#>   .. .. [list output truncated]
#>   ..$ dimensions:List of 2
#>   .. ..$ nrow: int 543150
#>   .. ..$ ncol: int 6
#>   ..$ dates     :List of 2
#>   .. ..$ min: chr "2020-01-22"
#>   .. ..$ max: chr "2020-04-15"
#>   [list output truncated]

names(dd)

#>  [1] "beoutbreakprepared_data"     "descartes_mobility_data"    
#>  [3] "usa_facts_data"              "jhu_data"                   
#>  [5] "jhu_us_data"                 "nytimes_county_data"        
#>  [7] "nytimes_state_data"          "kff_icu_beds"               
#>  [9] "us_county_geo_details"       "covidtracker_data"          
#> [11] "us_healthcare_capacity"      "government_policy_timeline" 
#> [13] "healthdata_projections_data" "eval_date"