list and detail available sars2pack datasets

list and detail available sars2pack datasets

available_datasets()

dataset_details()

Value

A data.frame in which each row represents an available dataset.

name: the short name of the dataset

accessor: accessor function

Details

dataset_details: returns a pre-computed set of column names and types, dimensions of the datasets, and for datasets with a date (time course), the min and max dates included in the dataset. Each dataset is an item in the list. See examples for details and for viewing suggestions.

Functions

dataset_details:

Examples

res = available_datasets()
res

#> # A tibble: 27 x 8
#>    name    accessor  data_type geographical geospatial region resolution url    
#>    <chr>   <chr>     <list>    <lgl>        <lgl>      <list> <list>     <chr>  
#>  1 Extens… us_hospi… <chr [1]> TRUE         TRUE       <chr … <chr [1]>  https:…
#>  2 US cou… us_count… <chr [1]> TRUE         FALSE      <chr … <chr [3]>  https:…
#>  3 Corona… coronane… <chr [1]> TRUE         FALSE      <chr … <chr [2]>  https:…
#>  4 Our Wo… owid_data <chr [4]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  5 GISAID… cov_glue… <chr [1]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  6 Newick… cov_glue… <chr [1]> FALSE        FALSE      <chr … <chr [1]>  https:…
#>  7 US CDC… cdc_exce… <chr [3]> TRUE         FALSE      <chr … <chr [1]>  https:…
#>  8 Publis… param_es… <chr [1]> FALSE        FALSE      <NULL> <NULL>     https:…
#>  9 Aggreg… cdc_aggr… <chr [1]> TRUE         FALSE      <NULL> <chr [2]>  https:…
#> 10 Google… google_m… <chr [1]> TRUE         FALSE      <chr … <chr [3]>  https:…
#> # … with 17 more rows

# and how to use the accessor programmatically
get(res[1,]$accessor)()

#> Simple feature collection with 7581 features and 30 fields
#> Attribute-geometry relationship: 30 constant, 0 aggregate, 0 identity
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -176.6403 ymin: -14.29024 xmax: 145.7245 ymax: 71.29285
#> CRS:            EPSG:4326
#> # A tibble: 7,581 x 31
#>      fid id    name  address city  state zip   zip4  telephone type  status
#>    <dbl> <chr> <chr> <chr>   <chr> <chr> <chr> <chr> <chr>     <chr> <chr> 
#>  1     1 0005… CENT… 1025 N… HANF… CA    93230 NOT … NOT AVAI… GENE… CLOSED
#>  2     2 0053… LOS … 150 VI… WEST… CA    91362 NOT … NOT AVAI… GENE… OPEN  
#>  3     3 0011… EAST… 4060 W… LOS … CA    90023 NOT … NOT AVAI… GENE… OPEN  
#>  4     4 0017… SOUT… 6245 D… HOLL… CA    90028 NOT … (323) 46… GENE… OPEN  
#>  5     5 0023… KIND… 14148 … BALD… CA    91706 NOT … NOT AVAI… GENE… OPEN  
#>  6     6 0025… LAKE… 3700 E… LAKE… CA    90712 NOT … NOT AVAI… GENE… OPEN  
#>  7     7 0030… MONR… 323 SO… MONR… CA    91016 NOT … NOT AVAI… GENE… OPEN  
#>  8     8 0030… MONT… 900 SO… MONT… CA    91754 NOT … NOT AVAI… GENE… OPEN  
#>  9     9 0047… STAR… 4025 W… TORR… CA    90505 NOT … NOT AVAI… PSYC… OPEN  
#> 10    10 0009… CORO… 730 MA… CORO… CA    92879 NOT … NOT AVAI… GENE… OPEN  
#> # … with 7,571 more rows, and 20 more variables: population <dbl>,
#> #   county <chr>, countyfips <chr>, country <chr>, naics_code <dbl>,
#> #   naics_desc <chr>, source <chr>, sourcedate <chr>, val_method <chr>,
#> #   val_date <chr>, website <chr>, state_id <chr>, alt_name <chr>,
#> #   st_fips <chr>, owner <chr>, ttl_staff <dbl>, beds <dbl>, trauma <chr>,
#> #   helipad <chr>, geometry <POINT [°]>

dd = dataset_details()
str(dd,list.len=3)

#> List of 2
#>  $ datasets :List of 27
#>   ..$ us_hospital_details               :List of 2
#>   .. ..$ columns   :List of 31
#>   .. .. ..$ fid       : chr "numeric"
#>   .. .. ..$ id        : chr "character"
#>   .. .. ..$ name      : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 7581
#>   .. .. ..$ ncol: int 31
#>   ..$ us_county_health_rankings         :List of 2
#>   .. ..$ columns   :List of 9
#>   .. .. ..$ statecode    : chr "character"
#>   .. .. ..$ countycode   : chr "character"
#>   .. .. ..$ fips         : chr "character"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 2488126
#>   .. .. ..$ ncol: int 9
#>   ..$ coronanet_government_response_data:List of 2
#>   .. ..$ columns   :List of 40
#>   .. .. ..$ record_id          : chr "character"
#>   .. .. ..$ policy_id          : chr "numeric"
#>   .. .. ..$ recorded_date      : chr [1:2] "POSIXct" "POSIXt"
#>   .. .. .. [list output truncated]
#>   .. ..$ dimensions:List of 2
#>   .. .. ..$ nrow: int 13666
#>   .. .. ..$ ncol: int 40
#>   .. [list output truncated]
#>  $ eval_date: chr "2020-05-17"

names(dd$datasets)

#>  [1] "us_hospital_details"                "us_county_health_rankings"         
#>  [3] "coronanet_government_response_data" "owid_data"                         
#>  [5] "cov_glue_lineage_data"              "cov_glue_newick_data"              
#>  [7] "cdc_excess_deaths"                  "param_estimates_published"         
#>  [9] "cdc_aggregated_projections"         "google_mobility_data"              
#> [11] "ecdc_data"                          "beoutbreakprepared_data"           
#> [13] "eu_data_cache_data"                 "descartes_mobility_data"           
#> [15] "apple_mobility_data"                "usa_facts_data"                    
#> [17] "jhu_data"                           "jhu_us_data"                       
#> [19] "nytimes_county_data"                "nytimes_state_data"                
#> [21] "kff_icu_beds"                       "us_county_geo_details"             
#> [23] "covidtracker_data"                  "us_healthcare_capacity"            
#> [25] "government_policy_timeline"         "healthdata_projections_data"       
#> [27] "us_state_distancing_policy"        

# evaluated
dd$eval_date

#> [1] "2020-05-17"