Individual-level data contributed from around the world

Individual-level data contributed from around the world

beoutbreakprepared_data(quietly = TRUE)

Arguments

quietly
logical(1) defaults to TRUE. If FALSE, warnings generated during parsing will be displayed. These often relate to nonstandard date values that occur idiosyncratically.

Source

https://github.com/beoutbreakprepared/nCoV2019

Value

tidy data.frame of content

Note

This is individual level data, collected from diverse sources. Data may be messy and we have made limited attempts at clean up.

Citation

misckraemer2020epidemiological, author = nCoV-2019 Data Working Group, title = Epidemiological Data from the nCoV-2019 Outbreak: Early Descriptions from Publicly Available Data, howpublished = Accessed on yyyy-mm-dd from http://virological.org/t/epidemiological-data-from-the-ncov-2019-outbreak-early-descriptions-from-publicly-available-data/337, year = 2020

ARTICLEXu2020-wb, title = "Open access epidemiological data from the COVID-19 outbreak", author = "Xu, Bo and Kraemer, Moritz U G and Open COVID-19 Data Curation Group", journal = "The Lancet infectious diseases", volume = 20, number = 5, pages = "534", month = may, year = 2020, url = "http://dx.doi.org/10.1016/S1473-3099(20)30119-5", file = "All Papers/X/Xu et al. 2020 - Open access epidemiological data from the COVID-19 outbreak.pdf", language = "en", issn = "1473-3099, 1474-4457", pmid = "32087115", doi = "10.1016/S1473-3099(20)30119-5", pmc = "PMC7158984"

See also

Examples

res = beoutbreakprepared_data()
colnames(res)

#>  [1] "ID"                       "age"                     
#>  [3] "sex"                      "city"                    
#>  [5] "province"                 "country"                 
#>  [7] "latitude"                 "longitude"               
#>  [9] "geo_resolution"           "date_onset_symptoms"     
#> [11] "date_admission_hospital"  "date_confirmation"       
#> [13] "symptoms"                 "lives_in_Wuhan"          
#> [15] "travel_history_dates"     "travel_history_location" 
#> [17] "reported_market_exposure" "additional_information"  
#> [19] "chronic_disease_binary"   "chronic_disease"         
#> [21] "source"                   "sequence_available"      
#> [23] "outcome"                  "date_death_or_discharge" 
#> [25] "notes_for_discussion"     "location"                
#> [27] "admin3"                   "admin2"                  
#> [29] "admin1"                   "country_new"             
#> [31] "admin_id"                 "data_moderator_initials" 
#> [33] "travel_history_binary"   

head(res)

#> # A tibble: 6 x 33
#>   ID    age   sex   city  province country latitude longitude geo_resolution
#>   <chr> <chr> <chr> <chr> <chr>    <chr>      <dbl>     <dbl> <chr>         
#> 1 000-… NA    male  Shek… Hong Ko… China      22.4      114.  point         
#> 2 000-… 78    male  Vo E… Veneto   Italy      45.3       11.7 point         
#> 3 000-… 61    fema… NA    NA       Singap…     1.35     104.  admin0        
#> 4 000-… NA    NA    Zhen… Henan    China      34.6      113.  admin2        
#> 5 000-… NA    NA    Ping… Jiangxi  China      27.5      114.  admin2        
#> 6 000-… NA    NA    Yich… Jiangxi  China      28.3      115.  admin2        
#> # … with 24 more variables: date_onset_symptoms <date>,
#> #   date_admission_hospital <date>, date_confirmation <date>, symptoms <chr>,
#> #   lives_in_Wuhan <chr>, travel_history_dates <date>,
#> #   travel_history_location <chr>, reported_market_exposure <chr>,
#> #   additional_information <chr>, chronic_disease_binary <lgl>,
#> #   chronic_disease <chr>, source <chr>, sequence_available <chr>,
#> #   outcome <chr>, date_death_or_discharge <date>, notes_for_discussion <chr>,
#> #   location <chr>, admin3 <chr>, admin2 <chr>, admin1 <chr>,
#> #   country_new <chr>, admin_id <dbl>, data_moderator_initials <chr>,
#> #   travel_history_binary <lgl>