Projected deaths from COVID-19 models

The US CDC gathers projections from several groups around the world and aggregates them into a single data resource. See the reference below for details of the models.

cdc_aggregated_projections()

Value

a data.frame

Details

These models are not updated daily but more like weekly. This function will attempt to grab the latest version.

References

https://www.cdc.gov/coronavirus/2019-ncov/covid-data/forecasting-us.html

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = cdc_aggregated_projections()

#> Warning: download failed
#>   web resource path: ‘https://www.cdc.gov/coronavirus/2019-ncov/covid-data/files/2020-08-10-model-data.csv’
#>   local file path: ‘/tmp/RtmpvI96RS/BiocFileCache/91364b0b85e_2020-08-10-model-data.csv’
#>   reason: Timeout was reached: [www.cdc.gov] Resolving timed out after 10000 milliseconds

#> Warning: bfcadd() failed; resource removed
#>   rid: BFC12
#>   fpath: ‘https://www.cdc.gov/coronavirus/2019-ncov/covid-data/files/2020-08-10-model-data.csv’
#>   reason: download failed

#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()

head(res)

#> Error in head(res): object 'res' not found

dplyr::glimpse(res)

#> Error in dplyr::glimpse(res): object 'res' not found

# available models
table(res$model)

#> Error in table(res$model): object 'res' not found

# projection targets
table(res$target)

#> Error in table(res$target): object 'res' not found

min(res$forecast_date)

#> Error in eval(expr, envir, enclos): object 'res' not found

max(res$target_week_end_date)

#> Error in eval(expr, envir, enclos): object 'res' not found

library(dplyr)
library(ggplot2)

# FACET view
res_ny = res %>%
    dplyr::filter(location_name=='New York' & grepl('cum death', target)) %>%
    dplyr::filter(model!='UMass-MechBayes')

#> Error in dplyr::filter(., location_name == "New York" & grepl("cum death",     target)): object 'res' not found

res_ny %>%
    dplyr::filter(location_name=='New York') %>%
    ggplot(aes(x=model, y=point, color=model)) +
    geom_errorbar(aes(ymin= quantile_0.025, ymax =  quantile_0.975)) +
    facet_wrap(facets='target_week_end_date') +
    geom_point() +
    labs(y='Projected Deaths') +
    theme_bw() +
    theme(axis.text.x=element_blank()) +
    ggtitle('Projected New York deaths for week ending')

#> Error in dplyr::filter(., location_name == "New York"): object 'res_ny' not found

#'
# combined view
pd <- position_dodge(width = 3) # use this to offset points and error bars
res_ny %>%
    ggplot(aes(x=target_week_end_date, y=point, color=model)) +
    geom_errorbar(aes(ymin= quantile_0.025, ymax =  quantile_0.975), position=pd) +
    geom_point(position=pd) +
    labs(y='Projected Deaths') +
    geom_line(position=pd) +
    theme_bw() +
    theme(legend.position='bottom') +
    ggtitle('Projected New York deaths for week ending')

#> Error in ggplot(., aes(x = target_week_end_date, y = point, color = model)): object 'res_ny' not found