Time series cases, hospitalizations, and testing from Corona Data Scraper project

Corona Data Scraper pulls COVID-19 Coronavirus case data from verified sources, finds the corresponding GeoJSON features, and adds population data. All sources are cited right in the same row as the data. Covers testing, hospitalization, deaths, cases, icu. Includes population and geospatial coordinates.

coronadatascraper_data()

Source

https://coronadatascraper.com/timeseries.csv

References

https://coronadatascraper.com/#home

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = coronadatascraper_data()
colnames(res)

#>  [1] "name"                 "level"                "city"                
#>  [4] "county"               "state"                "country"             
#>  [7] "population"           "lat"                  "long"                
#> [10] "url"                  "aggregate"            "tz"                  
#> [13] "cases"                "deaths"               "recovered"           
#> [16] "active"               "tested"               "hospitalized"        
#> [19] "hospitalized_current" "discharged"           "icu"                 
#> [22] "icu_current"          "growthFactor"         "date"                

head(res)

#>                          name  level city  county    state country population
#> 1: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#> 2: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#> 3: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#> 4: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#> 5: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#> 6: Antwerp, Flanders, Belgium county      Antwerp Flanders Belgium    1847486
#>        lat   long                         url aggregate              tz cases
#> 1: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#> 2: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#> 3: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#> 4: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#> 5: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#> 6: 51.2485 4.7175 https://epistat.wiv-isp.be/           Europe/Brussels     4
#>    deaths recovered active tested hospitalized hospitalized_current discharged
#> 1:     NA        NA     NA     NA           NA                   NA         NA
#> 2:     NA        NA     NA     NA           NA                   NA         NA
#> 3:     NA        NA     NA     NA           NA                   NA         NA
#> 4:     NA        NA     NA     NA           NA                   NA         NA
#> 5:     NA        NA     NA     NA           NA                   NA         NA
#> 6:     NA        NA     NA     NA           NA                   NA         NA
#>    icu icu_current growthFactor       date
#> 1:  NA          NA           NA 2020-01-22
#> 2:  NA          NA            1 2020-01-23
#> 3:  NA          NA            1 2020-01-24
#> 4:  NA          NA            1 2020-01-25
#> 5:  NA          NA            1 2020-01-26
#> 6:  NA          NA            1 2020-01-27

dplyr::glimpse(res)

#> Rows: 428,748
#> Columns: 24
#> $ name                 <chr> "Antwerp, Flanders, Belgium", "Antwerp, Flanders…
#> $ level                <chr> "county", "county", "county", "county", "county"…
#> $ city                 <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ county               <chr> "Antwerp", "Antwerp", "Antwerp", "Antwerp", "Ant…
#> $ state                <chr> "Flanders", "Flanders", "Flanders", "Flanders", …
#> $ country              <chr> "Belgium", "Belgium", "Belgium", "Belgium", "Bel…
#> $ population           <int> 1847486, 1847486, 1847486, 1847486, 1847486, 184…
#> $ lat                  <dbl> 51.2485, 51.2485, 51.2485, 51.2485, 51.2485, 51.…
#> $ long                 <dbl> 4.7175, 4.7175, 4.7175, 4.7175, 4.7175, 4.7175, …
#> $ url                  <chr> "https://epistat.wiv-isp.be/", "https://epistat.…
#> $ aggregate            <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ tz                   <chr> "Europe/Brussels", "Europe/Brussels", "Europe/Br…
#> $ cases                <int> 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, …
#> $ deaths               <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ recovered            <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ active               <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ tested               <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ hospitalized         <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ hospitalized_current <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ discharged           <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ icu                  <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ icu_current          <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ growthFactor         <dbl> NA, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ date                 <date> 2020-01-22, 2020-01-23, 2020-01-24, 2020-01-25,…

length(unique(res$names))

#> [1] 0

length(unique(res$county))

#> [1] 1884

length(unique(res$state))

#> [1] 788

length(unique(res$country))

#> [1] 191