CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 40
#>   record_id           policy_id entry_type correct_type update_type update_level
#>   <chr>                   <dbl> <chr>      <chr>        <chr>       <chr>       
#> 1 R_3nCudDknlhUIjpONA   2975738 new_entry  original     NA          NA          
#> 2 R_3OoySQxTamo89oXNA   7024226 new_entry  original     NA          NA          
#> 3 R_3kLn6IU4rQGHtCbNA    889484 new_entry  original     NA          NA          
#> 4 R_1rAmwrNMSZPgFhrNA    129322 new_entry  original     NA          NA          
#> 5 R_21oIyCpRWpNISIDDv   6721435 new_entry  correction   NA          NA          
#> 6 R_21oIyCpRWpNISIDDx   6721435 new_entry  correction   NA          NA          
#> # … with 34 more variables: description <chr>, date_announced <date>,
#> #   date_start <date>, date_end <date>, country <chr>, iso3c <chr>,
#> #   iso2c <chr>, init_country_level <chr>, domestic_policy <dbl>,
#> #   province <chr>, ISO_L2 <chr>, city <chr>, type <chr>, type_sub_cat <chr>,
#> #   type_text <chr>, institution_status <chr>, target_country <chr>,
#> #   target_geog_level <chr>, target_region <chr>, target_province <chr>,
#> #   target_city <chr>, target_other <lgl>, target_who_what <chr>,
#> #   target_direction <chr>, travel_mechanism <chr>, compliance <chr>,
#> #   enforcer <chr>, index_high_est <dbl>, index_med_est <dbl>,
#> #   index_low_est <dbl>, index_country_rank <dbl>, link <chr>,
#> #   date_updated <date>, recorded_date <dttm>

colnames(res)

#>  [1] "record_id"          "policy_id"          "entry_type"        
#>  [4] "correct_type"       "update_type"        "update_level"      
#>  [7] "description"        "date_announced"     "date_start"        
#> [10] "date_end"           "country"            "iso3c"             
#> [13] "iso2c"              "init_country_level" "domestic_policy"   
#> [16] "province"           "ISO_L2"             "city"              
#> [19] "type"               "type_sub_cat"       "type_text"         
#> [22] "institution_status" "target_country"     "target_geog_level" 
#> [25] "target_region"      "target_province"    "target_city"       
#> [28] "target_other"       "target_who_what"    "target_direction"  
#> [31] "travel_mechanism"   "compliance"         "enforcer"          
#> [34] "index_high_est"     "index_med_est"      "index_low_est"     
#> [37] "index_country_rank" "link"               "date_updated"      
#> [40] "recorded_date"     

dplyr::glimpse(res)

#> Rows: 58,283
#> Columns: 40
#> $ record_id          <chr> "R_3nCudDknlhUIjpONA", "R_3OoySQxTamo89oXNA", "R_3k…
#> $ policy_id          <dbl> 2975738, 7024226, 889484, 129322, 6721435, 6721435,…
#> $ entry_type         <chr> "new_entry", "new_entry", "new_entry", "new_entry",…
#> $ correct_type       <chr> "original", "original", "original", "original", "co…
#> $ update_type        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ update_level       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ description        <chr> "March 6, Afghanistan \"Measures have been taken in…
#> $ date_announced     <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-28, 20…
#> $ date_start         <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-28, 20…
#> $ date_end           <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2020-04-21…
#> $ country            <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afgha…
#> $ iso3c              <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "A…
#> $ iso2c              <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF…
#> $ init_country_level <chr> "National", "National", "National", "National", "Mu…
#> $ domestic_policy    <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ province           <chr> NA, NA, NA, NA, "Herat", "Herat", "Herat", "Samanga…
#> $ ISO_L2             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ city               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ type               <chr> "Anti-Disinformation Measures", "Anti-Disinformatio…
#> $ type_sub_cat       <chr> NA, NA, NA, NA, "Preschool or childcare facilities …
#> $ type_text          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ institution_status <chr> NA, NA, NA, NA, NA, NA, NA, "Primary Schools closed…
#> $ target_country     <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afgha…
#> $ target_geog_level  <chr> "One or more countries, but not all countries", "On…
#> $ target_region      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ target_province    <chr> NA, NA, "Balkh; Kabul; Nangarhar; Kunar; Parwan; Ka…
#> $ target_city        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ target_other       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ target_who_what    <chr> "All Residents (Citizen Residents + Foreign Residen…
#> $ target_direction   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ travel_mechanism   <chr> NA, NA, "Not Applicable", NA, NA, NA, NA, NA, NA, N…
#> $ compliance         <chr> "Voluntary/Recommended but No Penalties", "Mandator…
#> $ enforcer           <chr> "Ministry/Department of Health,Other (Please specif…
#> $ index_high_est     <dbl> 53.02945, 53.86454, NA, NA, 53.66498, 53.66498, 53.…
#> $ index_med_est      <dbl> 50.60497, 52.36461, NA, NA, 50.91633, 50.91633, 50.…
#> $ index_low_est      <dbl> 48.28019, 50.98443, NA, NA, 48.83211, 48.83211, 48.…
#> $ index_country_rank <dbl> 137, 141, NA, NA, 137, 137, 137, 146, 146, 146, 146…
#> $ link               <chr> "https://www.etilaatroz.com/94246/fears-rumors-and-…
#> $ date_updated       <date> 2020-08-21, 2020-09-13, 2020-09-12, 2020-09-05, 20…
#> $ recorded_date      <dttm> 2020-08-21 18:32:29, 2020-09-13 18:56:27, 2020-09-…

summary(res)

#>   record_id           policy_id          entry_type        correct_type      
#>  Length:58283       Min.   :0.000e+00   Length:58283       Length:58283      
#>  Class :character   1st Qu.:2.465e+06   Class :character   Class :character  
#>  Mode  :character   Median :4.982e+06   Mode  :character   Mode  :character  
#>                     Mean   :1.079e+09                                        
#>                     3rd Qu.:7.443e+06                                        
#>                     Max.   :6.258e+13                                        
#>                                                                              
#>  update_type        update_level       description        date_announced      
#>  Length:58283       Length:58283       Length:58283       Min.   :2019-12-31  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-20  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-04-14  
#>                                                           Mean   :2020-05-06  
#>                                                           3rd Qu.:2020-06-07  
#>                                                           Max.   :2022-12-22  
#>                                                           NA's   :12          
#>    date_start            date_end            country         
#>  Min.   :2019-12-31   Min.   :2020-01-06   Length:58283      
#>  1st Qu.:2020-03-22   1st Qu.:2020-04-30   Class :character  
#>  Median :2020-04-16   Median :2020-06-05   Mode  :character  
#>  Mean   :2020-05-09   Mean   :2020-07-08                     
#>  3rd Qu.:2020-06-08   3rd Qu.:2020-09-01                     
#>  Max.   :2024-05-29   Max.   :2024-05-29                     
#>                       NA's   :27333                          
#>     iso3c              iso2c           init_country_level domestic_policy
#>  Length:58283       Length:58283       Length:58283       Min.   :0.000  
#>  Class :character   Class :character   Class :character   1st Qu.:1.000  
#>  Mode  :character   Mode  :character   Mode  :character   Median :1.000  
#>                                                           Mean   :0.935  
#>                                                           3rd Qu.:1.000  
#>                                                           Max.   :1.000  
#>                                                                          
#>    province            ISO_L2              city               type          
#>  Length:58283       Length:58283       Length:58283       Length:58283      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  type_sub_cat        type_text         institution_status target_country    
#>  Length:58283       Length:58283       Length:58283       Length:58283      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_geog_level  target_region      target_province    target_city       
#>  Length:58283       Length:58283       Length:58283       Length:58283      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_other   target_who_what    target_direction   travel_mechanism  
#>  Mode:logical   Length:58283       Length:58283       Length:58283      
#>  NA's:58283     Class :character   Class :character   Class :character  
#>                 Mode  :character   Mode  :character   Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>   compliance          enforcer         index_high_est  index_med_est  
#>  Length:58283       Length:58283       Min.   :48.29   Min.   :45.42  
#>  Class :character   Class :character   1st Qu.:52.69   1st Qu.:51.45  
#>  Mode  :character   Mode  :character   Median :54.00   Median :52.68  
#>                                        Mean   :54.42   Mean   :52.59  
#>                                        3rd Qu.:55.59   3rd Qu.:53.84  
#>                                        Max.   :70.74   Max.   :57.59  
#>                                        NA's   :25331   NA's   :25331  
#>  index_low_est   index_country_rank     link            date_updated       
#>  Min.   :37.03   Min.   :  1.0      Length:58283       Min.   :2020-03-26  
#>  1st Qu.:50.01   1st Qu.: 77.0      Class :character   1st Qu.:2020-06-01  
#>  Median :51.31   Median :137.0      Mode  :character   Median :2020-08-13  
#>  Mean   :50.84   Mean   :126.9                         Mean   :2020-08-27  
#>  3rd Qu.:52.48   3rd Qu.:179.0                         3rd Qu.:2020-12-04  
#>  Max.   :56.26   Max.   :194.0                         Max.   :2021-02-16  
#>  NA's   :25331   NA's   :25331                                             
#>  recorded_date                
#>  Min.   :2020-03-26 05:04:08  
#>  1st Qu.:2020-06-01 16:38:47  
#>  Median :2020-08-13 18:19:38  
#>  Mean   :2020-08-27 18:44:01  
#>  3rd Qu.:2020-12-04 13:37:07  
#>  Max.   :2021-02-16 23:53:13  
#>