CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 40
#>   record_id policy_id recorded_date       date_updated date_announced date_start
#>   <chr>         <dbl> <dttm>              <date>       <date>         <date>    
#> 1 1000775Cn   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14     2020-03-14
#> 2 1000775Cf   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14     2020-03-14
#> 3 1003875NA   1003875 2020-04-15 11:04:50 2020-04-15   2020-04-07     2020-04-07
#> 4 1004681NA   2575122 2020-04-03 08:33:25 2020-04-10   2020-03-22     2020-03-25
#> 5 4866080NA   4866080 2020-03-30 05:21:44 2020-05-04   2020-03-20     2020-03-21
#> 6 1011795NA   4951252 2020-04-03 03:29:36 2020-04-03   2020-03-08     2020-03-08
#> # … with 34 more variables: date_end <date>, entry_type <chr>,
#> #   update_type <chr>, event_description <chr>, domestic_policy <dbl>,
#> #   type <chr>, type_sub_cat <chr>, type_text <dbl>, index_high_est <dbl>,
#> #   index_med_est <dbl>, index_low_est <dbl>, index_country_rank <dbl>,
#> #   country <chr>, init_country_level <chr>, province <chr>, city <chr>,
#> #   corr_entry_type <chr>, type_new_admin_coop <lgl>, source_file_Type <chr>,
#> #   source_file_2_Type <lgl>, target_country <chr>, target_geog_level <chr>,
#> #   target_region <chr>, target_province <chr>, target_city <chr>,
#> #   target_other <lgl>, target_who_what <chr>, target_direction <chr>,
#> #   travel_mechanism <chr>, compliance <chr>, enforcer <chr>, link <chr>,
#> #   iso3c <chr>, iso2c <chr>

colnames(res)

#>  [1] "record_id"           "policy_id"           "recorded_date"      
#>  [4] "date_updated"        "date_announced"      "date_start"         
#>  [7] "date_end"            "entry_type"          "update_type"        
#> [10] "event_description"   "domestic_policy"     "type"               
#> [13] "type_sub_cat"        "type_text"           "index_high_est"     
#> [16] "index_med_est"       "index_low_est"       "index_country_rank" 
#> [19] "country"             "init_country_level"  "province"           
#> [22] "city"                "corr_entry_type"     "type_new_admin_coop"
#> [25] "source_file_Type"    "source_file_2_Type"  "target_country"     
#> [28] "target_geog_level"   "target_region"       "target_province"    
#> [31] "target_city"         "target_other"        "target_who_what"    
#> [34] "target_direction"    "travel_mechanism"    "compliance"         
#> [37] "enforcer"            "link"                "iso3c"              
#> [40] "iso2c"              

dplyr::glimpse(res)

#> Rows: 13,666
#> Columns: 40
#> $ record_id           <chr> "1000775Cn", "1000775Cf", "1003875NA", "1004681NA…
#> $ policy_id           <dbl> 1342691, 1342691, 1003875, 2575122, 4866080, 4951…
#> $ recorded_date       <dttm> 2020-04-04 15:43:15, 2020-04-04 15:43:15, 2020-0…
#> $ date_updated        <date> 2020-04-10, 2020-04-10, 2020-04-15, 2020-04-10, …
#> $ date_announced      <date> 2020-03-14, 2020-03-14, 2020-04-07, 2020-03-22, …
#> $ date_start          <date> 2020-03-14, 2020-03-14, 2020-04-07, 2020-03-25, …
#> $ date_end            <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2020-05-…
#> $ entry_type          <chr> "update", "update", "new_entry", "update", "corre…
#> $ update_type         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ event_description   <chr> "Chile extends mandatory quarantine for people fr…
#> $ domestic_policy     <dbl> 0, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0…
#> $ type                <chr> "Quarantine/Lockdown", "Quarantine/Lockdown", "Ot…
#> $ type_sub_cat        <chr> "Self-Quarantine (i.e. quarantine at home)", "Qua…
#> $ type_text           <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ index_high_est      <dbl> 48.29759, 48.29759, 56.92399, 48.24469, 49.66255,…
#> $ index_med_est       <dbl> 47.21572, 47.21572, 50.71799, 47.80725, 47.86248,…
#> $ index_low_est       <dbl> 45.99840, 45.99840, 44.64792, 47.34911, 46.47806,…
#> $ index_country_rank  <dbl> 89, 89, 186, 65, 79, 104, 100, 175, 31, 46, NA, 1…
#> $ country             <chr> "Chile", "Chile", "Guatemala", "Kenya", "Italy", …
#> $ init_country_level  <chr> "National", "National", "National", "National", "…
#> $ province            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ city                <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ corr_entry_type     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ type_new_admin_coop <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ source_file_Type    <chr> "application/pdf", "application/pdf", "applicatio…
#> $ source_file_2_Type  <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ target_country      <chr> "China,France,Germany,Iran,Italy,Japan,North Kore…
#> $ target_geog_level   <chr> "One or more countries, but not all countries", "…
#> $ target_region       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ target_province     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ target_city         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ target_other        <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ target_who_what     <chr> "All Travelers (Citizen Travelers + Foreign Trave…
#> $ target_direction    <chr> "Inbound", "Inbound", NA, "Inbound", NA, NA, NA, …
#> $ travel_mechanism    <chr> "All kinds of transport", "All kinds of transport…
#> $ compliance          <chr> "Mandatory with Fines", "Mandatory with Fines", "…
#> $ enforcer            <chr> "National Government,Ministry/Department of Healt…
#> $ link                <chr> "https://www.gob.cl/noticias/covid-19-ministerio-…
#> $ iso3c               <chr> "CHL", "CHL", "GTM", "KEN", "ITA", "USA", "PER", …
#> $ iso2c               <chr> "CL", "CL", "GT", "KE", "IT", "US", "PE", "IL", "…

summary(res)

#>   record_id           policy_id         recorded_date                
#>  Length:13666       Min.   :2.559e+03   Min.   :2020-03-28 05:24:44  
#>  Class :character   1st Qu.:2.566e+06   1st Qu.:2020-04-03 00:47:10  
#>  Mode  :character   Median :5.115e+06   Median :2020-04-07 02:50:43  
#>                     Mean   :4.584e+09   Mean   :2020-04-10 07:51:02  
#>                     3rd Qu.:7.575e+06   3rd Qu.:2020-04-17 15:13:48  
#>                     Max.   :6.258e+13   Max.   :2020-04-30 23:56:08  
#>                                                                      
#>   date_updated        date_announced         date_start        
#>  Min.   :2020-03-28   Min.   :2019-12-31   Min.   :2019-12-31  
#>  1st Qu.:2020-04-10   1st Qu.:2020-03-12   1st Qu.:2020-03-13  
#>  Median :2020-04-28   Median :2020-03-20   Median :2020-03-20  
#>  Mean   :2020-04-22   Mean   :2020-03-18   Mean   :2020-03-19  
#>  3rd Qu.:2020-05-04   3rd Qu.:2020-04-01   3rd Qu.:2020-04-02  
#>  Max.   :2020-05-05   Max.   :2020-04-29   Max.   :2021-01-01  
#>                       NA's   :22                               
#>     date_end           entry_type        update_type        event_description 
#>  Min.   :2020-01-28   Length:13666       Length:13666       Length:13666      
#>  1st Qu.:2020-04-04   Class :character   Class :character   Class :character  
#>  Median :2020-04-15   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :2020-04-20                                                           
#>  3rd Qu.:2020-04-30                                                           
#>  Max.   :2022-03-13                                                           
#>  NA's   :9606                                                                 
#>  domestic_policy      type           type_sub_cat         type_text    
#>  Min.   :0.0000   Length:13666       Length:13666       Min.   : 4.00  
#>  1st Qu.:1.0000   Class :character   Class :character   1st Qu.:14.00  
#>  Median :1.0000   Mode  :character   Mode  :character   Median :21.00  
#>  Mean   :0.7724                                         Mean   :20.47  
#>  3rd Qu.:1.0000                                         3rd Qu.:28.00  
#>  Max.   :1.0000                                         Max.   :58.00  
#>                                                         NA's   :13553  
#>  index_high_est  index_med_est   index_low_est   index_country_rank
#>  Min.   :44.91   Min.   :41.90   Min.   :32.80   Min.   :  1.0     
#>  1st Qu.:47.86   1st Qu.:47.11   1st Qu.:46.17   1st Qu.: 58.0     
#>  Median :48.60   Median :47.68   Median :46.71   Median :108.0     
#>  Mean   :49.10   Mean   :47.88   Mean   :46.62   Mean   :108.5     
#>  3rd Qu.:50.03   3rd Qu.:48.81   3rd Qu.:47.87   3rd Qu.:159.0     
#>  Max.   :60.47   Max.   :50.95   Max.   :50.11   Max.   :189.0     
#>  NA's   :803     NA's   :803     NA's   :803     NA's   :803       
#>    country          init_country_level   province             city          
#>  Length:13666       Length:13666       Length:13666       Length:13666      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  corr_entry_type    type_new_admin_coop source_file_Type   source_file_2_Type
#>  Length:13666       Mode:logical        Length:13666       Mode:logical      
#>  Class :character   NA's:13666          Class :character   NA's:13666        
#>  Mode  :character                       Mode  :character                     
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>  target_country     target_geog_level  target_region      target_province   
#>  Length:13666       Length:13666       Length:13666       Length:13666      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_city        target_other   target_who_what    target_direction  
#>  Length:13666       Mode:logical   Length:13666       Length:13666      
#>  Class :character   NA's:13666     Class :character   Class :character  
#>  Mode  :character                  Mode  :character   Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  travel_mechanism    compliance          enforcer             link          
#>  Length:13666       Length:13666       Length:13666       Length:13666      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     iso3c              iso2c          
#>  Length:13666       Length:13666      
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>