CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 39
#>   record_id policy_id entry_type correct_type update_type update_level
#>   <chr>         <dbl> <chr>      <chr>        <chr>       <chr>       
#> 1 R_1gdyC2…   3889409 new_entry  original     <NA>        <NA>        
#> 2 R_vp4xmb…   1429889 new_entry  original     <NA>        <NA>        
#> 3 R_1KxVwu…   1846710 new_entry  original     <NA>        <NA>        
#> 4 R_yz7pUp…   9724038 new_entry  original     <NA>        <NA>        
#> 5 R_22kWz6…   1919243 new_entry  correction   <NA>        <NA>        
#> 6 R_2Ytcwe…   1924743 new_entry  original     <NA>        <NA>        
#> # … with 33 more variables: description <chr>, date_announced <date>,
#> #   date_start <date>, date_end <date>, country <chr>, iso3c <chr>,
#> #   iso2c <chr>, init_country_level <chr>, domestic_policy <dbl>,
#> #   province <chr>, city <chr>, type <chr>, type_sub_cat <chr>,
#> #   type_text <chr>, school_status <chr>, target_country <chr>,
#> #   target_geog_level <chr>, target_region <chr>, target_province <chr>,
#> #   target_city <chr>, target_other <lgl>, target_who_what <chr>,
#> #   target_direction <chr>, travel_mechanism <chr>, compliance <chr>,
#> #   enforcer <chr>, index_high_est <dbl>, index_med_est <dbl>,
#> #   index_low_est <dbl>, index_country_rank <dbl>, link <chr>,
#> #   date_updated <date>, recorded_date <dttm>

colnames(res)

#>  [1] "record_id"          "policy_id"          "entry_type"        
#>  [4] "correct_type"       "update_type"        "update_level"      
#>  [7] "description"        "date_announced"     "date_start"        
#> [10] "date_end"           "country"            "iso3c"             
#> [13] "iso2c"              "init_country_level" "domestic_policy"   
#> [16] "province"           "city"               "type"              
#> [19] "type_sub_cat"       "type_text"          "school_status"     
#> [22] "target_country"     "target_geog_level"  "target_region"     
#> [25] "target_province"    "target_city"        "target_other"      
#> [28] "target_who_what"    "target_direction"   "travel_mechanism"  
#> [31] "compliance"         "enforcer"           "index_high_est"    
#> [34] "index_med_est"      "index_low_est"      "index_country_rank"
#> [37] "link"               "date_updated"       "recorded_date"     

dplyr::glimpse(res)

#> Rows: 35,205
#> Columns: 39
#> $ record_id          <chr> "R_1gdyC25xLGlP5H1Bw", "R_vp4xmb4cOiQ3nb3NA", "R_1…
#> $ policy_id          <dbl> 3889409, 1429889, 1846710, 9724038, 1919243, 19247…
#> $ entry_type         <chr> "new_entry", "new_entry", "new_entry", "new_entry"…
#> $ correct_type       <chr> "original", "original", "original", "original", "c…
#> $ update_type        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ update_level       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ description        <chr> "US state Michigan creates new task force for econ…
#> $ date_announced     <date> 2020-05-29, 2020-06-09, 2020-05-28, 2020-07-08, 2…
#> $ date_start         <date> 2024-05-29, 2020-06-09, 2020-05-28, 2020-07-08, 2…
#> $ date_end           <date> 2024-05-29, 2023-01-01, 2021-12-31, 2021-12-08, 2…
#> $ country            <chr> "United States of America", "Germany", "Finland", …
#> $ iso3c              <chr> "USA", "DEU", "FIN", "ARG", "EGY", "IND", "IND", "…
#> $ iso2c              <chr> "US", "DE", "FI", "AR", "EG", "IN", "IN", "DE", "N…
#> $ init_country_level <chr> "Provincial", "Provincial", "National", NA, "Natio…
#> $ domestic_policy    <dbl> 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1,…
#> $ province           <chr> "Michigan", "Hesse", NA, NA, NA, "Kerala", "Kerala…
#> $ city               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type               <chr> "New Task Force, Bureau or Administrative Configur…
#> $ type_sub_cat       <chr> "New Task Force or Bureau (i.e. establishment of a…
#> $ type_text          <chr> NA, "Economic aid by the state government of Hesse…
#> $ school_status      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "Prima…
#> $ target_country     <chr> "United States of America", "Germany", "Finland", …
#> $ target_geog_level  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_region      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province    <chr> "Michigan", "Hesse", NA, NA, NA, "Kerala", "Kerala…
#> $ target_city        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_other       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what    <chr> "All Residents (Citizen Residents + Foreign Reside…
#> $ target_direction   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ travel_mechanism   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ compliance         <chr> "Mandatory (Unspecified/Implied)", "Voluntary/Reco…
#> $ enforcer           <chr> "New (COVID-19 specific) Task Force", "Provincial/…
#> $ index_high_est     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 52.558…
#> $ index_med_est      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 51.641…
#> $ index_low_est      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 50.893…
#> $ index_country_rank <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 63, 63…
#> $ link               <chr> "https://www.michigan.gov/whitmer/0,9309,7-387-904…
#> $ date_updated       <date> 2020-07-05, 2020-06-22, 2020-06-12, 2020-07-17, 2…
#> $ recorded_date      <dttm> 2020-07-05 12:40:29, 2020-06-22 09:20:26, 2020-06…

summary(res)

#>   record_id           policy_id          entry_type        correct_type      
#>  Length:35205       Min.   :2.087e+03   Length:35205       Length:35205      
#>  Class :character   1st Qu.:2.466e+06   Class :character   Class :character  
#>  Mode  :character   Median :5.035e+06   Mode  :character   Mode  :character  
#>                     Mean   :1.783e+09                                        
#>                     3rd Qu.:7.475e+06                                        
#>                     Max.   :6.258e+13                                        
#>                                                                              
#>  update_type        update_level       description        date_announced      
#>  Length:35205       Length:35205       Length:35205       Min.   :2019-12-31  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-18  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-04-02  
#>                                                           Mean   :2020-04-15  
#>                                                           3rd Qu.:2020-05-06  
#>                                                           Max.   :2030-02-28  
#>                                                           NA's   :14          
#>    date_start            date_end            country         
#>  Min.   :2019-12-31   Min.   :2020-01-10   Length:35205      
#>  1st Qu.:2020-03-18   1st Qu.:2020-04-24   Class :character  
#>  Median :2020-04-03   Median :2020-05-22   Mode  :character  
#>  Mean   :2020-04-16   Mean   :2020-06-01                     
#>  3rd Qu.:2020-05-08   3rd Qu.:2020-06-30                     
#>  Max.   :2024-05-29   Max.   :2024-05-29                     
#>                       NA's   :20045                          
#>     iso3c              iso2c           init_country_level domestic_policy 
#>  Length:35205       Length:35205       Length:35205       Min.   :0.0000  
#>  Class :character   Class :character   Class :character   1st Qu.:0.0000  
#>  Mode  :character   Mode  :character   Mode  :character   Median :0.0000  
#>                                                           Mean   :0.4421  
#>                                                           3rd Qu.:1.0000  
#>                                                           Max.   :1.0000  
#>                                                                           
#>    province             city               type           type_sub_cat      
#>  Length:35205       Length:35205       Length:35205       Length:35205      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>   type_text         school_status      target_country     target_geog_level 
#>  Length:35205       Length:35205       Length:35205       Length:35205      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_region      target_province    target_city        target_other  
#>  Length:35205       Length:35205       Length:35205       Mode:logical  
#>  Class :character   Class :character   Class :character   NA's:35205    
#>  Mode  :character   Mode  :character   Mode  :character                 
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_who_what    target_direction   travel_mechanism    compliance       
#>  Length:35205       Length:35205       Length:35205       Length:35205      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    enforcer         index_high_est  index_med_est   index_low_est  
#>  Length:35205       Min.   :48.29   Min.   :45.42   Min.   :37.03  
#>  Class :character   1st Qu.:52.64   1st Qu.:51.44   1st Qu.:50.06  
#>  Mode  :character   Median :53.91   Median :52.57   Median :51.31  
#>                     Mean   :54.24   Mean   :52.53   Mean   :50.87  
#>                     3rd Qu.:55.38   3rd Qu.:53.77   3rd Qu.:52.40  
#>                     Max.   :70.74   Max.   :57.59   Max.   :56.26  
#>                     NA's   :10806   NA's   :10806   NA's   :10806  
#>  index_country_rank     link            date_updated       
#>  Min.   :  1.0      Length:35205       Min.   :2020-03-28  
#>  1st Qu.: 76.0      Class :character   1st Qu.:2020-04-20  
#>  Median :132.0      Mode  :character   Median :2020-06-12  
#>  Mean   :124.4                         Mean   :2020-06-06  
#>  3rd Qu.:176.0                         3rd Qu.:2020-07-19  
#>  Max.   :194.0                         Max.   :2020-08-20  
#>  NA's   :10806                                             
#>  recorded_date                
#>  Min.   :2020-03-28 11:14:33  
#>  1st Qu.:2020-04-20 06:41:12  
#>  Median :2020-06-12 16:36:31  
#>  Mean   :2020-06-06 20:45:35  
#>  3rd Qu.:2020-07-19 17:17:41  
#>  Max.   :2020-08-20 21:57:26  
#>