CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 38
#>   record_id policy_id recorded_date       date_updated date_announced date_start
#>   <chr>         <dbl> <dttm>              <date>       <date>         <date>    
#> 1 1000509Dv   1000509 2020-06-17 12:53:43 2020-06-17   2020-03-19     2020-03-19
#> 2 1000509Dw   1000509 2020-06-17 12:53:43 2020-06-17   2020-03-19     2020-03-19
#> 3 1000509Cn   1000509 2020-06-17 12:53:43 2020-06-17   2020-03-19     2020-03-19
#> 4 1000775Em   1342691 2020-04-04 15:43:15 2020-04-04   2020-03-14     2020-03-14
#> 5 1000775Ea   1342691 2020-04-04 15:43:15 2020-04-04   2020-03-14     2020-03-14
#> 6 1003875NA   1003875 2020-04-15 11:04:50 2020-04-15   2020-04-07     2020-04-07
#> # … with 32 more variables: date_end <date>, entry_type <chr>,
#> #   update_type <chr>, event_description <chr>, domestic_policy <dbl>,
#> #   type <chr>, type_sub_cat <chr>, type_text <dbl>, index_high_est <dbl>,
#> #   index_med_est <dbl>, index_low_est <dbl>, index_country_rank <dbl>,
#> #   correct_type <chr>, country <chr>, init_country_level <chr>,
#> #   province <chr>, city <chr>, source_corr_type <chr>, target_country <chr>,
#> #   target_geog_level <chr>, target_region <chr>, target_province <chr>,
#> #   target_city <chr>, target_other <lgl>, target_who_what <chr>,
#> #   target_direction <chr>, travel_mechanism <chr>, compliance <chr>,
#> #   enforcer <chr>, link <chr>, iso3c <chr>, iso2c <chr>

colnames(res)

#>  [1] "record_id"          "policy_id"          "recorded_date"     
#>  [4] "date_updated"       "date_announced"     "date_start"        
#>  [7] "date_end"           "entry_type"         "update_type"       
#> [10] "event_description"  "domestic_policy"    "type"              
#> [13] "type_sub_cat"       "type_text"          "index_high_est"    
#> [16] "index_med_est"      "index_low_est"      "index_country_rank"
#> [19] "correct_type"       "country"            "init_country_level"
#> [22] "province"           "city"               "source_corr_type"  
#> [25] "target_country"     "target_geog_level"  "target_region"     
#> [28] "target_province"    "target_city"        "target_other"      
#> [31] "target_who_what"    "target_direction"   "travel_mechanism"  
#> [34] "compliance"         "enforcer"           "link"              
#> [37] "iso3c"              "iso2c"             

dplyr::glimpse(res)

#> Rows: 15,982
#> Columns: 38
#> $ record_id          <chr> "1000509Dv", "1000509Dw", "1000509Cn", "1000775Em"…
#> $ policy_id          <dbl> 1000509, 1000509, 1000509, 1342691, 1342691, 10038…
#> $ recorded_date      <dttm> 2020-06-17 12:53:43, 2020-06-17 12:53:43, 2020-06…
#> $ date_updated       <date> 2020-06-17, 2020-06-17, 2020-06-17, 2020-04-04, 2…
#> $ date_announced     <date> 2020-03-19, 2020-03-19, 2020-03-19, 2020-03-14, 2…
#> $ date_start         <date> 2020-03-19, 2020-03-19, 2020-03-19, 2020-03-14, 2…
#> $ date_end           <date> 2020-04-27, 2020-04-27, 2020-04-27, NA, NA, NA, N…
#> $ entry_type         <chr> "update", "update", "update", "update", "update", …
#> $ update_type        <chr> "End of Policy", "End of Policy", "End of Policy",…
#> $ event_description  <chr> "On March 19, the government of Croatia ordered th…
#> $ domestic_policy    <dbl> 1, 1, 1, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ type               <chr> "Restriction and Regulation of Government Services…
#> $ type_sub_cat       <chr> "Public libraries", "Public museums/galleries", "N…
#> $ type_text          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ index_high_est     <dbl> 56.11148, 56.11148, 56.11148, 52.03633, 52.03633, …
#> $ index_med_est      <dbl> 55.05597, 55.05597, 55.05597, 51.22552, 51.22552, …
#> $ index_low_est      <dbl> 54.02054, 54.02054, 54.02054, 50.20661, 50.20661, …
#> $ index_country_rank <dbl> 192, 192, 192, 113, 113, 181, NA, 95, 144, NA, 182…
#> $ correct_type       <chr> "original", "original", "original", "original", "o…
#> $ country            <chr> "Croatia", "Croatia", "Croatia", "Chile", "Chile",…
#> $ init_country_level <chr> "National", "National", "National", "National", "N…
#> $ province           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "Bern", NA…
#> $ city               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ source_corr_type   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_country     <chr> "Croatia", "Croatia", "Croatia", "China,France,Ger…
#> $ target_geog_level  <chr> NA, NA, NA, "One or more countries, but not all co…
#> $ target_region      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "Bern", NA…
#> $ target_city        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, "Duqm", NA, NA…
#> $ target_other       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what    <chr> NA, NA, NA, "All Travelers (Citizen Travelers + Fo…
#> $ target_direction   <chr> NA, NA, NA, "Inbound", "Inbound", NA, NA, NA, "Inb…
#> $ travel_mechanism   <chr> NA, NA, NA, "All kinds of transport", "All kinds o…
#> $ compliance         <chr> "Mandatory (Unspecified/Implied)", "Mandatory (Uns…
#> $ enforcer           <chr> "National Government", "National Government", "Nat…
#> $ link               <chr> "https://hr.usembassy.gov/health-alert-u-s-embassy…
#> $ iso3c              <chr> "HRV", "HRV", "HRV", "CHL", "CHL", "GTM", "TUR", "…
#> $ iso2c              <chr> "HR", "HR", "HR", "CL", "CL", "GT", "TR", "EG", "K…

summary(res)

#>   record_id           policy_id         recorded_date                
#>  Length:15982       Min.   :2.347e+03   Min.   :2020-03-28 05:38:17  
#>  Class :character   1st Qu.:2.534e+06   1st Qu.:2020-04-04 22:07:54  
#>  Mode  :character   Median :5.120e+06   Median :2020-04-21 06:29:26  
#>                     Mean   :3.921e+09   Mean   :2020-04-28 01:57:44  
#>                     3rd Qu.:7.528e+06   3rd Qu.:2020-05-17 12:44:37  
#>                     Max.   :6.258e+13   Max.   :2020-06-22 22:34:13  
#>                                                                      
#>   date_updated        date_announced         date_start        
#>  Min.   :2020-03-28   Min.   :2019-12-31   Min.   :2019-12-31  
#>  1st Qu.:2020-04-04   1st Qu.:2020-03-16   1st Qu.:2020-03-16  
#>  Median :2020-04-21   Median :2020-03-27   Median :2020-03-28  
#>  Mean   :2020-04-27   Mean   :2020-04-01   Mean   :2020-04-02  
#>  3rd Qu.:2020-05-17   3rd Qu.:2020-04-16   3rd Qu.:2020-04-18  
#>  Max.   :2020-06-22   Max.   :2020-06-22   Max.   :2021-01-01  
#>                       NA's   :14                               
#>     date_end           entry_type        update_type        event_description 
#>  Min.   :2020-01-28   Length:15982       Length:15982       Length:15982      
#>  1st Qu.:2020-04-11   Class :character   Class :character   Class :character  
#>  Median :2020-04-26   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :2020-05-05                                                           
#>  3rd Qu.:2020-05-18                                                           
#>  Max.   :2023-01-01                                                           
#>  NA's   :11366                                                                
#>  domestic_policy      type           type_sub_cat         type_text    
#>  Min.   :0.0000   Length:15982       Length:15982       Min.   : 2.00  
#>  1st Qu.:1.0000   Class :character   Class :character   1st Qu.:14.00  
#>  Median :1.0000   Mode  :character   Mode  :character   Median :20.00  
#>  Mean   :0.8453                                         Mean   :20.13  
#>  3rd Qu.:1.0000                                         3rd Qu.:28.00  
#>  Max.   :1.0000                                         Max.   :85.00  
#>                                                         NA's   :15849  
#>  index_high_est  index_med_est   index_low_est   index_country_rank
#>  Min.   :48.29   Min.   :45.42   Min.   :37.19   Min.   :  1.0     
#>  1st Qu.:52.47   1st Qu.:51.26   1st Qu.:49.82   1st Qu.: 71.0     
#>  Median :53.54   Median :52.29   Median :51.12   Median :123.0     
#>  Mean   :53.85   Mean   :52.23   Mean   :50.62   Mean   :118.7     
#>  3rd Qu.:54.97   3rd Qu.:53.46   3rd Qu.:52.13   3rd Qu.:168.0     
#>  Max.   :70.74   Max.   :57.59   Max.   :56.26   Max.   :194.0     
#>  NA's   :2871    NA's   :2871    NA's   :2871    NA's   :2871      
#>  correct_type         country          init_country_level   province        
#>  Length:15982       Length:15982       Length:15982       Length:15982      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>      city           source_corr_type   target_country     target_geog_level 
#>  Length:15982       Length:15982       Length:15982       Length:15982      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_region      target_province    target_city        target_other  
#>  Length:15982       Length:15982       Length:15982       Mode:logical  
#>  Class :character   Class :character   Class :character   NA's:15982    
#>  Mode  :character   Mode  :character   Mode  :character                 
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_who_what    target_direction   travel_mechanism    compliance       
#>  Length:15982       Length:15982       Length:15982       Length:15982      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    enforcer             link              iso3c              iso2c          
#>  Length:15982       Length:15982       Length:15982       Length:15982      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>