CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 39
#>   record_id policy_id entry_type correct_type update_type update_level
#>   <chr>         <dbl> <chr>      <chr>        <chr>       <chr>       
#> 1 R_3nCudD…   2975738 new_entry  original     NA          NA          
#> 2 R_3OoySQ…   7024226 new_entry  original     NA          NA          
#> 3 R_3kLn6I…    889484 new_entry  original     NA          NA          
#> 4 R_1rAmwr…    129322 new_entry  original     NA          NA          
#> 5 R_21oIyC…   6721435 new_entry  correction   NA          NA          
#> 6 R_21oIyC…   6721435 new_entry  correction   NA          NA          
#> # … with 33 more variables: description <chr>, date_announced <date>,
#> #   date_start <date>, date_end <date>, country <chr>, iso3c <chr>,
#> #   iso2c <chr>, init_country_level <chr>, domestic_policy <dbl>,
#> #   province <chr>, city <chr>, type <chr>, type_sub_cat <chr>,
#> #   type_text <chr>, institution_status <chr>, target_country <chr>,
#> #   target_geog_level <chr>, target_region <chr>, target_province <chr>,
#> #   target_city <chr>, target_other <lgl>, target_who_what <chr>,
#> #   target_direction <chr>, travel_mechanism <chr>, compliance <chr>,
#> #   enforcer <chr>, index_high_est <dbl>, index_med_est <dbl>,
#> #   index_low_est <dbl>, index_country_rank <dbl>, link <chr>,
#> #   date_updated <date>, recorded_date <dttm>

colnames(res)

#>  [1] "record_id"          "policy_id"          "entry_type"        
#>  [4] "correct_type"       "update_type"        "update_level"      
#>  [7] "description"        "date_announced"     "date_start"        
#> [10] "date_end"           "country"            "iso3c"             
#> [13] "iso2c"              "init_country_level" "domestic_policy"   
#> [16] "province"           "city"               "type"              
#> [19] "type_sub_cat"       "type_text"          "institution_status"
#> [22] "target_country"     "target_geog_level"  "target_region"     
#> [25] "target_province"    "target_city"        "target_other"      
#> [28] "target_who_what"    "target_direction"   "travel_mechanism"  
#> [31] "compliance"         "enforcer"           "index_high_est"    
#> [34] "index_med_est"      "index_low_est"      "index_country_rank"
#> [37] "link"               "date_updated"       "recorded_date"     

dplyr::glimpse(res)

#> Rows: 51,060
#> Columns: 39
#> $ record_id          <chr> "R_3nCudDknlhUIjpONA", "R_3OoySQxTamo89oXNA", "R_3…
#> $ policy_id          <dbl> 2975738, 7024226, 889484, 129322, 6721435, 6721435…
#> $ entry_type         <chr> "new_entry", "new_entry", "new_entry", "new_entry"…
#> $ correct_type       <chr> "original", "original", "original", "original", "c…
#> $ update_type        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "End of Po…
#> $ update_level       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "Strengthe…
#> $ description        <chr> "March 6, Afghanistan \"Measures have been taken i…
#> $ date_announced     <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-28, 2…
#> $ date_start         <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-28, 2…
#> $ date_end           <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2020-09-0…
#> $ country            <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afgh…
#> $ iso3c              <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "…
#> $ iso2c              <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", "A…
#> $ init_country_level <chr> "National", "National", "National", "National", "M…
#> $ domestic_policy    <dbl> 0, 0, 1, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0,…
#> $ province           <chr> NA, NA, NA, NA, "Herat", "Herat", "Herat", "Samang…
#> $ city               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type               <chr> "Anti-Disinformation Measures", "Anti-Disinformati…
#> $ type_sub_cat       <chr> NA, NA, NA, NA, "Preschool or childcare facilities…
#> $ type_text          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ institution_status <chr> NA, NA, NA, NA, NA, NA, NA, "Primary Schools close…
#> $ target_country     <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afgh…
#> $ target_geog_level  <chr> NA, NA, "One or more geographical or administrativ…
#> $ target_region      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province    <chr> NA, NA, "Balkh; Kabul; Nangarhar; Kunar; Parwan; K…
#> $ target_city        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_other       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what    <chr> NA, NA, "All Residents (Citizen Residents + Foreig…
#> $ target_direction   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ travel_mechanism   <chr> NA, NA, "Not Applicable", NA, NA, NA, NA, NA, NA, …
#> $ compliance         <chr> "Voluntary/Recommended but No Penalties", "Mandato…
#> $ enforcer           <chr> "Ministry/Department of Health,Other (Please speci…
#> $ index_high_est     <dbl> 53.02945, 53.86454, NA, NA, 53.66498, 53.66498, 53…
#> $ index_med_est      <dbl> 50.60497, 52.36461, NA, NA, 50.91633, 50.91633, 50…
#> $ index_low_est      <dbl> 48.28019, 50.98443, NA, NA, 48.83211, 48.83211, 48…
#> $ index_country_rank <dbl> 137, 141, NA, NA, 137, 137, 137, 146, 146, 146, 14…
#> $ link               <chr> "https://www.etilaatroz.com/94246/fears-rumors-and…
#> $ date_updated       <date> 2020-08-21, 2020-09-13, 2020-09-12, 2020-09-05, 2…
#> $ recorded_date      <dttm> 2020-08-21 18:32:29, 2020-09-13 18:56:27, 2020-09…

summary(res)

#>   record_id           policy_id          entry_type        correct_type      
#>  Length:51060       Min.   :0.000e+00   Length:51060       Length:51060      
#>  Class :character   1st Qu.:2.465e+06   Class :character   Class :character  
#>  Mode  :character   Median :4.987e+06   Mode  :character   Mode  :character  
#>                     Mean   :1.231e+09                                        
#>                     3rd Qu.:7.424e+06                                        
#>                     Max.   :6.258e+13                                        
#>                                                                              
#>  update_type        update_level       description        date_announced      
#>  Length:51060       Length:51060       Length:51060       Min.   :2019-12-31  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-20  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-04-11  
#>                                                           Mean   :2020-05-01  
#>                                                           3rd Qu.:2020-06-01  
#>                                                           Max.   :2020-12-30  
#>                                                           NA's   :14          
#>    date_start            date_end            country         
#>  Min.   :2019-12-31   Min.   :2020-01-10   Length:51060      
#>  1st Qu.:2020-03-21   1st Qu.:2020-04-30   Class :character  
#>  Median :2020-04-14   Median :2020-06-02   Mode  :character  
#>  Mean   :2020-05-03   Mean   :2020-06-28                     
#>  3rd Qu.:2020-06-05   3rd Qu.:2020-08-18                     
#>  Max.   :2024-05-29   Max.   :2024-05-29                     
#>                       NA's   :26323                          
#>     iso3c              iso2c           init_country_level domestic_policy 
#>  Length:51060       Length:51060       Length:51060       Min.   :0.0000  
#>  Class :character   Class :character   Class :character   1st Qu.:0.0000  
#>  Mode  :character   Mode  :character   Mode  :character   Median :0.0000  
#>                                                           Mean   :0.4005  
#>                                                           3rd Qu.:1.0000  
#>                                                           Max.   :1.0000  
#>                                                                           
#>    province             city               type           type_sub_cat      
#>  Length:51060       Length:51060       Length:51060       Length:51060      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>   type_text         institution_status target_country     target_geog_level 
#>  Length:51060       Length:51060       Length:51060       Length:51060      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_region      target_province    target_city        target_other  
#>  Length:51060       Length:51060       Length:51060       Mode:logical  
#>  Class :character   Class :character   Class :character   NA's:51060    
#>  Mode  :character   Mode  :character   Mode  :character                 
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_who_what    target_direction   travel_mechanism    compliance       
#>  Length:51060       Length:51060       Length:51060       Length:51060      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    enforcer         index_high_est  index_med_est   index_low_est  
#>  Length:51060       Min.   :48.29   Min.   :45.42   Min.   :37.03  
#>  Class :character   1st Qu.:52.66   1st Qu.:51.45   1st Qu.:50.01  
#>  Mode  :character   Median :54.00   Median :52.66   Median :51.32  
#>                     Mean   :54.38   Mean   :52.59   Mean   :50.86  
#>                     3rd Qu.:55.54   3rd Qu.:53.85   3rd Qu.:52.50  
#>                     Max.   :70.74   Max.   :57.59   Max.   :56.26  
#>                     NA's   :21258   NA's   :21258   NA's   :21258  
#>  index_country_rank     link            date_updated       
#>  Min.   :  1        Length:51060       Min.   :2020-03-28  
#>  1st Qu.: 76        Class :character   1st Qu.:2020-05-17  
#>  Median :135        Mode  :character   Median :2020-07-24  
#>  Mean   :126                           Mean   :2020-07-30  
#>  3rd Qu.:179                           3rd Qu.:2020-09-29  
#>  Max.   :194                           Max.   :2020-12-28  
#>  NA's   :21258                                             
#>  recorded_date                
#>  Min.   :2020-03-28 11:14:33  
#>  1st Qu.:2020-05-17 18:26:00  
#>  Median :2020-07-24 09:00:14  
#>  Mean   :2020-07-30 17:17:12  
#>  3rd Qu.:2020-09-29 15:56:06  
#>  Max.   :2020-12-28 23:58:20  
#>