CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
#> Warning: One or more parsing issues, see `problems()` for details
head(res)
#> # A tibble: 6 × 65
#>   record_id       policy_id entry_type update_type update_level update_level_var
#>   <chr>           <chr>     <chr>      <chr>       <chr>        <chr>           
#> 1 R_3FKlACY6Djr1… 2975738   new_entry  NA          NA           NA              
#> 2 R_2arWTemRbMLK… 7024226   new_entry  NA          NA           NA              
#> 3 R_1reUk0mSCc6q… 1850843   new_entry  NA          NA           NA              
#> 4 R_pMJfXPNXIYRV… 889484    new_entry  NA          NA           NA              
#> 5 R_1eJC2cHJL6SX… 769460    new_entry  NA          NA           NA              
#> 6 R_2y9GMTXGfnZh… 129322    new_entry  NA          NA           NA              
#> # … with 59 more variables: description <chr>, date_announced <date>,
#> #   date_start <date>, date_end <date>, date_end_spec <chr>, country <chr>,
#> #   iso3c <chr>, iso2c <chr>, init_country_level <chr>, domestic_policy <dbl>,
#> #   province <chr>, ISO_L2 <chr>, city <chr>, type <chr>, type_sub_cat <chr>,
#> #   type_new_admin_coop <chr>, type_vac_cat <chr>, type_vac_mix <chr>,
#> #   type_vac_reg <chr>, type_vac_purchase <chr>, type_vac_group <chr>,
#> #   type_vac_group_rank <dbl>, type_vac_who_pays <chr>, …
colnames(res)
#>  [1] "record_id"               "policy_id"              
#>  [3] "entry_type"              "update_type"            
#>  [5] "update_level"            "update_level_var"       
#>  [7] "description"             "date_announced"         
#>  [9] "date_start"              "date_end"               
#> [11] "date_end_spec"           "country"                
#> [13] "iso3c"                   "iso2c"                  
#> [15] "init_country_level"      "domestic_policy"        
#> [17] "province"                "ISO_L2"                 
#> [19] "city"                    "type"                   
#> [21] "type_sub_cat"            "type_new_admin_coop"    
#> [23] "type_vac_cat"            "type_vac_mix"           
#> [25] "type_vac_reg"            "type_vac_purchase"      
#> [27] "type_vac_group"          "type_vac_group_rank"    
#> [29] "type_vac_who_pays"       "type_vac_dist_admin"    
#> [31] "type_vac_loc"            "type_vac_cost_num"      
#> [33] "type_vac_cost_scale"     "type_vac_cost_unit"     
#> [35] "type_vac_cost_gov_perc"  "type_vac_amt_num"       
#> [37] "type_vac_amt_scale"      "type_vac_amt_unit"      
#> [39] "type_vac_amt_gov_perc"   "type_text"              
#> [41] "institution_cat"         "institution_status"     
#> [43] "institution_conditions"  "target_init_same"       
#> [45] "target_country"          "target_geog_level"      
#> [47] "target_region"           "target_province"        
#> [49] "target_city"             "target_intl_org"        
#> [51] "target_other"            "target_who_what"        
#> [53] "target_who_gen"          "target_direction"       
#> [55] "travel_mechanism"        "compliance"             
#> [57] "enforcer"                "dist_index_high_est"    
#> [59] "dist_index_med_est"      "dist_index_low_est"     
#> [61] "dist_index_country_rank" "pdf_link"               
#> [63] "link"                    "date_updated"           
#> [65] "recorded_date"          
dplyr::glimpse(res)
#> Rows: 105,519
#> Columns: 65
#> $ record_id               <chr> "R_3FKlACY6Djr1sjcNA", "R_2arWTemRbMLKdk0NA", …
#> $ policy_id               <chr> "2975738", "7024226", "1850843", "889484", "76…
#> $ entry_type              <chr> "new_entry", "new_entry", "new_entry", "new_en…
#> $ update_type             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ update_level            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ update_level_var        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ description             <chr> "March 6, Afghanistan \"Measures have been tak…
#> $ date_announced          <date> 2020-03-06, 2020-03-16, 2020-05-14, 2020-05-2…
#> $ date_start              <date> 2020-03-06, 2020-03-16, 2020-05-14, 2020-05-2…
#> $ date_end                <date> 2022-02-24, 2020-03-16, 2020-05-15, NA, 2020-…
#> $ date_end_spec           <chr> "The policy has an unlimited time span (e.g. r…
#> $ country                 <chr> "Afghanistan", "Afghanistan", "Afghanistan", "…
#> $ iso3c                   <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG…
#> $ iso2c                   <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF"…
#> $ init_country_level      <chr> "National", "National", "National", "National"…
#> $ domestic_policy         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ province                <chr> NA, NA, NA, NA, NA, NA, "Herat", "Herat", "Her…
#> $ ISO_L2                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ city                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type                    <chr> "Anti-Disinformation Measures", "Anti-Disinfor…
#> $ type_sub_cat            <chr> NA, NA, NA, NA, NA, NA, "Preschool or childcar…
#> $ type_new_admin_coop     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cat            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_mix            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_reg            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_purchase       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_group          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_group_rank     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_who_pays       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_dist_admin     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_loc            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_num       <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_scale     <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_unit      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_gov_perc  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_num        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_scale      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_unit       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_gov_perc   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_text               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ institution_cat         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ institution_status      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, "Primary S…
#> $ institution_conditions  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_init_same        <chr> "Yes", "Yes", "Yes", "No", "Yes", "Yes", NA, N…
#> $ target_country          <chr> "Afghanistan", "Afghanistan", "Afghanistan", "…
#> $ target_geog_level       <chr> "One or more countries, but not all countries"…
#> $ target_region           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province         <chr> NA, NA, NA, "Balkh; Kabul; Nangarhar; Kunar; P…
#> $ target_city             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_intl_org         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_other            <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what         <chr> "All Residents (Citizen Residents + Foreign Re…
#> $ target_who_gen          <chr> "No special population targeted", "No special …
#> $ target_direction        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ travel_mechanism        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ compliance              <chr> "Voluntary/Recommended but No Penalties", "Man…
#> $ enforcer                <chr> "Ministry/Department of Health,Other (Please s…
#> $ dist_index_high_est     <dbl> 52.91608, 52.68646, 84.67092, 85.83168, 85.831…
#> $ dist_index_med_est      <dbl> 44.30324, 45.55852, 77.76448, 79.82725, 79.827…
#> $ dist_index_low_est      <dbl> 34.85850, 36.72021, 70.08431, 73.13763, 73.137…
#> $ dist_index_country_rank <dbl> 123, 60, 125, 131, 131, 135, 121, 121, 121, 10…
#> $ pdf_link                <chr> "https://tummgmt.eu.qualtrics.com/Q/File.php?F…
#> $ link                    <chr> "https://www.etilaatroz.com/94246/fears-rumors…
#> $ date_updated            <date> 2021-08-15, 2021-08-15, 2021-12-11, 2021-08-1…
#> $ recorded_date           <dttm> 2021-08-15 16:53:01, 2021-08-15 16:57:33, 202…
summary(res)
#>   record_id          policy_id          entry_type        update_type       
#>  Length:105519      Length:105519      Length:105519      Length:105519     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  update_level       update_level_var   description        date_announced      
#>  Length:105519      Length:105519      Length:105519      Min.   :2019-12-31  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-28  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-05-17  
#>                                                           Mean   :2020-07-08  
#>                                                           3rd Qu.:2020-09-10  
#>                                                           Max.   :2022-12-28  
#>                                                           NA's   :9           
#>    date_start            date_end          date_end_spec     
#>  Min.   :2019-12-31   Min.   :2020-01-01   Length:105519     
#>  1st Qu.:2020-03-29   1st Qu.:2020-04-30   Class :character  
#>  Median :2020-05-18   Median :2020-07-06   Mode  :character  
#>  Mean   :2020-07-10   Mean   :2020-09-15                     
#>  3rd Qu.:2020-09-14   3rd Qu.:2020-12-06                     
#>  Max.   :2024-05-29   Max.   :2027-07-01                     
#>                       NA's   :44102                          
#>    country             iso3c              iso2c           init_country_level
#>  Length:105519      Length:105519      Length:105519      Length:105519     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  domestic_policy    province            ISO_L2              city          
#>  Min.   :0.0000   Length:105519      Length:105519      Length:105519     
#>  1st Qu.:0.0000   Class :character   Class :character   Class :character  
#>  Median :1.0000   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :0.7098                                                           
#>  3rd Qu.:1.0000                                                           
#>  Max.   :1.0000                                                           
#>                                                                           
#>      type           type_sub_cat       type_new_admin_coop type_vac_cat      
#>  Length:105519      Length:105519      Length:105519       Length:105519     
#>  Class :character   Class :character   Class :character    Class :character  
#>  Mode  :character   Mode  :character   Mode  :character    Mode  :character  
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>  type_vac_mix       type_vac_reg       type_vac_purchase  type_vac_group    
#>  Length:105519      Length:105519      Length:105519      Length:105519     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  type_vac_group_rank type_vac_who_pays  type_vac_dist_admin type_vac_loc      
#>  Min.   :1.00        Length:105519      Length:105519       Length:105519     
#>  1st Qu.:1.00        Class :character   Class :character    Class :character  
#>  Median :1.00        Mode  :character   Mode  :character    Mode  :character  
#>  Mean   :1.94                                                                 
#>  3rd Qu.:2.00                                                                 
#>  Max.   :9.00                                                                 
#>  NA's   :105185                                                               
#>  type_vac_cost_num  type_vac_cost_scale type_vac_cost_unit
#>  Min.   :0.00e+00   Mode:logical        Mode:logical      
#>  1st Qu.:3.00e+00   NA's:105519         NA's:105519       
#>  Median :1.00e+02                                         
#>  Mean   :1.91e+06                                         
#>  3rd Qu.:2.00e+02                                         
#>  Max.   :1.10e+08                                         
#>  NA's   :105461                                           
#>  type_vac_cost_gov_perc type_vac_amt_num   type_vac_amt_scale
#>  Min.   :  0.00         Min.   :     0.1   Length:105519     
#>  1st Qu.:100.00         1st Qu.:     7.0   Class :character  
#>  Median :100.00         Median :    44.7   Mode  :character  
#>  Mean   : 90.77         Mean   :  5818.3                     
#>  3rd Qu.:100.00         3rd Qu.:   350.7                     
#>  Max.   :100.00         Max.   :940800.0                     
#>  NA's   :105417         NA's   :105115                       
#>  type_vac_amt_unit  type_vac_amt_gov_perc  type_text         institution_cat   
#>  Length:105519      Min.   :  0.00        Length:105519      Length:105519     
#>  Class :character   1st Qu.:100.00        Class :character   Class :character  
#>  Mode  :character   Median :100.00        Mode  :character   Mode  :character  
#>                     Mean   : 93.28                                             
#>                     3rd Qu.:100.00                                             
#>                     Max.   :100.00                                             
#>                     NA's   :105266                                             
#>  institution_status institution_conditions target_init_same  
#>  Length:105519      Length:105519          Length:105519     
#>  Class :character   Class :character       Class :character  
#>  Mode  :character   Mode  :character       Mode  :character  
#>                                                              
#>                                                              
#>                                                              
#>                                                              
#>  target_country     target_geog_level  target_region      target_province   
#>  Length:105519      Length:105519      Length:105519      Length:105519     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_city        target_intl_org    target_other   target_who_what   
#>  Length:105519      Length:105519      Mode:logical   Length:105519     
#>  Class :character   Class :character   NA's:105519    Class :character  
#>  Mode  :character   Mode  :character                  Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_who_gen     target_direction   travel_mechanism    compliance       
#>  Length:105519      Length:105519      Length:105519      Length:105519     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    enforcer         dist_index_high_est dist_index_med_est dist_index_low_est
#>  Length:105519      Min.   :13.30       Min.   : 9.893     Min.   : 3.146    
#>  Class :character   1st Qu.:58.64       1st Qu.:55.567     1st Qu.:52.387    
#>  Mode  :character   Median :68.19       Median :65.017     Median :61.868    
#>                     Mean   :66.86       Mean   :63.593     Mean   :60.229    
#>                     3rd Qu.:77.11       3rd Qu.:73.685     3rd Qu.:70.132    
#>                     Max.   :98.75       Max.   :98.298     Max.   :97.683    
#>                     NA's   :19229       NA's   :19229      NA's   :19229     
#>  dist_index_country_rank   pdf_link             link          
#>  Min.   :  1.00          Length:105519      Length:105519     
#>  1st Qu.: 60.00          Class :character   Class :character  
#>  Median :100.00          Mode  :character   Mode  :character  
#>  Mean   : 90.62                                               
#>  3rd Qu.:126.00                                               
#>  Max.   :139.00                                               
#>  NA's   :19229                                                
#>   date_updated        recorded_date                
#>  Min.   :2020-03-28   Min.   :2020-03-28 06:37:36  
#>  1st Qu.:2020-11-04   1st Qu.:2020-11-04 15:48:42  
#>  Median :2021-04-22   Median :2021-04-22 06:01:55  
#>  Mean   :2021-03-20   Mean   :2021-03-21 10:44:50  
#>  3rd Qu.:2021-08-09   3rd Qu.:2021-08-09 12:02:19  
#>  Max.   :2022-02-19   Max.   :2022-02-19 20:44:27  
#>