CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 × 65
#>   record_id       policy_id entry_type update_type update_level update_level_var
#>   <chr>           <chr>     <chr>      <chr>       <chr>        <chr>           
#> 1 R_3nCudDknlhUI… 2975738   new_entry  NA          NA           NA              
#> 2 R_3OoySQxTamo8… 7024226   new_entry  NA          NA           NA              
#> 3 R_3kLn6IU4rQGH… 889484    new_entry  NA          NA           NA              
#> 4 R_1rAmwrNMSZPg… 129322    new_entry  NA          NA           NA              
#> 5 R_21oIyCpRWpNI… 6721435Ea new_entry  NA          NA           NA              
#> 6 R_21oIyCpRWpNI… 6721435Ec new_entry  NA          NA           NA              
#> # … with 59 more variables: description <chr>, date_announced <date>,
#> #   date_start <date>, date_end <date>, date_end_spec <chr>, country <chr>,
#> #   iso3c <chr>, iso2c <chr>, init_country_level <chr>, domestic_policy <dbl>,
#> #   province <chr>, ISO_L2 <chr>, city <chr>, type <chr>, type_sub_cat <chr>,
#> #   type_new_admin_coop <chr>, type_vac_cat <chr>, type_vac_mix <chr>,
#> #   type_vac_reg <chr>, type_vac_purchase <chr>, type_vac_group <chr>,
#> #   type_vac_group_rank <dbl>, type_vac_who_pays <chr>,
#> #   type_vac_dist_admin <chr>, type_vac_loc <chr>, type_vac_cost_num <chr>,
#> #   type_vac_cost_scale <chr>, type_vac_cost_unit <chr>,
#> #   type_vac_cost_gov_perc <dbl>, type_vac_amt_num <dbl>,
#> #   type_vac_amt_scale <chr>, type_vac_amt_unit <chr>,
#> #   type_vac_amt_gov_perc <dbl>, type_text <chr>, institution_cat <chr>,
#> #   institution_status <chr>, institution_conditions <chr>,
#> #   target_init_same <chr>, target_country <chr>, target_geog_level <chr>,
#> #   target_region <chr>, target_province <chr>, target_city <chr>,
#> #   target_intl_org <chr>, target_other <lgl>, target_who_what <chr>,
#> #   target_who_gen <chr>, target_direction <chr>, travel_mechanism <chr>,
#> #   compliance <chr>, enforcer <chr>, dist_index_high_est <dbl>,
#> #   dist_index_med_est <dbl>, dist_index_low_est <dbl>,
#> #   dist_index_country_rank <dbl>, pdf_link <chr>, link <chr>,
#> #   date_updated <date>, recorded_date <dttm>

colnames(res)

#>  [1] "record_id"               "policy_id"              
#>  [3] "entry_type"              "update_type"            
#>  [5] "update_level"            "update_level_var"       
#>  [7] "description"             "date_announced"         
#>  [9] "date_start"              "date_end"               
#> [11] "date_end_spec"           "country"                
#> [13] "iso3c"                   "iso2c"                  
#> [15] "init_country_level"      "domestic_policy"        
#> [17] "province"                "ISO_L2"                 
#> [19] "city"                    "type"                   
#> [21] "type_sub_cat"            "type_new_admin_coop"    
#> [23] "type_vac_cat"            "type_vac_mix"           
#> [25] "type_vac_reg"            "type_vac_purchase"      
#> [27] "type_vac_group"          "type_vac_group_rank"    
#> [29] "type_vac_who_pays"       "type_vac_dist_admin"    
#> [31] "type_vac_loc"            "type_vac_cost_num"      
#> [33] "type_vac_cost_scale"     "type_vac_cost_unit"     
#> [35] "type_vac_cost_gov_perc"  "type_vac_amt_num"       
#> [37] "type_vac_amt_scale"      "type_vac_amt_unit"      
#> [39] "type_vac_amt_gov_perc"   "type_text"              
#> [41] "institution_cat"         "institution_status"     
#> [43] "institution_conditions"  "target_init_same"       
#> [45] "target_country"          "target_geog_level"      
#> [47] "target_region"           "target_province"        
#> [49] "target_city"             "target_intl_org"        
#> [51] "target_other"            "target_who_what"        
#> [53] "target_who_gen"          "target_direction"       
#> [55] "travel_mechanism"        "compliance"             
#> [57] "enforcer"                "dist_index_high_est"    
#> [59] "dist_index_med_est"      "dist_index_low_est"     
#> [61] "dist_index_country_rank" "pdf_link"               
#> [63] "link"                    "date_updated"           
#> [65] "recorded_date"          

dplyr::glimpse(res)

#> Rows: 82,652
#> Columns: 65
#> $ record_id               <chr> "R_3nCudDknlhUIjpONA", "R_3OoySQxTamo89oXNA", …
#> $ policy_id               <chr> "2975738", "7024226", "889484", "129322", "672…
#> $ entry_type              <chr> "new_entry", "new_entry", "new_entry", "new_en…
#> $ update_type             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ update_level            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ update_level_var        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ description             <chr> "March 6, Afghanistan \"Measures have been tak…
#> $ date_announced          <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-2…
#> $ date_start              <date> 2020-03-06, 2020-03-16, 2020-05-28, 2020-06-2…
#> $ date_end                <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2020-…
#> $ date_end_spec           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "The p…
#> $ country                 <chr> "Afghanistan", "Afghanistan", "Afghanistan", "…
#> $ iso3c                   <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG…
#> $ iso2c                   <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF"…
#> $ init_country_level      <chr> "National", "National", "National", "National"…
#> $ domestic_policy         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ province                <chr> NA, NA, NA, NA, "Herat", "Herat", "Herat", "Sa…
#> $ ISO_L2                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ city                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type                    <chr> "Anti-Disinformation Measures", "Anti-Disinfor…
#> $ type_sub_cat            <chr> NA, NA, NA, NA, "Preschool or childcare facili…
#> $ type_new_admin_coop     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cat            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_mix            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_reg            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_purchase       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_group          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_group_rank     <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_who_pays       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_dist_admin     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_loc            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_num       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_scale     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_unit      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_cost_gov_perc  <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_num        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_scale      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_unit       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_vac_amt_gov_perc   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ type_text               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ institution_cat         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ institution_status      <chr> NA, NA, NA, NA, NA, NA, NA, "Primary Schools c…
#> $ institution_conditions  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_init_same        <chr> "Yes", "Yes", "No", "Yes", "Yes", "Yes", "Yes"…
#> $ target_country          <chr> "Afghanistan", "Afghanistan", "Afghanistan", "…
#> $ target_geog_level       <chr> "One or more countries, but not all countries"…
#> $ target_region           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province         <chr> NA, NA, "Balkh; Kabul; Nangarhar; Kunar; Parwa…
#> $ target_city             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_intl_org         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_other            <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what         <chr> "All Residents (Citizen Residents + Foreign Re…
#> $ target_who_gen          <chr> "No special population targeted", "No special …
#> $ target_direction        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ travel_mechanism        <chr> NA, NA, "Not Applicable", NA, NA, NA, NA, NA, …
#> $ compliance              <chr> "Voluntary/Recommended but No Penalties", "Man…
#> $ enforcer                <chr> "Ministry/Department of Health,Other (Please s…
#> $ dist_index_high_est     <dbl> 52.91608, 52.68646, 85.83168, 86.46325, 53.131…
#> $ dist_index_med_est      <dbl> 44.30324, 45.55852, 79.82725, 79.73033, 44.864…
#> $ dist_index_low_est      <dbl> 34.85850, 36.72021, 73.13763, 71.94793, 34.432…
#> $ dist_index_country_rank <dbl> 123, 60, 131, 135, 121, 121, 121, 106, 106, 10…
#> $ pdf_link                <chr> "https://tummgmt.eu.qualtrics.com/Q/File.php?F…
#> $ link                    <chr> "https://www.etilaatroz.com/94246/fears-rumors…
#> $ date_updated            <date> 2020-08-21, 2020-09-13, 2020-09-12, 2020-09-0…
#> $ recorded_date           <dttm> 2020-08-21 18:32:29, 2020-09-13 18:56:27, 202…

summary(res)

#> Warning: One or more parsing issues, see `problems()` for details

#>   record_id          policy_id          entry_type        update_type       
#>  Length:82652       Length:82652       Length:82652       Length:82652      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  update_level       update_level_var   description        date_announced      
#>  Length:82652       Length:82652       Length:82652       Min.   :2019-12-31  
#>  Class :character   Class :character   Class :character   1st Qu.:2020-03-25  
#>  Mode  :character   Mode  :character   Mode  :character   Median :2020-04-30  
#>                                                           Mean   :2020-06-06  
#>                                                           3rd Qu.:2020-07-13  
#>                                                           Max.   :2022-12-22  
#>                                                           NA's   :11          
#>    date_start            date_end          date_end_spec     
#>  Min.   :2019-12-31   Min.   :2020-01-06   Length:82652      
#>  1st Qu.:2020-03-25   1st Qu.:2020-04-28   Class :character  
#>  Median :2020-05-04   Median :2020-06-15   Mode  :character  
#>  Mean   :2020-06-08   Mean   :2020-08-07                     
#>  3rd Qu.:2020-07-16   3rd Qu.:2020-10-11                     
#>  Max.   :2024-05-29   Max.   :2024-05-29                     
#>                       NA's   :36214                          
#>    country             iso3c              iso2c           init_country_level
#>  Length:82652       Length:82652       Length:82652       Length:82652      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  domestic_policy    province            ISO_L2              city          
#>  Min.   :0.0000   Length:82652       Length:82652       Length:82652      
#>  1st Qu.:1.0000   Class :character   Class :character   Class :character  
#>  Median :1.0000   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :0.9133                                                           
#>  3rd Qu.:1.0000                                                           
#>  Max.   :1.0000                                                           
#>  NA's   :135                                                              
#>      type           type_sub_cat       type_new_admin_coop type_vac_cat      
#>  Length:82652       Length:82652       Length:82652        Length:82652      
#>  Class :character   Class :character   Class :character    Class :character  
#>  Mode  :character   Mode  :character   Mode  :character    Mode  :character  
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>  type_vac_mix       type_vac_reg       type_vac_purchase  type_vac_group    
#>  Length:82652       Length:82652       Length:82652       Length:82652      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  type_vac_group_rank type_vac_who_pays  type_vac_dist_admin type_vac_loc      
#>  Min.   :1.00        Length:82652       Length:82652        Length:82652      
#>  1st Qu.:1.00        Class :character   Class :character    Class :character  
#>  Median :1.00        Mode  :character   Mode  :character    Mode  :character  
#>  Mean   :1.68                                                                 
#>  3rd Qu.:2.00                                                                 
#>  Max.   :9.00                                                                 
#>  NA's   :82492                                                                
#>  type_vac_cost_num  type_vac_cost_scale type_vac_cost_unit
#>  Length:82652       Length:82652        Length:82652      
#>  Class :character   Class :character    Class :character  
#>  Mode  :character   Mode  :character    Mode  :character  
#>                                                           
#>                                                           
#>                                                           
#>                                                           
#>  type_vac_cost_gov_perc type_vac_amt_num   type_vac_amt_scale
#>  Min.   :  0.00         Min.   :     0.1   Length:82652      
#>  1st Qu.:100.00         1st Qu.:     8.2   Class :character  
#>  Median :100.00         Median :    41.0   Mode  :character  
#>  Mean   : 90.16         Mean   :  7079.1                     
#>  3rd Qu.:100.00         3rd Qu.:   176.6                     
#>  Max.   :100.00         Max.   :940800.0                     
#>  NA's   :82597          NA's   :82448                        
#>  type_vac_amt_unit  type_vac_amt_gov_perc  type_text         institution_cat   
#>  Length:82652       Min.   :  0.00        Length:82652       Length:82652      
#>  Class :character   1st Qu.:100.00        Class :character   Class :character  
#>  Mode  :character   Median :100.00        Mode  :character   Mode  :character  
#>                     Mean   : 94.03                                             
#>                     3rd Qu.:100.00                                             
#>                     Max.   :100.00                                             
#>                     NA's   :82518                                              
#>  institution_status institution_conditions target_init_same  
#>  Length:82652       Length:82652           Length:82652      
#>  Class :character   Class :character       Class :character  
#>  Mode  :character   Mode  :character       Mode  :character  
#>                                                              
#>                                                              
#>                                                              
#>                                                              
#>  target_country     target_geog_level  target_region      target_province   
#>  Length:82652       Length:82652       Length:82652       Length:82652      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_city        target_intl_org    target_other   target_who_what   
#>  Length:82652       Length:82652       Mode:logical   Length:82652      
#>  Class :character   Class :character   NA's:82652     Class :character  
#>  Mode  :character   Mode  :character                  Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  target_who_gen     target_direction   travel_mechanism    compliance       
#>  Length:82652       Length:82652       Length:82652       Length:82652      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    enforcer         dist_index_high_est dist_index_med_est dist_index_low_est
#>  Length:82652       Min.   :13.30       Min.   : 9.893     Min.   : 3.146    
#>  Class :character   1st Qu.:58.75       1st Qu.:55.606     1st Qu.:52.415    
#>  Mode  :character   Median :67.77       Median :64.546     Median :61.448    
#>                     Mean   :66.48       Mean   :63.205     Mean   :59.842    
#>                     3rd Qu.:76.22       3rd Qu.:72.868     3rd Qu.:69.389    
#>                     Max.   :98.75       Max.   :98.298     Max.   :97.683    
#>                     NA's   :9703        NA's   :9703       NA's   :9703      
#>  dist_index_country_rank   pdf_link             link          
#>  Min.   :  1.00          Length:82652       Length:82652      
#>  1st Qu.: 61.00          Class :character   Class :character  
#>  Median :100.00          Mode  :character   Mode  :character  
#>  Mean   : 89.82                                               
#>  3rd Qu.:123.00                                               
#>  Max.   :139.00                                               
#>  NA's   :9703                                                 
#>   date_updated        recorded_date                
#>  Min.   :2020-03-28   Min.   :2020-03-28 06:37:36  
#>  1st Qu.:2020-08-06   1st Qu.:2020-08-06 14:25:49  
#>  Median :2021-01-25   Median :2021-01-25 15:47:02  
#>  Mean   :2020-12-22   Mean   :2020-12-22 22:06:57  
#>  3rd Qu.:2021-04-30   3rd Qu.:2021-04-30 20:02:56  
#>  Max.   :2021-07-21   Max.   :2021-07-21 22:23:57  
#>