CoronaNet government policy response database

This dataset contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types.

coronanet_government_response_data()

Source

https://coronanet-project.org/download

Value

record_id Unique identifier for each policy record

entry_type Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description A short description of the policy change

type The category of the policy

country The country initiating the policy

init_country_level Whether the policy came from the national level or a sub-national unit

index_prov The ID of the sub-national unit

target_country Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_who_what Who the policy is targeted at

recorded_date When the record was entered into our data

target_direction Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism If a travel policy, what kind of transportation it affects

compliance Whether the policy is voluntary or mandatory

enforcer What unit in the country is responsible for enforcement

date_announced When the policy goes into effect

link A link to at least one source for the policy

ISO_A3 3-digit ISO country codes

ISO_A2 2-digit ISO country codes

severity_index_5perc 5% posterior low estimate (i.e. lower bound of uncertainty interval) for severity index

severity_index_median posterior median estimate (point estimate) for severity index, which comes from a Bayesian latent variable model aggregating across policy types to measure country-level policy severity (see paper on our website)

severity_index_5perc 95% posterior high estimate (i.e. upper bound of uncertainty interval) for severity index

References

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” SocArXiv. April 12. doi:10.31235/osf.io/dkvxy.

https://coronanet-project.org/working_papers

See also

Examples

res = coronanet_government_response_data()
head(res)

#> # A tibble: 6 x 36
#>   record_id policy_id recorded_date       date_updated date_announced date_start
#>   <chr>         <dbl> <dttm>              <date>       <date>         <date>    
#> 1 1000775Bc   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14     2020-03-14
#> 2 1000775Ay   1342691 2020-04-04 15:43:15 2020-04-10   2020-03-14     2020-03-14
#> 3 1003875NA   1003875 2020-04-15 11:04:50 2020-04-15   2020-04-07     2020-04-07
#> 4 1004681NA   2575122 2020-04-03 08:33:25 2020-04-10   2020-03-22     2020-03-25
#> 5 4866080NA   4866080 2020-03-30 05:21:44 2020-05-04   2020-03-20     2020-03-21
#> 6 1011795NA   4951252 2020-04-03 03:29:36 2020-04-03   2020-03-08     2020-03-08
#> # … with 30 more variables: date_end <date>, entry_type <chr>,
#> #   event_description <chr>, domestic_policy <dbl>, type <chr>,
#> #   type_sub_cat <chr>, type_text <dbl>, index_high_est <dbl>,
#> #   index_med_est <dbl>, index_low_est <dbl>, index_country_rank <dbl>,
#> #   country <chr>, init_country_level <chr>, province <chr>, city <chr>,
#> #   source_file_Type <chr>, target_country <chr>, target_geog_level <chr>,
#> #   target_region <chr>, target_province <chr>, target_city <chr>,
#> #   target_other <lgl>, target_who_what <chr>, target_direction <chr>,
#> #   travel_mechanism <chr>, compliance <chr>, enforcer <chr>, link <chr>,
#> #   iso3c <chr>, iso2c <chr>

colnames(res)

#>  [1] "record_id"          "policy_id"          "recorded_date"     
#>  [4] "date_updated"       "date_announced"     "date_start"        
#>  [7] "date_end"           "entry_type"         "event_description" 
#> [10] "domestic_policy"    "type"               "type_sub_cat"      
#> [13] "type_text"          "index_high_est"     "index_med_est"     
#> [16] "index_low_est"      "index_country_rank" "country"           
#> [19] "init_country_level" "province"           "city"              
#> [22] "source_file_Type"   "target_country"     "target_geog_level" 
#> [25] "target_region"      "target_province"    "target_city"       
#> [28] "target_other"       "target_who_what"    "target_direction"  
#> [31] "travel_mechanism"   "compliance"         "enforcer"          
#> [34] "link"               "iso3c"              "iso2c"             

dplyr::glimpse(res)

#> Rows: 12,063
#> Columns: 36
#> $ record_id          <chr> "1000775Bc", "1000775Ay", "1003875NA", "1004681NA"…
#> $ policy_id          <dbl> 1342691, 1342691, 1003875, 2575122, 4866080, 49512…
#> $ recorded_date      <dttm> 2020-04-04 15:43:15, 2020-04-04 15:43:15, 2020-04…
#> $ date_updated       <date> 2020-04-10, 2020-04-10, 2020-04-15, 2020-04-10, 2…
#> $ date_announced     <date> 2020-03-14, 2020-03-14, 2020-04-07, 2020-03-22, 2…
#> $ date_start         <date> 2020-03-14, 2020-03-14, 2020-04-07, 2020-03-25, 2…
#> $ date_end           <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
#> $ entry_type         <chr> "update", "update", "new_entry", "update", "correc…
#> $ event_description  <chr> "Chile extends mandatory quarantine for people fro…
#> $ domestic_policy    <dbl> 0, 0, NA, 0, NA, NA, 1, NA, NA, NA, 1, NA, 1, 1, 1…
#> $ type               <chr> "Quarantine/Lockdown", "Quarantine/Lockdown", "Oth…
#> $ type_sub_cat       <chr> "Self-Quarantine (i.e. quarantine at home)", "Quar…
#> $ type_text          <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ index_high_est     <dbl> 52.03633, 52.03633, 56.17884, 54.14313, 56.65772, …
#> $ index_med_est      <dbl> 51.22552, 51.22552, 54.83055, 53.08173, 53.66517, …
#> $ index_low_est      <dbl> 50.20661, 50.20661, 53.59422, 52.28079, 50.42090, …
#> $ index_country_rank <dbl> 113, 113, 181, 144, 168, 126, 148, 169, 30, 74, 11…
#> $ country            <chr> "Chile", "Chile", "Guatemala", "Kenya", "Italy", "…
#> $ init_country_level <chr> "National", "National", "National", "National", "N…
#> $ province           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ city               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ source_file_Type   <chr> "application/pdf", "application/pdf", "application…
#> $ target_country     <chr> "China,France,Germany,Iran,Italy,Japan,North Korea…
#> $ target_geog_level  <chr> "One or more countries, but not all countries", "O…
#> $ target_region      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_province    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_city        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_other       <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ target_who_what    <chr> "All Travelers (Citizen Travelers + Foreign Travel…
#> $ target_direction   <chr> "Inbound", "Inbound", NA, "Inbound", NA, NA, NA, N…
#> $ travel_mechanism   <chr> "All kinds of transport (all below except visa res…
#> $ compliance         <chr> "Mandatory with Fines", "Mandatory with Fines", "M…
#> $ enforcer           <chr> "National Government,Ministry/Department of Health…
#> $ link               <chr> "https://www.gob.cl/noticias/covid-19-ministerio-d…
#> $ iso3c              <chr> "CHL", "CHL", "GTM", "KEN", "ITA", "USA", "PER", "…
#> $ iso2c              <chr> "CL", "CL", "GT", "KE", "IT", "US", "PE", "IL", "G…

summary(res)

#>   record_id           policy_id         recorded_date                
#>  Length:12063       Min.   :5.036e+03   Min.   :2020-03-28 04:32:36  
#>  Class :character   1st Qu.:2.590e+06   1st Qu.:2020-04-02 15:18:39  
#>  Mode  :character   Median :5.135e+06   Median :2020-04-05 04:24:30  
#>                     Mean   :5.193e+09   Mean   :2020-04-07 23:04:52  
#>                     3rd Qu.:7.582e+06   3rd Qu.:2020-04-13 00:46:09  
#>                     Max.   :6.258e+13   Max.   :2020-04-24 04:38:20  
#>                                                                      
#>   date_updated        date_announced         date_start        
#>  Min.   :2020-03-28   Min.   :2019-12-31   Min.   :2019-12-31  
#>  1st Qu.:2020-04-10   1st Qu.:2020-03-11   1st Qu.:2020-03-12  
#>  Median :2020-05-04   Median :2020-03-18   Median :2020-03-19  
#>  Mean   :2020-04-26   Mean   :2020-03-15   Mean   :2020-03-17  
#>  3rd Qu.:2020-05-08   3rd Qu.:2020-03-29   3rd Qu.:2020-03-30  
#>  Max.   :2020-05-20   Max.   :2020-04-23   Max.   :2021-01-01  
#>                       NA's   :22                               
#>     date_end           entry_type        event_description  domestic_policy
#>  Min.   :2020-01-28   Length:12063       Length:12063       Min.   :0.00   
#>  1st Qu.:2020-04-03   Class :character   Class :character   1st Qu.:0.00   
#>  Median :2020-04-14   Mode  :character   Mode  :character   Median :1.00   
#>  Mean   :2020-04-17                                         Mean   :0.52   
#>  3rd Qu.:2020-04-26                                         3rd Qu.:1.00   
#>  Max.   :2022-03-13                                         Max.   :1.00   
#>  NA's   :8407                                               NA's   :6657   
#>      type           type_sub_cat         type_text     index_high_est 
#>  Length:12063       Length:12063       Min.   : 4.00   Min.   :48.29  
#>  Class :character   Class :character   1st Qu.:14.00   1st Qu.:52.32  
#>  Mode  :character   Mode  :character   Median :21.00   Median :53.34  
#>                                        Mean   :20.88   Mean   :53.52  
#>                                        3rd Qu.:30.00   3rd Qu.:54.54  
#>                                        Max.   :58.00   Max.   :65.91  
#>                                        NA's   :11963   NA's   :107    
#>  index_med_est   index_low_est   index_country_rank   country         
#>  Min.   :45.42   Min.   :38.00   Min.   :  1.0      Length:12063      
#>  1st Qu.:51.22   1st Qu.:50.14   1st Qu.: 82.0      Class :character  
#>  Median :52.20   Median :51.11   Median :132.0      Mode  :character  
#>  Mean   :52.15   Mean   :50.79   Mean   :126.9                        
#>  3rd Qu.:53.31   3rd Qu.:52.13   3rd Qu.:176.0                        
#>  Max.   :57.58   Max.   :56.26   Max.   :194.0                        
#>  NA's   :107     NA's   :107     NA's   :107                          
#>  init_country_level   province             city           source_file_Type  
#>  Length:12063       Length:12063       Length:12063       Length:12063      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_country     target_geog_level  target_region      target_province   
#>  Length:12063       Length:12063       Length:12063       Length:12063      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  target_city        target_other   target_who_what    target_direction  
#>  Length:12063       Mode:logical   Length:12063       Length:12063      
#>  Class :character   NA's:12063     Class :character   Class :character  
#>  Mode  :character                  Mode  :character   Mode  :character  
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  travel_mechanism    compliance          enforcer             link          
#>  Length:12063       Length:12063       Length:12063       Length:12063      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     iso3c              iso2c          
#>  Length:12063       Length:12063      
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>