US State-level data including test results

Get data from the COVID tracking project https://covidtracking.com/ including daily historical data on testing results, hospitalizations, and deaths.

covidtracker_data()

Value

A tidy tbl_df

Details

From the COVID Tracking Project website:

The COVID Tracking Project collects information from 50 US states, the District of Columbia, and 5 other US territories to provide the most comprehensive testing data we can collect for the novel coronavirus, SARS-CoV-2. We attempt to report positive and negative results, pending tests, and total people tested for each state or district currently reporting that data.

See also

Examples

res = covidtracker_data()
colnames(res)

#> [1] "date"         "state"        "positive"     "negative"     "pending"     
#> [6] "hospitalized" "death"        "dateChecked"  "fips"        

dim(res)

#> [1] 3321    9

dplyr::glimpse(res)

#> Rows: 3,321
#> Columns: 9
#> $ date         <date> 2020-05-03, 2020-05-03, 2020-05-03, 2020-05-03, 2020-05…
#> $ state        <chr> "AK", "AL", "AR", "AS", "AZ", "CA", "CO", "CT", "DC", "D…
#> $ positive     <dbl> 368, 7725, 3431, 0, 8640, 53616, 16225, 29287, 5016, 520…
#> $ negative     <dbl> 21210, 84775, 49459, 57, 72479, 662135, 63681, 73206, 18…
#> $ pending      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1226, NA, NA, NA…
#> $ hospitalized <dbl> NA, 1035, 427, NA, 1348, NA, 2793, 7758, NA, NA, 6244, 5…
#> $ death        <dbl> 9, 290, 76, 0, 362, 2215, 832, 2436, 251, 177, 1403, 117…
#> $ dateChecked  <dttm> 2020-05-03 20:00:00, 2020-05-03 20:00:00, 2020-05-03 20…
#> $ fips         <chr> "00002", "00001", "00005", "00060", "00004", "00006", "0…