United States Consumer Spending

Aggregated and anonymized purchase data from consumer credit and debit card spending. Spending is reported based on the ZIP code where the cardholder lives, not the ZIP code where transactions occurred.

econ_tracker_consumer_spending_city_data()

econ_tracker_consumer_spending_county_data()

econ_tracker_consumer_spending_state_data()

econ_tracker_consumer_spending_national_data()

Source

Affinity Solutions via Opportunity Insight econ tracker

Update Frequency: Weekly

Details

Data are Seasonally adjusted change since January 2020. Data is indexed in 2019 and 2020 as the change relative to the January index period. We then seasonally adjust by dividing year-over-year, which represents the difference between the change since January observed in 2020 compared to the change since January observed since 2019. We account for differences in the dates of federal holidays between 2019 and 2020 by shifting the 2019 reference data to align the holidays before performing the year-over-year division.

Geographies: National, State, County, Metro

Breakdowns:

By Industry. Industries are constructed by grouping merchant codes that are used by Affinity Solutions to identify the category of merchant and merchant activity.

Apparel and General Merchandise

Entertainment and Recreation

Grocery

Health Care

Resturants and Hotels

Transportation

By Consumer Zip Code Income. Transactions are linked to zip codes where the consumer lives and zip codes are classified into income categories based on measurements of median household income and population provided by the American Community Survey (2014 - 2018).

High Income (median household income greater than $78,000 per year)

Middle Income (median household income between $46,000 per year and $78,000 per year)

Low Income (median household income less than $46,000 per year)

Column descriptions

spend_all: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_acf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in accomodation and food service (ACF) MCCs, 7 day moving average, 7 day moving average.

spend_aer: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in arts, entertainment, and recreation (AER) MCCs, 7 day moving average.

spend_apg: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in general merchandise stores (GEN) and apparel and accessories (AAP) MCCs, 7 day moving average.

spend_grf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in grocery and food store (GRF) MCCs, 7 day moving average.

spend_hcs: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in health care and social assistance (HCS) MCCs, 7 day moving average.

spend_tws: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in transportation and warehousing (TWS) MCCs, 7 day moving average.

spend_all_inchigh: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with high (top quartile) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_incmiddle: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with middle (middle two quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_inclow: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with low (bottom quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_q2: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the second quartile (i.e. second lowest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_q3: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the third quartile (i.e. second highest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

Note

The raw data contains discontinuous breaks caused by entry or exit of credit card providers from the sample. In order to reliably identify and correct these breaks, we require at least 3 weeks of data. The most recent 3 weeks of data are therefore marked 'provisional' and are subject to non-negligible changes as new data is posted. For breaks found prior to the last 3 weeks, we correct for it using a method outlined in the paper. Otherwise we substitute the national mean for more recent breaks while we gather enough data to implement the corrections outlined in the paper. Additionally, at the county-level when are there more than one structural breaks the data is too noisy to correct for these breaks and counties with multiple breaks are dropped from the sample. Lastly, Affinity Solutions suppresses any cut of the data with fewer than five transactions. For more details refer to the accompanying paper.

References

https://opportunityinsights.org/wp-content/uploads/2020/05/tracker_paper.pdf

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = econ_tracker_consumer_spending_city_data()
res

#>        cityid    cityname stateabbrev  statename statefips   lat     lon
#>     1:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     2:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     3:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     4:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     5:      1 Los Angeles          CA California     00006 34.05 -118.24
#>    ---                                                                  
#> 31718:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 31719:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 31720:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 31721:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 31722:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#>        city_pop2019 freq spend_all spend_aap spend_acf spend_aer spend_apg
#>     1:     10039107    d        NA        NA        NA        NA        NA
#>     2:     10039107    d        NA        NA        NA        NA        NA
#>     3:     10039107    d        NA        NA        NA        NA        NA
#>     4:     10039107    d        NA        NA        NA        NA        NA
#>     5:     10039107    d        NA        NA        NA        NA        NA
#>    ---                                                                    
#> 31718:       651552    d    0.1190    0.0349     0.117   -0.3460    0.0969
#> 31719:       651552    d    0.0927    0.0705     0.156   -0.3790    0.1290
#> 31720:       651552    d    0.0878    0.0554     0.149   -0.3390    0.1410
#> 31721:       651552    d    0.0801    0.0190     0.161   -0.3570    0.1270
#> 31722:       651552    w    0.2150        NA     0.247   -0.0862    0.2780
#>        spend_durables spend_nondurables spend_grf spend_gen spend_hic spend_hcs
#>     1:             NA                NA        NA        NA        NA        NA
#>     2:             NA                NA        NA        NA        NA        NA
#>     3:             NA                NA        NA        NA        NA        NA
#>     4:             NA                NA        NA        NA        NA        NA
#>     5:             NA                NA        NA        NA        NA        NA
#>    ---                                                                         
#> 31718:          0.196           -0.0128    0.0966   -0.0649   0.04900        NA
#> 31719:          0.153           -0.0554    0.0716   -0.1170  -0.00473        NA
#> 31720:          0.110           -0.0425    0.0732   -0.0828  -0.04920        NA
#> 31721:          0.120            0.0484    0.0644    0.2150  -0.04100        NA
#> 31722:          0.259            0.1110    0.0597    0.3400   0.13300        NA
#>        spend_inpersonmisc spend_remoteservices spend_sgh spend_tws
#>     1:                 NA                   NA        NA        NA
#>     2:                 NA                   NA        NA        NA
#>     3:                 NA                   NA        NA        NA
#>     4:                 NA                   NA        NA        NA
#>     5:                 NA                   NA        NA        NA
#>    ---                                                            
#> 31718:             0.2310                0.312        NA        NA
#> 31719:             0.1840                0.180        NA        NA
#> 31720:             0.1560                0.173        NA        NA
#> 31721:             0.0577                0.102        NA        NA
#> 31722:             0.1850                0.372        NA        NA
#>        spend_retail_w_grocery spend_retail_no_grocery provisional       date
#>     1:                     NA                      NA           0 2020-01-01
#>     2:                     NA                      NA           0 2020-01-02
#>     3:                     NA                      NA           0 2020-01-03
#>     4:                     NA                      NA           0 2020-01-04
#>     5:                     NA                      NA           0 2020-01-05
#>    ---                                                                      
#> 31718:                 0.0791                0.074800           1 2021-09-09
#> 31719:                 0.0369                0.026000           1 2021-09-10
#> 31720:                 0.0193               -0.000175           1 2021-09-11
#> 31721:                 0.0698                0.073800           1 2021-09-12
#> 31722:                 0.1640                0.241000           1 2021-09-19

res = econ_tracker_consumer_spending_county_data()
res

#>          countyfips countyname cityid cityname    cz     czname statename
#>       1:      01001    Autauga     NA          11101 Montgomery   Alabama
#>       2:      01001    Autauga     NA          11101 Montgomery   Alabama
#>       3:      01001    Autauga     NA          11101 Montgomery   Alabama
#>       4:      01001    Autauga     NA          11101 Montgomery   Alabama
#>       5:      01001    Autauga     NA          11101 Montgomery   Alabama
#>      ---                                                                 
#> 1062846:      56045     Weston     NA          34601   Gillette   Wyoming
#> 1062847:      56045     Weston     NA          34601   Gillette   Wyoming
#> 1062848:      56045     Weston     NA          34601   Gillette   Wyoming
#> 1062849:      56045     Weston     NA          34601   Gillette   Wyoming
#> 1062850:      56045     Weston     NA          34601   Gillette   Wyoming
#>          statefips stateabbrev county_pop2019 freq spend_all provisional
#>       1:     00001          AL          55869    d        NA           0
#>       2:     00001          AL          55869    d        NA           0
#>       3:     00001          AL          55869    d        NA           0
#>       4:     00001          AL          55869    d        NA           0
#>       5:     00001          AL          55869    d        NA           0
#>      ---                                                                
#> 1062846:     00056          WY           6927    d    0.3200           0
#> 1062847:     00056          WY           6927    d    0.1750           0
#> 1062848:     00056          WY           6927    d    0.1730           0
#> 1062849:     00056          WY           6927    d    0.1700           0
#> 1062850:     00056          WY           6927    w    0.0904           0
#>                date
#>       1: 2020-01-01
#>       2: 2020-01-02
#>       3: 2020-01-03
#>       4: 2020-01-04
#>       5: 2020-01-05
#>      ---           
#> 1062846: 2021-09-09
#> 1062847: 2021-09-10
#> 1062848: 2021-09-11
#> 1062849: 2021-09-12
#> 1062850: 2021-09-19

res = econ_tracker_consumer_spending_state_data()
res

#>        statefips statename stateabbrev state_pop2019 freq spend_all spend_aap
#>     1:     00001   Alabama          AL       4903185    d        NA        NA
#>     2:     00001   Alabama          AL       4903185    d        NA        NA
#>     3:     00001   Alabama          AL       4903185    d        NA        NA
#>     4:     00001   Alabama          AL       4903185    d        NA        NA
#>     5:     00001   Alabama          AL       4903185    d        NA        NA
#>    ---                                                                       
#> 31718:     00056   Wyoming          WY        578759    d     0.105    0.0488
#> 31719:     00056   Wyoming          WY        578759    d     0.092    0.0134
#> 31720:     00056   Wyoming          WY        578759    d     0.109   -0.0891
#> 31721:     00056   Wyoming          WY        578759    d     0.101   -0.1000
#> 31722:     00056   Wyoming          WY        578759    w     0.153   -0.0792
#>        spend_acf spend_aer spend_apg spend_durables spend_nondurables spend_grf
#>     1:        NA        NA        NA             NA                NA        NA
#>     2:        NA        NA        NA             NA                NA        NA
#>     3:        NA        NA        NA             NA                NA        NA
#>     4:        NA        NA        NA             NA                NA        NA
#>     5:        NA        NA        NA             NA                NA        NA
#>    ---                                                                         
#> 31718:    0.0787  -0.14900     0.223          0.231            0.0660    0.0553
#> 31719:    0.0629  -0.12500     0.199          0.213            0.0644    0.0495
#> 31720:    0.0393  -0.00892     0.144          0.242            0.0597    0.0376
#> 31721:    0.0478  -0.10200     0.133          0.246            0.0523    0.0327
#> 31722:    0.1340  -0.00628     0.323          0.172            0.1430    0.0765
#>        spend_gen spend_hic spend_hcs spend_inpersonmisc spend_remoteservices
#>     1:        NA        NA        NA                 NA                   NA
#>     2:        NA        NA        NA                 NA                   NA
#>     3:        NA        NA        NA                 NA                   NA
#>     4:        NA        NA        NA                 NA                   NA
#>     5:        NA        NA        NA                 NA                   NA
#>    ---                                                                      
#> 31718:     0.328    0.3580        NA            -0.0125              -0.0996
#> 31719:     0.343    0.2830        NA             0.0326              -0.1410
#> 31720:     0.424    0.2770        NA             0.1250              -0.0890
#> 31721:     0.405    0.3050        NA             0.0892              -0.0591
#> 31722:     0.848    0.0659        NA             0.0503               0.1560
#>        spend_sgh spend_tws spend_retail_w_grocery spend_retail_no_grocery
#>     1:        NA        NA                     NA                      NA
#>     2:        NA        NA                     NA                      NA
#>     3:        NA        NA                     NA                      NA
#>     4:        NA        NA                     NA                      NA
#>     5:        NA        NA                     NA                      NA
#>    ---                                                                   
#> 31718:     0.207    0.0945                  0.129                   0.175
#> 31719:     0.271    0.0991                  0.124                   0.171
#> 31720:     0.234    0.1090                  0.128                   0.187
#> 31721:     0.161    0.0818                  0.128                   0.191
#> 31722:     0.223    0.0927                  0.149                   0.193
#>        spend_all_incmiddle spend_all_q1 spend_all_q2 spend_all_q3 spend_all_q4
#>     1:                  NA           NA           NA           NA           NA
#>     2:                  NA           NA           NA           NA           NA
#>     3:                  NA           NA           NA           NA           NA
#>     4:                  NA           NA           NA           NA           NA
#>     5:                  NA           NA           NA           NA           NA
#>    ---                                                                        
#> 31718:              0.0920        0.148        0.125       0.0663        0.155
#> 31719:              0.0879        0.089        0.120       0.0622        0.129
#> 31720:              0.0951        0.119        0.142       0.0587        0.193
#> 31721:              0.0838        0.107        0.118       0.0563        0.180
#> 31722:              0.1450        0.153        0.222       0.0860        0.184
#>        provisional       date
#>     1:           0 2020-01-01
#>     2:           0 2020-01-02
#>     3:           0 2020-01-03
#>     4:           0 2020-01-04
#>     5:           0 2020-01-05
#>    ---                       
#> 31718:           1 2021-09-09
#> 31719:           1 2021-09-10
#> 31720:           1 2021-09-11
#> 31721:           1 2021-09-12
#> 31722:           1 2021-09-19

res = econ_tracker_consumer_spending_national_data()
res

#>      freq spend_all spend_aap spend_acf spend_aer spend_apg spend_durables
#>   1:    d        NA        NA        NA        NA        NA             NA
#>   2:    d        NA        NA        NA        NA        NA             NA
#>   3:    d        NA        NA        NA        NA        NA             NA
#>   4:    d        NA        NA        NA        NA        NA             NA
#>   5:    d        NA        NA        NA        NA        NA             NA
#>  ---                                                                      
#> 977:    d     0.167     0.277     0.138   -0.0386     0.286          0.271
#> 978:    d     0.157     0.258     0.142   -0.0378     0.270          0.259
#> 979:    d     0.161     0.243     0.130   -0.0659     0.254          0.262
#> 980:    d     0.154     0.231     0.121   -0.0930     0.249          0.254
#> 981:    w     0.195     0.284     0.136   -0.0222     0.309          0.275
#>      spend_nondurables spend_grf spend_gen spend_hic spend_hcs
#>   1:                NA        NA        NA        NA        NA
#>   2:                NA        NA        NA        NA        NA
#>   3:                NA        NA        NA        NA        NA
#>   4:                NA        NA        NA        NA        NA
#>   5:                NA        NA        NA        NA        NA
#>  ---                                                          
#> 977:             0.226     0.199     0.285     0.308    0.1780
#> 978:             0.213     0.192     0.252     0.290    0.0906
#> 979:             0.211     0.177     0.281     0.264    0.1850
#> 980:             0.214     0.188     0.266     0.258    0.1830
#> 981:             0.242     0.230     0.333     0.306    0.1680
#>      spend_inpersonmisc spend_remoteservices spend_sgh spend_tws
#>   1:                 NA                   NA        NA        NA
#>   2:                 NA                   NA        NA        NA
#>   3:                 NA                   NA        NA        NA
#>   4:                 NA                   NA        NA        NA
#>   5:                 NA                   NA        NA        NA
#>  ---                                                            
#> 977:                 NA               0.1020     0.483   -0.0799
#> 978:                 NA               0.0587     0.455   -0.0869
#> 979:                 NA               0.0960     0.453   -0.0831
#> 980:                 NA               0.0890     0.436   -0.0787
#> 981:                 NA               0.1610     0.426   -0.0470
#>      spend_retail_w_grocery spend_retail_no_grocery spend_all_incmiddle
#>   1:                     NA                      NA                  NA
#>   2:                     NA                      NA                  NA
#>   3:                     NA                      NA                  NA
#>   4:                     NA                      NA                  NA
#>   5:                     NA                      NA                  NA
#>  ---                                                                   
#> 977:                  0.259                   0.286               0.168
#> 978:                  0.244                   0.268               0.158
#> 979:                  0.242                   0.273               0.162
#> 980:                  0.240                   0.264               0.156
#> 981:                  0.273                   0.293               0.188
#>      spend_all_q1 spend_all_q2 spend_all_q3 spend_all_q4 spend_19_aap
#>   1:           NA           NA           NA           NA           NA
#>   2:           NA           NA           NA           NA           NA
#>   3:           NA           NA           NA           NA           NA
#>   4:           NA           NA           NA           NA           NA
#>   5:           NA           NA           NA           NA           NA
#>  ---                                                                 
#> 977:        0.194        0.153        0.180        0.150        0.572
#> 978:        0.180        0.142        0.172        0.141        0.559
#> 979:        0.185        0.146        0.175        0.143        0.540
#> 980:        0.176        0.142        0.168        0.138        0.522
#> 981:        0.226        0.176        0.198        0.187        0.561
#>      spend_19_acf spend_19_aer spend_19_all spend_19_apg spend_19_durables
#>   1:           NA           NA           NA           NA                NA
#>   2:           NA           NA           NA           NA                NA
#>   3:           NA           NA           NA           NA                NA
#>   4:           NA           NA           NA           NA                NA
#>   5:           NA           NA           NA           NA                NA
#>  ---                                                                      
#> 977:        0.423        0.194        0.378        0.586             0.542
#> 978:        0.416        0.218        0.360        0.572             0.529
#> 979:        0.418        0.214        0.347        0.558             0.513
#> 980:        0.416        0.181        0.333        0.547             0.491
#> 981:        0.444        0.271        0.380        0.591             0.503
#>      spend_19_gen spend_19_grf spend_19_hcs spend_19_hic spend_19_inpersonmisc
#>   1:           NA           NA           NA           NA                    NA
#>   2:           NA           NA           NA           NA                    NA
#>   3:           NA           NA           NA           NA                    NA
#>   4:           NA           NA           NA           NA                    NA
#>   5:           NA           NA           NA           NA                    NA
#>  ---                                                                          
#> 977:        0.623        0.363        0.204        0.813                 0.359
#> 978:        0.606        0.332        0.177        0.797                 0.310
#> 979:        0.587        0.318        0.145        0.775                 0.285
#> 980:        0.570        0.328        0.106        0.751                 0.262
#> 981:        0.619        0.325        0.315        0.691                 0.284
#>      spend_19_nondurables spend_19_remoteservices spend_19_sgh spend_19_tws
#>   1:                   NA                      NA           NA           NA
#>   2:                   NA                      NA           NA           NA
#>   3:                   NA                      NA           NA           NA
#>   4:                   NA                      NA           NA           NA
#>   5:                   NA                      NA           NA           NA
#>  ---                                                                       
#> 977:                0.419                   0.278        0.805       0.1160
#> 978:                0.393                   0.258        0.775       0.0896
#> 979:                0.377                   0.241        0.767       0.0874
#> 980:                0.375                   0.220        0.740       0.0974
#> 981:                0.393                   0.304        0.704       0.1340
#>      spend_19_all_q4 spend_19_all_q1 spend_19_all_incmiddle spend_19_all_q2
#>   1:              NA              NA                     NA              NA
#>   2:              NA              NA                     NA              NA
#>   3:              NA              NA                     NA              NA
#>   4:              NA              NA                     NA              NA
#>   5:              NA              NA                     NA              NA
#>  ---                                                                       
#> 977:           0.321           0.462                  0.390           0.392
#> 978:           0.304           0.439                  0.373           0.373
#> 979:           0.293           0.422                  0.359           0.359
#> 980:           0.280           0.407                  0.345           0.345
#> 981:           0.318           0.471                  0.394           0.398
#>      spend_19_all_q3 spend_aap_q1 spend_acf_q1 spend_aer_q1 spend_apg_q1
#>   1:              NA           NA           NA           NA           NA
#>   2:              NA           NA           NA           NA           NA
#>   3:              NA           NA           NA           NA           NA
#>   4:              NA           NA           NA           NA           NA
#>   5:              NA           NA           NA           NA           NA
#>  ---                                                                    
#> 977:           0.388        0.400        0.180           NA        0.381
#> 978:           0.373        0.387        0.181           NA        0.372
#> 979:           0.359        0.359        0.174           NA        0.349
#> 980:           0.345        0.336        0.164           NA        0.342
#> 981:           0.390        0.410        0.175           NA        0.411
#>      spend_durables_q1 spend_gen_q1 spend_grf_q1 spend_hcs_q1 spend_hic_q1
#>   1:                NA           NA           NA           NA           NA
#>   2:                NA           NA           NA           NA           NA
#>   3:                NA           NA           NA           NA           NA
#>   4:                NA           NA           NA           NA           NA
#>   5:                NA           NA           NA           NA           NA
#>  ---                                                                      
#> 977:             0.270        0.369        0.161       0.1210        0.308
#> 978:             0.252        0.358        0.151       0.0667        0.275
#> 979:             0.261        0.356        0.129       0.1710        0.263
#> 980:             0.258        0.346        0.137       0.1580        0.266
#> 981:             0.268        0.412        0.189       0.1360        0.265
#>      spend_inpersonmisc_q1 spend_nondurables_q1 spend_remoteservices_q1
#>   1:                    NA                   NA                      NA
#>   2:                    NA                   NA                      NA
#>   3:                    NA                   NA                      NA
#>   4:                    NA                   NA                      NA
#>   5:                    NA                   NA                      NA
#>  ---                                                                   
#> 977:                0.1370                0.228                  0.1220
#> 978:                0.0748                0.220                  0.0805
#> 979:                0.0961                0.203                  0.1230
#> 980:                0.0680                0.204                  0.0999
#> 981:                0.1780                0.245                  0.1920
#>      spend_sgh_q1 spend_tws_q1 spend_retail_no_grocery_q1
#>   1:           NA           NA                         NA
#>   2:           NA           NA                         NA
#>   3:           NA           NA                         NA
#>   4:           NA           NA                         NA
#>   5:           NA           NA                         NA
#>  ---                                                     
#> 977:        0.460     -0.02540                      0.317
#> 978:        0.441     -0.02690                      0.302
#> 979:        0.443     -0.02300                      0.304
#> 980:        0.400     -0.01720                      0.298
#> 981:        0.414      0.00255                      0.327
#>      spend_retail_w_grocery_q1 spend_aap_q2 spend_acf_q2 spend_aer_q2
#>   1:                        NA           NA           NA           NA
#>   2:                        NA           NA           NA           NA
#>   3:                        NA           NA           NA           NA
#>   4:                        NA           NA           NA           NA
#>   5:                        NA           NA           NA           NA
#>  ---                                                                 
#> 977:                     0.261        0.195        0.121     -0.01760
#> 978:                     0.248        0.166        0.128     -0.03360
#> 979:                     0.241        0.153        0.119     -0.05060
#> 980:                     0.240        0.151        0.116     -0.07780
#> 981:                     0.278        0.166        0.128     -0.00979
#>      spend_apg_q2 spend_durables_q2 spend_gen_q2 spend_grf_q2 spend_hcs_q2
#>   1:           NA                NA           NA           NA           NA
#>   2:           NA                NA           NA           NA           NA
#>   3:           NA                NA           NA           NA           NA
#>   4:           NA                NA           NA           NA           NA
#>   5:           NA                NA           NA           NA           NA
#>  ---                                                                      
#> 977:        0.262             0.232        0.302        0.160       0.1560
#> 978:        0.238             0.219        0.288        0.156       0.0598
#> 979:        0.227             0.223        0.285        0.138       0.1520
#> 980:        0.227             0.220        0.293        0.151       0.1540
#> 981:        0.274             0.235        0.372        0.183       0.1440
#>      spend_hic_q2 spend_inpersonmisc_q2 spend_nondurables_q2
#>   1:           NA                    NA                   NA
#>   2:           NA                    NA                   NA
#>   3:           NA                    NA                   NA
#>   4:           NA                    NA                   NA
#>   5:           NA                    NA                   NA
#>  ---                                                        
#> 977:        0.206                0.0621                0.183
#> 978:        0.202                0.0174                0.175
#> 979:        0.174                0.0395                0.162
#> 980:        0.183                0.0171                0.173
#> 981:        0.251                    NA                0.204
#>      spend_remoteservices_q2 spend_sgh_q2 spend_tws_q2
#>   1:                      NA           NA           NA
#>   2:                      NA           NA           NA
#>   3:                      NA           NA           NA
#>   4:                      NA           NA           NA
#>   5:                      NA           NA           NA
#>  ---                                                  
#> 977:                  0.0689        0.477      -0.0472
#> 978:                  0.0472        0.444      -0.0550
#> 979:                  0.0815        0.437      -0.0440
#> 980:                  0.0686        0.429      -0.0368
#> 981:                  0.1230        0.368      -0.0414
#>      spend_retail_no_grocery_q2 spend_retail_w_grocery_q2 spend_aap_q3
#>   1:                         NA                        NA           NA
#>   2:                         NA                        NA           NA
#>   3:                         NA                        NA           NA
#>   4:                         NA                        NA           NA
#>   5:                         NA                        NA           NA
#>  ---                                                                  
#> 977:                      0.243                     0.215        0.281
#> 978:                      0.230                     0.205        0.265
#> 979:                      0.225                     0.196        0.254
#> 980:                      0.224                     0.199        0.246
#> 981:                      0.259                     0.234        0.290
#>      spend_acf_q3 spend_aer_q3 spend_apg_q3 spend_durables_q3 spend_gen_q3
#>   1:           NA           NA           NA                NA           NA
#>   2:           NA           NA           NA                NA           NA
#>   3:           NA           NA           NA                NA           NA
#>   4:           NA           NA           NA                NA           NA
#>   5:           NA           NA           NA                NA           NA
#>  ---                                                                      
#> 977:        0.140      -0.0109        0.299             0.262        0.294
#> 978:        0.146       0.0030        0.284             0.266        0.269
#> 979:        0.132      -0.0390        0.267             0.266        0.284
#> 980:        0.123      -0.0728        0.265             0.254        0.283
#> 981:        0.148      -0.0217        0.323             0.266        0.354
#>      spend_grf_q3 spend_hcs_q3 spend_hic_q3 spend_inpersonmisc_q3
#>   1:           NA           NA           NA                    NA
#>   2:           NA           NA           NA                    NA
#>   3:           NA           NA           NA                    NA
#>   4:           NA           NA           NA                    NA
#>   5:           NA           NA           NA                    NA
#>  ---                                                             
#> 977:        0.206        0.175        0.273                0.0946
#> 978:        0.198        0.106        0.260                0.0503
#> 979:        0.181        0.202        0.218                0.0717
#> 980:        0.191        0.200        0.215                0.0506
#> 981:        0.233        0.175        0.265                0.0961
#>      spend_nondurables_q3 spend_remoteservices_q3 spend_sgh_q3 spend_tws_q3
#>   1:                   NA                      NA           NA           NA
#>   2:                   NA                      NA           NA           NA
#>   3:                   NA                      NA           NA           NA
#>   4:                   NA                      NA           NA           NA
#>   5:                   NA                      NA           NA           NA
#>  ---                                                                       
#> 977:                0.236                  0.1060        0.445      -0.0694
#> 978:                0.226                  0.0358        0.419      -0.0723
#> 979:                0.220                  0.0788        0.421      -0.0679
#> 980:                0.227                  0.0842        0.405      -0.0654
#> 981:                0.247                  0.1380        0.406      -0.0401
#>      spend_retail_no_grocery_q3 spend_retail_w_grocery_q3 spend_aap_q4
#>   1:                         NA                        NA           NA
#>   2:                         NA                        NA           NA
#>   3:                         NA                        NA           NA
#>   4:                         NA                        NA           NA
#>   5:                         NA                        NA           NA
#>  ---                                                                  
#> 977:                      0.293                     0.265        0.244
#> 978:                      0.280                     0.254        0.227
#> 979:                      0.279                     0.247        0.210
#> 980:                      0.272                     0.246        0.194
#> 981:                      0.291                     0.272        0.265
#>      spend_acf_q4 spend_aer_q4 spend_apg_q4 spend_durables_q4 spend_gen_q4
#>   1:           NA           NA           NA                NA           NA
#>   2:           NA           NA           NA                NA           NA
#>   3:           NA           NA           NA                NA           NA
#>   4:           NA           NA           NA                NA           NA
#>   5:           NA           NA           NA                NA           NA
#>  ---                                                                      
#> 977:       0.1180      -0.1030        0.224             0.280        0.213
#> 978:       0.1220      -0.0955        0.206             0.261        0.172
#> 979:       0.1100      -0.1210        0.191             0.261        0.175
#> 980:       0.0976      -0.1480        0.184             0.252        0.173
#> 981:       0.1080      -0.0607        0.245             0.288        0.223
#>      spend_grf_q4 spend_hcs_q4 spend_hic_q4 spend_inpersonmisc_q4
#>   1:           NA           NA           NA                    NA
#>   2:           NA           NA           NA                    NA
#>   3:           NA           NA           NA                    NA
#>   4:           NA           NA           NA                    NA
#>   5:           NA           NA           NA                    NA
#>  ---                                                             
#> 977:        0.239        0.188        0.389                0.0490
#> 978:        0.232        0.108        0.377                0.0249
#> 979:        0.224        0.170        0.348                0.0434
#> 980:        0.235        0.173        0.327                0.0307
#> 981:        0.283        0.168        0.378                0.0990
#>      spend_nondurables_q4 spend_remoteservices_q4 spend_sgh_q4 spend_tws_q4
#>   1:                   NA                      NA           NA           NA
#>   2:                   NA                      NA           NA           NA
#>   3:                   NA                      NA           NA           NA
#>   4:                   NA                      NA           NA           NA
#>   5:                   NA                      NA           NA           NA
#>  ---                                                                       
#> 977:                0.233                  0.0641        0.481      -0.1520
#> 978:                0.218                  0.0493        0.467      -0.1630
#> 979:                0.214                  0.0777        0.459      -0.1640
#> 980:                0.219                  0.0764        0.442      -0.1590
#> 981:                0.249                  0.1590        0.462      -0.0994
#>      spend_retail_no_grocery_q4 spend_retail_w_grocery_q4 spend_s_aap
#>   1:                         NA                        NA          NA
#>   2:                         NA                        NA          NA
#>   3:                         NA                        NA          NA
#>   4:                         NA                        NA          NA
#>   5:                         NA                        NA          NA
#>  ---                                                                 
#> 977:                      0.276                     0.265       0.531
#> 978:                      0.255                     0.248       0.519
#> 979:                      0.254                     0.245       0.501
#> 980:                      0.245                     0.242       0.483
#> 981:                      0.278                     0.280       0.521
#>      spend_s_acf spend_s_aer spend_s_all spend_s_apg spend_s_durables
#>   1:          NA          NA          NA          NA               NA
#>   2:          NA          NA          NA          NA               NA
#>   3:          NA          NA          NA          NA               NA
#>   4:          NA          NA          NA          NA               NA
#>   5:          NA          NA          NA          NA               NA
#>  ---                                                                 
#> 977:       0.315          NA       0.299       0.528            0.460
#> 978:       0.308          NA       0.282       0.515            0.447
#> 979:       0.310          NA       0.269       0.501            0.432
#> 980:       0.308          NA       0.256       0.490            0.411
#> 981:       0.334          NA       0.301       0.532            0.422
#>      spend_s_gen spend_s_grf spend_s_hcs spend_s_hic spend_s_inpersonmisc
#>   1:          NA          NA          NA          NA                   NA
#>   2:          NA          NA          NA          NA                   NA
#>   3:          NA          NA          NA          NA                   NA
#>   4:          NA          NA          NA          NA                   NA
#>   5:          NA          NA          NA          NA                   NA
#>  ---                                                                     
#> 977:       0.549       0.313      0.1160       0.761                0.292
#> 978:       0.532       0.283      0.0903       0.746                0.246
#> 979:       0.514       0.269      0.0604       0.724                0.222
#> 980:       0.497       0.279      0.0244       0.701                0.200
#> 981:       0.544       0.277          NA       0.642                0.222
#>      spend_s_nondurables spend_s_remoteservices spend_s_sgh spend_s_tws
#>   1:                  NA                     NA          NA          NA
#>   2:                  NA                     NA          NA          NA
#>   3:                  NA                     NA          NA          NA
#>   4:                  NA                     NA          NA          NA
#>   5:                  NA                     NA          NA          NA
#>  ---                                                                   
#> 977:               0.368                  0.187       0.762     0.03650
#> 978:               0.343                  0.169       0.732     0.01190
#> 979:               0.328                  0.153       0.724     0.00999
#> 980:               0.326                  0.134       0.698     0.01960
#> 981:               0.343                  0.211       0.663          NA
#>      spend_s_all_incmiddle spend_s_all_q1 spend_s_all_q2 spend_s_all_q3
#>   1:                    NA             NA             NA             NA
#>   2:                    NA             NA             NA             NA
#>   3:                    NA             NA             NA             NA
#>   4:                    NA             NA             NA             NA
#>   5:                    NA             NA             NA             NA
#>  ---                                                                   
#> 977:                 0.299          0.352          0.291          0.305
#> 978:                 0.283          0.332          0.274          0.291
#> 979:                 0.270          0.316          0.261          0.278
#> 980:                 0.257          0.302          0.248          0.264
#> 981:                 0.303          0.361          0.297          0.307
#>      spend_s_all_q4 spend_s_retail_no_grocery spend_s_retail_w_grocery
#>   1:             NA                        NA                       NA
#>   2:             NA                        NA                       NA
#>   3:             NA                        NA                       NA
#>   4:             NA                        NA                       NA
#>   5:             NA                        NA                       NA
#>  ---                                                                  
#> 977:          0.271                     0.480                    0.425
#> 978:          0.254                     0.465                    0.404
#> 979:          0.244                     0.449                    0.389
#> 980:          0.232                     0.428                    0.379
#> 981:          0.268                     0.441                    0.387
#>      spend_19_retail_no_grocery spend_19_retail_w_grocery provisional
#>   1:                         NA                        NA           0
#>   2:                         NA                        NA           0
#>   3:                         NA                        NA           0
#>   4:                         NA                        NA           0
#>   5:                         NA                        NA           0
#>  ---                                                                 
#> 977:                      0.551                     0.488           1
#> 978:                      0.535                     0.467           1
#> 979:                      0.518                     0.451           1
#> 980:                      0.497                     0.440           1
#> 981:                      0.510                     0.448           1
#>            date
#>   1: 2019-01-07
#>   2: 2019-01-08
#>   3: 2019-01-09
#>   4: 2019-01-10
#>   5: 2019-01-11
#>  ---           
#> 977: 2021-09-09
#> 978: 2021-09-10
#> 979: 2021-09-11
#> 980: 2021-09-12
#> 981: 2021-09-19