United States Consumer Spending

Aggregated and anonymized purchase data from consumer credit and debit card spending. Spending is reported based on the ZIP code where the cardholder lives, not the ZIP code where transactions occurred.

econ_tracker_consumer_spending_city_data()

econ_tracker_consumer_spending_county_data()

econ_tracker_consumer_spending_state_data()

econ_tracker_consumer_spending_national_data()

Source

Affinity Solutions via Opportunity Insight econ tracker

Update Frequency: Weekly

Details

Data are Seasonally adjusted change since January 2020. Data is indexed in 2019 and 2020 as the change relative to the January index period. We then seasonally adjust by dividing year-over-year, which represents the difference between the change since January observed in 2020 compared to the change since January observed since 2019. We account for differences in the dates of federal holidays between 2019 and 2020 by shifting the 2019 reference data to align the holidays before performing the year-over-year division.

Geographies: National, State, County, Metro

Breakdowns:

By Industry. Industries are constructed by grouping merchant codes that are used by Affinity Solutions to identify the category of merchant and merchant activity.

Apparel and General Merchandise

Entertainment and Recreation

Grocery

Health Care

Resturants and Hotels

Transportation

By Consumer Zip Code Income. Transactions are linked to zip codes where the consumer lives and zip codes are classified into income categories based on measurements of median household income and population provided by the American Community Survey (2014 - 2018).

High Income (median household income greater than $78,000 per year)

Middle Income (median household income between $46,000 per year and $78,000 per year)

Low Income (median household income less than $46,000 per year)

Column descriptions

spend_all: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_acf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in accomodation and food service (ACF) MCCs, 7 day moving average, 7 day moving average.

spend_aer: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in arts, entertainment, and recreation (AER) MCCs, 7 day moving average.

spend_apg: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in general merchandise stores (GEN) and apparel and accessories (AAP) MCCs, 7 day moving average.

spend_grf: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in grocery and food store (GRF) MCCs, 7 day moving average.

spend_hcs: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in health care and social assistance (HCS) MCCs, 7 day moving average.

spend_tws: Seasonally adjusted credit/debit card spending relative to January 4-31 2020 in transportation and warehousing (TWS) MCCs, 7 day moving average.

spend_all_inchigh: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with high (top quartile) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_incmiddle: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with middle (middle two quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_inclow: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes with low (bottom quartiles) median income, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_q2: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the second quartile (i.e. second lowest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

spend_all_q3: Seasonally adjusted credit/debit card spending by consumers living in ZIP codes in the third quartile (i.e. second highest) of median incomes, relative to January 4-31 2020 in all merchant category codes (MCC), 7 day moving average.

Note

The raw data contains discontinuous breaks caused by entry or exit of credit card providers from the sample. In order to reliably identify and correct these breaks, we require at least 3 weeks of data. The most recent 3 weeks of data are therefore marked 'provisional' and are subject to non-negligible changes as new data is posted. For breaks found prior to the last 3 weeks, we correct for it using a method outlined in the paper. Otherwise we substitute the national mean for more recent breaks while we gather enough data to implement the corrections outlined in the paper. Additionally, at the county-level when are there more than one structural breaks the data is too noisy to correct for these breaks and counties with multiple breaks are dropped from the sample. Lastly, Affinity Solutions suppresses any cut of the data with fewer than five transactions. For more details refer to the accompanying paper.

References

https://opportunityinsights.org/wp-content/uploads/2020/05/tracker_paper.pdf

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = econ_tracker_consumer_spending_city_data()
res

#>        cityid    cityname stateabbrev  statename statefips   lat     lon
#>     1:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     2:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     3:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     4:      1 Los Angeles          CA California     00006 34.05 -118.24
#>     5:      1 Los Angeles          CA California     00006 34.05 -118.24
#>    ---                                                                  
#> 24578:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 24579:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 24580:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 24581:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#> 24582:     53       Tulsa          OK   Oklahoma     00040 36.15  -95.99
#>        city_pop2019 freq spend_acf spend_aer spend_all spend_apg spend_grf
#>     1:     10039107    d        NA        NA        NA        NA        NA
#>     2:     10039107    d        NA        NA        NA        NA        NA
#>     3:     10039107    d        NA        NA        NA        NA        NA
#>     4:     10039107    d        NA        NA        NA        NA        NA
#>     5:     10039107    d        NA        NA        NA        NA        NA
#>    ---                                                                    
#> 24578:       651552    d     0.148    -0.208     0.142     0.101    0.0230
#> 24579:       651552    d     0.170    -0.213     0.157     0.098    0.0687
#> 24580:       651552    d     0.190    -0.310     0.150     0.118    0.1310
#> 24581:       651552    d     0.136    -0.366     0.157     0.214    0.1860
#> 24582:       651552    w     0.145    -0.325     0.139     0.214    0.0798
#>        spend_hcs spend_tws spend_retail_w_grocery spend_retail_no_grocery
#>     1:        NA        NA                     NA                      NA
#>     2:        NA        NA                     NA                      NA
#>     3:        NA        NA                     NA                      NA
#>     4:        NA        NA                     NA                      NA
#>     5:        NA        NA                     NA                      NA
#>    ---                                                                   
#> 24578:     0.208    -0.146                  0.103                   0.164
#> 24579:     0.321    -0.163                  0.132                   0.180
#> 24580:     0.326    -0.145                  0.160                   0.180
#> 24581:     0.342    -0.197                  0.212                   0.229
#> 24582:     0.207    -0.206                  0.213                   0.304
#>        provisional       date
#>     1:           0 2020-01-01
#>     2:           0 2020-01-02
#>     3:           0 2020-01-03
#>     4:           0 2020-01-04
#>     5:           0 2020-01-05
#>    ---                       
#> 24578:           1 2021-04-22
#> 24579:           1 2021-04-23
#> 24580:           1 2021-04-24
#> 24581:           1 2021-04-25
#> 24582:           1 2021-05-02

res = econ_tracker_consumer_spending_county_data()
res

#>         countyfips countyname cityid cityname    cz     czname statename
#>      1:      01001    Autauga     NA          11101 Montgomery   Alabama
#>      2:      01001    Autauga     NA          11101 Montgomery   Alabama
#>      3:      01001    Autauga     NA          11101 Montgomery   Alabama
#>      4:      01001    Autauga     NA          11101 Montgomery   Alabama
#>      5:      01001    Autauga     NA          11101 Montgomery   Alabama
#>     ---                                                                 
#> 822996:      56045     Weston     NA          34601   Gillette   Wyoming
#> 822997:      56045     Weston     NA          34601   Gillette   Wyoming
#> 822998:      56045     Weston     NA          34601   Gillette   Wyoming
#> 822999:      56045     Weston     NA          34601   Gillette   Wyoming
#> 823000:      56045     Weston     NA          34601   Gillette   Wyoming
#>         statefips stateabbrev county_pop2019 freq spend_all provisional
#>      1:     00001          AL          55869    d        NA           0
#>      2:     00001          AL          55869    d        NA           0
#>      3:     00001          AL          55869    d        NA           0
#>      4:     00001          AL          55869    d        NA           0
#>      5:     00001          AL          55869    d        NA           0
#>     ---                                                                
#> 822996:     00056          WY           6927    d    0.0358           0
#> 822997:     00056          WY           6927    d    0.1910           0
#> 822998:     00056          WY           6927    d    0.0991           0
#> 822999:     00056          WY           6927    d    0.0393           0
#> 823000:     00056          WY           6927    w   -0.0598           0
#>               date
#>      1: 2020-01-01
#>      2: 2020-01-02
#>      3: 2020-01-03
#>      4: 2020-01-04
#>      5: 2020-01-05
#>     ---           
#> 822996: 2021-04-22
#> 822997: 2021-04-23
#> 822998: 2021-04-24
#> 822999: 2021-04-25
#> 823000: 2021-05-02

res = econ_tracker_consumer_spending_state_data()
res

#>        statefips statename stateabbrev state_pop2019 freq spend_acf spend_aer
#>     1:     00001   Alabama          AL       4903185    d        NA        NA
#>     2:     00001   Alabama          AL       4903185    d        NA        NA
#>     3:     00001   Alabama          AL       4903185    d        NA        NA
#>     4:     00001   Alabama          AL       4903185    d        NA        NA
#>     5:     00001   Alabama          AL       4903185    d        NA        NA
#>    ---                                                                       
#> 24578:     00056   Wyoming          WY        578759    d    0.1240   -0.3680
#> 24579:     00056   Wyoming          WY        578759    d    0.1040   -0.3130
#> 24580:     00056   Wyoming          WY        578759    d    0.0866   -0.2360
#> 24581:     00056   Wyoming          WY        578759    d    0.0722   -0.0627
#> 24582:     00056   Wyoming          WY        578759    w    0.0790   -0.1910
#>        spend_all spend_apg spend_grf spend_hcs spend_tws spend_all_inchigh
#>     1:        NA        NA        NA        NA        NA                NA
#>     2:        NA        NA        NA        NA        NA                NA
#>     3:        NA        NA        NA        NA        NA                NA
#>     4:        NA        NA        NA        NA        NA                NA
#>     5:        NA        NA        NA        NA        NA                NA
#>    ---                                                                    
#> 24578:    0.1530     0.251    0.0311     0.144    -0.108           0.12000
#> 24579:    0.1180     0.202    0.0212     0.219    -0.127           0.03890
#> 24580:    0.1180     0.188    0.0328     0.178    -0.129           0.04510
#> 24581:    0.1340     0.206    0.0579     0.189    -0.118           0.07510
#> 24582:    0.0999     0.176    0.0610     0.163    -0.152           0.00693
#>        spend_all_inclow spend_all_incmiddle spend_retail_w_grocery
#>     1:               NA                  NA                     NA
#>     2:               NA                  NA                     NA
#>     3:               NA                  NA                     NA
#>     4:               NA                  NA                     NA
#>     5:               NA                  NA                     NA
#>    ---                                                            
#> 24578:            0.234               0.150                  0.201
#> 24579:            0.216               0.124                  0.166
#> 24580:            0.224               0.123                  0.167
#> 24581:            0.216               0.140                  0.182
#> 24582:            0.182               0.113                  0.157
#>        spend_retail_no_grocery provisional       date
#>     1:                      NA           0 2020-01-01
#>     2:                      NA           0 2020-01-02
#>     3:                      NA           0 2020-01-03
#>     4:                      NA           0 2020-01-04
#>     5:                      NA           0 2020-01-05
#>    ---                                               
#> 24578:                   0.319           1 2021-04-22
#> 24579:                   0.261           1 2021-04-23
#> 24580:                   0.252           1 2021-04-24
#> 24581:                   0.257           1 2021-04-25
#> 24582:                   0.217           1 2021-05-02

res = econ_tracker_consumer_spending_national_data()
res

#>      freq spend_acf spend_aer spend_all spend_apg spend_grf spend_hcs spend_tws
#>   1:    d        NA        NA        NA        NA        NA        NA        NA
#>   2:    d        NA        NA        NA        NA        NA        NA        NA
#>   3:    d        NA        NA        NA        NA        NA        NA        NA
#>   4:    d        NA        NA        NA        NA        NA        NA        NA
#>   5:    d        NA        NA        NA        NA        NA        NA        NA
#>  ---                                                                           
#> 837:    d  0.012300    -0.236    0.1380     0.178     0.139    0.1500    -0.247
#> 838:    d  0.006050    -0.319    0.1240     0.143     0.131    0.0944    -0.250
#> 839:    d -0.004150    -0.336    0.1160     0.137     0.151    0.0823    -0.260
#> 840:    d -0.005470    -0.351    0.1110     0.137     0.185    0.0745    -0.264
#> 841:    w -0.000587    -0.328    0.0959     0.173     0.170    0.0573    -0.246
#>      spend_all_inchigh spend_all_inclow spend_all_incmiddle spend_all_q2
#>   1:                NA               NA                  NA           NA
#>   2:                NA               NA                  NA           NA
#>   3:                NA               NA                  NA           NA
#>   4:                NA               NA                  NA           NA
#>   5:                NA               NA                  NA           NA
#>  ---                                                                    
#> 837:            0.0903            0.217              0.1420        0.151
#> 838:            0.0784            0.203              0.1260        0.136
#> 839:            0.0701            0.195              0.1180        0.125
#> 840:            0.0662            0.189              0.1140        0.121
#> 841:            0.0581            0.153              0.0997        0.102
#>      spend_all_q3 spend_retail_no_grocery spend_retail_w_grocery provisional
#>   1:           NA                      NA                     NA           0
#>   2:           NA                      NA                     NA           0
#>   3:           NA                      NA                     NA           0
#>   4:           NA                      NA                     NA           0
#>   5:           NA                      NA                     NA           0
#>  ---                                                                        
#> 837:       0.1350                   0.326                  0.264           1
#> 838:       0.1190                   0.292                  0.238           1
#> 839:       0.1120                   0.271                  0.232           1
#> 840:       0.1080                   0.259                  0.237           1
#> 841:       0.0976                   0.270                  0.239           1
#>            date
#>   1: 2019-01-07
#>   2: 2019-01-08
#>   3: 2019-01-09
#>   4: 2019-01-10
#>   5: 2019-01-11
#>  ---           
#> 837: 2021-04-22
#> 838: 2021-04-23
#> 839: 2021-04-24
#> 840: 2021-04-25
#> 841: 2021-05-02