Econ Tracker Unemployment Insurance Claims

Unemployment insurance claims data from the Department of Labor (national and state-level) and numerous individual state agencies (county-level).

econ_tracker_unemp_city_data()

econ_tracker_unemp_county_data()

econ_tracker_unemp_state_data()

econ_tracker_unemp_national_data()

Source

https://github.com/OpportunityInsights/EconomicTracker

Details

All datasets contain the following columns in addition to the location details. Dates specify the last day of the week ending for the reported data.

initclaims_rate_regular: Number of initial claims per 100 people in the 2019 labor force, Regular UI only

initclaims_count_regular: Count of initial claims, Regular UI only

initclaims_rate_pua: Number of initial claims per 100 people in the 2019 labor force, PUA (Pandemic Unemployment Assistance) only

initclaims_count_pua: Count of initial claims, PUA (Pandemic Unemployment Assistance) only

initclaims_rate_combined: Number of initial claims per 100 people in the 2019 labor force, combining Regular and PUA claims

initclaims_count_combined: Count of initial claims, combining Regular and PUA claims

contclaims_rate_regular: Number of continued claims per 100 people in the 2019 labor force, Regular UI only

contclaims_count_regular: Count of continued claims, Regular UI only

contclaims_rate_pua: Number of continued claims per 100 people in the 2019 labor force, PUA (Pandemic Unemployment Assistance) only

contclaims_count_pua: Count of continued claims, PUA (Pandemic Unemployment Assistance) only

contclaims_rate_peuc: Number of continued claims per 100 people in the 2019 labor force, PEUC (Pandemic Emergency Unemployment Compensation) only

contclaims_count_peuc: Count of continued claims, PEUC (Pandemic Emergency Unemployment Compensation) only

contclaims_rate_combined: Number of continued claims per 100 people in the 2019 labor force, combining Regular, PUA and PEUC claims

contclaims_count_combined: Count of continued claims, combining Regular, PUA and PEUC claims

References

https://github.com/OpportunityInsights/EconomicTracker

https://tracktherecovery.org/

See also

Examples

# City Level data

res = econ_tracker_unemp_city_data()
head(res)

#>    cityid initclaims_count_regular initclaims_rate_regular       date
#> 1:     49                      405                   0.159 2020-01-04
#> 2:      1                    13268                   0.259 2020-01-11
#> 3:      2                    22479                   0.562 2020-01-11
#> 4:      3                     4980                   0.192 2020-01-11
#> 5:      6                     3365                   0.212 2020-01-11
#> 6:      8                     2187                   0.193 2020-01-11

# County Level data

res = econ_tracker_unemp_county_data()
head(res)

#>    countyfips initclaims_count_regular initclaims_rate_regular       date
#> 1:      01001                       21                  0.0802 2020-03-14
#> 2:      01001                       99                  0.3780 2020-03-21
#> 3:      01001                      843                  3.2200 2020-03-28
#> 4:      01001                     1088                  4.1600 2020-04-04
#> 5:      01001                      812                  3.1000 2020-04-11
#> 6:      01001                      675                  2.5800 2020-04-18
#>    countyname cityid cityname    cz     czname statename statefips stateabbrev
#> 1:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#> 2:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#> 3:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#> 4:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#> 5:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#> 6:    Autauga     NA          11101 Montgomery   Alabama     00001          AL
#>    county_pop2019
#> 1:          55869
#> 2:          55869
#> 3:          55869
#> 4:          55869
#> 5:          55869
#> 6:          55869

# State Level data

res = econ_tracker_unemp_state_data()
head(res)

#>    statefips initclaims_count_regular contclaims_count_regular
#> 1:     00001                     4578                    18523
#> 2:     00001                     3629                    21143
#> 3:     00001                     2483                    17402
#> 4:     00001                     2129                    18390
#> 5:     00001                     2170                    17284
#> 6:     00001                     2176                    16745
#>    initclaims_count_pua contclaims_count_pua contclaims_count_peuc
#> 1:                    0                    0                     0
#> 2:                    0                    0                     0
#> 3:                    0                    0                     0
#> 4:                    0                    0                     0
#> 5:                    0                    0                     0
#> 6:                    0                    0                     0
#>    initclaims_count_combined contclaims_count_combined initclaims_rate_regular
#> 1:                      4578                     18523                  0.2040
#> 2:                      3629                     21143                  0.1620
#> 3:                      2483                     17402                  0.1110
#> 4:                      2129                     18390                  0.0950
#> 5:                      2170                     17284                  0.0968
#> 6:                      2176                     16745                  0.0971
#>    contclaims_rate_regular initclaims_rate_pua contclaims_rate_pua
#> 1:                   0.826                   0                   0
#> 2:                   0.943                   0                   0
#> 3:                   0.776                   0                   0
#> 4:                   0.820                   0                   0
#> 5:                   0.771                   0                   0
#> 6:                   0.747                   0                   0
#>    contclaims_rate_peuc initclaims_rate_combined contclaims_rate_combined
#> 1:                    0                   0.2040                    0.826
#> 2:                    0                   0.1620                    0.943
#> 3:                    0                   0.1110                    0.776
#> 4:                    0                   0.0950                    0.820
#> 5:                    0                   0.0968                    0.771
#> 6:                    0                   0.0971                    0.747
#>          date statename stateabbrev state_pop2019
#> 1: 2020-01-04   Alabama          AL       4903185
#> 2: 2020-01-11   Alabama          AL       4903185
#> 3: 2020-01-18   Alabama          AL       4903185
#> 4: 2020-01-25   Alabama          AL       4903185
#> 5: 2020-02-01   Alabama          AL       4903185
#> 6: 2020-02-08   Alabama          AL       4903185

# National Level data

res = econ_tracker_unemp_national_data()
head(res)

#>    initclaims_count_regular contclaims_count_regular initclaims_count_pua
#> 1:                   334431                  2128086                    0
#> 2:                   336873                  2234999                    0
#> 3:                   278345                  2119848                    0
#> 4:                   225929                  2059122                    0
#> 5:                   222475                  2130470                    0
#> 6:                   217544                  2064993                    0
#>    contclaims_count_pua contclaims_count_peuc initclaims_count_combined
#> 1:                    0                     0                    334431
#> 2:                    0                     0                    336873
#> 3:                    0                     0                    278345
#> 4:                    0                     0                    225929
#> 5:                    0                     0                    222475
#> 6:                    0                     0                    217544
#>    contclaims_count_combined initclaims_rate_regular contclaims_rate_regular
#> 1:                   2128086                   0.205                    1.30
#> 2:                   2234999                   0.207                    1.37
#> 3:                   2119848                   0.171                    1.30
#> 4:                   2059122                   0.139                    1.26
#> 5:                   2130470                   0.136                    1.31
#> 6:                   2064993                   0.133                    1.27
#>    initclaims_rate_pua contclaims_rate_pua contclaims_rate_peuc
#> 1:                   0                   0                    0
#> 2:                   0                   0                    0
#> 3:                   0                   0                    0
#> 4:                   0                   0                    0
#> 5:                   0                   0                    0
#> 6:                   0                   0                    0
#>    initclaims_rate_combined contclaims_rate_combined       date
#> 1:                    0.205                     1.30 2020-01-04
#> 2:                    0.207                     1.37 2020-01-11
#> 3:                    0.171                     1.30 2020-01-18
#> 4:                    0.139                     1.26 2020-01-25
#> 5:                    0.136                     1.31 2020-02-01
#> 6:                    0.133                     1.27 2020-02-08