Get enriched jhu dataset enriched with country-level metadata

The purpose of this function is to enrich the JHU dataset with additional country-level metadata and then get the data into a form that can be easily written to excel or used as a "wide-format" tabular dataset.

enriched_jhu_data(cols_to_remove = .cols_to_remove, warn = FALSE)

Arguments

cols_to_remove
a character vector of column names from country_metadata() to remove.

warn
logical(1) defaults to FALSE, sent to countrycode::countrycode to determine whether failed matches should be reported.

Value

A list of three data.frames named deaths, confirmed, and recovered.

Examples

res = enriched_jhu_data()
colnames(res)

#>  [1] "name"           "topLevelDomain" "alpha2Code"     "alpha3Code"    
#>  [5] "capital"        "region"         "subregion"      "population"    
#>  [9] "area"           "gini"           "borders"        "numericCode"   
#> [13] "cioc"           "ProvinceState"  "CountryRegion"  "Lat"           
#> [17] "Long"           "date"           "count"          "subset"        

library(dplyr)
head(res)

#> # A tibble: 6 x 20
#>   name  topLevelDomain alpha2Code alpha3Code capital region subregion population
#>   <chr> <list>         <chr>      <chr>      <chr>   <chr>  <chr>          <int>
#> 1 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 2 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 3 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 4 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 5 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> 6 Afgh… <chr [1]>      AF         AFG        Kabul   Asia   Southern…   27657145
#> # … with 12 more variables: area <dbl>, gini <dbl>, borders <list>,
#> #   numericCode <chr>, cioc <chr>, ProvinceState <chr>, CountryRegion <chr>,
#> #   Lat <dbl>, Long <dbl>, date <date>, count <dbl>, subset <chr>

dplyr::glimpse(res)

#> Rows: 162,495
#> Columns: 20
#> $ name           <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afghanis…
#> $ topLevelDomain <list> [".af", ".af", ".af", ".af", ".af", ".af", ".af", ".a…
#> $ alpha2Code     <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", …
#> $ alpha3Code     <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG"…
#> $ capital        <chr> "Kabul", "Kabul", "Kabul", "Kabul", "Kabul", "Kabul", …
#> $ region         <chr> "Asia", "Asia", "Asia", "Asia", "Asia", "Asia", "Asia"…
#> $ subregion      <chr> "Southern Asia", "Southern Asia", "Southern Asia", "So…
#> $ population     <int> 27657145, 27657145, 27657145, 27657145, 27657145, 2765…
#> $ area           <dbl> 652230, 652230, 652230, 652230, 652230, 652230, 652230…
#> $ gini           <dbl> 27.8, 27.8, 27.8, 27.8, 27.8, 27.8, 27.8, 27.8, 27.8, …
#> $ borders        <list> [<"IRN", "PAK", "TKM", "UZB", "TJK", "CHN">, <"IRN", …
#> $ numericCode    <chr> "004", "004", "004", "004", "004", "004", "004", "004"…
#> $ cioc           <chr> "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG", "AFG"…
#> $ ProvinceState  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ CountryRegion  <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afghanis…
#> $ Lat            <dbl> 33.93911, 33.93911, 33.93911, 33.93911, 33.93911, 33.9…
#> $ Long           <dbl> 67.70995, 67.70995, 67.70995, 67.70995, 67.70995, 67.7…
#> $ date           <date> 2020-01-22, 2020-01-23, 2020-01-24, 2020-01-25, 2020-…
#> $ count          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ subset         <chr> "confirmed", "confirmed", "confirmed", "confirmed", "c…