Google search trends for COVID-19 and illness-related terms

This aggregated, anonymized dataset shows trends in search patterns for symptoms and is intended to help researchers to better understand the impact of COVID-19. Public health experts indicated that trends in search patterns might be helpful in broadly understanding how COVID-19 impacts communities and even in detecting outbreaks earlier. You shouldn’t assume that the data is a recording of real-world clinical events, or use this data for medical diagnostic, prognostic, or treatment purposes.

google_search_trends_data()

Source

https://github.com/google-research/open-covid-19-data/tree/master/data/exports/search_trends_symptoms_dataset

Details

This data reflects the volume of Google searches for a broad set of symptoms, signs and health conditions. To keep things simple in this documentation, we will refer to all of these collectively as symptoms. The data covers hundreds of symptoms such as fever, difficulty breathing, and stress—based on the following:

a symptom’s prevalence in Google’s searches

data quality and privacy considerations

For each day, we count the searches mapped to each of these symptoms and organize the data by geographic region. The resulting dataset is a daily or weekly time series for each region showing the relative frequency of searches for each symptom.

A single search query can be mapped to more than one symptom. For example, we map a search for “acid reflux and coughing up mucus” to three symptoms: Cough, Gastroesophageal reflux disease, and Heartburn.

The dataset covers the recent period and we’ll gradually expand its range as part of regular updates. Each update will bring the coverage to within three days of the day of the update.

Although we are releasing the dataset in English, we count searches in other languages. In each supported country, we include the languages needed to cover the majority of symptom search queries. For example, in the United States we support Spanish and English.

The data represents a sample of our users and might not represent the exact behavior of a wider population.

Preserving privacy

For this dataset, we use differential privacy, which adds artificial noise to our datasets while enabling high quality results without identifying any individual person.

To further protect people’s privacy, we ensure that no personal information or individual search queries are included in the dataset, and we don’t link any search-based health inferences to an individual user. More information about the privacy methods used to generate the dataset can be found in this report.

How we process the data

The data shows the relative popularity of symptoms in searches within a geographical region.

When the daily volume of the data for a given region does not meet quality or privacy thresholds, we do the following:

Try to provide a given symptom at the weekly resolution. If we cannot meet our quality or privacy thresholds at the weekly resolution, we do not provide the data for the symptom in that region. As a result, for a given region, some symptoms will appear exclusively in the daily time series, some exclusively in the weekly time series, and some might not appear at all (if the search data for those symptoms is very scarce).

To normalize and scale the daily and the weekly time series (processed separately), we do the following for each region:

First, the algorithm counts the number of searches for each symptom in that region for that day/week. Next, we divide this count by the total number of Search users in the region for that day/week to calculate relative popularity (which can be interpreted as the probability that a user in this region will search for the given symptom on that day/week). We refer to this ratio as the normalized popularity of a symptom. We then find the maximum value of the normalized popularity across the entire published time range for that region, over all symptoms using the chosen time resolution (day/week). We scale this maximum value to 100. All the other values are mapped to proportionally smaller values (linear scaling) in the range 0-100. Finally, we store the scaling factor and use it to scale values (for the same region and time resolution) in subsequent releases. In future updates, when a symptom popularity exceeds the previously-observed maximum value (found in step 3), then the new scaled value will be larger than 100. In each region, we scale all the normalized (daily / weekly) popularities using the same scaling factor. In a single region, you can compare the relative popularity of two (or more) symptoms (at the same time resolution) over any time interval. However, you should not compare the values of symptom popularity across regions or time resolutions — the region and time resolution specific scalings make these comparisons meaningless.

Note

Use of this dataset implies acceptance of Google's terms of use https://policies.google.com/terms.

References

Google LLC "Google COVID-19 Search Trends symptoms dataset". http://goo.gle/covid19symptomdataset

Author

Sean Davis seandavi@gmail.com

Examples

res = google_search_trends_data()
dim(res)

#> [1] 25955532       10

head(res)

#>    open_covid_region_code country_region_code country_region sub_region_1
#> 1:                  US-AK                  US  United States       Alaska
#> 2:                  US-AK                  US  United States       Alaska
#> 3:                  US-AK                  US  United States       Alaska
#> 4:                  US-AK                  US  United States       Alaska
#> 5:                  US-AK                  US  United States       Alaska
#> 6:                  US-AK                  US  United States       Alaska
#>    sub_region_1_code sub_region_2 sub_region_2_code       date
#> 1:             US-AK           NA                NA 2017-08-21
#> 2:             US-AK           NA                NA 2017-08-22
#> 3:             US-AK           NA                NA 2017-08-23
#> 4:             US-AK           NA                NA 2017-08-24
#> 5:             US-AK           NA                NA 2017-08-25
#> 6:             US-AK           NA                NA 2017-08-26
#>                     variable value
#> 1: symptom:Abdominal obesity  1.43
#> 2: symptom:Abdominal obesity  1.55
#> 3: symptom:Abdominal obesity  1.76
#> 4: symptom:Abdominal obesity  2.04
#> 5: symptom:Abdominal obesity  1.40
#> 6: symptom:Abdominal obesity  1.76

dplyr::glimpse(res)

#> Rows: 25,955,532
#> Columns: 10
#> $ open_covid_region_code <chr> "US-AK", "US-AK", "US-AK", "US-AK", "US-AK", "…
#> $ country_region_code    <chr> "US", "US", "US", "US", "US", "US", "US", "US"…
#> $ country_region         <chr> "United States", "United States", "United Stat…
#> $ sub_region_1           <chr> "Alaska", "Alaska", "Alaska", "Alaska", "Alask…
#> $ sub_region_1_code      <chr> "US-AK", "US-AK", "US-AK", "US-AK", "US-AK", "…
#> $ sub_region_2           <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ sub_region_2_code      <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ date                   <date> 2017-08-21, 2017-08-22, 2017-08-23, 2017-08-2…
#> $ variable               <fct> symptom:Abdominal obesity, symptom:Abdominal o…
#> $ value                  <dbl> 1.43, 1.55, 1.76, 2.04, 1.40, 1.76, 1.62, 1.49…

table(res$country_region)

#> 
#> United States 
#>      25955532 

table(res$sub_region_1)

#> 
#>              Alabama               Alaska              Arizona 
#>               508932               508932               508932 
#>             Arkansas           California             Colorado 
#>               508932               508932               508932 
#>          Connecticut             Delaware District of Columbia 
#>               508932               508932               508932 
#>              Florida              Georgia               Hawaii 
#>               508932               508932               508932 
#>                Idaho             Illinois              Indiana 
#>               508932               508932               508932 
#>                 Iowa               Kansas             Kentucky 
#>               508932               508932               508932 
#>            Louisiana                Maine             Maryland 
#>               508932               508932               508932 
#>        Massachusetts             Michigan            Minnesota 
#>               508932               508932               508932 
#>          Mississippi             Missouri              Montana 
#>               508932               508932               508932 
#>             Nebraska               Nevada        New Hampshire 
#>               508932               508932               508932 
#>           New Jersey           New Mexico             New York 
#>               508932               508932               508932 
#>       North Carolina         North Dakota                 Ohio 
#>               508932               508932               508932 
#>             Oklahoma               Oregon         Pennsylvania 
#>               508932               508932               508932 
#>         Rhode Island       South Carolina         South Dakota 
#>               508932               508932               508932 
#>            Tennessee                Texas                 Utah 
#>               508932               508932               508932 
#>              Vermont             Virginia           Washington 
#>               508932               508932               508932 
#>        West Virginia            Wisconsin              Wyoming 
#>               508932               508932               508932 

if(require(timetk)) {

  plot_time_series(res[sub_region_1 %in% c('Maryland','Texas','New York',
                                           'California','Georgia','Arizona',
                                           'Illinois','New Jersey','Florida') & 
                         variable %in% c(
                           'symptom:Cough','symptom:Fever',
                           'symptom:Shortness of breath')][,year := lubridate::year(date)],
                   date,value,.smooth_period='month',
                   .color_var=sub_region_1,.facet_vars = 'variable')

}

#> Loading required package: timetk

#> Warning: there is no package called ‘timetk’