Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#> # A tibble: 6 x 36
#>   country iso3c date       `C1_School clos… C1_Flag `C2_Workplace c… C2_Flag
#>   <chr>   <chr> <date>                <dbl>   <dbl>            <dbl>   <dbl>
#> 1 Aruba   ABW   2020-01-01                0      NA                0      NA
#> 2 Aruba   ABW   2020-01-02                0      NA                0      NA
#> 3 Aruba   ABW   2020-01-03                0      NA                0      NA
#> 4 Aruba   ABW   2020-01-04                0      NA                0      NA
#> 5 Aruba   ABW   2020-01-05                0      NA                0      NA
#> 6 Aruba   ABW   2020-01-06                0      NA                0      NA
#> # … with 29 more variables: `C3_Cancel public events` <dbl>, C3_Flag <dbl>,
#> #   `C4_Restrictions on gatherings` <dbl>, C4_Flag <dbl>, `C5_Close public
#> #   transport` <dbl>, C5_Flag <dbl>, `C6_Stay at home requirements` <dbl>,
#> #   C6_Flag <dbl>, `C7_Restrictions on internal movement` <dbl>, C7_Flag <dbl>,
#> #   `C8_International travel controls` <dbl>, `E1_Income support` <dbl>,
#> #   E1_Flag <dbl>, `E2_Debt/contract relief` <dbl>, `E3_Fiscal measures` <dbl>,
#> #   `E4_International support` <dbl>, `H1_Public information campaigns` <dbl>,
#> #   H1_Flag <dbl>, `H2_Testing policy` <dbl>, `H3_Contact tracing` <dbl>,
#> #   `H4_Emergency investment in healthcare` <dbl>, `H5_Investment in
#> #   vaccines` <dbl>, M1_Wildcard <lgl>, ConfirmedCases <dbl>,
#> #   ConfirmedDeaths <dbl>, StringencyIndex <dbl>,
#> #   StringencyIndexForDisplay <dbl>, LegacyStringencyIndex <dbl>,
#> #   LegacyStringencyIndexForDisplay <dbl>

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "C1_School closing"                    
#>  [5] "C1_Flag"                              
#>  [6] "C2_Workplace closing"                 
#>  [7] "C2_Flag"                              
#>  [8] "C3_Cancel public events"              
#>  [9] "C3_Flag"                              
#> [10] "C4_Restrictions on gatherings"        
#> [11] "C4_Flag"                              
#> [12] "C5_Close public transport"            
#> [13] "C5_Flag"                              
#> [14] "C6_Stay at home requirements"         
#> [15] "C6_Flag"                              
#> [16] "C7_Restrictions on internal movement" 
#> [17] "C7_Flag"                              
#> [18] "C8_International travel controls"     
#> [19] "E1_Income support"                    
#> [20] "E1_Flag"                              
#> [21] "E2_Debt/contract relief"              
#> [22] "E3_Fiscal measures"                   
#> [23] "E4_International support"             
#> [24] "H1_Public information campaigns"      
#> [25] "H1_Flag"                              
#> [26] "H2_Testing policy"                    
#> [27] "H3_Contact tracing"                   
#> [28] "H4_Emergency investment in healthcare"
#> [29] "H5_Investment in vaccines"            
#> [30] "M1_Wildcard"                          
#> [31] "ConfirmedCases"                       
#> [32] "ConfirmedDeaths"                      
#> [33] "StringencyIndex"                      
#> [34] "StringencyIndexForDisplay"            
#> [35] "LegacyStringencyIndex"                
#> [36] "LegacyStringencyIndexForDisplay"      

dplyr::glimpse(res)

#> Rows: 23,209
#> Columns: 36
#> $ country                                 <chr> "Aruba", "Aruba", "Aruba", "A…
#> $ iso3c                                   <chr> "ABW", "ABW", "ABW", "ABW", "…
#> $ date                                    <date> 2020-01-01, 2020-01-02, 2020…
#> $ `C1_School closing`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C2_Workplace closing`                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C2_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C3_Cancel public events`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C3_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C4_Restrictions on gatherings`         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C4_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C5_Close public transport`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C5_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C6_Stay at home requirements`          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C6_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C7_Restrictions on internal movement`  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C7_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C8_International travel controls`      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E1_Income support`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ E1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `E2_Debt/contract relief`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E3_Fiscal measures`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E4_International support`              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H1_Public information campaigns`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ H1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `H2_Testing policy`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H3_Contact tracing`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H4_Emergency investment in healthcare` <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H5_Investment in vaccines`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ M1_Wildcard                             <lgl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedCases                          <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedDeaths                         <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ StringencyIndex                         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyIndexForDisplay               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ LegacyStringencyIndex                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ LegacyStringencyIndexForDisplay         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…

summary(res)

#>    country             iso3c                date            C1_School closing
#>  Length:23209       Length:23209       Min.   :2020-01-01   Min.   :0.000    
#>  Class :character   Class :character   1st Qu.:2020-02-04   1st Qu.:0.000    
#>  Mode  :character   Mode  :character   Median :2020-03-10   Median :0.000    
#>                                        Mean   :2020-03-09   Mean   :1.338    
#>                                        3rd Qu.:2020-04-14   3rd Qu.:3.000    
#>                                        Max.   :2020-05-20   Max.   :3.000    
#>                                                             NA's   :565      
#>     C1_Flag      C2_Workplace closing    C2_Flag      C3_Cancel public events
#>  Min.   :0.000   Min.   :0.0000       Min.   :0.000   Min.   :0.0000         
#>  1st Qu.:1.000   1st Qu.:0.0000       1st Qu.:1.000   1st Qu.:0.0000         
#>  Median :1.000   Median :0.0000       Median :1.000   Median :0.0000         
#>  Mean   :0.925   Mean   :0.9257       Mean   :0.828   Mean   :0.8999         
#>  3rd Qu.:1.000   3rd Qu.:2.0000       3rd Qu.:1.000   3rd Qu.:2.0000         
#>  Max.   :1.000   Max.   :3.0000       Max.   :1.000   Max.   :2.0000         
#>  NA's   :12840   NA's   :711          NA's   :14392   NA's   :702            
#>     C3_Flag      C4_Restrictions on gatherings    C4_Flag     
#>  Min.   :0.000   Min.   :0.000                 Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:0.000                 1st Qu.:1.000  
#>  Median :1.000   Median :0.000                 Median :1.000  
#>  Mean   :0.911   Mean   :1.383                 Mean   :0.906  
#>  3rd Qu.:1.000   3rd Qu.:3.000                 3rd Qu.:1.000  
#>  Max.   :1.000   Max.   :4.000                 Max.   :1.000  
#>  NA's   :12762   NA's   :707                   NA's   :14167  
#>  C5_Close public transport    C5_Flag      C6_Stay at home requirements
#>  Min.   :0.000             Min.   :0.000   Min.   :0.0000              
#>  1st Qu.:0.000             1st Qu.:1.000   1st Qu.:0.0000              
#>  Median :0.000             Median :1.000   Median :0.0000              
#>  Mean   :0.461             Mean   :0.776   Mean   :0.6709              
#>  3rd Qu.:1.000             3rd Qu.:1.000   3rd Qu.:2.0000              
#>  Max.   :2.000             Max.   :1.000   Max.   :3.0000              
#>  NA's   :748               NA's   :16419   NA's   :690                 
#>     C6_Flag      C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000   Min.   :0.0000                       Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:0.0000                       1st Qu.:0.000  
#>  Median :1.000   Median :0.0000                       Median :1.000  
#>  Mean   :0.801   Mean   :0.6796                       Mean   :0.714  
#>  3rd Qu.:1.000   3rd Qu.:2.0000                       3rd Qu.:1.000  
#>  Max.   :1.000   Max.   :2.0000                       Max.   :1.000  
#>  NA's   :15220   NA's   :734                          NA's   :14756  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.0000    Min.   :0.000  
#>  1st Qu.:0.000                    1st Qu.:0.0000    1st Qu.:0.000  
#>  Median :2.000                    Median :0.0000    Median :1.000  
#>  Mean   :1.887                    Mean   :0.3014    Mean   :0.547  
#>  3rd Qu.:4.000                    3rd Qu.:0.0000    3rd Qu.:1.000  
#>  Max.   :4.000                    Max.   :2.0000    Max.   :1.000  
#>  NA's   :708                      NA's   :830       NA's   :18293  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.0000          Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.0000          1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :0.0000          Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.3917          Mean   :3.979e+08   Mean   :3.812e+07       
#>  3rd Qu.:0.0000          3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.0000          Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :844             NA's   :1341        NA's   :1186            
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.0000   
#>  1st Qu.:0.000                   1st Qu.:1.000   1st Qu.:0.0000   
#>  Median :2.000                   Median :1.000   Median :1.0000   
#>  Mean   :1.211                   Mean   :0.982   Mean   :0.8192   
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:1.0000   
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.0000   
#>  NA's   :831                     NA's   :8704    NA's   :1002     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.0000     Min.   :0.000e+00                    
#>  1st Qu.:0.0000     1st Qu.:0.000e+00                    
#>  Median :0.0000     Median :0.000e+00                    
#>  Mean   :0.7367     Mean   :2.856e+07                    
#>  3rd Qu.:2.0000     3rd Qu.:0.000e+00                    
#>  Max.   :2.0000     Max.   :2.424e+11                    
#>  NA's   :1083       NA's   :1433                         
#>  H5_Investment in vaccines M1_Wildcard    ConfirmedCases    ConfirmedDeaths  
#>  Min.   :0.000e+00         Mode:logical   Min.   :      0   Min.   :    0.0  
#>  1st Qu.:0.000e+00         NA's:23209     1st Qu.:      4   1st Qu.:    0.0  
#>  Median :0.000e+00                        Median :    107   Median :    2.0  
#>  Mean   :2.648e+05                        Mean   :   9102   Mean   :  605.7  
#>  3rd Qu.:0.000e+00                        3rd Qu.:   1389   3rd Qu.:   30.0  
#>  Max.   :1.390e+09                        Max.   :1467884   Max.   :88754.0  
#>  NA's   :1295                             NA's   :8299      NA's   :8299     
#>  StringencyIndex  StringencyIndexForDisplay LegacyStringencyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   : 0.00        
#>  1st Qu.:  0.00   1st Qu.:  0.00            1st Qu.: 0.00        
#>  Median : 19.44   Median : 19.44            Median :22.86        
#>  Mean   : 37.82   Mean   : 38.54            Mean   :39.29        
#>  3rd Qu.: 78.96   3rd Qu.: 79.49            3rd Qu.:80.00        
#>  Max.   :100.00   Max.   :100.00            Max.   :97.14        
#>  NA's   :804      NA's   :366               NA's   :776          
#>  LegacyStringencyIndexForDisplay
#>  Min.   : 0.00                  
#>  1st Qu.: 0.00                  
#>  Median :22.86                  
#>  Mean   :40.00                  
#>  3rd Qu.:80.00                  
#>  Max.   :97.14                  
#>  NA's   :353