Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Value

a data.frame

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)
#>    country iso3c date RegionCode Jurisdiction       Date C1_School closing
#> 1:   Aruba   ABW                    NAT_TOTAL 2020-01-01                 0
#> 2:   Aruba   ABW                    NAT_TOTAL 2020-01-02                 0
#> 3:   Aruba   ABW                    NAT_TOTAL 2020-01-03                 0
#> 4:   Aruba   ABW                    NAT_TOTAL 2020-01-04                 0
#> 5:   Aruba   ABW                    NAT_TOTAL 2020-01-05                 0
#> 6:   Aruba   ABW                    NAT_TOTAL 2020-01-06                 0
#>    C1_Flag C2_Workplace closing C2_Flag C3_Cancel public events C3_Flag
#> 1:      NA                    0      NA                       0      NA
#> 2:      NA                    0      NA                       0      NA
#> 3:      NA                    0      NA                       0      NA
#> 4:      NA                    0      NA                       0      NA
#> 5:      NA                    0      NA                       0      NA
#> 6:      NA                    0      NA                       0      NA
#>    C4_Restrictions on gatherings C4_Flag C5_Close public transport C5_Flag
#> 1:                             0      NA                         0      NA
#> 2:                             0      NA                         0      NA
#> 3:                             0      NA                         0      NA
#> 4:                             0      NA                         0      NA
#> 5:                             0      NA                         0      NA
#> 6:                             0      NA                         0      NA
#>    C6_Stay at home requirements C6_Flag C7_Restrictions on internal movement
#> 1:                            0      NA                                    0
#> 2:                            0      NA                                    0
#> 3:                            0      NA                                    0
#> 4:                            0      NA                                    0
#> 5:                            0      NA                                    0
#> 6:                            0      NA                                    0
#>    C7_Flag C8_International travel controls E1_Income support E1_Flag
#> 1:      NA                                0                 0      NA
#> 2:      NA                                0                 0      NA
#> 3:      NA                                0                 0      NA
#> 4:      NA                                0                 0      NA
#> 5:      NA                                0                 0      NA
#> 6:      NA                                0                 0      NA
#>    E2_Debt/contract relief E3_Fiscal measures E4_International support
#> 1:                       0                  0                        0
#> 2:                       0                  0                        0
#> 3:                       0                  0                        0
#> 4:                       0                  0                        0
#> 5:                       0                  0                        0
#> 6:                       0                  0                        0
#>    H1_Public information campaigns H1_Flag H2_Testing policy H3_Contact tracing
#> 1:                               0      NA                 0                  0
#> 2:                               0      NA                 0                  0
#> 3:                               0      NA                 0                  0
#> 4:                               0      NA                 0                  0
#> 5:                               0      NA                 0                  0
#> 6:                               0      NA                 0                  0
#>    H4_Emergency investment in healthcare H5_Investment in vaccines
#> 1:                                     0                         0
#> 2:                                     0                         0
#> 3:                                     0                         0
#> 4:                                     0                         0
#> 5:                                     0                         0
#> 6:                                     0                         0
#>    H6_Facial Coverings H6_Flag H7_Vaccination policy H7_Flag
#> 1:                   0      NA                     0      NA
#> 2:                   0      NA                     0      NA
#> 3:                   0      NA                     0      NA
#> 4:                   0      NA                     0      NA
#> 5:                   0      NA                     0      NA
#> 6:                   0      NA                     0      NA
#>    H8_Protection of elderly people H8_Flag M1_Wildcard
#> 1:                               0      NA          NA
#> 2:                               0      NA          NA
#> 3:                               0      NA          NA
#> 4:                               0      NA          NA
#> 5:                               0      NA          NA
#> 6:                               0      NA          NA
#>    V1_Vaccine Prioritisation (summary) V2A_Vaccine Availability (summary)
#> 1:                                   0                                  0
#> 2:                                   0                                  0
#> 3:                                   0                                  0
#> 4:                                   0                                  0
#> 5:                                   0                                  0
#> 6:                                   0                                  0
#>    V2B_Vaccine age eligibility/availability age floor (general population summary)
#> 1:                                                                                
#> 2:                                                                                
#> 3:                                                                                
#> 4:                                                                                
#> 5:                                                                                
#> 6:                                                                                
#>    V2C_Vaccine age eligibility/availability age floor (at risk summary)
#> 1:                                                                     
#> 2:                                                                     
#> 3:                                                                     
#> 4:                                                                     
#> 5:                                                                     
#> 6:                                                                     
#>    V2D_Medically/ clinically vulnerable (Non-elderly) V2E_Education
#> 1:                                                 NA            NA
#> 2:                                                 NA            NA
#> 3:                                                 NA            NA
#> 4:                                                 NA            NA
#> 5:                                                 NA            NA
#> 6:                                                 NA            NA
#>    V2F_Frontline workers  (non healthcare) V2G_Frontline workers  (healthcare)
#> 1:                                      NA                                  NA
#> 2:                                      NA                                  NA
#> 3:                                      NA                                  NA
#> 4:                                      NA                                  NA
#> 5:                                      NA                                  NA
#> 6:                                      NA                                  NA
#>    V3_Vaccine Financial Support (summary) V4_Mandatory Vaccination (summary)
#> 1:                                      0                                 NA
#> 2:                                      0                                 NA
#> 3:                                      0                                 NA
#> 4:                                      0                                 NA
#> 5:                                      0                                 NA
#> 6:                                      0                                 NA
#>    ConfirmedCases ConfirmedDeaths StringencyIndex StringencyIndexForDisplay
#> 1:             NA              NA               0                         0
#> 2:             NA              NA               0                         0
#> 3:             NA              NA               0                         0
#> 4:             NA              NA               0                         0
#> 5:             NA              NA               0                         0
#> 6:             NA              NA               0                         0
#>    StringencyLegacyIndex StringencyLegacyIndexForDisplay
#> 1:                     0                               0
#> 2:                     0                               0
#> 3:                     0                               0
#> 4:                     0                               0
#> 5:                     0                               0
#> 6:                     0                               0
#>    GovernmentResponseIndex GovernmentResponseIndexForDisplay
#> 1:                       0                                 0
#> 2:                       0                                 0
#> 3:                       0                                 0
#> 4:                       0                                 0
#> 5:                       0                                 0
#> 6:                       0                                 0
#>    ContainmentHealthIndex ContainmentHealthIndexForDisplay EconomicSupportIndex
#> 1:                      0                                0                    0
#> 2:                      0                                0                    0
#> 3:                      0                                0                    0
#> 4:                      0                                0                    0
#> 5:                      0                                0                    0
#> 6:                      0                                0                    0
#>    EconomicSupportIndexForDisplay
#> 1:                              0
#> 2:                              0
#> 3:                              0
#> 4:                              0
#> 5:                              0
#> 6:                              0
colnames(res)
#>  [1] "country"                                                                        
#>  [2] "iso3c"                                                                          
#>  [3] "date"                                                                           
#>  [4] "RegionCode"                                                                     
#>  [5] "Jurisdiction"                                                                   
#>  [6] "Date"                                                                           
#>  [7] "C1_School closing"                                                              
#>  [8] "C1_Flag"                                                                        
#>  [9] "C2_Workplace closing"                                                           
#> [10] "C2_Flag"                                                                        
#> [11] "C3_Cancel public events"                                                        
#> [12] "C3_Flag"                                                                        
#> [13] "C4_Restrictions on gatherings"                                                  
#> [14] "C4_Flag"                                                                        
#> [15] "C5_Close public transport"                                                      
#> [16] "C5_Flag"                                                                        
#> [17] "C6_Stay at home requirements"                                                   
#> [18] "C6_Flag"                                                                        
#> [19] "C7_Restrictions on internal movement"                                           
#> [20] "C7_Flag"                                                                        
#> [21] "C8_International travel controls"                                               
#> [22] "E1_Income support"                                                              
#> [23] "E1_Flag"                                                                        
#> [24] "E2_Debt/contract relief"                                                        
#> [25] "E3_Fiscal measures"                                                             
#> [26] "E4_International support"                                                       
#> [27] "H1_Public information campaigns"                                                
#> [28] "H1_Flag"                                                                        
#> [29] "H2_Testing policy"                                                              
#> [30] "H3_Contact tracing"                                                             
#> [31] "H4_Emergency investment in healthcare"                                          
#> [32] "H5_Investment in vaccines"                                                      
#> [33] "H6_Facial Coverings"                                                            
#> [34] "H6_Flag"                                                                        
#> [35] "H7_Vaccination policy"                                                          
#> [36] "H7_Flag"                                                                        
#> [37] "H8_Protection of elderly people"                                                
#> [38] "H8_Flag"                                                                        
#> [39] "M1_Wildcard"                                                                    
#> [40] "V1_Vaccine Prioritisation (summary)"                                            
#> [41] "V2A_Vaccine Availability (summary)"                                             
#> [42] "V2B_Vaccine age eligibility/availability age floor (general population summary)"
#> [43] "V2C_Vaccine age eligibility/availability age floor (at risk summary)"           
#> [44] "V2D_Medically/ clinically vulnerable (Non-elderly)"                             
#> [45] "V2E_Education"                                                                  
#> [46] "V2F_Frontline workers  (non healthcare)"                                        
#> [47] "V2G_Frontline workers  (healthcare)"                                            
#> [48] "V3_Vaccine Financial Support (summary)"                                         
#> [49] "V4_Mandatory Vaccination (summary)"                                             
#> [50] "ConfirmedCases"                                                                 
#> [51] "ConfirmedDeaths"                                                                
#> [52] "StringencyIndex"                                                                
#> [53] "StringencyIndexForDisplay"                                                      
#> [54] "StringencyLegacyIndex"                                                          
#> [55] "StringencyLegacyIndexForDisplay"                                                
#> [56] "GovernmentResponseIndex"                                                        
#> [57] "GovernmentResponseIndexForDisplay"                                              
#> [58] "ContainmentHealthIndex"                                                         
#> [59] "ContainmentHealthIndexForDisplay"                                               
#> [60] "EconomicSupportIndex"                                                           
#> [61] "EconomicSupportIndexForDisplay"                                                 
dplyr::glimpse(res)
#> Rows: 276,636
#> Columns: 61
#> $ country                                                                           <chr> …
#> $ iso3c                                                                             <chr> …
#> $ date                                                                              <chr> …
#> $ RegionCode                                                                        <chr> …
#> $ Jurisdiction                                                                      <chr> …
#> $ Date                                                                              <date> …
#> $ `C1_School closing`                                                               <dbl> …
#> $ C1_Flag                                                                           <int> …
#> $ `C2_Workplace closing`                                                            <dbl> …
#> $ C2_Flag                                                                           <int> …
#> $ `C3_Cancel public events`                                                         <dbl> …
#> $ C3_Flag                                                                           <int> …
#> $ `C4_Restrictions on gatherings`                                                   <dbl> …
#> $ C4_Flag                                                                           <int> …
#> $ `C5_Close public transport`                                                       <dbl> …
#> $ C5_Flag                                                                           <int> …
#> $ `C6_Stay at home requirements`                                                    <dbl> …
#> $ C6_Flag                                                                           <int> …
#> $ `C7_Restrictions on internal movement`                                            <dbl> …
#> $ C7_Flag                                                                           <int> …
#> $ `C8_International travel controls`                                                <dbl> …
#> $ `E1_Income support`                                                               <dbl> …
#> $ E1_Flag                                                                           <int> …
#> $ `E2_Debt/contract relief`                                                         <dbl> …
#> $ `E3_Fiscal measures`                                                              <dbl> …
#> $ `E4_International support`                                                        <dbl> …
#> $ `H1_Public information campaigns`                                                 <dbl> …
#> $ H1_Flag                                                                           <int> …
#> $ `H2_Testing policy`                                                               <dbl> …
#> $ `H3_Contact tracing`                                                              <dbl> …
#> $ `H4_Emergency investment in healthcare`                                           <dbl> …
#> $ `H5_Investment in vaccines`                                                       <dbl> …
#> $ `H6_Facial Coverings`                                                             <dbl> …
#> $ H6_Flag                                                                           <int> …
#> $ `H7_Vaccination policy`                                                           <dbl> …
#> $ H7_Flag                                                                           <int> …
#> $ `H8_Protection of elderly people`                                                 <dbl> …
#> $ H8_Flag                                                                           <int> …
#> $ M1_Wildcard                                                                       <lgl> …
#> $ `V1_Vaccine Prioritisation (summary)`                                             <int> …
#> $ `V2A_Vaccine Availability (summary)`                                              <int> …
#> $ `V2B_Vaccine age eligibility/availability age floor (general population summary)` <chr> …
#> $ `V2C_Vaccine age eligibility/availability age floor (at risk summary)`            <chr> …
#> $ `V2D_Medically/ clinically vulnerable (Non-elderly)`                              <int> …
#> $ V2E_Education                                                                     <int> …
#> $ `V2F_Frontline workers  (non healthcare)`                                         <int> …
#> $ `V2G_Frontline workers  (healthcare)`                                             <int> …
#> $ `V3_Vaccine Financial Support (summary)`                                          <int> …
#> $ `V4_Mandatory Vaccination (summary)`                                              <int> …
#> $ ConfirmedCases                                                                    <int> …
#> $ ConfirmedDeaths                                                                   <int> …
#> $ StringencyIndex                                                                   <dbl> …
#> $ StringencyIndexForDisplay                                                         <dbl> …
#> $ StringencyLegacyIndex                                                             <dbl> …
#> $ StringencyLegacyIndexForDisplay                                                   <dbl> …
#> $ GovernmentResponseIndex                                                           <dbl> …
#> $ GovernmentResponseIndexForDisplay                                                 <dbl> …
#> $ ContainmentHealthIndex                                                            <dbl> …
#> $ ContainmentHealthIndexForDisplay                                                  <dbl> …
#> $ EconomicSupportIndex                                                              <dbl> …
#> $ EconomicSupportIndexForDisplay                                                    <dbl> …
summary(res)
#>    country             iso3c               date            RegionCode       
#>  Length:276636      Length:276636      Length:276636      Length:276636     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Jurisdiction            Date            C1_School closing    C1_Flag     
#>  Length:276636      Min.   :2020-01-01   Min.   :0.000     Min.   :0.00   
#>  Class :character   1st Qu.:2020-08-08   1st Qu.:1.000     1st Qu.:1.00   
#>  Mode  :character   Median :2021-03-08   Median :1.000     Median :1.00   
#>                     Mean   :2021-02-26   Mean   :1.554     Mean   :0.78   
#>                     3rd Qu.:2021-09-18   3rd Qu.:3.000     3rd Qu.:1.00   
#>                     Max.   :2022-03-30   Max.   :3.000     Max.   :1.00   
#>                                          NA's   :10243     NA's   :59617  
#>  C2_Workplace closing    C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.00         Min.   :0.00    Min.   :0.000           Min.   :0.00   
#>  1st Qu.:1.00         1st Qu.:1.00    1st Qu.:1.000           1st Qu.:1.00   
#>  Median :1.00         Median :1.00    Median :1.000           Median :1.00   
#>  Mean   :1.35         Mean   :0.79    Mean   :1.304           Mean   :0.85   
#>  3rd Qu.:2.00         3rd Qu.:1.00    3rd Qu.:2.000           3rd Qu.:1.00   
#>  Max.   :3.00         Max.   :1.00    Max.   :2.000           Max.   :1.00   
#>  NA's   :10483        NA's   :68912   NA's   :10619           NA's   :58286  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.0     Min.   :0.000            
#>  1st Qu.:0.000                 1st Qu.:1.0     1st Qu.:0.000            
#>  Median :3.000                 Median :1.0     Median :0.000            
#>  Mean   :2.353                 Mean   :0.8     Mean   :0.512            
#>  3rd Qu.:4.000                 3rd Qu.:1.0     3rd Qu.:1.000            
#>  Max.   :4.000                 Max.   :1.0     Max.   :2.000            
#>  NA's   :10700                 NA's   :79015   NA's   :10810            
#>     C5_Flag       C6_Stay at home requirements    C6_Flag      
#>  Min.   :0.00     Min.   :0.000                Min.   :0.00    
#>  1st Qu.:0.00     1st Qu.:0.000                1st Qu.:1.00    
#>  Median :1.00     Median :1.000                Median :1.00    
#>  Mean   :0.71     Mean   :0.899                Mean   :0.75    
#>  3rd Qu.:1.00     3rd Qu.:2.000                3rd Qu.:1.00    
#>  Max.   :1.00     Max.   :3.000                Max.   :1.00    
#>  NA's   :164433   NA's   :10612                NA's   :118593  
#>  C7_Restrictions on internal movement    C7_Flag      
#>  Min.   :0.000                        Min.   :0.00    
#>  1st Qu.:0.000                        1st Qu.:0.00    
#>  Median :1.000                        Median :1.00    
#>  Mean   :0.922                        Mean   :0.74    
#>  3rd Qu.:2.000                        3rd Qu.:1.00    
#>  Max.   :2.000                        Max.   :1.00    
#>  NA's   :10661                        NA's   :114001  
#>  C8_International travel controls E1_Income support    E1_Flag      
#>  Min.   :0.000                    Min.   :0.000     Min.   :0.00    
#>  1st Qu.:2.000                    1st Qu.:0.000     1st Qu.:0.00    
#>  Median :3.000                    Median :1.000     Median :0.00    
#>  Mean   :2.491                    Mean   :0.783     Mean   :0.49    
#>  3rd Qu.:3.000                    3rd Qu.:1.000     3rd Qu.:1.00    
#>  Max.   :4.000                    Max.   :2.000     Max.   :1.00    
#>  NA's   :11468                    NA's   :23667     NA's   :125277  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.00            Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.00            1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.00            Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.91            Mean   :1.219e+08   Mean   :6.164e+06       
#>  3rd Qu.:2.00            3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.00            Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :23755           NA's   :129887      NA's   :129991          
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:2.000                   1st Qu.:1.000   1st Qu.:2.000    
#>  Median :2.000                   Median :1.000   Median :2.000    
#>  Mean   :1.857                   Mean   :0.994   Mean   :2.128    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:3.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :10626                   NA's   :23852   NA's   :11615    
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.000      Min.   :0.000e+00                    
#>  1st Qu.:1.000      1st Qu.:0.000e+00                    
#>  Median :2.000      Median :0.000e+00                    
#>  Mean   :1.426      Mean   :6.058e+06                    
#>  3rd Qu.:2.000      3rd Qu.:0.000e+00                    
#>  Max.   :2.000      Max.   :2.424e+11                    
#>  NA's   :13119      NA's   :125537                       
#>  H5_Investment in vaccines H6_Facial Coverings    H6_Flag     
#>  Min.   :0.000e+00         Min.   :0.000       Min.   :0.00   
#>  1st Qu.:0.000e+00         1st Qu.:2.000       1st Qu.:1.00   
#>  Median :0.000e+00         Median :3.000       Median :1.00   
#>  Mean   :9.361e+05         Mean   :2.406       Mean   :0.89   
#>  3rd Qu.:0.000e+00         3rd Qu.:3.000       3rd Qu.:1.00   
#>  Max.   :1.004e+11         Max.   :4.000       Max.   :1.00   
#>  NA's   :69932             NA's   :10791       NA's   :48800  
#>  H7_Vaccination policy    H7_Flag       H8_Protection of elderly people
#>  Min.   :0.000         Min.   :0        Min.   :0.000                  
#>  1st Qu.:0.000         1st Qu.:1        1st Qu.:0.000                  
#>  Median :2.000         Median :1        Median :2.000                  
#>  Mean   :2.213         Mean   :1        Mean   :1.459                  
#>  3rd Qu.:5.000         3rd Qu.:1        3rd Qu.:2.000                  
#>  Max.   :5.000         Max.   :1        Max.   :3.000                  
#>  NA's   :10960         NA's   :140808   NA's   :21216                  
#>     H8_Flag      M1_Wildcard    V1_Vaccine Prioritisation (summary)
#>  Min.   :0.00    Mode:logical   Min.   :0.00                       
#>  1st Qu.:1.00    NA's:276636    1st Qu.:0.00                       
#>  Median :1.00                   Median :1.00                       
#>  Mean   :0.87                   Mean   :0.83                       
#>  3rd Qu.:1.00                   3rd Qu.:2.00                       
#>  Max.   :1.00                   Max.   :2.00                       
#>  NA's   :87898                  NA's   :89759                      
#>  V2A_Vaccine Availability (summary)
#>  Min.   :0                         
#>  1st Qu.:0                         
#>  Median :1                         
#>  Mean   :1                         
#>  3rd Qu.:2                         
#>  Max.   :3                         
#>  NA's   :90092                     
#>  V2B_Vaccine age eligibility/availability age floor (general population summary)
#>  Length:276636                                                                  
#>  Class :character                                                               
#>  Mode  :character                                                               
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>  V2C_Vaccine age eligibility/availability age floor (at risk summary)
#>  Length:276636                                                       
#>  Class :character                                                    
#>  Mode  :character                                                    
#>                                                                      
#>                                                                      
#>                                                                      
#>                                                                      
#>  V2D_Medically/ clinically vulnerable (Non-elderly) V2E_Education   
#>  Min.   :0.00                                       Min.   :0.00    
#>  1st Qu.:1.00                                       1st Qu.:0.00    
#>  Median :2.00                                       Median :2.00    
#>  Mean   :1.46                                       Mean   :1.37    
#>  3rd Qu.:2.00                                       3rd Qu.:2.00    
#>  Max.   :2.00                                       Max.   :2.00    
#>  NA's   :179053                                     NA's   :179053  
#>  V2F_Frontline workers  (non healthcare) V2G_Frontline workers  (healthcare)
#>  Min.   :0.00                            Min.   :0.00                       
#>  1st Qu.:2.00                            1st Qu.:2.00                       
#>  Median :2.00                            Median :2.00                       
#>  Mean   :1.62                            Mean   :1.75                       
#>  3rd Qu.:2.00                            3rd Qu.:2.00                       
#>  Max.   :2.00                            Max.   :2.00                       
#>  NA's   :179053                          NA's   :179053                     
#>  V3_Vaccine Financial Support (summary) V4_Mandatory Vaccination (summary)
#>  Min.   :0.00                           Min.   :0.00                      
#>  1st Qu.:0.00                           1st Qu.:0.00                      
#>  Median :1.00                           Median :0.00                      
#>  Mean   :2.48                           Mean   :0.28                      
#>  3rd Qu.:5.00                           3rd Qu.:1.00                      
#>  Max.   :5.00                           Max.   :1.00                      
#>  NA's   :90107                          NA's   :238659                    
#>  ConfirmedCases     ConfirmedDeaths  StringencyIndex  StringencyIndexForDisplay
#>  Min.   :       0   Min.   :     0   Min.   :  0.00   Min.   :  0.00           
#>  1st Qu.:    1179   1st Qu.:    10   1st Qu.: 37.50   1st Qu.: 37.04           
#>  Median :   34950   Median :   624   Median : 52.78   Median : 52.78           
#>  Mean   :  579967   Mean   : 11319   Mean   : 50.77   Mean   : 50.73           
#>  3rd Qu.:  310260   3rd Qu.:  5872   3rd Qu.: 66.67   3rd Qu.: 66.67           
#>  Max.   :79995485   Max.   :977687   Max.   :100.00   Max.   :100.00           
#>  NA's   :26825      NA's   :26913    NA's   :11011    NA's   :10086            
#>  StringencyLegacyIndex StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00        Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 48.81        1st Qu.: 48.81                  1st Qu.:44.27          
#>  Median : 63.10        Median : 63.10                  Median :54.69          
#>  Mean   : 58.96        Mean   : 58.92                  Mean   :51.26          
#>  3rd Qu.: 73.81        3rd Qu.: 73.81                  3rd Qu.:63.80          
#>  Max.   :100.00        Max.   :100.00                  Max.   :91.67          
#>  NA's   :10904         NA's   :9982                    NA's   :12808          
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   : 0.00         
#>  1st Qu.:44.06                     1st Qu.:45.83         
#>  Median :54.69                     Median :57.14         
#>  Mean   :51.25                     Mean   :53.32         
#>  3rd Qu.:63.75                     3rd Qu.:66.07         
#>  Max.   :91.67                     Max.   :94.05         
#>  NA's   :11876                     NA's   :11188         
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   : 0.00                    Min.   :  0.00      
#>  1st Qu.:45.83                    1st Qu.:  0.00      
#>  Median :57.14                    Median : 37.50      
#>  Mean   :53.30                    Mean   : 38.52      
#>  3rd Qu.:66.07                    3rd Qu.: 62.50      
#>  Max.   :94.05                    Max.   :100.00      
#>  NA's   :10261                    NA's   :23879       
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 37.50                
#>  Mean   : 38.48                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :22954