Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

Examples

library(dplyr)
res = government_policy_timeline()

#> New names:
#> * `` -> ...35

head(res)

#> # A tibble: 6 x 35
#>   country iso3c date       `S1_School clos… S1_IsGeneral S1_Notes
#>   <chr>   <chr> <date>                <dbl>        <dbl> <chr>   
#> 1 Afghan… AFG   2020-01-01                0            1 https:/…
#> 2 Angola  AGO   2020-01-01                0           NA <NA>    
#> 3 United… ARE   2020-01-01               NA           NA <NA>    
#> 4 Argent… ARG   2020-01-01                0           NA <NA>    
#> 5 Armenia ARM   2020-01-01               NA           NA <NA>    
#> 6 Austra… AUS   2020-01-01                0           NA <NA>    
#> # … with 29 more variables: `S2_Workplace closing` <dbl>, S2_IsGeneral <dbl>,
#> #   S2_Notes <chr>, `S3_Cancel public events` <dbl>, S3_IsGeneral <dbl>,
#> #   S3_Notes <chr>, `S4_Close public transport` <dbl>, S4_IsGeneral <dbl>,
#> #   S4_Notes <chr>, `S5_Public information campaigns` <dbl>,
#> #   S5_IsGeneral <dbl>, S5_Notes <chr>, `S6_Restrictions on internal
#> #   movement` <dbl>, S6_IsGeneral <dbl>, S6_Notes <chr>, `S7_International
#> #   travel controls` <dbl>, S7_Notes <chr>, `S8_Fiscal measures` <dbl>,
#> #   S8_Notes <chr>, `S9_Monetary measures` <dbl>, S9_Notes <chr>,
#> #   `S10_Emergency investment in health care` <dbl>, S10_Notes <chr>,
#> #   `S11_Investment in Vaccines` <dbl>, S11_Notes <chr>, ConfirmedCases <dbl>,
#> #   ConfirmedDeaths <dbl>, StringencyIndex <dbl>, ...35 <lgl>

colnames(res)

#>  [1] "country"                                
#>  [2] "iso3c"                                  
#>  [3] "date"                                   
#>  [4] "S1_School closing"                      
#>  [5] "S1_IsGeneral"                           
#>  [6] "S1_Notes"                               
#>  [7] "S2_Workplace closing"                   
#>  [8] "S2_IsGeneral"                           
#>  [9] "S2_Notes"                               
#> [10] "S3_Cancel public events"                
#> [11] "S3_IsGeneral"                           
#> [12] "S3_Notes"                               
#> [13] "S4_Close public transport"              
#> [14] "S4_IsGeneral"                           
#> [15] "S4_Notes"                               
#> [16] "S5_Public information campaigns"        
#> [17] "S5_IsGeneral"                           
#> [18] "S5_Notes"                               
#> [19] "S6_Restrictions on internal movement"   
#> [20] "S6_IsGeneral"                           
#> [21] "S6_Notes"                               
#> [22] "S7_International travel controls"       
#> [23] "S7_Notes"                               
#> [24] "S8_Fiscal measures"                     
#> [25] "S8_Notes"                               
#> [26] "S9_Monetary measures"                   
#> [27] "S9_Notes"                               
#> [28] "S10_Emergency investment in health care"
#> [29] "S10_Notes"                              
#> [30] "S11_Investment in Vaccines"             
#> [31] "S11_Notes"                              
#> [32] "ConfirmedCases"                         
#> [33] "ConfirmedDeaths"                        
#> [34] "StringencyIndex"                        
#> [35] "...35"                                  

dplyr::glimpse(res)

#> Rows: 10,561
#> Columns: 35
#> $ country                                   <chr> "Afghanistan", "Angola", "U…
#> $ iso3c                                     <chr> "AFG", "AGO", "ARE", "ARG",…
#> $ date                                      <date> 2020-01-01, 2020-01-01, 20…
#> $ `S1_School closing`                       <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S1_IsGeneral                              <dbl> 1, NA, NA, NA, NA, NA, NA, …
#> $ S1_Notes                                  <chr> "https://en.unesco.org/them…
#> $ `S2_Workplace closing`                    <dbl> NA, 0, NA, 0, NA, 0, 0, NA,…
#> $ S2_IsGeneral                              <dbl> NA, NA, NA, NA, NA, NA, NA,…
#> $ S2_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S3_Cancel public events`                 <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S3_IsGeneral                              <dbl> 1, 1, NA, NA, NA, NA, NA, N…
#> $ S3_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S4_Close public transport`               <dbl> NA, 0, NA, 0, NA, 0, 0, NA,…
#> $ S4_IsGeneral                              <dbl> NA, 1, NA, NA, NA, NA, NA, …
#> $ S4_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S5_Public information campaigns`         <dbl> 0, 0, NA, 0, NA, 1, 0, NA, …
#> $ S5_IsGeneral                              <dbl> 1, 1, NA, NA, NA, 0, NA, NA…
#> $ S5_Notes                                  <chr> NA, NA, NA, NA, NA, "Target…
#> $ `S6_Restrictions on internal movement`    <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S6_IsGeneral                              <dbl> 1, 1, NA, NA, NA, NA, NA, N…
#> $ S6_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S7_International travel controls`        <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S7_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S8_Fiscal measures`                      <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S8_Notes                                  <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S9_Monetary measures`                    <dbl> NA, 0.00, NA, 55.00, NA, 0.…
#> $ S9_Notes                                  <chr> NA, NA, NA, "http://www.bcr…
#> $ `S10_Emergency investment in health care` <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S10_Notes                                 <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ `S11_Investment in Vaccines`              <dbl> 0, 0, NA, 0, NA, 0, 0, NA, …
#> $ S11_Notes                                 <chr> NA, NA, NA, NA, NA, NA, NA,…
#> $ ConfirmedCases                            <dbl> 0, NA, 0, NA, 0, 0, 0, 0, 0…
#> $ ConfirmedDeaths                           <dbl> 0, NA, 0, NA, 0, 0, 0, 0, 0…
#> $ StringencyIndex                           <dbl> 0, 0, NA, 0, NA, 7, 0, NA, …
#> $ ...35                                     <lgl> NA, NA, NA, NA, NA, NA, NA,…