Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#> # A tibble: 6 x 44
#>   country iso3c date  RegionCode Date       `C1_School clos… C1_Flag
#>   <chr>   <chr> <chr> <chr>      <date>                <dbl>   <dbl>
#> 1 Aruba   ABW   NA    NA         2020-01-01                0      NA
#> 2 Aruba   ABW   NA    NA         2020-01-02                0      NA
#> 3 Aruba   ABW   NA    NA         2020-01-03                0      NA
#> 4 Aruba   ABW   NA    NA         2020-01-04                0      NA
#> 5 Aruba   ABW   NA    NA         2020-01-05                0      NA
#> 6 Aruba   ABW   NA    NA         2020-01-06                0      NA
#> # … with 37 more variables: `C2_Workplace closing` <dbl>, C2_Flag <dbl>,
#> #   `C3_Cancel public events` <dbl>, C3_Flag <dbl>, `C4_Restrictions on
#> #   gatherings` <dbl>, C4_Flag <dbl>, `C5_Close public transport` <dbl>,
#> #   C5_Flag <dbl>, `C6_Stay at home requirements` <dbl>, C6_Flag <dbl>,
#> #   `C7_Restrictions on internal movement` <dbl>, C7_Flag <dbl>,
#> #   `C8_International travel controls` <dbl>, `E1_Income support` <dbl>,
#> #   E1_Flag <dbl>, `E2_Debt/contract relief` <dbl>, `E3_Fiscal measures` <dbl>,
#> #   `E4_International support` <dbl>, `H1_Public information campaigns` <dbl>,
#> #   H1_Flag <dbl>, `H2_Testing policy` <dbl>, `H3_Contact tracing` <dbl>,
#> #   `H4_Emergency investment in healthcare` <dbl>, `H5_Investment in
#> #   vaccines` <dbl>, M1_Wildcard <lgl>, ConfirmedCases <dbl>,
#> #   ConfirmedDeaths <dbl>, StringencyIndex <dbl>,
#> #   StringencyIndexForDisplay <dbl>, StringencyLegacyIndex <dbl>,
#> #   StringencyLegacyIndexForDisplay <dbl>, GovernmentResponseIndex <dbl>,
#> #   GovernmentResponseIndexForDisplay <dbl>, ContainmentHealthIndex <dbl>,
#> #   ContainmentHealthIndexForDisplay <dbl>, EconomicSupportIndex <dbl>,
#> #   EconomicSupportIndexForDisplay <dbl>

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "RegionCode"                           
#>  [5] "Date"                                 
#>  [6] "C1_School closing"                    
#>  [7] "C1_Flag"                              
#>  [8] "C2_Workplace closing"                 
#>  [9] "C2_Flag"                              
#> [10] "C3_Cancel public events"              
#> [11] "C3_Flag"                              
#> [12] "C4_Restrictions on gatherings"        
#> [13] "C4_Flag"                              
#> [14] "C5_Close public transport"            
#> [15] "C5_Flag"                              
#> [16] "C6_Stay at home requirements"         
#> [17] "C6_Flag"                              
#> [18] "C7_Restrictions on internal movement" 
#> [19] "C7_Flag"                              
#> [20] "C8_International travel controls"     
#> [21] "E1_Income support"                    
#> [22] "E1_Flag"                              
#> [23] "E2_Debt/contract relief"              
#> [24] "E3_Fiscal measures"                   
#> [25] "E4_International support"             
#> [26] "H1_Public information campaigns"      
#> [27] "H1_Flag"                              
#> [28] "H2_Testing policy"                    
#> [29] "H3_Contact tracing"                   
#> [30] "H4_Emergency investment in healthcare"
#> [31] "H5_Investment in vaccines"            
#> [32] "M1_Wildcard"                          
#> [33] "ConfirmedCases"                       
#> [34] "ConfirmedDeaths"                      
#> [35] "StringencyIndex"                      
#> [36] "StringencyIndexForDisplay"            
#> [37] "StringencyLegacyIndex"                
#> [38] "StringencyLegacyIndexForDisplay"      
#> [39] "GovernmentResponseIndex"              
#> [40] "GovernmentResponseIndexForDisplay"    
#> [41] "ContainmentHealthIndex"               
#> [42] "ContainmentHealthIndexForDisplay"     
#> [43] "EconomicSupportIndex"                 
#> [44] "EconomicSupportIndexForDisplay"       

dplyr::glimpse(res)

#> Rows: 60,973
#> Columns: 44
#> $ country                                 <chr> "Aruba", "Aruba", "Aruba", "A…
#> $ iso3c                                   <chr> "ABW", "ABW", "ABW", "ABW", "…
#> $ date                                    <chr> NA, NA, NA, NA, NA, NA, NA, N…
#> $ RegionCode                              <chr> NA, NA, NA, NA, NA, NA, NA, N…
#> $ Date                                    <date> 2020-01-01, 2020-01-02, 2020…
#> $ `C1_School closing`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C2_Workplace closing`                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C2_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C3_Cancel public events`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C3_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C4_Restrictions on gatherings`         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C4_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C5_Close public transport`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C5_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C6_Stay at home requirements`          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C6_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C7_Restrictions on internal movement`  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C7_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C8_International travel controls`      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E1_Income support`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ E1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `E2_Debt/contract relief`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E3_Fiscal measures`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E4_International support`              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H1_Public information campaigns`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ H1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `H2_Testing policy`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H3_Contact tracing`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H4_Emergency investment in healthcare` <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H5_Investment in vaccines`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ M1_Wildcard                             <lgl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedCases                          <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedDeaths                         <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ StringencyIndex                         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyIndexForDisplay               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyLegacyIndex                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyLegacyIndexForDisplay         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ GovernmentResponseIndex                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ GovernmentResponseIndexForDisplay       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ ContainmentHealthIndex                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ ContainmentHealthIndexForDisplay        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ EconomicSupportIndex                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ EconomicSupportIndexForDisplay          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…

summary(res)

#>    country             iso3c               date            RegionCode       
#>  Length:60973       Length:60973       Length:60973       Length:60973      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>       Date            C1_School closing    C1_Flag      C2_Workplace closing
#>  Min.   :2020-01-01   Min.   :0.000     Min.   :0.000   Min.   :0.000       
#>  1st Qu.:2020-03-04   1st Qu.:0.000     1st Qu.:1.000   1st Qu.:0.000       
#>  Median :2020-05-06   Median :2.000     Median :1.000   Median :1.000       
#>  Mean   :2020-05-06   Mean   :1.755     Mean   :0.838   Mean   :1.231       
#>  3rd Qu.:2020-07-08   3rd Qu.:3.000     3rd Qu.:1.000   3rd Qu.:2.000       
#>  Max.   :2020-09-09   Max.   :3.000     Max.   :1.000   Max.   :3.000       
#>                       NA's   :1740      NA's   :22949   NA's   :1842        
#>     C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.000   Min.   :0.000           Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:0.000           1st Qu.:1.000  
#>  Median :1.000   Median :2.000           Median :1.000  
#>  Mean   :0.808   Mean   :1.191           Mean   :0.887  
#>  3rd Qu.:1.000   3rd Qu.:2.000           3rd Qu.:1.000  
#>  Max.   :1.000   Max.   :2.000           Max.   :1.000  
#>  NA's   :25145   NA's   :1798            NA's   :22103  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.000   Min.   :0.0000           
#>  1st Qu.:0.000                 1st Qu.:1.000   1st Qu.:0.0000           
#>  Median :3.000                 Median :1.000   Median :0.0000           
#>  Mean   :2.019                 Mean   :0.857   Mean   :0.5262           
#>  3rd Qu.:4.000                 3rd Qu.:1.000   3rd Qu.:1.0000           
#>  Max.   :4.000                 Max.   :1.000   Max.   :2.0000           
#>  NA's   :1809                  NA's   :25178   NA's   :1787             
#>     C5_Flag      C6_Stay at home requirements    C6_Flag     
#>  Min.   :0.00    Min.   :0.0000               Min.   :0.000  
#>  1st Qu.:0.00    1st Qu.:0.0000               1st Qu.:1.000  
#>  Median :1.00    Median :1.0000               Median :1.000  
#>  Mean   :0.67    Mean   :0.8603               Mean   :0.777  
#>  3rd Qu.:1.00    3rd Qu.:2.0000               3rd Qu.:1.000  
#>  Max.   :1.00    Max.   :3.0000               Max.   :1.000  
#>  NA's   :38343   NA's   :1836                 NA's   :29043  
#>  C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.0000                       Min.   :0.000  
#>  1st Qu.:0.0000                       1st Qu.:0.000  
#>  Median :1.0000                       Median :1.000  
#>  Mean   :0.8546                       Mean   :0.695  
#>  3rd Qu.:2.0000                       3rd Qu.:1.000  
#>  Max.   :2.0000                       Max.   :1.000  
#>  NA's   :1796                         NA's   :29477  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.0000    Min.   :0.000  
#>  1st Qu.:1.000                    1st Qu.:0.0000    1st Qu.:0.000  
#>  Median :3.000                    Median :1.0000    Median :0.000  
#>  Mean   :2.417                    Mean   :0.7991    Mean   :0.427  
#>  3rd Qu.:4.000                    3rd Qu.:2.0000    3rd Qu.:1.000  
#>  Max.   :4.000                    Max.   :2.0000    Max.   :1.000  
#>  NA's   :1803                     NA's   :2168      NA's   :31210  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.0000          Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.0000          1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.0000          Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.8325          Mean   :2.112e+08   Mean   :1.870e+07       
#>  3rd Qu.:2.0000          3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.0000          Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :3063            NA's   :13199       NA's   :13005           
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:1.000                   1st Qu.:1.000   1st Qu.:1.000    
#>  Median :2.000                   Median :1.000   Median :1.000    
#>  Mean   :1.545                   Mean   :0.991   Mean   :1.356    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:2.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :2113                    NA's   :13093   NA's   :2292     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.000      Min.   :0.000e+00                    
#>  1st Qu.:0.000      1st Qu.:0.000e+00                    
#>  Median :1.000      Median :0.000e+00                    
#>  Mean   :1.118      Mean   :1.351e+07                    
#>  3rd Qu.:2.000      3rd Qu.:0.000e+00                    
#>  Max.   :2.000      Max.   :2.424e+11                    
#>  NA's   :2310       NA's   :13684                        
#>  H5_Investment in vaccines M1_Wildcard    ConfirmedCases    ConfirmedDeaths 
#>  Min.   :0.000e+00         Mode:logical   Min.   :      0   Min.   :     0  
#>  1st Qu.:0.000e+00         NA's:60973     1st Qu.:     68   1st Qu.:     1  
#>  Median :0.000e+00                        Median :   1720   Median :    38  
#>  Mean   :4.367e+05                        Mean   :  45723   Mean   :  2038  
#>  3rd Qu.:0.000e+00                        3rd Qu.:  17850   3rd Qu.:   479  
#>  Max.   :3.435e+09                        Max.   :6300671   Max.   :189208  
#>  NA's   :13281                            NA's   :12093     NA's   :12222   
#>  StringencyIndex  StringencyIndexForDisplay StringencyLegacyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   :  0.00       
#>  1st Qu.: 11.11   1st Qu.: 13.89            1st Qu.: 16.67       
#>  Median : 56.48   Median : 56.48            Median : 65.00       
#>  Mean   : 47.60   Mean   : 47.72            Mean   : 53.11       
#>  3rd Qu.: 75.46   3rd Qu.: 75.00            3rd Qu.: 80.95       
#>  Max.   :100.00   Max.   :100.00            Max.   :100.00       
#>  NA's   :1862     NA's   :861               NA's   :1862         
#>  StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 17.86                  1st Qu.:14.10          
#>  Median : 65.00                  Median :57.05          
#>  Mean   : 53.25                  Mean   :46.61          
#>  3rd Qu.: 80.95                  3rd Qu.:70.51          
#>  Max.   :100.00                  Max.   :96.15          
#>  NA's   :861                     NA's   :2465           
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   :  0.00        
#>  1st Qu.:14.10                     1st Qu.: 16.67        
#>  Median :57.05                     Median : 58.71        
#>  Mean   :46.78                     Mean   : 48.38        
#>  3rd Qu.:70.19                     3rd Qu.: 73.11        
#>  Max.   :96.15                     Max.   :100.00        
#>  NA's   :1471                      NA's   :2296          
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   :  0.00                   Min.   :  0.00      
#>  1st Qu.: 16.67                   1st Qu.:  0.00      
#>  Median : 58.71                   Median : 37.50      
#>  Mean   : 48.54                   Mean   : 37.11      
#>  3rd Qu.: 73.11                   3rd Qu.: 62.50      
#>  Max.   :100.00                   Max.   :100.00      
#>  NA's   :1302                     NA's   :3085        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 37.50                
#>  Mean   : 37.36                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :2120