Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#> # A tibble: 6 x 42
#>   country iso3c date       `C1_School clos… C1_Flag `C2_Workplace c… C2_Flag
#>   <chr>   <chr> <date>                <dbl>   <dbl>            <dbl>   <dbl>
#> 1 Aruba   ABW   2020-01-01                0      NA                0      NA
#> 2 Aruba   ABW   2020-01-02                0      NA                0      NA
#> 3 Aruba   ABW   2020-01-03                0      NA                0      NA
#> 4 Aruba   ABW   2020-01-04                0      NA                0      NA
#> 5 Aruba   ABW   2020-01-05                0      NA                0      NA
#> 6 Aruba   ABW   2020-01-06                0      NA                0      NA
#> # … with 35 more variables: `C3_Cancel public events` <dbl>, C3_Flag <dbl>,
#> #   `C4_Restrictions on gatherings` <dbl>, C4_Flag <dbl>, `C5_Close public
#> #   transport` <dbl>, C5_Flag <dbl>, `C6_Stay at home requirements` <dbl>,
#> #   C6_Flag <dbl>, `C7_Restrictions on internal movement` <dbl>, C7_Flag <dbl>,
#> #   `C8_International travel controls` <dbl>, `E1_Income support` <dbl>,
#> #   E1_Flag <dbl>, `E2_Debt/contract relief` <dbl>, `E3_Fiscal measures` <dbl>,
#> #   `E4_International support` <dbl>, `H1_Public information campaigns` <dbl>,
#> #   H1_Flag <dbl>, `H2_Testing policy` <dbl>, `H3_Contact tracing` <dbl>,
#> #   `H4_Emergency investment in healthcare` <dbl>, `H5_Investment in
#> #   vaccines` <dbl>, M1_Wildcard <lgl>, ConfirmedCases <dbl>,
#> #   ConfirmedDeaths <dbl>, StringencyIndex <dbl>,
#> #   StringencyIndexForDisplay <dbl>, StringencyLegacyIndex <dbl>,
#> #   StringencyLegacyIndexForDisplay <dbl>, GovernmentResponseIndex <dbl>,
#> #   GovernmentResponseIndexForDisplay <dbl>, ContainmentHealthIndex <dbl>,
#> #   ContainmentHealthIndexForDisplay <dbl>, EconomicSupportIndex <dbl>,
#> #   EconomicSupportIndexForDisplay <dbl>

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "C1_School closing"                    
#>  [5] "C1_Flag"                              
#>  [6] "C2_Workplace closing"                 
#>  [7] "C2_Flag"                              
#>  [8] "C3_Cancel public events"              
#>  [9] "C3_Flag"                              
#> [10] "C4_Restrictions on gatherings"        
#> [11] "C4_Flag"                              
#> [12] "C5_Close public transport"            
#> [13] "C5_Flag"                              
#> [14] "C6_Stay at home requirements"         
#> [15] "C6_Flag"                              
#> [16] "C7_Restrictions on internal movement" 
#> [17] "C7_Flag"                              
#> [18] "C8_International travel controls"     
#> [19] "E1_Income support"                    
#> [20] "E1_Flag"                              
#> [21] "E2_Debt/contract relief"              
#> [22] "E3_Fiscal measures"                   
#> [23] "E4_International support"             
#> [24] "H1_Public information campaigns"      
#> [25] "H1_Flag"                              
#> [26] "H2_Testing policy"                    
#> [27] "H3_Contact tracing"                   
#> [28] "H4_Emergency investment in healthcare"
#> [29] "H5_Investment in vaccines"            
#> [30] "M1_Wildcard"                          
#> [31] "ConfirmedCases"                       
#> [32] "ConfirmedDeaths"                      
#> [33] "StringencyIndex"                      
#> [34] "StringencyIndexForDisplay"            
#> [35] "StringencyLegacyIndex"                
#> [36] "StringencyLegacyIndexForDisplay"      
#> [37] "GovernmentResponseIndex"              
#> [38] "GovernmentResponseIndexForDisplay"    
#> [39] "ContainmentHealthIndex"               
#> [40] "ContainmentHealthIndexForDisplay"     
#> [41] "EconomicSupportIndex"                 
#> [42] "EconomicSupportIndexForDisplay"       

dplyr::glimpse(res)

#> Rows: 40,586
#> Columns: 42
#> $ country                                 <chr> "Aruba", "Aruba", "Aruba", "A…
#> $ iso3c                                   <chr> "ABW", "ABW", "ABW", "ABW", "…
#> $ date                                    <date> 2020-01-01, 2020-01-02, 2020…
#> $ `C1_School closing`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C2_Workplace closing`                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C2_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C3_Cancel public events`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C3_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C4_Restrictions on gatherings`         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C4_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C5_Close public transport`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C5_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C6_Stay at home requirements`          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C6_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C7_Restrictions on internal movement`  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ C7_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `C8_International travel controls`      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E1_Income support`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ E1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `E2_Debt/contract relief`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E3_Fiscal measures`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `E4_International support`              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H1_Public information campaigns`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ H1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ `H2_Testing policy`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H3_Contact tracing`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H4_Emergency investment in healthcare` <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ `H5_Investment in vaccines`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ M1_Wildcard                             <lgl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedCases                          <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ ConfirmedDeaths                         <dbl> NA, NA, NA, NA, NA, NA, NA, N…
#> $ StringencyIndex                         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyIndexForDisplay               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyLegacyIndex                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ StringencyLegacyIndexForDisplay         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ GovernmentResponseIndex                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ GovernmentResponseIndexForDisplay       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ ContainmentHealthIndex                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ ContainmentHealthIndexForDisplay        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ EconomicSupportIndex                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ EconomicSupportIndexForDisplay          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…

summary(res)

#>    country             iso3c                date            C1_School closing
#>  Length:40586       Length:40586       Min.   :2020-01-01   Min.   :0.000    
#>  Class :character   Class :character   1st Qu.:2020-02-25   1st Qu.:0.000    
#>  Mode  :character   Mode  :character   Median :2020-04-21   Median :2.000    
#>                                        Mean   :2020-04-21   Mean   :1.671    
#>                                        3rd Qu.:2020-06-16   3rd Qu.:3.000    
#>                                        Max.   :2020-08-10   Max.   :3.000    
#>                                                             NA's   :1786     
#>     C1_Flag      C2_Workplace closing    C2_Flag      C3_Cancel public events
#>  Min.   :0.000   Min.   :0.000        Min.   :0.000   Min.   :0.000          
#>  1st Qu.:1.000   1st Qu.:0.000        1st Qu.:1.000   1st Qu.:0.000          
#>  Median :1.000   Median :1.000        Median :1.000   Median :2.000          
#>  Mean   :0.899   Mean   :1.169        Mean   :0.802   Mean   :1.178          
#>  3rd Qu.:1.000   3rd Qu.:2.000        3rd Qu.:1.000   3rd Qu.:2.000          
#>  Max.   :1.000   Max.   :3.000        Max.   :1.000   Max.   :2.000          
#>  NA's   :16931   NA's   :1941         NA's   :19251   NA's   :1803           
#>     C3_Flag      C4_Restrictions on gatherings    C4_Flag     
#>  Min.   :0.000   Min.   :0.000                 Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:0.000                 1st Qu.:1.000  
#>  Median :1.000   Median :2.000                 Median :1.000  
#>  Mean   :0.906   Mean   :1.911                 Mean   :0.879  
#>  3rd Qu.:1.000   3rd Qu.:4.000                 3rd Qu.:1.000  
#>  Max.   :1.000   Max.   :4.000                 Max.   :1.000  
#>  NA's   :16289   NA's   :1819                  NA's   :18257  
#>  C5_Close public transport    C5_Flag      C6_Stay at home requirements
#>  Min.   :0.0000            Min.   :0.000   Min.   :0.0000              
#>  1st Qu.:0.0000            1st Qu.:0.000   1st Qu.:0.0000              
#>  Median :0.0000            Median :1.000   Median :0.0000              
#>  Mean   :0.5672            Mean   :0.733   Mean   :0.8536              
#>  3rd Qu.:1.0000            3rd Qu.:1.000   3rd Qu.:2.0000              
#>  Max.   :2.0000            Max.   :1.000   Max.   :3.0000              
#>  NA's   :2084              NA's   :25577   NA's   :2003                
#>     C6_Flag      C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000   Min.   :0.0000                       Min.   :0.000  
#>  1st Qu.:1.000   1st Qu.:0.0000                       1st Qu.:0.000  
#>  Median :1.000   Median :0.0000                       Median :1.000  
#>  Mean   :0.759   Mean   :0.8459                       Mean   :0.628  
#>  3rd Qu.:1.000   3rd Qu.:2.0000                       3rd Qu.:1.000  
#>  Max.   :1.000   Max.   :2.0000                       Max.   :1.000  
#>  NA's   :21508   NA's   :2108                         NA's   :21934  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.0000    Min.   :0.000  
#>  1st Qu.:0.000                    1st Qu.:0.0000    1st Qu.:0.000  
#>  Median :3.000                    Median :0.0000    Median :1.000  
#>  Mean   :2.401                    Mean   :0.5885    Mean   :0.554  
#>  3rd Qu.:4.000                    3rd Qu.:1.0000    3rd Qu.:1.000  
#>  Max.   :4.000                    Max.   :2.0000    Max.   :1.000  
#>  NA's   :1734                     NA's   :2070      NA's   :24343  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.0000          Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.0000          1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :0.0000          Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.7217          Mean   :2.980e+08   Mean   :2.333e+07       
#>  3rd Qu.:2.0000          3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.0000          Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :2033            NA's   :2264        NA's   :2244            
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.00    Min.   :0.0      
#>  1st Qu.:1.000                   1st Qu.:1.00    1st Qu.:0.0      
#>  Median :2.000                   Median :1.00    Median :1.0      
#>  Mean   :1.499                   Mean   :0.99    Mean   :1.2      
#>  3rd Qu.:2.000                   3rd Qu.:1.00    3rd Qu.:2.0      
#>  Max.   :2.000                   Max.   :1.00    Max.   :3.0      
#>  NA's   :1712                    NA's   :10005   NA's   :2006     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.00       Min.   :0.000e+00                    
#>  1st Qu.:0.00       1st Qu.:0.000e+00                    
#>  Median :1.00       Median :0.000e+00                    
#>  Mean   :1.03       Mean   :1.634e+07                    
#>  3rd Qu.:2.00       3rd Qu.:0.000e+00                    
#>  Max.   :2.00       Max.   :2.424e+11                    
#>  NA's   :1894       NA's   :2318                         
#>  H5_Investment in vaccines M1_Wildcard    ConfirmedCases    ConfirmedDeaths 
#>  Min.   :0.000e+00         Mode:logical   Min.   :      0   Min.   :     0  
#>  1st Qu.:0.000e+00         NA's:40586     1st Qu.:     35   1st Qu.:     0  
#>  Median :0.000e+00                        Median :    842   Median :    13  
#>  Mean   :4.849e+05                        Mean   :  32652   Mean   :  1642  
#>  3rd Qu.:0.000e+00                        3rd Qu.:   7777   3rd Qu.:   164  
#>  Max.   :3.435e+09                        Max.   :4823891   Max.   :158256  
#>  NA's   :2323                             NA's   :10553     NA's   :10553   
#>  StringencyIndex  StringencyIndexForDisplay StringencyLegacyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   :  0.00       
#>  1st Qu.: 11.11   1st Qu.: 11.11            1st Qu.: 14.29       
#>  Median : 52.78   Median : 52.78            Median : 63.10       
#>  Mean   : 46.63   Mean   : 46.84            Mean   : 51.82       
#>  3rd Qu.: 77.78   3rd Qu.: 77.78            3rd Qu.: 83.33       
#>  Max.   :100.00   Max.   :100.00            Max.   :100.00       
#>  NA's   :2108     NA's   :1217              NA's   :2016         
#>  StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 14.29                  1st Qu.: 9.62          
#>  Median : 63.10                  Median :53.21          
#>  Mean   : 52.08                  Mean   :43.89          
#>  3rd Qu.: 83.33                  3rd Qu.:69.87          
#>  Max.   :100.00                  Max.   :96.15          
#>  NA's   :1118                    NA's   :2193           
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   :  0.00        
#>  1st Qu.:10.26                     1st Qu.: 11.36        
#>  Median :53.53                     Median : 54.55        
#>  Mean   :44.19                     Mean   : 46.42        
#>  3rd Qu.:69.87                     3rd Qu.: 75.00        
#>  Max.   :96.15                     Max.   :100.00        
#>  NA's   :1308                      NA's   :2132          
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   :  0.00                   Min.   :  0.00      
#>  1st Qu.: 12.12                   1st Qu.:  0.00      
#>  Median : 55.30                   Median : 25.00      
#>  Mean   : 46.69                   Mean   : 30.38      
#>  3rd Qu.: 75.00                   3rd Qu.: 62.50      
#>  Max.   :100.00                   Max.   :100.00      
#>  NA's   :1248                     NA's   :2153        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 25.00                
#>  Mean   : 30.85                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :1270