Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Value

a data.frame

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#>    country iso3c date RegionCode Jurisdiction       Date C1_School closing
#> 1:   Aruba   ABW                    NAT_TOTAL 2020-01-01                 0
#> 2:   Aruba   ABW                    NAT_TOTAL 2020-01-02                 0
#> 3:   Aruba   ABW                    NAT_TOTAL 2020-01-03                 0
#> 4:   Aruba   ABW                    NAT_TOTAL 2020-01-04                 0
#> 5:   Aruba   ABW                    NAT_TOTAL 2020-01-05                 0
#> 6:   Aruba   ABW                    NAT_TOTAL 2020-01-06                 0
#>    C1_Flag C2_Workplace closing C2_Flag C3_Cancel public events C3_Flag
#> 1:      NA                    0      NA                       0      NA
#> 2:      NA                    0      NA                       0      NA
#> 3:      NA                    0      NA                       0      NA
#> 4:      NA                    0      NA                       0      NA
#> 5:      NA                    0      NA                       0      NA
#> 6:      NA                    0      NA                       0      NA
#>    C4_Restrictions on gatherings C4_Flag C5_Close public transport C5_Flag
#> 1:                             0      NA                         0      NA
#> 2:                             0      NA                         0      NA
#> 3:                             0      NA                         0      NA
#> 4:                             0      NA                         0      NA
#> 5:                             0      NA                         0      NA
#> 6:                             0      NA                         0      NA
#>    C6_Stay at home requirements C6_Flag C7_Restrictions on internal movement
#> 1:                            0      NA                                    0
#> 2:                            0      NA                                    0
#> 3:                            0      NA                                    0
#> 4:                            0      NA                                    0
#> 5:                            0      NA                                    0
#> 6:                            0      NA                                    0
#>    C7_Flag C8_International travel controls E1_Income support E1_Flag
#> 1:      NA                                0                 0      NA
#> 2:      NA                                0                 0      NA
#> 3:      NA                                0                 0      NA
#> 4:      NA                                0                 0      NA
#> 5:      NA                                0                 0      NA
#> 6:      NA                                0                 0      NA
#>    E2_Debt/contract relief E3_Fiscal measures E4_International support
#> 1:                       0                  0                        0
#> 2:                       0                  0                        0
#> 3:                       0                  0                        0
#> 4:                       0                  0                        0
#> 5:                       0                  0                        0
#> 6:                       0                  0                        0
#>    H1_Public information campaigns H1_Flag H2_Testing policy H3_Contact tracing
#> 1:                               0      NA                 0                  0
#> 2:                               0      NA                 0                  0
#> 3:                               0      NA                 0                  0
#> 4:                               0      NA                 0                  0
#> 5:                               0      NA                 0                  0
#> 6:                               0      NA                 0                  0
#>    H4_Emergency investment in healthcare H5_Investment in vaccines
#> 1:                                     0                         0
#> 2:                                     0                         0
#> 3:                                     0                         0
#> 4:                                     0                         0
#> 5:                                     0                         0
#> 6:                                     0                         0
#>    H6_Facial Coverings H6_Flag H7_Vaccination policy H7_Flag
#> 1:                   0      NA                     0      NA
#> 2:                   0      NA                     0      NA
#> 3:                   0      NA                     0      NA
#> 4:                   0      NA                     0      NA
#> 5:                   0      NA                     0      NA
#> 6:                   0      NA                     0      NA
#>    H8_Protection of elderly people H8_Flag M1_Wildcard ConfirmedCases
#> 1:                               0      NA          NA             NA
#> 2:                               0      NA          NA             NA
#> 3:                               0      NA          NA             NA
#> 4:                               0      NA          NA             NA
#> 5:                               0      NA          NA             NA
#> 6:                               0      NA          NA             NA
#>    ConfirmedDeaths StringencyIndex StringencyIndexForDisplay
#> 1:              NA               0                         0
#> 2:              NA               0                         0
#> 3:              NA               0                         0
#> 4:              NA               0                         0
#> 5:              NA               0                         0
#> 6:              NA               0                         0
#>    StringencyLegacyIndex StringencyLegacyIndexForDisplay
#> 1:                     0                               0
#> 2:                     0                               0
#> 3:                     0                               0
#> 4:                     0                               0
#> 5:                     0                               0
#> 6:                     0                               0
#>    GovernmentResponseIndex GovernmentResponseIndexForDisplay
#> 1:                       0                                 0
#> 2:                       0                                 0
#> 3:                       0                                 0
#> 4:                       0                                 0
#> 5:                       0                                 0
#> 6:                       0                                 0
#>    ContainmentHealthIndex ContainmentHealthIndexForDisplay EconomicSupportIndex
#> 1:                      0                                0                    0
#> 2:                      0                                0                    0
#> 3:                      0                                0                    0
#> 4:                      0                                0                    0
#> 5:                      0                                0                    0
#> 6:                      0                                0                    0
#>    EconomicSupportIndexForDisplay
#> 1:                              0
#> 2:                              0
#> 3:                              0
#> 4:                              0
#> 5:                              0
#> 6:                              0

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "RegionCode"                           
#>  [5] "Jurisdiction"                         
#>  [6] "Date"                                 
#>  [7] "C1_School closing"                    
#>  [8] "C1_Flag"                              
#>  [9] "C2_Workplace closing"                 
#> [10] "C2_Flag"                              
#> [11] "C3_Cancel public events"              
#> [12] "C3_Flag"                              
#> [13] "C4_Restrictions on gatherings"        
#> [14] "C4_Flag"                              
#> [15] "C5_Close public transport"            
#> [16] "C5_Flag"                              
#> [17] "C6_Stay at home requirements"         
#> [18] "C6_Flag"                              
#> [19] "C7_Restrictions on internal movement" 
#> [20] "C7_Flag"                              
#> [21] "C8_International travel controls"     
#> [22] "E1_Income support"                    
#> [23] "E1_Flag"                              
#> [24] "E2_Debt/contract relief"              
#> [25] "E3_Fiscal measures"                   
#> [26] "E4_International support"             
#> [27] "H1_Public information campaigns"      
#> [28] "H1_Flag"                              
#> [29] "H2_Testing policy"                    
#> [30] "H3_Contact tracing"                   
#> [31] "H4_Emergency investment in healthcare"
#> [32] "H5_Investment in vaccines"            
#> [33] "H6_Facial Coverings"                  
#> [34] "H6_Flag"                              
#> [35] "H7_Vaccination policy"                
#> [36] "H7_Flag"                              
#> [37] "H8_Protection of elderly people"      
#> [38] "H8_Flag"                              
#> [39] "M1_Wildcard"                          
#> [40] "ConfirmedCases"                       
#> [41] "ConfirmedDeaths"                      
#> [42] "StringencyIndex"                      
#> [43] "StringencyIndexForDisplay"            
#> [44] "StringencyLegacyIndex"                
#> [45] "StringencyLegacyIndexForDisplay"      
#> [46] "GovernmentResponseIndex"              
#> [47] "GovernmentResponseIndexForDisplay"    
#> [48] "ContainmentHealthIndex"               
#> [49] "ContainmentHealthIndexForDisplay"     
#> [50] "EconomicSupportIndex"                 
#> [51] "EconomicSupportIndexForDisplay"       

dplyr::glimpse(res)

#> Rows: 209,352
#> Columns: 51
#> $ country                                 <chr> "Aruba", "Aruba", "Aruba", "Ar…
#> $ iso3c                                   <chr> "ABW", "ABW", "ABW", "ABW", "A…
#> $ date                                    <chr> "", "", "", "", "", "", "", ""…
#> $ RegionCode                              <chr> "", "", "", "", "", "", "", ""…
#> $ Jurisdiction                            <chr> "NAT_TOTAL", "NAT_TOTAL", "NAT…
#> $ Date                                    <date> 2020-01-01, 2020-01-02, 2020-…
#> $ `C1_School closing`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C1_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C2_Workplace closing`                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C2_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C3_Cancel public events`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C3_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C4_Restrictions on gatherings`         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C4_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C5_Close public transport`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C5_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C6_Stay at home requirements`          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C6_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C7_Restrictions on internal movement`  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C7_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C8_International travel controls`      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E1_Income support`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ E1_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `E2_Debt/contract relief`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E3_Fiscal measures`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E4_International support`              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H1_Public information campaigns`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H1_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H2_Testing policy`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H3_Contact tracing`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H4_Emergency investment in healthcare` <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H5_Investment in vaccines`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H6_Facial Coverings`                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H6_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H7_Vaccination policy`                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H7_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H8_Protection of elderly people`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H8_Flag                                 <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ M1_Wildcard                             <lgl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ConfirmedCases                          <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ConfirmedDeaths                         <int> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ StringencyIndex                         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyIndexForDisplay               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyLegacyIndex                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyLegacyIndexForDisplay         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ GovernmentResponseIndex                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ GovernmentResponseIndexForDisplay       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ ContainmentHealthIndex                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ ContainmentHealthIndexForDisplay        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ EconomicSupportIndex                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ EconomicSupportIndexForDisplay          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …

summary(res)

#>    country             iso3c               date            RegionCode       
#>  Length:209352      Length:209352      Length:209352      Length:209352     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Jurisdiction            Date            C1_School closing    C1_Flag     
#>  Length:209352      Min.   :2020-01-01   Min.   :0.000     Min.   :0.00   
#>  Class :character   1st Qu.:2020-06-16   1st Qu.:1.000     1st Qu.:1.00   
#>  Mode  :character   Median :2020-12-01   Median :2.000     Median :1.00   
#>                     Mean   :2020-12-01   Mean   :1.636     Mean   :0.77   
#>                     3rd Qu.:2021-05-18   3rd Qu.:3.000     3rd Qu.:1.00   
#>                     Max.   :2021-11-01   Max.   :3.000     Max.   :1.00   
#>                                          NA's   :7257      NA's   :44685  
#>  C2_Workplace closing    C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.000        Min.   :0.00    Min.   :0.000           Min.   :0.00   
#>  1st Qu.:1.000        1st Qu.:1.00    1st Qu.:1.000           1st Qu.:1.00   
#>  Median :2.000        Median :1.00    Median :2.000           Median :1.00   
#>  Mean   :1.389        Mean   :0.78    Mean   :1.333           Mean   :0.85   
#>  3rd Qu.:2.000        3rd Qu.:1.00    3rd Qu.:2.000           3rd Qu.:1.00   
#>  Max.   :3.000        Max.   :1.00    Max.   :2.000           Max.   :1.00   
#>  NA's   :7365         NA's   :52680   NA's   :7287            NA's   :42288  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.00    Min.   :0.000            
#>  1st Qu.:1.000                 1st Qu.:1.00    1st Qu.:0.000            
#>  Median :3.000                 Median :1.00    Median :0.000            
#>  Mean   :2.448                 Mean   :0.81    Mean   :0.554            
#>  3rd Qu.:4.000                 3rd Qu.:1.00    3rd Qu.:1.000            
#>  Max.   :4.000                 Max.   :1.00    Max.   :2.000            
#>  NA's   :7378                  NA's   :57741   NA's   :7385             
#>     C5_Flag       C6_Stay at home requirements    C6_Flag     
#>  Min.   :0.0      Min.   :0.000                Min.   :0.00   
#>  1st Qu.:0.0      1st Qu.:0.000                1st Qu.:1.00   
#>  Median :1.0      Median :1.000                Median :1.00   
#>  Mean   :0.7      Mean   :0.959                Mean   :0.76   
#>  3rd Qu.:1.0      3rd Qu.:2.000                3rd Qu.:1.00   
#>  Max.   :1.0      Max.   :3.000                Max.   :1.00   
#>  NA's   :116943   NA's   :7350                 NA's   :80821  
#>  C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000                        Min.   :0.00   
#>  1st Qu.:0.000                        1st Qu.:0.00   
#>  Median :1.000                        Median :1.00   
#>  Mean   :0.937                        Mean   :0.72   
#>  3rd Qu.:2.000                        3rd Qu.:1.00   
#>  Max.   :2.000                        Max.   :1.00   
#>  NA's   :7479                         NA's   :83109  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.000     Min.   :0.00   
#>  1st Qu.:2.000                    1st Qu.:0.000     1st Qu.:0.00   
#>  Median :3.000                    Median :1.000     Median :0.00   
#>  Mean   :2.502                    Mean   :0.836     Mean   :0.47   
#>  3rd Qu.:3.000                    3rd Qu.:1.000     3rd Qu.:1.00   
#>  Max.   :4.000                    Max.   :2.000     Max.   :1.00   
#>  NA's   :9344                     NA's   :6926      NA's   :82316  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.000           Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.000           1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.000           Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.954           Mean   :1.220e+08   Mean   :6.169e+06       
#>  3rd Qu.:2.000           3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.000           Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :7748            NA's   :62716       NA's   :62820           
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:2.000                   1st Qu.:1.000   1st Qu.:1.000    
#>  Median :2.000                   Median :1.000   Median :2.000    
#>  Mean   :1.823                   Mean   :0.993   Mean   :1.977    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:3.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :7444                    NA's   :20378   NA's   :7603     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.000      Min.   :0.000e+00                    
#>  1st Qu.:1.000      1st Qu.:0.000e+00                    
#>  Median :2.000      Median :0.000e+00                    
#>  Mean   :1.412      Mean   :6.215e+06                    
#>  3rd Qu.:2.000      3rd Qu.:0.000e+00                    
#>  Max.   :2.000      Max.   :2.424e+11                    
#>  NA's   :7847       NA's   :62993                        
#>  H5_Investment in vaccines H6_Facial Coverings    H6_Flag     
#>  Min.   :0.000e+00         Min.   :0.000       Min.   :0.00   
#>  1st Qu.:0.000e+00         1st Qu.:2.000       1st Qu.:1.00   
#>  Median :0.000e+00         Median :3.000       Median :1.00   
#>  Mean   :5.094e+05         Mean   :2.252       Mean   :0.88   
#>  3rd Qu.:0.000e+00         3rd Qu.:3.000       3rd Qu.:1.00   
#>  Max.   :1.575e+10         Max.   :4.000       Max.   :1.00   
#>  NA's   :45790             NA's   :7439        NA's   :44613  
#>  H7_Vaccination policy    H7_Flag       H8_Protection of elderly people
#>  Min.   :0.000         Min.   :0        Min.   :0.000                  
#>  1st Qu.:0.000         1st Qu.:1        1st Qu.:0.000                  
#>  Median :0.000         Median :1        Median :2.000                  
#>  Mean   :1.448         Mean   :1        Mean   :1.509                  
#>  3rd Qu.:3.000         3rd Qu.:1        3rd Qu.:2.000                  
#>  Max.   :5.000         Max.   :1        Max.   :3.000                  
#>  NA's   :7245          NA's   :131293   NA's   :8186                   
#>     H8_Flag      M1_Wildcard    ConfirmedCases     ConfirmedDeaths 
#>  Min.   :0.00    Mode:logical   Min.   :       0   Min.   :     0  
#>  1st Qu.:1.00    NA's:209352    1st Qu.:     866   1st Qu.:     7  
#>  Median :1.00                   Median :   19746   Median :   374  
#>  Mean   :0.87                   Mean   :  392242   Mean   :  8841  
#>  3rd Qu.:1.00                   3rd Qu.:  209536   3rd Qu.:  4409  
#>  Max.   :1.00                   Max.   :45953186   Max.   :745668  
#>  NA's   :59572                  NA's   :10710      NA's   :10785   
#>  StringencyIndex  StringencyIndexForDisplay StringencyLegacyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   :  0.00       
#>  1st Qu.: 37.96   1st Qu.: 37.96            1st Qu.: 50.00       
#>  Median : 54.63   Median : 54.63            Median : 64.29       
#>  Mean   : 51.86   Mean   : 51.83            Mean   : 59.63       
#>  3rd Qu.: 68.98   3rd Qu.: 68.98            3rd Qu.: 75.71       
#>  Max.   :100.00   Max.   :100.00            Max.   :100.00       
#>  NA's   :7418     NA's   :6310              NA's   :7388         
#>  StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 50.00                  1st Qu.:44.27          
#>  Median : 64.29                  Median :55.94          
#>  Mean   : 59.60                  Mean   :51.04          
#>  3rd Qu.: 75.71                  3rd Qu.:64.58          
#>  Max.   :100.00                  Max.   :91.15          
#>  NA's   :6280                    NA's   :7868           
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   : 0.00         
#>  1st Qu.:44.27                     1st Qu.:45.83         
#>  Median :55.94                     Median :57.68         
#>  Mean   :51.04                     Mean   :52.53         
#>  3rd Qu.:64.58                     3rd Qu.:66.07         
#>  Max.   :91.15                     Max.   :93.45         
#>  NA's   :6750                      NA's   :7650          
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   : 0.00                    Min.   :  0.00      
#>  1st Qu.:45.83                    1st Qu.: 12.50      
#>  Median :57.62                    Median : 37.50      
#>  Mean   :52.54                    Mean   : 40.57      
#>  3rd Qu.:66.07                    3rd Qu.: 62.50      
#>  Max.   :93.45                    Max.   :100.00      
#>  NA's   :6532                     NA's   :7982        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 37.50                
#>  Mean   : 40.54                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :6871