Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Value

a data.frame

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#> # A tibble: 6 x 51
#>   country iso3c date  RegionCode Jurisdiction Date       `C1_School closing`
#>   <chr>   <chr> <chr> <chr>      <chr>        <date>                   <dbl>
#> 1 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-01                   0
#> 2 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-02                   0
#> 3 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-03                   0
#> 4 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-04                   0
#> 5 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-05                   0
#> 6 Aruba   ABW   NA    NA         NAT_TOTAL    2020-01-06                   0
#> # … with 44 more variables: C1_Flag <dbl>, C2_Workplace closing <dbl>,
#> #   C2_Flag <dbl>, C3_Cancel public events <dbl>, C3_Flag <dbl>,
#> #   C4_Restrictions on gatherings <dbl>, C4_Flag <dbl>,
#> #   C5_Close public transport <dbl>, C5_Flag <dbl>,
#> #   C6_Stay at home requirements <dbl>, C6_Flag <dbl>,
#> #   C7_Restrictions on internal movement <dbl>, C7_Flag <dbl>,
#> #   C8_International travel controls <dbl>, E1_Income support <dbl>,
#> #   E1_Flag <dbl>, E2_Debt/contract relief <dbl>, E3_Fiscal measures <dbl>,
#> #   E4_International support <dbl>, H1_Public information campaigns <dbl>,
#> #   H1_Flag <dbl>, H2_Testing policy <dbl>, H3_Contact tracing <dbl>,
#> #   H4_Emergency investment in healthcare <dbl>,
#> #   H5_Investment in vaccines <dbl>, H6_Facial Coverings <dbl>, H6_Flag <dbl>,
#> #   H7_Vaccination policy <dbl>, H7_Flag <dbl>,
#> #   H8_Protection of elderly people <dbl>, H8_Flag <dbl>, M1_Wildcard <lgl>,
#> #   ConfirmedCases <dbl>, ConfirmedDeaths <dbl>, StringencyIndex <dbl>,
#> #   StringencyIndexForDisplay <dbl>, StringencyLegacyIndex <dbl>,
#> #   StringencyLegacyIndexForDisplay <dbl>, GovernmentResponseIndex <dbl>,
#> #   GovernmentResponseIndexForDisplay <dbl>, ContainmentHealthIndex <dbl>,
#> #   ContainmentHealthIndexForDisplay <dbl>, EconomicSupportIndex <dbl>,
#> #   EconomicSupportIndexForDisplay <dbl>

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "RegionCode"                           
#>  [5] "Jurisdiction"                         
#>  [6] "Date"                                 
#>  [7] "C1_School closing"                    
#>  [8] "C1_Flag"                              
#>  [9] "C2_Workplace closing"                 
#> [10] "C2_Flag"                              
#> [11] "C3_Cancel public events"              
#> [12] "C3_Flag"                              
#> [13] "C4_Restrictions on gatherings"        
#> [14] "C4_Flag"                              
#> [15] "C5_Close public transport"            
#> [16] "C5_Flag"                              
#> [17] "C6_Stay at home requirements"         
#> [18] "C6_Flag"                              
#> [19] "C7_Restrictions on internal movement" 
#> [20] "C7_Flag"                              
#> [21] "C8_International travel controls"     
#> [22] "E1_Income support"                    
#> [23] "E1_Flag"                              
#> [24] "E2_Debt/contract relief"              
#> [25] "E3_Fiscal measures"                   
#> [26] "E4_International support"             
#> [27] "H1_Public information campaigns"      
#> [28] "H1_Flag"                              
#> [29] "H2_Testing policy"                    
#> [30] "H3_Contact tracing"                   
#> [31] "H4_Emergency investment in healthcare"
#> [32] "H5_Investment in vaccines"            
#> [33] "H6_Facial Coverings"                  
#> [34] "H6_Flag"                              
#> [35] "H7_Vaccination policy"                
#> [36] "H7_Flag"                              
#> [37] "H8_Protection of elderly people"      
#> [38] "H8_Flag"                              
#> [39] "M1_Wildcard"                          
#> [40] "ConfirmedCases"                       
#> [41] "ConfirmedDeaths"                      
#> [42] "StringencyIndex"                      
#> [43] "StringencyIndexForDisplay"            
#> [44] "StringencyLegacyIndex"                
#> [45] "StringencyLegacyIndexForDisplay"      
#> [46] "GovernmentResponseIndex"              
#> [47] "GovernmentResponseIndexForDisplay"    
#> [48] "ContainmentHealthIndex"               
#> [49] "ContainmentHealthIndexForDisplay"     
#> [50] "EconomicSupportIndex"                 
#> [51] "EconomicSupportIndexForDisplay"       

dplyr::glimpse(res)

#> Rows: 147,525
#> Columns: 51
#> $ country                                 <chr> "Aruba", "Aruba", "Aruba", "Ar…
#> $ iso3c                                   <chr> "ABW", "ABW", "ABW", "ABW", "A…
#> $ date                                    <chr> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ RegionCode                              <chr> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ Jurisdiction                            <chr> "NAT_TOTAL", "NAT_TOTAL", "NAT…
#> $ Date                                    <date> 2020-01-01, 2020-01-02, 2020-…
#> $ `C1_School closing`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C2_Workplace closing`                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C2_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C3_Cancel public events`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C3_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C4_Restrictions on gatherings`         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C4_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C5_Close public transport`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C5_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C6_Stay at home requirements`          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C6_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C7_Restrictions on internal movement`  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ C7_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `C8_International travel controls`      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E1_Income support`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ E1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `E2_Debt/contract relief`               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E3_Fiscal measures`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `E4_International support`              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H1_Public information campaigns`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H1_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H2_Testing policy`                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H3_Contact tracing`                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H4_Emergency investment in healthcare` <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H5_Investment in vaccines`             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ `H6_Facial Coverings`                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H6_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H7_Vaccination policy`                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H7_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ `H8_Protection of elderly people`       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ H8_Flag                                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ M1_Wildcard                             <lgl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ConfirmedCases                          <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ ConfirmedDeaths                         <dbl> NA, NA, NA, NA, NA, NA, NA, NA…
#> $ StringencyIndex                         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyIndexForDisplay               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyLegacyIndex                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ StringencyLegacyIndexForDisplay         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ GovernmentResponseIndex                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ GovernmentResponseIndexForDisplay       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ ContainmentHealthIndex                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ ContainmentHealthIndexForDisplay        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ EconomicSupportIndex                    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ EconomicSupportIndexForDisplay          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …

summary(res)

#>    country             iso3c               date            RegionCode       
#>  Length:147525      Length:147525      Length:147525      Length:147525     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Jurisdiction            Date            C1_School closing    C1_Flag     
#>  Length:147525      Min.   :2020-01-01   Min.   :0.000     Min.   :0.00   
#>  Class :character   1st Qu.:2020-05-11   1st Qu.:1.000     1st Qu.:1.00   
#>  Mode  :character   Median :2020-09-19   Median :2.000     Median :1.00   
#>                     Mean   :2020-09-19   Mean   :1.769     Mean   :0.75   
#>                     3rd Qu.:2021-01-28   3rd Qu.:3.000     3rd Qu.:1.00   
#>                     Max.   :2021-06-08   Max.   :3.000     Max.   :1.00   
#>                                          NA's   :5236      NA's   :33721  
#>  C2_Workplace closing    C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.000        Min.   :0.00    Min.   :0.000           Min.   :0.00   
#>  1st Qu.:1.000        1st Qu.:1.00    1st Qu.:1.000           1st Qu.:1.00   
#>  Median :2.000        Median :1.00    Median :2.000           Median :1.00   
#>  Mean   :1.438        Mean   :0.77    Mean   :1.368           Mean   :0.86   
#>  3rd Qu.:2.000        3rd Qu.:1.00    3rd Qu.:2.000           3rd Qu.:1.00   
#>  Max.   :3.000        Max.   :1.00    Max.   :2.000           Max.   :1.00   
#>  NA's   :5328         NA's   :39677   NA's   :5229            NA's   :32770  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.0     Min.   :0.000            
#>  1st Qu.:1.000                 1st Qu.:1.0     1st Qu.:0.000            
#>  Median :3.000                 Median :1.0     Median :0.000            
#>  Mean   :2.551                 Mean   :0.8     Mean   :0.544            
#>  3rd Qu.:4.000                 3rd Qu.:1.0     3rd Qu.:1.000            
#>  Max.   :4.000                 Max.   :1.0     Max.   :2.000            
#>  NA's   :5235                  NA's   :39903   NA's   :5284             
#>     C5_Flag      C6_Stay at home requirements    C6_Flag     
#>  Min.   :0.00    Min.   :0                    Min.   :0.00   
#>  1st Qu.:0.00    1st Qu.:0                    1st Qu.:1.00   
#>  Median :1.00    Median :1                    Median :1.00   
#>  Mean   :0.66    Mean   :1                    Mean   :0.75   
#>  3rd Qu.:1.00    3rd Qu.:2                    3rd Qu.:1.00   
#>  Max.   :1.00    Max.   :3                    Max.   :1.00   
#>  NA's   :85341   NA's   :5321                 NA's   :54330  
#>  C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000                        Min.   :0.00   
#>  1st Qu.:0.000                        1st Qu.:0.00   
#>  Median :1.000                        Median :1.00   
#>  Mean   :0.944                        Mean   :0.68   
#>  3rd Qu.:2.000                        3rd Qu.:1.00   
#>  Max.   :2.000                        Max.   :1.00   
#>  NA's   :5319                         NA's   :62237  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.000     Min.   :0.00   
#>  1st Qu.:1.000                    1st Qu.:0.000     1st Qu.:0.00   
#>  Median :3.000                    Median :1.000     Median :0.00   
#>  Mean   :2.478                    Mean   :0.848     Mean   :0.46   
#>  3rd Qu.:3.000                    3rd Qu.:2.000     3rd Qu.:1.00   
#>  Max.   :4.000                    Max.   :2.000     Max.   :1.00   
#>  NA's   :6680                     NA's   :7761      NA's   :64022  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.000           Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.000           1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.000           Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.974           Mean   :1.392e+08   Mean   :7.918e+06       
#>  3rd Qu.:2.000           3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.000           Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :7980            NA's   :33242       NA's   :33241           
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:2.000                   1st Qu.:1.000   1st Qu.:1.000    
#>  Median :2.000                   Median :1.000   Median :2.000    
#>  Mean   :1.773                   Mean   :0.991   Mean   :1.777    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:3.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :5651                    NA's   :17740   NA's   :5811     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.00       Min.   :0.000e+00                    
#>  1st Qu.:1.00       1st Qu.:0.000e+00                    
#>  Median :2.00       Median :0.000e+00                    
#>  Mean   :1.34       Mean   :7.733e+06                    
#>  3rd Qu.:2.00       3rd Qu.:0.000e+00                    
#>  Max.   :2.00       Max.   :2.424e+11                    
#>  NA's   :6126       NA's   :33500                        
#>  H5_Investment in vaccines H6_Facial Coverings    H6_Flag     
#>  Min.   :0.000e+00         Min.   :0.000       Min.   :0.00   
#>  1st Qu.:0.000e+00         1st Qu.:0.000       1st Qu.:1.00   
#>  Median :0.000e+00         Median :3.000       Median :1.00   
#>  Mean   :6.905e+05         Mean   :2.106       Mean   :0.85   
#>  3rd Qu.:0.000e+00         3rd Qu.:3.000       3rd Qu.:1.00   
#>  Max.   :1.575e+10         Max.   :4.000       Max.   :1.00   
#>  NA's   :33315             NA's   :7449        NA's   :42828  
#>  H7_Vaccination policy    H7_Flag       H8_Protection of elderly people
#>  Min.   :0.000         Min.   :0        Min.   :0.000                  
#>  1st Qu.:0.000         1st Qu.:1        1st Qu.:0.000                  
#>  Median :0.000         Median :1        Median :2.000                  
#>  Mean   :0.749         Mean   :1        Mean   :1.398                  
#>  3rd Qu.:1.000         3rd Qu.:1        3rd Qu.:2.000                  
#>  Max.   :5.000         Max.   :1        Max.   :3.000                  
#>  NA's   :24324         NA's   :115928   NA's   :6782                   
#>     H8_Flag      M1_Wildcard    ConfirmedCases     ConfirmedDeaths 
#>  Min.   :0.00    Mode:logical   Min.   :       0   Min.   :     0  
#>  1st Qu.:1.00    NA's:147525    1st Qu.:     741   1st Qu.:    10  
#>  Median :1.00                   Median :   16103   Median :   316  
#>  Mean   :0.86                   Mean   :  273018   Mean   :  6472  
#>  3rd Qu.:1.00                   3rd Qu.:  145759   3rd Qu.:  3130  
#>  Max.   :1.00                   Max.   :33362600   Max.   :597628  
#>  NA's   :50480                  NA's   :9645       NA's   :9724    
#>  StringencyIndex  StringencyIndexForDisplay StringencyLegacyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   :  0.00       
#>  1st Qu.: 38.89   1st Qu.: 38.89            1st Qu.: 50.00       
#>  Median : 57.41   Median : 57.41            Median : 65.48       
#>  Mean   : 52.72   Mean   : 52.73            Mean   : 59.37       
#>  3rd Qu.: 71.30   3rd Qu.: 71.30            3rd Qu.: 77.38       
#>  Max.   :100.00   Max.   :100.00            Max.   :100.00       
#>  NA's   :5637     NA's   :4935              NA's   :5569         
#>  StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 50.00                  1st Qu.:40.62          
#>  Median : 65.48                  Median :56.25          
#>  Mean   : 59.38                  Mean   :49.41          
#>  3rd Qu.: 77.38                  3rd Qu.:65.10          
#>  Max.   :100.00                  Max.   :89.84          
#>  NA's   :4867                    NA's   :7726           
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   : 0.00         
#>  1st Qu.:40.62                     1st Qu.:41.67         
#>  Median :56.25                     Median :57.26         
#>  Mean   :49.44                     Mean   :50.63         
#>  3rd Qu.:65.10                     3rd Qu.:66.07         
#>  Max.   :89.84                     Max.   :91.96         
#>  NA's   :7005                      NA's   :6012          
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   : 0.00                    Min.   :  0.00      
#>  1st Qu.:41.67                    1st Qu.:  0.00      
#>  Median :57.32                    Median : 37.50      
#>  Mean   :50.67                    Mean   : 41.51      
#>  3rd Qu.:66.07                    3rd Qu.: 62.50      
#>  Max.   :91.96                    Max.   :100.00      
#>  NA's   :5293                     NA's   :8103        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 37.50                
#>  Mean   : 41.52                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :7391