Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Value

a data.frame

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)
#>    country iso3c date RegionCode Jurisdiction       Date C1_School closing
#> 1:   Aruba   ABW                    NAT_TOTAL 2020-01-01                 0
#> 2:   Aruba   ABW                    NAT_TOTAL 2020-01-02                 0
#> 3:   Aruba   ABW                    NAT_TOTAL 2020-01-03                 0
#> 4:   Aruba   ABW                    NAT_TOTAL 2020-01-04                 0
#> 5:   Aruba   ABW                    NAT_TOTAL 2020-01-05                 0
#> 6:   Aruba   ABW                    NAT_TOTAL 2020-01-06                 0
#>    C1_Flag C2_Workplace closing C2_Flag C3_Cancel public events C3_Flag
#> 1:      NA                    0      NA                       0      NA
#> 2:      NA                    0      NA                       0      NA
#> 3:      NA                    0      NA                       0      NA
#> 4:      NA                    0      NA                       0      NA
#> 5:      NA                    0      NA                       0      NA
#> 6:      NA                    0      NA                       0      NA
#>    C4_Restrictions on gatherings C4_Flag C5_Close public transport C5_Flag
#> 1:                             0      NA                         0      NA
#> 2:                             0      NA                         0      NA
#> 3:                             0      NA                         0      NA
#> 4:                             0      NA                         0      NA
#> 5:                             0      NA                         0      NA
#> 6:                             0      NA                         0      NA
#>    C6_Stay at home requirements C6_Flag C7_Restrictions on internal movement
#> 1:                            0      NA                                    0
#> 2:                            0      NA                                    0
#> 3:                            0      NA                                    0
#> 4:                            0      NA                                    0
#> 5:                            0      NA                                    0
#> 6:                            0      NA                                    0
#>    C7_Flag C8_International travel controls E1_Income support E1_Flag
#> 1:      NA                                0                 0      NA
#> 2:      NA                                0                 0      NA
#> 3:      NA                                0                 0      NA
#> 4:      NA                                0                 0      NA
#> 5:      NA                                0                 0      NA
#> 6:      NA                                0                 0      NA
#>    E2_Debt/contract relief E3_Fiscal measures E4_International support
#> 1:                       0                  0                        0
#> 2:                       0                  0                        0
#> 3:                       0                  0                        0
#> 4:                       0                  0                        0
#> 5:                       0                  0                        0
#> 6:                       0                  0                        0
#>    H1_Public information campaigns H1_Flag H2_Testing policy H3_Contact tracing
#> 1:                               0      NA                 0                  0
#> 2:                               0      NA                 0                  0
#> 3:                               0      NA                 0                  0
#> 4:                               0      NA                 0                  0
#> 5:                               0      NA                 0                  0
#> 6:                               0      NA                 0                  0
#>    H4_Emergency investment in healthcare H5_Investment in vaccines
#> 1:                                     0                         0
#> 2:                                     0                         0
#> 3:                                     0                         0
#> 4:                                     0                         0
#> 5:                                     0                         0
#> 6:                                     0                         0
#>    H6_Facial Coverings H6_Flag H7_Vaccination policy H7_Flag
#> 1:                   0      NA                     0      NA
#> 2:                   0      NA                     0      NA
#> 3:                   0      NA                     0      NA
#> 4:                   0      NA                     0      NA
#> 5:                   0      NA                     0      NA
#> 6:                   0      NA                     0      NA
#>    H8_Protection of elderly people H8_Flag M1_Wildcard
#> 1:                               0      NA          NA
#> 2:                               0      NA          NA
#> 3:                               0      NA          NA
#> 4:                               0      NA          NA
#> 5:                               0      NA          NA
#> 6:                               0      NA          NA
#>    V1_Vaccine Prioritisation (summary) V2A_Vaccine Availability (summary)
#> 1:                                   0                                  0
#> 2:                                   0                                  0
#> 3:                                   0                                  0
#> 4:                                   0                                  0
#> 5:                                   0                                  0
#> 6:                                   0                                  0
#>    V2B_Vaccine age eligibility/availability age floor (general population summary)
#> 1:                                                                                
#> 2:                                                                                
#> 3:                                                                                
#> 4:                                                                                
#> 5:                                                                                
#> 6:                                                                                
#>    V2C_Vaccine age eligibility/availability age floor (at risk summary)
#> 1:                                                                     
#> 2:                                                                     
#> 3:                                                                     
#> 4:                                                                     
#> 5:                                                                     
#> 6:                                                                     
#>    V2D_Medically/ clinically vulnerable (Non-elderly) V2E_Education
#> 1:                                                 NA            NA
#> 2:                                                 NA            NA
#> 3:                                                 NA            NA
#> 4:                                                 NA            NA
#> 5:                                                 NA            NA
#> 6:                                                 NA            NA
#>    V2F_Frontline workers  (non healthcare) V2G_Frontline workers  (healthcare)
#> 1:                                      NA                                  NA
#> 2:                                      NA                                  NA
#> 3:                                      NA                                  NA
#> 4:                                      NA                                  NA
#> 5:                                      NA                                  NA
#> 6:                                      NA                                  NA
#>    V3_Vaccine Financial Support (summary) ConfirmedCases ConfirmedDeaths
#> 1:                                      0             NA              NA
#> 2:                                      0             NA              NA
#> 3:                                      0             NA              NA
#> 4:                                      0             NA              NA
#> 5:                                      0             NA              NA
#> 6:                                      0             NA              NA
#>    StringencyIndex StringencyIndexForDisplay StringencyLegacyIndex
#> 1:               0                         0                     0
#> 2:               0                         0                     0
#> 3:               0                         0                     0
#> 4:               0                         0                     0
#> 5:               0                         0                     0
#> 6:               0                         0                     0
#>    StringencyLegacyIndexForDisplay GovernmentResponseIndex
#> 1:                               0                       0
#> 2:                               0                       0
#> 3:                               0                       0
#> 4:                               0                       0
#> 5:                               0                       0
#> 6:                               0                       0
#>    GovernmentResponseIndexForDisplay ContainmentHealthIndex
#> 1:                                 0                      0
#> 2:                                 0                      0
#> 3:                                 0                      0
#> 4:                                 0                      0
#> 5:                                 0                      0
#> 6:                                 0                      0
#>    ContainmentHealthIndexForDisplay EconomicSupportIndex
#> 1:                                0                    0
#> 2:                                0                    0
#> 3:                                0                    0
#> 4:                                0                    0
#> 5:                                0                    0
#> 6:                                0                    0
#>    EconomicSupportIndexForDisplay
#> 1:                              0
#> 2:                              0
#> 3:                              0
#> 4:                              0
#> 5:                              0
#> 6:                              0
colnames(res)
#>  [1] "country"                                                                        
#>  [2] "iso3c"                                                                          
#>  [3] "date"                                                                           
#>  [4] "RegionCode"                                                                     
#>  [5] "Jurisdiction"                                                                   
#>  [6] "Date"                                                                           
#>  [7] "C1_School closing"                                                              
#>  [8] "C1_Flag"                                                                        
#>  [9] "C2_Workplace closing"                                                           
#> [10] "C2_Flag"                                                                        
#> [11] "C3_Cancel public events"                                                        
#> [12] "C3_Flag"                                                                        
#> [13] "C4_Restrictions on gatherings"                                                  
#> [14] "C4_Flag"                                                                        
#> [15] "C5_Close public transport"                                                      
#> [16] "C5_Flag"                                                                        
#> [17] "C6_Stay at home requirements"                                                   
#> [18] "C6_Flag"                                                                        
#> [19] "C7_Restrictions on internal movement"                                           
#> [20] "C7_Flag"                                                                        
#> [21] "C8_International travel controls"                                               
#> [22] "E1_Income support"                                                              
#> [23] "E1_Flag"                                                                        
#> [24] "E2_Debt/contract relief"                                                        
#> [25] "E3_Fiscal measures"                                                             
#> [26] "E4_International support"                                                       
#> [27] "H1_Public information campaigns"                                                
#> [28] "H1_Flag"                                                                        
#> [29] "H2_Testing policy"                                                              
#> [30] "H3_Contact tracing"                                                             
#> [31] "H4_Emergency investment in healthcare"                                          
#> [32] "H5_Investment in vaccines"                                                      
#> [33] "H6_Facial Coverings"                                                            
#> [34] "H6_Flag"                                                                        
#> [35] "H7_Vaccination policy"                                                          
#> [36] "H7_Flag"                                                                        
#> [37] "H8_Protection of elderly people"                                                
#> [38] "H8_Flag"                                                                        
#> [39] "M1_Wildcard"                                                                    
#> [40] "V1_Vaccine Prioritisation (summary)"                                            
#> [41] "V2A_Vaccine Availability (summary)"                                             
#> [42] "V2B_Vaccine age eligibility/availability age floor (general population summary)"
#> [43] "V2C_Vaccine age eligibility/availability age floor (at risk summary)"           
#> [44] "V2D_Medically/ clinically vulnerable (Non-elderly)"                             
#> [45] "V2E_Education"                                                                  
#> [46] "V2F_Frontline workers  (non healthcare)"                                        
#> [47] "V2G_Frontline workers  (healthcare)"                                            
#> [48] "V3_Vaccine Financial Support (summary)"                                         
#> [49] "ConfirmedCases"                                                                 
#> [50] "ConfirmedDeaths"                                                                
#> [51] "StringencyIndex"                                                                
#> [52] "StringencyIndexForDisplay"                                                      
#> [53] "StringencyLegacyIndex"                                                          
#> [54] "StringencyLegacyIndexForDisplay"                                                
#> [55] "GovernmentResponseIndex"                                                        
#> [56] "GovernmentResponseIndexForDisplay"                                              
#> [57] "ContainmentHealthIndex"                                                         
#> [58] "ContainmentHealthIndexForDisplay"                                               
#> [59] "EconomicSupportIndex"                                                           
#> [60] "EconomicSupportIndexForDisplay"                                                 
dplyr::glimpse(res)
#> Rows: 240,240
#> Columns: 60
#> $ country                                                                           <chr> …
#> $ iso3c                                                                             <chr> …
#> $ date                                                                              <chr> …
#> $ RegionCode                                                                        <chr> …
#> $ Jurisdiction                                                                      <chr> …
#> $ Date                                                                              <date> …
#> $ `C1_School closing`                                                               <dbl> …
#> $ C1_Flag                                                                           <int> …
#> $ `C2_Workplace closing`                                                            <dbl> …
#> $ C2_Flag                                                                           <int> …
#> $ `C3_Cancel public events`                                                         <dbl> …
#> $ C3_Flag                                                                           <int> …
#> $ `C4_Restrictions on gatherings`                                                   <dbl> …
#> $ C4_Flag                                                                           <int> …
#> $ `C5_Close public transport`                                                       <dbl> …
#> $ C5_Flag                                                                           <int> …
#> $ `C6_Stay at home requirements`                                                    <dbl> …
#> $ C6_Flag                                                                           <int> …
#> $ `C7_Restrictions on internal movement`                                            <dbl> …
#> $ C7_Flag                                                                           <int> …
#> $ `C8_International travel controls`                                                <dbl> …
#> $ `E1_Income support`                                                               <dbl> …
#> $ E1_Flag                                                                           <int> …
#> $ `E2_Debt/contract relief`                                                         <dbl> …
#> $ `E3_Fiscal measures`                                                              <dbl> …
#> $ `E4_International support`                                                        <dbl> …
#> $ `H1_Public information campaigns`                                                 <dbl> …
#> $ H1_Flag                                                                           <int> …
#> $ `H2_Testing policy`                                                               <dbl> …
#> $ `H3_Contact tracing`                                                              <dbl> …
#> $ `H4_Emergency investment in healthcare`                                           <dbl> …
#> $ `H5_Investment in vaccines`                                                       <dbl> …
#> $ `H6_Facial Coverings`                                                             <dbl> …
#> $ H6_Flag                                                                           <int> …
#> $ `H7_Vaccination policy`                                                           <dbl> …
#> $ H7_Flag                                                                           <int> …
#> $ `H8_Protection of elderly people`                                                 <dbl> …
#> $ H8_Flag                                                                           <int> …
#> $ M1_Wildcard                                                                       <lgl> …
#> $ `V1_Vaccine Prioritisation (summary)`                                             <int> …
#> $ `V2A_Vaccine Availability (summary)`                                              <int> …
#> $ `V2B_Vaccine age eligibility/availability age floor (general population summary)` <chr> …
#> $ `V2C_Vaccine age eligibility/availability age floor (at risk summary)`            <chr> …
#> $ `V2D_Medically/ clinically vulnerable (Non-elderly)`                              <int> …
#> $ V2E_Education                                                                     <int> …
#> $ `V2F_Frontline workers  (non healthcare)`                                         <int> …
#> $ `V2G_Frontline workers  (healthcare)`                                             <int> …
#> $ `V3_Vaccine Financial Support (summary)`                                          <int> …
#> $ ConfirmedCases                                                                    <int> …
#> $ ConfirmedDeaths                                                                   <int> …
#> $ StringencyIndex                                                                   <dbl> …
#> $ StringencyIndexForDisplay                                                         <dbl> …
#> $ StringencyLegacyIndex                                                             <dbl> …
#> $ StringencyLegacyIndexForDisplay                                                   <dbl> …
#> $ GovernmentResponseIndex                                                           <dbl> …
#> $ GovernmentResponseIndexForDisplay                                                 <dbl> …
#> $ ContainmentHealthIndex                                                            <dbl> …
#> $ ContainmentHealthIndexForDisplay                                                  <dbl> …
#> $ EconomicSupportIndex                                                              <dbl> …
#> $ EconomicSupportIndexForDisplay                                                    <dbl> …
summary(res)
#>    country             iso3c               date            RegionCode       
#>  Length:240240      Length:240240      Length:240240      Length:240240     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Jurisdiction            Date            C1_School closing    C1_Flag     
#>  Length:240240      Min.   :2020-01-01   Min.   :0.000     Min.   :0.00   
#>  Class :character   1st Qu.:2020-07-11   1st Qu.:1.000     1st Qu.:1.00   
#>  Mode  :character   Median :2021-01-19   Median :1.000     Median :1.00   
#>                     Mean   :2021-01-19   Mean   :1.565     Mean   :0.77   
#>                     3rd Qu.:2021-07-31   3rd Qu.:3.000     3rd Qu.:1.00   
#>                     Max.   :2022-02-08   Max.   :3.000     Max.   :1.00   
#>                                          NA's   :5755      NA's   :49950  
#>  C2_Workplace closing    C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.000        Min.   :0.00    Min.   :0.00            Min.   :0.00   
#>  1st Qu.:1.000        1st Qu.:1.00    1st Qu.:1.00            1st Qu.:1.00   
#>  Median :1.000        Median :1.00    Median :1.00            Median :1.00   
#>  Mean   :1.357        Mean   :0.78    Mean   :1.31            Mean   :0.85   
#>  3rd Qu.:2.000        3rd Qu.:1.00    3rd Qu.:2.00            3rd Qu.:1.00   
#>  Max.   :3.000        Max.   :1.00    Max.   :2.00            Max.   :1.00   
#>  NA's   :5799         NA's   :58331   NA's   :5816            NA's   :47808  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.0     Min.   :0.000            
#>  1st Qu.:0.000                 1st Qu.:1.0     1st Qu.:0.000            
#>  Median :3.000                 Median :1.0     Median :0.000            
#>  Mean   :2.388                 Mean   :0.8     Mean   :0.509            
#>  3rd Qu.:4.000                 3rd Qu.:1.0     3rd Qu.:1.000            
#>  Max.   :4.000                 Max.   :1.0     Max.   :2.000            
#>  NA's   :5992                  NA's   :65975   NA's   :5904             
#>     C5_Flag       C6_Stay at home requirements    C6_Flag     
#>  Min.   :0.00     Min.   :0.000                Min.   :0.00   
#>  1st Qu.:0.00     1st Qu.:0.000                1st Qu.:0.00   
#>  Median :1.00     Median :1.000                Median :1.00   
#>  Mean   :0.69     Mean   :0.911                Mean   :0.75   
#>  3rd Qu.:1.00     3rd Qu.:2.000                3rd Qu.:1.00   
#>  Max.   :1.00     Max.   :3.000                Max.   :1.00   
#>  NA's   :142009   NA's   :5910                 NA's   :98016  
#>  C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000                        Min.   :0.00   
#>  1st Qu.:0.000                        1st Qu.:0.00   
#>  Median :1.000                        Median :1.00   
#>  Mean   :0.918                        Mean   :0.73   
#>  3rd Qu.:2.000                        3rd Qu.:1.00   
#>  Max.   :2.000                        Max.   :1.00   
#>  NA's   :10005                        NA's   :98873  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.000     Min.   :0.00   
#>  1st Qu.:1.000                    1st Qu.:0.000     1st Qu.:0.00   
#>  Median :3.000                    Median :1.000     Median :0.00   
#>  Mean   :2.464                    Mean   :0.803     Mean   :0.48   
#>  3rd Qu.:3.000                    3rd Qu.:1.000     3rd Qu.:1.00   
#>  Max.   :4.000                    Max.   :2.000     Max.   :1.00   
#>  NA's   :7128                     NA's   :6488      NA's   :98087  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.000           Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.000           1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.000           Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.926           Mean   :1.220e+08   Mean   :6.169e+06       
#>  3rd Qu.:2.000           3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.000           Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :6472            NA's   :93604       NA's   :93708           
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:2.000                   1st Qu.:1.000   1st Qu.:1.000    
#>  Median :2.000                   Median :1.000   Median :2.000    
#>  Mean   :1.838                   Mean   :0.994   Mean   :2.063    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:3.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :10004                   NA's   :23132   NA's   :10166    
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.000      Min.   :0.000e+00                    
#>  1st Qu.:1.000      1st Qu.:0.000e+00                    
#>  Median :2.000      Median :0.000e+00                    
#>  Mean   :1.408      Mean   :6.215e+06                    
#>  3rd Qu.:2.000      3rd Qu.:0.000e+00                    
#>  Max.   :2.000      Max.   :2.424e+11                    
#>  NA's   :10881      NA's   :93881                        
#>  H5_Investment in vaccines H6_Facial Coverings    H6_Flag     
#>  Min.   :0.000e+00         Min.   :0.000       Min.   :0.00   
#>  1st Qu.:0.000e+00         1st Qu.:2.000       1st Qu.:1.00   
#>  Median :0.000e+00         Median :3.000       Median :1.00   
#>  Mean   :4.596e+05         Mean   :2.297       Mean   :0.88   
#>  3rd Qu.:0.000e+00         3rd Qu.:3.000       3rd Qu.:1.00   
#>  Max.   :1.575e+10         Max.   :4.000       Max.   :1.00   
#>  NA's   :57609             NA's   :9971        NA's   :47553  
#>  H7_Vaccination policy    H7_Flag       H8_Protection of elderly people
#>  Min.   :0.000         Min.   :0        Min.   :0.000                  
#>  1st Qu.:0.000         1st Qu.:1        1st Qu.:0.000                  
#>  Median :0.000         Median :1        Median :2.000                  
#>  Mean   :1.851         Mean   :1        Mean   :1.487                  
#>  3rd Qu.:4.000         3rd Qu.:1        3rd Qu.:2.000                  
#>  Max.   :5.000         Max.   :1        Max.   :3.000                  
#>  NA's   :9963          NA's   :134159   NA's   :10733                  
#>     H8_Flag      M1_Wildcard    V1_Vaccine Prioritisation (summary)
#>  Min.   :0.00    Mode:logical   Min.   :0.00                       
#>  1st Qu.:1.00    NA's:240240    1st Qu.:0.00                       
#>  Median :1.00                   Median :0.00                       
#>  Mean   :0.88                   Mean   :0.71                       
#>  3rd Qu.:1.00                   3rd Qu.:1.00                       
#>  Max.   :1.00                   Max.   :2.00                       
#>  NA's   :69168                  NA's   :99921                      
#>  V2A_Vaccine Availability (summary)
#>  Min.   :0.00                      
#>  1st Qu.:0.00                      
#>  Median :0.00                      
#>  Mean   :0.81                      
#>  3rd Qu.:1.00                      
#>  Max.   :3.00                      
#>  NA's   :99956                     
#>  V2B_Vaccine age eligibility/availability age floor (general population summary)
#>  Length:240240                                                                  
#>  Class :character                                                               
#>  Mode  :character                                                               
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>  V2C_Vaccine age eligibility/availability age floor (at risk summary)
#>  Length:240240                                                       
#>  Class :character                                                    
#>  Mode  :character                                                    
#>                                                                      
#>                                                                      
#>                                                                      
#>                                                                      
#>  V2D_Medically/ clinically vulnerable (Non-elderly) V2E_Education   
#>  Min.   :0.0                                        Min.   :0.00    
#>  1st Qu.:1.0                                        1st Qu.:0.00    
#>  Median :2.0                                        Median :2.00    
#>  Mean   :1.4                                        Mean   :1.28    
#>  3rd Qu.:2.0                                        3rd Qu.:2.00    
#>  Max.   :2.0                                        Max.   :2.00    
#>  NA's   :157906                                     NA's   :157906  
#>  V2F_Frontline workers  (non healthcare) V2G_Frontline workers  (healthcare)
#>  Min.   :0.00                            Min.   :0.00                       
#>  1st Qu.:2.00                            1st Qu.:2.00                       
#>  Median :2.00                            Median :2.00                       
#>  Mean   :1.57                            Mean   :1.73                       
#>  3rd Qu.:2.00                            3rd Qu.:2.00                       
#>  Max.   :2.00                            Max.   :2.00                       
#>  NA's   :157906                          NA's   :157906                     
#>  V3_Vaccine Financial Support (summary) ConfirmedCases     ConfirmedDeaths 
#>  Min.   :0.00                           Min.   :       0   Min.   :     0  
#>  1st Qu.:0.00                           1st Qu.:    1046   1st Qu.:     9  
#>  Median :0.00                           Median :   30514   Median :   582  
#>  Mean   :2.23                           Mean   :  504700   Mean   : 10614  
#>  3rd Qu.:5.00                           3rd Qu.:  277748   3rd Qu.:  5498  
#>  Max.   :5.00                           Max.   :76511821   Max.   :902635  
#>  NA's   :99985                          NA's   :10260      NA's   :10394   
#>  StringencyIndex  StringencyIndexForDisplay StringencyLegacyIndex
#>  Min.   :  0.00   Min.   :  0.00            Min.   :  0.00       
#>  1st Qu.: 37.96   1st Qu.: 37.96            1st Qu.: 47.14       
#>  Median : 52.78   Median : 52.78            Median : 63.10       
#>  Mean   : 50.82   Mean   : 50.81            Mean   : 58.35       
#>  3rd Qu.: 67.59   3rd Qu.: 67.59            3rd Qu.: 73.81       
#>  Max.   :100.00   Max.   :100.00            Max.   :100.00       
#>  NA's   :9999     NA's   :9158              NA's   :5892         
#>  StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 47.14                  1st Qu.:43.75          
#>  Median : 63.10                  Median :54.95          
#>  Mean   : 58.34                  Mean   :50.96          
#>  3rd Qu.: 73.81                  3rd Qu.:64.06          
#>  Max.   :100.00                  Max.   :91.15          
#>  NA's   :5051                    NA's   :10629          
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   : 0.00         
#>  1st Qu.:43.75                     1st Qu.:45.54         
#>  Median :54.95                     Median :56.90         
#>  Mean   :50.96                     Mean   :52.65         
#>  3rd Qu.:64.06                     3rd Qu.:65.77         
#>  Max.   :91.15                     Max.   :93.45         
#>  NA's   :9776                      NA's   :10341         
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   : 0.00                    Min.   :  0.00      
#>  1st Qu.:45.54                    1st Qu.:  0.00      
#>  Median :56.90                    Median : 37.50      
#>  Mean   :52.65                    Mean   : 39.23      
#>  3rd Qu.:65.77                    3rd Qu.: 62.50      
#>  Max.   :93.45                    Max.   :100.00      
#>  NA's   :9494                     NA's   :6949        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.0                 
#>  1st Qu.:  0.0                 
#>  Median : 37.5                 
#>  Mean   : 39.2                 
#>  3rd Qu.: 62.5                 
#>  Max.   :100.0                 
#>  NA's   :6088