Longitudinal tracking of government interventions to combat COVID-19

Governments are taking a wide range of measures in response to the COVID-19 outbreak. The Oxford COVID-19 Government Response Tracker (OxCGRT) aims to track and compare government responses to the coronavirus outbreak worldwide rigorously and consistently.

government_policy_timeline()

Source

https://www.bsg.ox.ac.uk/research/research-projects/oxford-covid-19-government-response-tracker

Value

a data.frame

Details

Systematic information on which governments have taken which measures, and when, can help decision-makers and citizens understand the robustness of governmental responses in a consistent way, aiding efforts to fight the pandemic. The OxCGRT systematically collects information on several different common policy responses governments have taken, scores the stringency of such measures, and aggregates these scores into a common Stringency Index.

Data is collected from public sources by a team of dozens of Oxford University students and staff from every part of the world.

OxCGRT collects publicly available information on 11 indicators of government response (S1-S11). The first seven indicators (S1-S7) take policies such as school closures, travel bans, etc. are recorded on an ordinal scale; the remainder (S8-S11) are financial indicators such as fiscal or monetary measures. For a full description of the data and how it is collected, see this working paper.

Recommended citation for database: Hale, Thomas and Samuel Webster (2020). Oxford COVID-19 Government Response Tracker. Data use policy: Creative Commons Attribution CC BY standard.

See also

Author

Sean Davis seandavi@gmail.com

Examples

library(dplyr)
res = government_policy_timeline()
head(res)

#> Error in gsub(finish, start, ..., fixed = TRUE): zero-length pattern

colnames(res)

#>  [1] "country"                              
#>  [2] "iso3c"                                
#>  [3] "date"                                 
#>  [4] "RegionCode"                           
#>  [5] "Jurisdiction"                         
#>  [6] "Date"                                 
#>  [7] "C1_School closing"                    
#>  [8] "C1_Flag"                              
#>  [9] "C2_Workplace closing"                 
#> [10] "C2_Flag"                              
#> [11] "C3_Cancel public events"              
#> [12] "C3_Flag"                              
#> [13] "C4_Restrictions on gatherings"        
#> [14] "C4_Flag"                              
#> [15] "C5_Close public transport"            
#> [16] "C5_Flag"                              
#> [17] "C6_Stay at home requirements"         
#> [18] "C6_Flag"                              
#> [19] "C7_Restrictions on internal movement" 
#> [20] "C7_Flag"                              
#> [21] "C8_International travel controls"     
#> [22] "E1_Income support"                    
#> [23] "E1_Flag"                              
#> [24] "E2_Debt/contract relief"              
#> [25] "E3_Fiscal measures"                   
#> [26] "E4_International support"             
#> [27] "H1_Public information campaigns"      
#> [28] "H1_Flag"                              
#> [29] "H2_Testing policy"                    
#> [30] "H3_Contact tracing"                   
#> [31] "H4_Emergency investment in healthcare"
#> [32] "H5_Investment in vaccines"            
#> [33] "H6_Facial Coverings"                  
#> [34] "H6_Flag"                              
#> [35] "H7_Vaccination policy"                
#> [36] "H7_Flag"                              
#> [37] "M1_Wildcard"                          
#> [38] "ConfirmedCases"                       
#> [39] "ConfirmedDeaths"                      
#> [40] "StringencyIndex"                      
#> [41] "StringencyIndexForDisplay"            
#> [42] "StringencyLegacyIndex"                
#> [43] "StringencyLegacyIndexForDisplay"      
#> [44] "GovernmentResponseIndex"              
#> [45] "GovernmentResponseIndexForDisplay"    
#> [46] "ContainmentHealthIndex"               
#> [47] "ContainmentHealthIndexForDisplay"     
#> [48] "EconomicSupportIndex"                 
#> [49] "EconomicSupportIndexForDisplay"       

dplyr::glimpse(res)

#> Rows: 115,080
#> Columns: 49

#> Error in gsub(finish, start, ..., fixed = TRUE): zero-length pattern

summary(res)

#>    country             iso3c               date            RegionCode       
#>  Length:115080      Length:115080      Length:115080      Length:115080     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  Jurisdiction            Date            C1_School closing    C1_Flag     
#>  Length:115080      Min.   :2020-01-01   Min.   :0.000     Min.   :0.000  
#>  Class :character   1st Qu.:2020-04-12   1st Qu.:1.000     1st Qu.:1.000  
#>  Mode  :character   Median :2020-07-24   Median :2.000     Median :1.000  
#>                     Mean   :2020-07-24   Mean   :1.815     Mean   :0.786  
#>                     3rd Qu.:2020-11-04   3rd Qu.:3.000     3rd Qu.:1.000  
#>                     Max.   :2021-02-14   Max.   :3.000     Max.   :1.000  
#>                                          NA's   :4159      NA's   :28912  
#>  C2_Workplace closing    C2_Flag      C3_Cancel public events    C3_Flag     
#>  Min.   :0.000        Min.   :0.00    Min.   :0.000           Min.   :0.000  
#>  1st Qu.:0.000        1st Qu.:1.00    1st Qu.:1.000           1st Qu.:1.000  
#>  Median :2.000        Median :1.00    Median :2.000           Median :1.000  
#>  Mean   :1.388        Mean   :0.77    Mean   :1.337           Mean   :0.868  
#>  3rd Qu.:2.000        3rd Qu.:1.00    3rd Qu.:2.000           3rd Qu.:1.000  
#>  Max.   :3.000        Max.   :1.00    Max.   :2.000           Max.   :1.000  
#>  NA's   :4195         NA's   :35817   NA's   :4187            NA's   :29468  
#>  C4_Restrictions on gatherings    C4_Flag      C5_Close public transport
#>  Min.   :0.000                 Min.   :0.00    Min.   :0.000            
#>  1st Qu.:0.000                 1st Qu.:1.00    1st Qu.:0.000            
#>  Median :3.000                 Median :1.00    Median :0.000            
#>  Mean   :2.422                 Mean   :0.81    Mean   :0.542            
#>  3rd Qu.:4.000                 3rd Qu.:1.00    3rd Qu.:1.000            
#>  Max.   :4.000                 Max.   :1.00    Max.   :2.000            
#>  NA's   :4180                  NA's   :34451   NA's   :4160             
#>     C5_Flag      C6_Stay at home requirements    C6_Flag     
#>  Min.   :0.00    Min.   :0.000                Min.   :0.00   
#>  1st Qu.:0.00    1st Qu.:0.000                1st Qu.:1.00   
#>  Median :1.00    Median :1.000                Median :1.00   
#>  Mean   :0.66    Mean   :0.943                Mean   :0.76   
#>  3rd Qu.:1.00    3rd Qu.:2.000                3rd Qu.:1.00   
#>  Max.   :1.00    Max.   :3.000                Max.   :1.00   
#>  NA's   :67577   NA's   :4168                 NA's   :45376  
#>  C7_Restrictions on internal movement    C7_Flag     
#>  Min.   :0.000                        Min.   :0.00   
#>  1st Qu.:0.000                        1st Qu.:0.00   
#>  Median :1.000                        Median :1.00   
#>  Mean   :0.939                        Mean   :0.68   
#>  3rd Qu.:2.000                        3rd Qu.:1.00   
#>  Max.   :2.000                        Max.   :1.00   
#>  NA's   :4142                         NA's   :49764  
#>  C8_International travel controls E1_Income support    E1_Flag     
#>  Min.   :0.000                    Min.   :0.000     Min.   :0.00   
#>  1st Qu.:2.000                    1st Qu.:0.000     1st Qu.:0.00   
#>  Median :3.000                    Median :1.000     Median :1.00   
#>  Mean   :2.471                    Mean   :0.817     Mean   :0.51   
#>  3rd Qu.:4.000                    3rd Qu.:2.000     3rd Qu.:1.00   
#>  Max.   :4.000                    Max.   :2.000     Max.   :1.00   
#>  NA's   :5117                     NA's   :5420      NA's   :54214  
#>  E2_Debt/contract relief E3_Fiscal measures  E4_International support
#>  Min.   :0.000           Min.   :0.000e+00   Min.   :0.000e+00       
#>  1st Qu.:0.000           1st Qu.:0.000e+00   1st Qu.:0.000e+00       
#>  Median :1.000           Median :0.000e+00   Median :0.000e+00       
#>  Mean   :0.948           Mean   :1.417e+08   Mean   :1.019e+07       
#>  3rd Qu.:2.000           3rd Qu.:0.000e+00   3rd Qu.:0.000e+00       
#>  Max.   :2.000           Max.   :1.958e+12   Max.   :8.344e+11       
#>  NA's   :5926            NA's   :26484       NA's   :26447           
#>  H1_Public information campaigns    H1_Flag      H2_Testing policy
#>  Min.   :0.000                   Min.   :0.000   Min.   :0.000    
#>  1st Qu.:2.000                   1st Qu.:1.000   1st Qu.:1.000    
#>  Median :2.000                   Median :1.000   Median :2.000    
#>  Mean   :1.723                   Mean   :0.991   Mean   :1.633    
#>  3rd Qu.:2.000                   3rd Qu.:1.000   3rd Qu.:2.000    
#>  Max.   :2.000                   Max.   :1.000   Max.   :3.000    
#>  NA's   :4364                    NA's   :16368   NA's   :4336     
#>  H3_Contact tracing H4_Emergency investment in healthcare
#>  Min.   :0.000      Min.   :0.000e+00                    
#>  1st Qu.:1.000      1st Qu.:0.000e+00                    
#>  Median :1.000      Median :0.000e+00                    
#>  Mean   :1.309      Mean   :9.233e+06                    
#>  3rd Qu.:2.000      3rd Qu.:0.000e+00                    
#>  Max.   :2.000      Max.   :2.424e+11                    
#>  NA's   :4855       NA's   :26921                        
#>  H5_Investment in vaccines H6_Facial Coverings    H6_Flag     
#>  Min.   :0.000e+00         Min.   :0.000       Min.   :0.00   
#>  1st Qu.:0.000e+00         1st Qu.:0.000       1st Qu.:1.00   
#>  Median :0.000e+00         Median :2.000       Median :1.00   
#>  Mean   :5.726e+05         Mean   :1.891       Mean   :0.85   
#>  3rd Qu.:0.000e+00         3rd Qu.:3.000       3rd Qu.:1.00   
#>  Max.   :1.575e+10         Max.   :4.000       Max.   :1.00   
#>  NA's   :26414             NA's   :4819        NA's   :39663  
#>  H7_Vaccination policy    H7_Flag       M1_Wildcard    ConfirmedCases    
#>  Min.   :0.00          Min.   :0.00     Mode:logical   Min.   :       0  
#>  1st Qu.:0.00          1st Qu.:1.00     NA's:115080    1st Qu.:     273  
#>  Median :0.00          Median :1.00                    Median :    7184  
#>  Mean   :0.14          Mean   :0.99                    Mean   :  153417  
#>  3rd Qu.:0.00          3rd Qu.:1.00                    3rd Qu.:   79235  
#>  Max.   :5.00          Max.   :1.00                    Max.   :27492023  
#>  NA's   :21220         NA's   :108865                  NA's   :10446     
#>  ConfirmedDeaths  StringencyIndex  StringencyIndexForDisplay
#>  Min.   :     0   Min.   :  0.00   Min.   :  0.00           
#>  1st Qu.:     4   1st Qu.: 35.19   1st Qu.: 35.19           
#>  Median :   135   Median : 58.33   Median : 58.33           
#>  Mean   :  4093   Mean   : 51.77   Mean   : 51.81           
#>  3rd Qu.:  1744   3rd Qu.: 72.69   3rd Qu.: 72.22           
#>  Max.   :480887   Max.   :100.00   Max.   :100.00           
#>  NA's   :10557    NA's   :4289     NA's   :3270             
#>  StringencyLegacyIndex StringencyLegacyIndexForDisplay GovernmentResponseIndex
#>  Min.   :  0.00        Min.   :  0.00                  Min.   : 0.00          
#>  1st Qu.: 45.24        1st Qu.: 45.24                  1st Qu.:36.67          
#>  Median : 66.67        Median : 66.67                  Median :55.56          
#>  Mean   : 58.14        Mean   : 58.17                  Mean   :47.57          
#>  3rd Qu.: 78.57        3rd Qu.: 78.57                  3rd Qu.:64.44          
#>  Max.   :100.00        Max.   :100.00                  Max.   :89.17          
#>  NA's   :4270          NA's   :3252                    NA's   :4473           
#>  GovernmentResponseIndexForDisplay ContainmentHealthIndex
#>  Min.   : 0.00                     Min.   : 0.00         
#>  1st Qu.:36.67                     1st Qu.:37.18         
#>  Median :55.56                     Median :56.41         
#>  Mean   :47.63                     Mean   :48.72         
#>  3rd Qu.:64.44                     3rd Qu.:66.03         
#>  Max.   :89.17                     Max.   :91.35         
#>  NA's   :3451                      NA's   :4342          
#>  ContainmentHealthIndexForDisplay EconomicSupportIndex
#>  Min.   : 0.00                    Min.   :  0.00      
#>  1st Qu.:37.50                    1st Qu.:  0.00      
#>  Median :56.41                    Median : 37.50      
#>  Mean   :48.78                    Mean   : 40.75      
#>  3rd Qu.:66.03                    3rd Qu.: 62.50      
#>  Max.   :91.35                    Max.   :100.00      
#>  NA's   :3320                     NA's   :5964        
#>  EconomicSupportIndexForDisplay
#>  Min.   :  0.00                
#>  1st Qu.:  0.00                
#>  Median : 37.50                
#>  Mean   : 40.74                
#>  3rd Qu.: 62.50                
#>  Max.   :100.00                
#>  NA's   :4941