US counties COVID-19 data from JHU CSSEGIS

This function access and munges the cumulative time series of confirmed, and deaths from the US data in the repository for the 2019 Novel Coronavirus Visual Dashboard operated by the Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE). Also, Supported by ESRI Living Atlas Team and the Johns Hopkins University Applied Physics Lab (JHU APL).

jhu_us_data()

Source

https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series,mGT, method=c('EG','TD'))

Value

A tidy data.frame (actually, a tbl_df) with columns:

UID: Universal Identifier

iso2: ISO 3166-1 alpha-2 code

iso3: ISO 3166-1 alpha-3 code

code3

FIPS: Federal Information Processing Standard Publication code

Admin2: County

ProvinceState: Province or state.

CountryRegion: US

Lat: Latitude

Long_: Longitude

Combined_Key: Comma-separated combination of columns Admin2, ProvinceState, and CountryRegion

date: Date

count: The cumulative count of cases for a given geographic area.

subset: either confirmed or deaths

Details

Data are updated daily by JHU. Each call to this function redownloads the data from github. No data cleansing is performed. Data are downloaded and then munged into long-form tidy data.frame.

Note

Uses https://raw.githubusercontent.com/CSSEGISandData/... as data source, then modifies column names and munges to long form table.

US States are treated different from other countries, so are not directly included right now.

Although numbers are meant to be cumulative, there are instances where a day's count might be less than the prior day due to a reclassification of a case. These are not currently corrected in the source data

See also

Examples

res = jhu_data()
colnames(res)

#> [1] "ProvinceState" "CountryRegion" "Lat"           "Long"         
#> [5] "date"          "count"         "subset"       

head(res)

#> # A tibble: 6 x 7
#>   ProvinceState CountryRegion   Lat  Long date       count subset   
#>   <chr>         <chr>         <dbl> <dbl> <date>     <dbl> <chr>    
#> 1 <NA>          Afghanistan      33    65 2020-01-22     0 confirmed
#> 2 <NA>          Afghanistan      33    65 2020-01-23     0 confirmed
#> 3 <NA>          Afghanistan      33    65 2020-01-24     0 confirmed
#> 4 <NA>          Afghanistan      33    65 2020-01-25     0 confirmed
#> 5 <NA>          Afghanistan      33    65 2020-01-26     0 confirmed
#> 6 <NA>          Afghanistan      33    65 2020-01-27     0 confirmed