plot the disease course and death projections

Using data from https://covid19.healthdata.org/projections, either render a plot or return a ggplot object for further manipulation.

plot_projection(
  location_name,
  projections = healthdata_projections_data(),
  metrics = unique(projections$metric)
)

Arguments

location_name
character(1), chosen from unique(projections$location_name) that is used to filter the projections data.frame to one location.

projections
supply a data.frame, typically from a call to healthdata_projections_data(). This data.frame must have columns that include date, location_name, metric, mean, upper, and lower.

metrics
character vector of metrics to include on plot. Get a list of these from the metric column.

Source

https://covid19.healthdata.org/projections

Details

This function is a convenience function to return a ggplot2 object that displays projections for a given location. Projections are typically coming from a call to healthdata_projections_data(). If assigned to a variable, the results can be further manipulated or passed into any function that accepts a ggplot2 object (like plotly, cowplot, etc.). See examples for details.

See also

Examples

library(sars2pack)

hdp = healthdata_projections_data()

#> Error in unzip(destfile, exdir = tmpd): object 'tmpd' not found

library(ggplot2)
# possible locations
sort(unique(hdp$location_name))

#> Error in unique(hdp$location_name): object 'hdp' not found

# create a ggplot2 object
nyp = plot_projection(projections = hdp,
    location_name='New York')

#> Error in unique(projections$metric): object 'hdp' not found

# and do the plot
print(nyp)

#> Error in print(nyp): object 'nyp' not found

# Zoomable html version using plotly
if(require(plotly)) {
  ggplotly(nyp)
}

#> Loading required package: plotly

#> 
#> Attaching package: ‘plotly’

#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot

#> The following object is masked from ‘package:MASS’:
#> 
#>     select

#> The following object is masked from ‘package:stats’:
#> 
#>     filter

#> The following object is masked from ‘package:graphics’:
#> 
#>     layout

#> Error in ggplotly(nyp): object 'nyp' not found

# California
calip = plot_projection(projections = hdp,
    location_name='California')

#> Error in unique(projections$metric): object 'hdp' not found

print(calip)

#> Error in print(calip): object 'calip' not found

# entire US
plot_projection(location_name = "United States of America", projections = hdp)

#> Error in unique(projections$metric): object 'hdp' not found

# Compare two countries
spain_proj = plot_projection(location_name = "Spain", projections = hdp)

#> Error in unique(projections$metric): object 'hdp' not found

portugal_proj = plot_projection(location_name = "Portugal", projections = hdp)

#> Error in unique(projections$metric): object 'hdp' not found

#'
print(spain_proj)

#> Error in print(spain_proj): object 'spain_proj' not found

print(portugal_proj)

#> Error in print(portugal_proj): object 'portugal_proj' not found

# plot both on same plot
if(require(cowplot)) {
  plot_grid(spain_proj, portugal_proj)
}

#> Loading required package: cowplot

#> 
#> ********************************************************

#> Note: As of version 1.0.0, cowplot does not change the

#>   default ggplot2 theme anymore. To recover the previous

#>   behavior, execute:
#>   theme_set(theme_cowplot())

#> ********************************************************

#> Error in plot_grid(spain_proj, portugal_proj): object 'spain_proj' not found