US hospital dataset

Includes details of beds, NAICS codes, and geolocation data for US states and territories.

us_hospital_details()

Source

https://hub.arcgis.com/datasets/geoplatform::hospitals

Value

a data.frame

Details

This dataset contains locations of Hospitals for 50 US states, Washington D.C., US territories of Puerto Rico, Guam, American Samoa, Northern Mariana Islands, Palau, and Virgin Islands. The dataset only includes hospital facilities based on data acquired from various state departments or federal sources which has been referenced in the SOURCE field. Hospital facilities which do not occur in these sources will be not present in the database. The source data was available in a variety of formats (pdfs, tables, webpages, etc.) which was cleaned and geocoded and then converted into a spatial database. The database does not contain nursing homes or health centers. Hospitals have been categorized into children, chronic disease, critical access, general acute care, long term care, military, psychiatric, rehabilitation, special, and women based on the range of the available values from the various sources after removing similarities. In this update the TRAUMA field was populated for 172 additional hospitals and helipad presence were verified for all hospitals.

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_hospital_details()
#> Warning: download failed
#>   web resource path: ‘https://opendata.arcgis.com/datasets/6ac5e325468c4cb9b905f1728d6fbf0f_0.csv’
#>   local file path: ‘~/.cache/sars2pack/83e34ffdcf8_6ac5e325468c4cb9b905f1728d6fbf0f_0.csv’
#>   reason: Not Found (HTTP 404).
#> Warning: bfcadd() failed; resource removed
#>   rid: BFC159
#>   fpath: ‘https://opendata.arcgis.com/datasets/6ac5e325468c4cb9b905f1728d6fbf0f_0.csv’
#>   reason: download failed
#> Error in .util_download(x, rid[i], proxy, config, "bfcadd()", ...): bfcadd() failed; see warnings()
head(res)
#> Error in head(res): object 'res' not found
colnames(res)
#> Error in is.data.frame(x): object 'res' not found
summary(res)
#> Error in summary(res): object 'res' not found