United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Value

a data.frame

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_state_distancing_policy()

#> Warning:  1 failed to parse.

colnames(res)

#>  [1] "pid"                   "location_id"           "statefips"            
#>  [4] "iso2c"                 "state"                 "statepolicy"          
#>  [7] "mandate"               "statewide"             "swgeo"                
#> [10] "swpop"                 "curfew"                "curfewstart"          
#> [13] "curfewend"             "ingathlim"             "outgathlim"           
#> [16] "ingathlimreligious"    "outgathlimreligious"   "businessrestrictlevel"
#> [19] "schoolrestrictlevel"   "publicmasklevel"       "dateissued"           
#> [22] "dateenacted"           "dateexpiry"            "dateended"            
#> [25] "extends"               "expands"               "eases"                
#> [28] "ends"                  "joins"                 "leaves"               
#> [31] "appliesto"             "policycodingnotes"     "policysource"         
#> [34] "policysourceid"        "lastupdated"           "lastupdatednotes"     
#> [37] "rereviewed"            "updatingsource1"       "updatingsource2"      

library(dplyr)
dplyr::glimpse(res)

#> Rows: 11,566
#> Columns: 39
#> $ pid                   <chr> "AL0001", "AL0002", "AL0080", "AL0013", "AL0015"…
#> $ location_id           <int> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523…
#> $ statefips             <chr> "00001", "00001", "00001", "00001", "00001", "00…
#> $ iso2c                 <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", …
#> $ state                 <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Ala…
#> $ statepolicy           <chr> "EmergDec", "SchoolClose", "BarRestrict", "GathR…
#> $ mandate               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, …
#> $ statewide             <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, T…
#> $ swgeo                 <int> 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ swpop                 <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ curfew                <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ curfewstart           <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ curfewend             <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ ingathlim             <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlim            <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ ingathlimreligious    <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlimreligious   <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ businessrestrictlevel <chr> "", "", "TakeawayOnly", "", "TakeawayOnly", "", …
#> $ schoolrestrictlevel   <chr> "", "NoInPerson", "", "", "", "NoInPerson", "NoI…
#> $ publicmasklevel       <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ dateissued            <date> 2020-03-13, 2020-03-13, 2020-03-17, 2020-03-17,…
#> $ dateenacted           <date> 2020-03-13, 2020-03-19, 2020-03-17, 2020-03-17,…
#> $ dateexpiry            <date> 2020-05-12, 2020-04-06, 2020-03-24, NA, 2020-03…
#> $ dateended             <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 202…
#> $ extends               <chr> "", "", "", "", "", "", "AL0002", "", "", "", ""…
#> $ expands               <chr> "", "", "", "", "", "AL0002", "", "", "", "", ""…
#> $ eases                 <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ ends                  <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ joins                 <chr> "", "", "", "", "", "", "", "AL0080", "AL0013", …
#> $ leaves                <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ appliesto             <chr> "", "", "Blount, St. Clair, Shelby, Tuscaloosa, …
#> $ policycodingnotes     <chr> "State of Public Health Emergency declared. Late…
#> $ policysource          <chr> "https://governor.alabama.gov/assets/2020/03/202…
#> $ policysourceid        <chr> "Alabama0001", "Alabama0002", "Alabama0004", "Al…
#> $ lastupdated           <date> 2021-01-08, 2021-01-06, 2021-01-06, 2021-01-06,…
#> $ lastupdatednotes      <chr> "Updated DateExpiry to 20200512", "Added SchoolR…
#> $ rereviewed            <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ updatingsource1       <chr> "https://covid19.alaska.gov/health-mandates/", "…
#> $ updatingsource2       <chr> "https://www.alabamapublichealth.gov/legal/order…

summary(res)

#>      pid             location_id   statefips            iso2c          
#>  Length:11566       Min.   :523   Length:11566       Length:11566      
#>  Class :character   1st Qu.:531   Class :character   Class :character  
#>  Mode  :character   Median :547   Mode  :character   Mode  :character  
#>                     Mean   :547                                        
#>                     3rd Qu.:561                                        
#>                     Max.   :573                                        
#>                                                                        
#>     state           statepolicy         mandate        statewide      
#>  Length:11566       Length:11566       Mode :logical   Mode :logical  
#>  Class :character   Class :character   FALSE:761       FALSE:4320     
#>  Mode  :character   Mode  :character   TRUE :10805     TRUE :7246     
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>      swgeo            swpop            curfew        curfewstart       
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.00000   Length:11566      
#>  1st Qu.:0.0000   1st Qu.:1.0000   1st Qu.:0.00000   Class :character  
#>  Median :1.0000   Median :1.0000   Median :0.00000   Mode  :character  
#>  Mean   :0.6443   Mean   :0.9807   Mean   :0.03493                     
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:0.00000                     
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1.00000                     
#>                                                                        
#>   curfewend           ingathlim         outgathlim      ingathlimreligious
#>  Length:11566       Min.   :    0.0   Min.   :    0.0   Min.   :   0.00   
#>  Class :character   1st Qu.:   10.0   1st Qu.:   10.0   1st Qu.:  10.00   
#>  Mode  :character   Median :   10.0   Median :   12.0   Median :  50.00   
#>                     Mean   :  104.1   Mean   :  122.5   Mean   :  70.16   
#>                     3rd Qu.:   25.0   3rd Qu.:   50.0   3rd Qu.: 100.00   
#>                     Max.   :10000.0   Max.   :10000.0   Max.   :1000.00   
#>                     NA's   :10243     NA's   :10249     NA's   :10973     
#>  outgathlimreligious businessrestrictlevel schoolrestrictlevel
#>  Min.   :   0.00     Length:11566          Length:11566       
#>  1st Qu.:  10.00     Class :character      Class :character   
#>  Median :  25.00     Mode  :character      Mode  :character   
#>  Mean   :  69.47                                              
#>  3rd Qu.: 100.00                                              
#>  Max.   :1000.00                                              
#>  NA's   :11168                                                
#>  publicmasklevel      dateissued          dateenacted        
#>  Length:11566       Min.   :2020-02-29   Min.   :2020-02-29  
#>  Class :character   1st Qu.:2020-05-28   1st Qu.:2020-06-01  
#>  Mode  :character   Median :2020-08-20   Median :2020-08-21  
#>                     Mean   :2020-08-24   Mean   :2020-08-25  
#>                     3rd Qu.:2020-11-22   3rd Qu.:2020-11-23  
#>                     Max.   :2021-02-16   Max.   :2021-02-23  
#>                                                              
#>    dateexpiry           dateended            extends         
#>  Min.   :2020-03-21   Min.   :2020-04-24   Length:11566      
#>  1st Qu.:2020-06-16   1st Qu.:2020-06-05   Class :character  
#>  Median :2020-09-20   Median :2020-08-07   Mode  :character  
#>  Mean   :2020-09-21   Mean   :2020-08-29                     
#>  3rd Qu.:2020-12-28   3rd Qu.:2020-10-02                     
#>  Max.   :2021-06-30   Max.   :2021-02-12                     
#>  NA's   :3186         NA's   :10373                          
#>    expands             eases               ends              joins          
#>  Length:11566       Length:11566       Length:11566       Length:11566      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     leaves           appliesto         policycodingnotes  policysource      
#>  Length:11566       Length:11566       Length:11566       Length:11566      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  policysourceid      lastupdated         lastupdatednotes     rereviewed    
#>  Length:11566       Min.   :2020-03-24   Length:11566       Min.   :0.0000  
#>  Class :character   1st Qu.:2020-11-10   Class :character   1st Qu.:1.0000  
#>  Mode  :character   Median :2020-12-04   Mode  :character   Median :1.0000  
#>                     Mean   :2020-11-25                      Mean   :0.8025  
#>                     3rd Qu.:2021-01-17                      3rd Qu.:1.0000  
#>                     Max.   :2021-02-17                      Max.   :1.0000  
#>                                                                             
#>  updatingsource1    updatingsource2   
#>  Length:11566       Length:11566      
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>