United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for a subset of the states in the United States.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Examples

res = us_state_distancing_policy()
colnames(res)

#>  [1] "location_id"       "StateFIPS"         "StatePostal"      
#>  [4] "StateName"         "StatePolicy"       "Mandate"          
#>  [7] "DateIssued"        "DateEnacted"       "DateExpiry"       
#> [10] "DateEased"         "DateEnded"         "PolicyCodingNotes"
#> [13] "PolicySource"      "StateWide"         "LastUpdated"      
#> [16] "LastUpdatedNotes" 

library(dplyr)
dplyr::glimpse(res)

#> Rows: 1,141
#> Columns: 16
#> $ location_id       <dbl> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523, 5…
#> $ StateFIPS         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, …
#> $ StatePostal       <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL…
#> $ StateName         <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Alabam…
#> $ StatePolicy       <chr> "EmergDec", "SchoolClose", "GathRestrict25", "GathR…
#> $ Mandate           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ DateIssued        <dbl> 20200313, 20200314, 20200319, 20200319, 20200319, 2…
#> $ DateEnacted       <dbl> 20200313, 20200318, 20200319, 20200319, 20200319, 2…
#> $ DateExpiry        <dbl> NA, NA, NA, 20200515, 20200515, 20200515, 20200403,…
#> $ DateEased         <dbl> NA, NA, NA, NA, 20200430, NA, NA, NA, 20200430, NA,…
#> $ DateEnded         <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, 20200430, NA, N…
#> $ PolicyCodingNotes <chr> "Public Health Emergency", "in effect 20200318. 202…
#> $ PolicySource      <chr> "https://governor.alabama.gov/newsroom/2020/03/thir…
#> $ StateWide         <dbl> 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, …
#> $ LastUpdated       <dbl> 20200324, 20200324, 20200420, 20200420, 20200420, 2…
#> $ LastUpdatedNotes  <chr> NA, NA, "Updated DateEnacted to 20200319 per state …

summary(res)

#>   location_id      StateFIPS    StatePostal         StateName        
#>  Min.   :523.0   Min.   : 1.0   Length:1141        Length:1141       
#>  1st Qu.:535.0   1st Qu.:16.0   Class :character   Class :character  
#>  Median :549.0   Median :30.0   Mode  :character   Mode  :character  
#>  Mean   :548.3   Mean   :29.3                                        
#>  3rd Qu.:561.0   3rd Qu.:42.0                                        
#>  Max.   :573.0   Max.   :56.0                                        
#>  NA's   :582     NA's   :582                                         
#>  StatePolicy           Mandate         DateIssued        DateEnacted      
#>  Length:1141        Min.   :0.0000   Min.   :20200229   Min.   :20200229  
#>  Class :character   1st Qu.:1.0000   1st Qu.:20200315   1st Qu.:20200316  
#>  Mode  :character   Median :1.0000   Median :20200319   Median :20200320  
#>                     Mean   :0.9106   Mean   :20200329   Mean   :20200332  
#>                     3rd Qu.:1.0000   3rd Qu.:20200325   3rd Qu.:20200326  
#>                     Max.   :1.0000   Max.   :20200430   Max.   :20200504  
#>                     NA's   :582      NA's   :583        NA's   :584       
#>    DateExpiry         DateEased          DateEnded        PolicyCodingNotes 
#>  Min.   :20200330   Min.   :20200420   Min.   :20200424   Length:1141       
#>  1st Qu.:20200430   1st Qu.:20200426   1st Qu.:20200426   Class :character  
#>  Median :20200508   Median :20200429   Median :20200428   Mode  :character  
#>  Mean   :20200484   Mean   :20200459   Mean   :20200448                     
#>  3rd Qu.:20200516   3rd Qu.:20200501   3rd Qu.:20200466                     
#>  Max.   :20200909   Max.   :20200511   Max.   :20200501                     
#>  NA's   :858        NA's   :1095       NA's   :1134                         
#>  PolicySource         StateWide       LastUpdated       LastUpdatedNotes  
#>  Length:1141        Min.   :0.0000   Min.   :20200324   Length:1141       
#>  Class :character   1st Qu.:1.0000   1st Qu.:20200327   Class :character  
#>  Mode  :character   Median :1.0000   Median :20200413   Mode  :character  
#>                     Mean   :0.9284   Mean   :20200398                     
#>                     3rd Qu.:1.0000   3rd Qu.:20200426                     
#>                     Max.   :1.0000   Max.   :20200501                     
#>                     NA's   :582      NA's   :582