United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for a subset of the states in the United States.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Examples

res = us_state_distancing_policy()

#> Warning: Missing column names filled in: 'X16' [16], 'X17' [17], 'X18' [18], 'X19' [19]

colnames(res)

#>  [1] "location_id"       "StateFIPS"         "StatePostal"      
#>  [4] "StateName"         "StatePolicy"       "Mandate"          
#>  [7] "DateIssued"        "DateEnacted"       "DateExpiry"       
#> [10] "DateEnded"         "PolicyCodingNotes" "PolicySource"     
#> [13] "StateWide"         "LastUpdated"       "LastUpdatedNotes" 
#> [16] "X16"               "X17"               "X18"              
#> [19] "X19"              

library(dplyr)
dplyr::glimpse(res)

#> Rows: 532
#> Columns: 19
#> $ location_id       <dbl> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523, 5…
#> $ StateFIPS         <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, …
#> $ StatePostal       <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL…
#> $ StateName         <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Alabam…
#> $ StatePolicy       <chr> "EmergDec", "SchoolClose", "GathRestrict25", "GathR…
#> $ Mandate           <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ DateIssued        <dbl> 20200313, 20200314, 20200319, 20200319, 20200319, 2…
#> $ DateEnacted       <dbl> 20200313, 20200318, 20200320, 20200320, 20200320, 2…
#> $ DateExpiry        <dbl> NA, NA, NA, NA, NA, NA, 20200403, 20200417, 2020041…
#> $ DateEnded         <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ PolicyCodingNotes <chr> "Public Health Emergency", "in effect 20200318. 202…
#> $ PolicySource      <chr> "https://governor.alabama.gov/newsroom/2020/03/thir…
#> $ StateWide         <dbl> 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, …
#> $ LastUpdated       <dbl> 20200324, 20200324, 20200324, 20200324, 20200324, 2…
#> $ LastUpdatedNotes  <chr> NA, NA, NA, NA, NA, NA, "Added Birmingham only Stay…
#> $ X16               <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ X17               <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ X18               <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ X19               <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,…