United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Value

a data.frame

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_state_distancing_policy()

#> Warning: Problem with `mutate()` input `DateExpiry`.
#> ℹ  1 failed to parse.
#> ℹ Input `DateExpiry` is `lubridate::ymd(.data$DateExpiry)`.

#> Warning:  1 failed to parse.

colnames(res)

#>  [1] "pid"                   "location_id"           "statefips"            
#>  [4] "iso2c"                 "state"                 "statepolicy"          
#>  [7] "mandate"               "statewide"             "swgeo"                
#> [10] "swpop"                 "curfew"                "curfewstart"          
#> [13] "curfewend"             "ingathlim"             "outgathlim"           
#> [16] "ingathlimreligious"    "outgathlimreligious"   "businessrestrictlevel"
#> [19] "schoolrestrictlevel"   "publicmasklevel"       "dateissued"           
#> [22] "dateenacted"           "dateexpiry"            "dateended"            
#> [25] "extends"               "expands"               "eases"                
#> [28] "ends"                  "joins"                 "leaves"               
#> [31] "appliesto"             "policycodingnotes"     "policysource"         
#> [34] "policysourceid"        "lastupdated"           "lastupdatednotes"     
#> [37] "rereviewed"            "updatingsource1"       "updatingsource2"      

library(dplyr)
dplyr::glimpse(res)

#> Rows: 9,607
#> Columns: 39
#> $ pid                   <chr> "AL0001", "AL0002", "AL0080", "AL0013", "AL0015…
#> $ location_id           <int> 523, 523, 523, 523, 523, 523, 523, 523, 523, 52…
#> $ statefips             <chr> "00001", "00001", "00001", "00001", "00001", "0…
#> $ iso2c                 <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL",…
#> $ state                 <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Al…
#> $ statepolicy           <chr> "EmergDec", "SchoolClose", "BarRestrict", "Gath…
#> $ mandate               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,…
#> $ statewide             <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, TRUE, T…
#> $ swgeo                 <int> 1, 1, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ swpop                 <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ curfew                <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ curfewstart           <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ curfewend             <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ ingathlim             <int> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA,…
#> $ outgathlim            <int> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA,…
#> $ ingathlimreligious    <int> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA,…
#> $ outgathlimreligious   <int> NA, NA, NA, 25, NA, NA, NA, 25, NA, NA, NA, NA,…
#> $ businessrestrictlevel <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ schoolrestrictlevel   <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ publicmasklevel       <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ dateissued            <date> 2020-03-13, 2020-03-13, 2020-03-17, 2020-03-17…
#> $ dateenacted           <date> 2020-03-13, 2020-03-19, 2020-03-17, 2020-03-17…
#> $ dateexpiry            <date> NA, 2020-04-06, 2020-03-24, NA, 2020-03-24, 20…
#> $ dateended             <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2020-0…
#> $ extends               <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ expands               <chr> "", "", "", "", "", "", "AL0080", "AL0013", "",…
#> $ eases                 <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ ends                  <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ joins                 <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ leaves                <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ appliesto             <chr> "", "", "", "", "", "", "", "", "", "", "", "",…
#> $ policycodingnotes     <chr> "Public Health Emergency declared. No clear exp…
#> $ policysource          <chr> "https://governor.alabama.gov/assets/2020/03/20…
#> $ policysourceid        <chr> "Alabama0001", "Alabama0002", "Alabama0004", "A…
#> $ lastupdated           <date> 2020-03-24, 2020-04-30, 2020-07-03, 2020-07-24…
#> $ lastupdatednotes      <chr> "", "Updated DateIssued to 20200313, DateEnacte…
#> $ rereviewed            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ updatingsource1       <chr> "https://covid19.alaska.gov/health-mandates/", …
#> $ updatingsource2       <chr> "https://www.alabamapublichealth.gov/legal/orde…

summary(res)

#>      pid             location_id     statefips            iso2c          
#>  Length:9607        Min.   :523.0   Length:9607        Length:9607       
#>  Class :character   1st Qu.:533.0   Class :character   Class :character  
#>  Mode  :character   Median :548.0   Mode  :character   Mode  :character  
#>                     Mean   :547.8                                        
#>                     3rd Qu.:561.0                                        
#>                     Max.   :573.0                                        
#>                                                                          
#>     state           statepolicy         mandate        statewide      
#>  Length:9607        Length:9607        Mode :logical   Mode :logical  
#>  Class :character   Class :character   FALSE:666       FALSE:3261     
#>  Mode  :character   Mode  :character   TRUE :8941      TRUE :6346     
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>      swgeo            swpop            curfew        curfewstart       
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.00000   Length:9607       
#>  1st Qu.:0.0000   1st Qu.:1.0000   1st Qu.:0.00000   Class :character  
#>  Median :1.0000   Median :1.0000   Median :0.00000   Mode  :character  
#>  Mean   :0.6793   Mean   :0.9796   Mean   :0.02061                     
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:0.00000                     
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1.00000                     
#>                                                                        
#>   curfewend           ingathlim         outgathlim      ingathlimreligious
#>  Length:9607        Min.   :    0.0   Min.   :    0.0   Min.   :   0.00   
#>  Class :character   1st Qu.:   10.0   1st Qu.:   10.0   1st Qu.:  10.00   
#>  Mode  :character   Median :   10.0   Median :   20.0   Median :  25.00   
#>                     Mean   :  117.8   Mean   :  136.4   Mean   :  76.98   
#>                     3rd Qu.:   25.0   3rd Qu.:   50.0   3rd Qu.: 100.00   
#>                     Max.   :10000.0   Max.   :10000.0   Max.   :3750.00   
#>                     NA's   :8471      NA's   :8475      NA's   :8958      
#>  outgathlimreligious businessrestrictlevel schoolrestrictlevel
#>  Min.   :   0.00     Length:9607           Length:9607        
#>  1st Qu.:  10.00     Class :character      Class :character   
#>  Median :  25.00     Mode  :character      Mode  :character   
#>  Mean   :  88.66                                              
#>  3rd Qu.: 100.00                                              
#>  Max.   :7500.00                                              
#>  NA's   :9126                                                 
#>  publicmasklevel      dateissued          dateenacted        
#>  Length:9607        Min.   :2020-02-29   Min.   :2020-02-29  
#>  Class :character   1st Qu.:2020-05-21   1st Qu.:2020-05-22  
#>  Mode  :character   Median :2020-07-17   Median :2020-07-17  
#>                     Mean   :2020-07-30   Mean   :2020-07-31  
#>                     3rd Qu.:2020-10-20   3rd Qu.:2020-10-20  
#>                     Max.   :2021-01-04   Max.   :2021-01-09  
#>                                                              
#>    dateexpiry           dateended            extends         
#>  Min.   :2020-03-21   Min.   :2020-04-24   Length:9607       
#>  1st Qu.:2020-06-01   1st Qu.:2020-05-30   Class :character  
#>  Median :2020-08-07   Median :2020-07-01   Mode  :character  
#>  Mean   :2020-08-24   Mean   :2020-07-17                     
#>  3rd Qu.:2020-11-16   3rd Qu.:2020-09-14                     
#>  Max.   :2021-06-30   Max.   :2020-12-21                     
#>  NA's   :3084         NA's   :8689                           
#>    expands             eases               ends              joins          
#>  Length:9607        Length:9607        Length:9607        Length:9607       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     leaves           appliesto         policycodingnotes  policysource      
#>  Length:9607        Length:9607        Length:9607        Length:9607       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  policysourceid      lastupdated         lastupdatednotes     rereviewed    
#>  Length:9607        Min.   :2020-03-24   Length:9607        Min.   :0.0000  
#>  Class :character   1st Qu.:2020-09-18   Class :character   1st Qu.:0.0000  
#>  Mode  :character   Median :2020-11-17   Mode  :character   Median :1.0000  
#>                     Mean   :2020-10-19                      Mean   :0.6303  
#>                     3rd Qu.:2020-12-03                      3rd Qu.:1.0000  
#>                     Max.   :2021-01-05                      Max.   :1.0000  
#>                                                                             
#>  updatingsource1    updatingsource2   
#>  Length:9607        Length:9607       
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>