United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_state_distancing_policy()

#> Warning: Problem with `mutate()` input `DateEnacted`.
#> ℹ  8 failed to parse.
#> ℹ Input `DateEnacted` is `lubridate::ymd(DateEnacted)`.

#> Warning:  8 failed to parse.

colnames(res)

#>  [1] "PID"                                                    
#>  [2] "Counties on the County Monitoring List must close malls"
#>  [3] "StateFIPS"                                              
#>  [4] "iso2c"                                                  
#>  [5] "state"                                                  
#>  [6] "StatePolicy"                                            
#>  [7] "Mandate"                                                
#>  [8] "StateWide"                                              
#>  [9] "SWGeo"                                                  
#> [10] "SWPop"                                                  
#> [11] "Encourages"                                             
#> [12] "InGathLim"                                              
#> [13] "OutGathLim"                                             
#> [14] "InGathLimReligious"                                     
#> [15] "OutGathLimReligious"                                    
#> [16] "PublicMaskLevel"                                        
#> [17] "DateIssued"                                             
#> [18] "DateEnacted"                                            
#> [19] "DateExpiry"                                             
#> [20] "DateEnded"                                              
#> [21] "Extends"                                                
#> [22] "Expands"                                                
#> [23] "Eases"                                                  
#> [24] "Ends"                                                   
#> [25] "Depends"                                                
#> [26] "Joins"                                                  
#> [27] "PolicyCodingNotes"                                      
#> [28] "PolicySource"                                           
#> [29] "PolicySourceID"                                         
#> [30] "LastUpdated"                                            
#> [31] "LastUpdatedNotes"                                       
#> [32] "UpdatingSource1"                                        
#> [33] "UpdatingSource2"                                        

library(dplyr)
dplyr::glimpse(res)

#> Rows: 5,979
#> Columns: 33
#> $ PID                                                       <chr> "AL0001", "…
#> $ `Counties on the County Monitoring List must close malls` <dbl> 523, 523, 5…
#> $ StateFIPS                                                 <chr> "00001", "0…
#> $ iso2c                                                     <chr> "AL", "AL",…
#> $ state                                                     <chr> "Alabama", …
#> $ StatePolicy                                               <chr> "EmergDec",…
#> $ Mandate                                                   <lgl> TRUE, TRUE,…
#> $ StateWide                                                 <lgl> TRUE, TRUE,…
#> $ SWGeo                                                     <dbl> 1, 1, 0, 0,…
#> $ SWPop                                                     <dbl> 1, 1, 1, 1,…
#> $ Encourages                                                <dbl> NA, NA, NA,…
#> $ InGathLim                                                 <dbl> NA, NA, NA,…
#> $ OutGathLim                                                <dbl> NA, NA, NA,…
#> $ InGathLimReligious                                        <dbl> NA, NA, NA,…
#> $ OutGathLimReligious                                       <dbl> NA, NA, NA,…
#> $ PublicMaskLevel                                           <chr> NA, NA, NA,…
#> $ DateIssued                                                <date> 2020-03-13…
#> $ DateEnacted                                               <date> 2020-03-13…
#> $ DateExpiry                                                <date> NA, 2020-0…
#> $ DateEnded                                                 <date> NA, NA, NA…
#> $ Extends                                                   <chr> NA, NA, NA,…
#> $ Expands                                                   <chr> NA, NA, NA,…
#> $ Eases                                                     <chr> NA, NA, NA,…
#> $ Ends                                                      <chr> NA, NA, NA,…
#> $ Depends                                                   <chr> NA, NA, NA,…
#> $ Joins                                                     <chr> NA, NA, NA,…
#> $ PolicyCodingNotes                                         <chr> "Public Hea…
#> $ PolicySource                                              <chr> "https://go…
#> $ PolicySourceID                                            <chr> "Alabama000…
#> $ LastUpdated                                               <date> 2020-03-24…
#> $ LastUpdatedNotes                                          <chr> NA, "Update…
#> $ UpdatingSource1                                           <chr> "https://co…
#> $ UpdatingSource2                                           <chr> "https://ww…

summary(res)

#>      PID            Counties on the County Monitoring List must close malls
#>  Length:5979        Min.   :523.0                                          
#>  Class :character   1st Qu.:534.0                                          
#>  Mode  :character   Median :550.0                                          
#>                     Mean   :548.3                                          
#>                     3rd Qu.:561.0                                          
#>                     Max.   :575.0                                          
#>                     NA's   :1                                              
#>   StateFIPS            iso2c              state           StatePolicy       
#>  Length:5979        Length:5979        Length:5979        Length:5979       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>   Mandate        StateWide           SWGeo            SWPop       
#>  Mode :logical   Mode :logical   Min.   :0.0000   Min.   :0.0000  
#>  FALSE:620       FALSE:1734      1st Qu.:0.0000   1st Qu.:1.0000  
#>  TRUE :5358      TRUE :4244      Median :1.0000   Median :1.0000  
#>  NA's :1         NA's :1         Mean   :0.7298   Mean   :0.9781  
#>                                  3rd Qu.:1.0000   3rd Qu.:1.0000  
#>                                  Max.   :1.0000   Max.   :1.0000  
#>                                  NA's   :1        NA's   :1       
#>    Encourages     InGathLim         OutGathLim      InGathLimReligious
#>  Min.   :1      Min.   :    0.0   Min.   :    0.0   Min.   :   0.0    
#>  1st Qu.:1      1st Qu.:   10.0   1st Qu.:   10.0   1st Qu.:  10.0    
#>  Median :1      Median :   10.0   Median :   20.0   Median :  25.0    
#>  Mean   :1      Mean   :  143.1   Mean   :  154.1   Mean   :  64.4    
#>  3rd Qu.:1      3rd Qu.:   50.0   3rd Qu.:   50.0   3rd Qu.:  50.0    
#>  Max.   :1      Max.   :10000.0   Max.   :10000.0   Max.   :1000.0    
#>  NA's   :5913   NA's   :5369      NA's   :5370      NA's   :5550      
#>  OutGathLimReligious PublicMaskLevel      DateIssued        
#>  Min.   :   0.00     Length:5979        Min.   :2020-02-29  
#>  1st Qu.:  10.00     Class :character   1st Qu.:2020-05-01  
#>  Median :  25.00     Mode  :character   Median :2020-06-05  
#>  Mean   :  61.75                        Mean   :2020-06-08  
#>  3rd Qu.:  50.00                        3rd Qu.:2020-07-16  
#>  Max.   :1000.00                        Max.   :2020-09-17  
#>  NA's   :5580                           NA's   :1           
#>   DateEnacted           DateExpiry           DateEnded         
#>  Min.   :2020-02-29   Min.   :2020-03-21   Min.   :2020-04-24  
#>  1st Qu.:2020-05-04   1st Qu.:2020-05-16   1st Qu.:2020-06-01  
#>  Median :2020-06-06   Median :2020-07-01   Median :2020-06-16  
#>  Mean   :2020-06-09   Mean   :2020-06-30   Mean   :2020-07-01  
#>  3rd Qu.:2020-07-17   3rd Qu.:2020-08-06   3rd Qu.:2020-07-27  
#>  Max.   :2020-09-30   Max.   :2021-06-30   Max.   :2020-09-14  
#>  NA's   :9            NA's   :2212         NA's   :5341        
#>    Extends            Expands             Eases               Ends          
#>  Length:5979        Length:5979        Length:5979        Length:5979       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    Depends             Joins           PolicyCodingNotes  PolicySource      
#>  Length:5979        Length:5979        Length:5979        Length:5979       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  PolicySourceID      LastUpdated         LastUpdatedNotes   UpdatingSource1   
#>  Length:5979        Min.   :2020-03-24   Length:5979        Length:5979       
#>  Class :character   1st Qu.:2020-06-13   Class :character   Class :character  
#>  Mode  :character   Median :2020-07-18   Mode  :character   Mode  :character  
#>                     Mean   :2020-07-13                                        
#>                     3rd Qu.:2020-08-12                                        
#>                     Max.   :2020-09-17                                        
#>                     NA's   :1                                                 
#>  UpdatingSource2   
#>  Length:5979       
#>  Class :character  
#>  Mode  :character  
#>                    
#>                    
#>                    
#>