United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Value

a data.frame

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_state_distancing_policy()
#> Warning:  1 failed to parse.
colnames(res)
#>  [1] "pid"                   "location_id"           "statefips"            
#>  [4] "iso2c"                 "state"                 "statepolicy"          
#>  [7] "mandate"               "statewide"             "swgeo"                
#> [10] "swpop"                 "curfew"                "curfewstart"          
#> [13] "curfewend"             "vaccineexempt"         "ingathlim"            
#> [16] "outgathlim"            "ingathlimreligious"    "outgathlimreligious"  
#> [19] "businessrestrictlevel" "schoolrestrictlevel"   "publicmasklevel"      
#> [22] "dateissued"            "dateenacted"           "dateexpiry"           
#> [25] "dateended"             "extends"               "expands"              
#> [28] "eases"                 "ends"                  "joins"                
#> [31] "leaves"                "appliesto"             "policycodingnotes"    
#> [34] "policysource"          "policysourceid"        "lastupdated"          
#> [37] "lastupdatednotes"      "rereviewed"            "updatingsource1"      
#> [40] "updatingsource2"      
library(dplyr)
dplyr::glimpse(res)
#> Rows: 16,512
#> Columns: 40
#> $ pid                   <chr> "AL0001", "AL0002", "AL0080", "AL0013", "AL0015"…
#> $ location_id           <int> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523…
#> $ statefips             <chr> "00001", "00001", "00001", "00001", "00001", "00…
#> $ iso2c                 <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", …
#> $ state                 <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Ala…
#> $ statepolicy           <chr> "EmergDec", "SchoolClose", "BarRestrict", "GathR…
#> $ mandate               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, …
#> $ statewide             <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, T…
#> $ swgeo                 <int> 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ swpop                 <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ curfew                <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ curfewstart           <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ curfewend             <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ vaccineexempt         <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ ingathlim             <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlim            <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ ingathlimreligious    <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlimreligious   <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ businessrestrictlevel <chr> "", "", "TakeawayOnly", "", "TakeawayOnly", "", …
#> $ schoolrestrictlevel   <chr> "", "NoInPerson", "", "", "", "NoInPerson", "NoI…
#> $ publicmasklevel       <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ dateissued            <date> 2020-03-13, 2020-03-13, 2020-03-17, 2020-03-17,…
#> $ dateenacted           <date> 2020-03-13, 2020-03-19, 2020-03-17, 2020-03-17,…
#> $ dateexpiry            <date> 2020-05-12, 2020-04-06, 2020-03-24, NA, 2020-03…
#> $ dateended             <date> 2021-07-07, NA, NA, NA, NA, NA, NA, NA, NA, NA,…
#> $ extends               <chr> "", "", "", "", "", "", "AL0002", "", "", "", ""…
#> $ expands               <chr> "", "", "", "", "", "AL0002", "", "", "", "", ""…
#> $ eases                 <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ ends                  <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ joins                 <chr> "", "", "", "", "", "", "", "AL0080", "AL0013", …
#> $ leaves                <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ appliesto             <chr> "", "", "Blount, St Clair, Shelby, Tuscaloosa, W…
#> $ policycodingnotes     <chr> "State of Public Health Emergency declared. Late…
#> $ policysource          <chr> "https://governor.alabama.gov/assets/2020/03/202…
#> $ policysourceid        <chr> "Alabama0001", "Alabama0002", "Alabama0004", "Al…
#> $ lastupdated           <date> 2021-07-06, 2021-01-06, 2021-01-06, 2021-01-06,…
#> $ lastupdatednotes      <chr> "Updated DateEnded", "Added SchoolRestrictLevel"…
#> $ rereviewed            <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ updatingsource1       <chr> "https://covid19.alaska.gov/health-mandates/", "…
#> $ updatingsource2       <chr> "https://www.alabamapublichealth.gov/legal/order…
summary(res)
#>      pid             location_id     statefips            iso2c          
#>  Length:16512       Min.   :523.0   Length:16512       Length:16512      
#>  Class :character   1st Qu.:531.0   Class :character   Class :character  
#>  Mode  :character   Median :547.0   Mode  :character   Mode  :character  
#>                     Mean   :546.8                                        
#>                     3rd Qu.:560.0                                        
#>                     Max.   :573.0                                        
#>                                                                          
#>     state           statepolicy         mandate        statewide      
#>  Length:16512       Length:16512       Mode :logical   Mode :logical  
#>  Class :character   Class :character   FALSE:1130      FALSE:6393     
#>  Mode  :character   Mode  :character   TRUE :15382     TRUE :10119    
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>      swgeo            swpop            curfew        curfewstart       
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.00000   Length:16512      
#>  1st Qu.:0.0000   1st Qu.:1.0000   1st Qu.:0.00000   Class :character  
#>  Median :1.0000   Median :1.0000   Median :0.00000   Mode  :character  
#>  Mean   :0.6282   Mean   :0.9835   Mean   :0.04578                     
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:0.00000                     
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1.00000                     
#>                                                                        
#>   curfewend         vaccineexempt     ingathlim         outgathlim     
#>  Length:16512       Min.   :1       Min.   :    0.0   Min.   :    0.0  
#>  Class :character   1st Qu.:1       1st Qu.:   10.0   1st Qu.:   10.0  
#>  Mode  :character   Median :1       Median :   10.0   Median :   20.0  
#>                     Mean   :1       Mean   :  101.2   Mean   :  123.4  
#>                     3rd Qu.:1       3rd Qu.:   50.0   3rd Qu.:   50.0  
#>                     Max.   :1       Max.   :10000.0   Max.   :10000.0  
#>                     NA's   :16012   NA's   :14832     NA's   :14858    
#>  ingathlimreligious outgathlimreligious businessrestrictlevel
#>  Min.   :   0.00    Min.   :   0.00     Length:16512         
#>  1st Qu.:  10.00    1st Qu.:  10.00     Class :character     
#>  Median :  50.00    Median :  25.00     Mode  :character     
#>  Mean   :  77.07    Mean   :  87.74                          
#>  3rd Qu.: 100.00    3rd Qu.: 100.00                          
#>  Max.   :1000.00    Max.   :1000.00                          
#>  NA's   :15868      NA's   :16078                            
#>  schoolrestrictlevel publicmasklevel      dateissued        
#>  Length:16512        Length:16512       Min.   :2020-02-29  
#>  Class :character    Class :character   1st Qu.:2020-06-22  
#>  Mode  :character    Mode  :character   Median :2020-11-06  
#>                                         Mean   :2020-10-28  
#>                                         3rd Qu.:2021-02-24  
#>                                         Max.   :2021-07-31  
#>                                                             
#>   dateenacted           dateexpiry           dateended         
#>  Min.   :2020-02-29   Min.   :2020-03-21   Min.   :2020-04-24  
#>  1st Qu.:2020-06-24   1st Qu.:2020-07-07   1st Qu.:2021-03-22  
#>  Median :2020-11-09   Median :2020-11-28   Median :2021-05-24  
#>  Mean   :2020-10-29   Mean   :2020-11-12   Mean   :2021-04-10  
#>  3rd Qu.:2021-02-26   3rd Qu.:2021-03-06   3rd Qu.:2021-06-11  
#>  Max.   :2021-08-01   Max.   :2021-12-31   Max.   :2021-07-30  
#>                       NA's   :4714         NA's   :8539        
#>    extends            expands             eases               ends          
#>  Length:16512       Length:16512       Length:16512       Length:16512      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     joins              leaves           appliesto         policycodingnotes 
#>  Length:16512       Length:16512       Length:16512       Length:16512      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  policysource       policysourceid      lastupdated         lastupdatednotes  
#>  Length:16512       Length:16512       Min.   :2020-03-24   Length:16512      
#>  Class :character   Class :character   1st Qu.:2021-01-28   Class :character  
#>  Mode  :character   Mode  :character   Median :2021-04-20   Mode  :character  
#>                                        Mean   :2021-03-30                     
#>                                        3rd Qu.:2021-06-01                     
#>                                        Max.   :2021-09-04                     
#>                                                                               
#>    rereviewed     updatingsource1    updatingsource2   
#>  Min.   :0.0000   Length:16512       Length:16512      
#>  1st Qu.:1.0000   Class :character   Class :character  
#>  Median :1.0000   Mode  :character   Mode  :character  
#>  Mean   :0.9551                                        
#>  3rd Qu.:1.0000                                        
#>  Max.   :1.0000                                        
#>