United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Value

a data.frame

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = us_state_distancing_policy()

#> Warning:  1 failed to parse.

colnames(res)

#>  [1] "pid"                   "location_id"           "statefips"            
#>  [4] "iso2c"                 "state"                 "statepolicy"          
#>  [7] "mandate"               "statewide"             "swgeo"                
#> [10] "swpop"                 "curfew"                "curfewstart"          
#> [13] "curfewend"             "vaccineexempt"         "ingathlim"            
#> [16] "outgathlim"            "ingathlimreligious"    "outgathlimreligious"  
#> [19] "businessrestrictlevel" "schoolrestrictlevel"   "publicmasklevel"      
#> [22] "dateissued"            "dateenacted"           "dateexpiry"           
#> [25] "dateended"             "extends"               "expands"              
#> [28] "eases"                 "ends"                  "joins"                
#> [31] "leaves"                "appliesto"             "policycodingnotes"    
#> [34] "policysource"          "policysourceid"        "lastupdated"          
#> [37] "lastupdatednotes"      "rereviewed"            "updatingsource1"      
#> [40] "updatingsource2"      

library(dplyr)
dplyr::glimpse(res)

#> Rows: 14,388
#> Columns: 40
#> $ pid                   <chr> "AL0001", "AL0002", "AL0080", "AL0013", "AL0015"…
#> $ location_id           <int> 523, 523, 523, 523, 523, 523, 523, 523, 523, 523…
#> $ statefips             <chr> "00001", "00001", "00001", "00001", "00001", "00…
#> $ iso2c                 <chr> "AL", "AL", "AL", "AL", "AL", "AL", "AL", "AL", …
#> $ state                 <chr> "Alabama", "Alabama", "Alabama", "Alabama", "Ala…
#> $ statepolicy           <chr> "EmergDec", "SchoolClose", "BarRestrict", "GathR…
#> $ mandate               <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, …
#> $ statewide             <lgl> TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, T…
#> $ swgeo                 <int> 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ swpop                 <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ curfew                <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ curfewstart           <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ curfewend             <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ vaccineexempt         <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ ingathlim             <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlim            <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ ingathlimreligious    <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ outgathlimreligious   <int> NA, NA, NA, 25, NA, NA, NA, NA, 25, NA, NA, NA, …
#> $ businessrestrictlevel <chr> "", "", "TakeawayOnly", "", "TakeawayOnly", "", …
#> $ schoolrestrictlevel   <chr> "", "NoInPerson", "", "", "", "NoInPerson", "NoI…
#> $ publicmasklevel       <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ dateissued            <date> 2020-03-13, 2020-03-13, 2020-03-17, 2020-03-17,…
#> $ dateenacted           <date> 2020-03-13, 2020-03-19, 2020-03-17, 2020-03-17,…
#> $ dateexpiry            <date> 2020-05-12, 2020-04-06, 2020-03-24, NA, 2020-03…
#> $ dateended             <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 202…
#> $ extends               <chr> "", "", "", "", "", "", "AL0002", "", "", "", ""…
#> $ expands               <chr> "", "", "", "", "", "AL0002", "", "", "", "", ""…
#> $ eases                 <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ ends                  <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ joins                 <chr> "", "", "", "", "", "", "", "AL0080", "AL0013", …
#> $ leaves                <chr> "", "", "", "", "", "", "", "", "", "", "", "", …
#> $ appliesto             <chr> "", "", "Blount, St Clair, Shelby, Tuscaloosa, W…
#> $ policycodingnotes     <chr> "State of Public Health Emergency declared. Late…
#> $ policysource          <chr> "https://governor.alabama.gov/assets/2020/03/202…
#> $ policysourceid        <chr> "Alabama0001", "Alabama0002", "Alabama0004", "Al…
#> $ lastupdated           <date> 2021-01-08, 2021-01-06, 2021-01-06, 2021-01-06,…
#> $ lastupdatednotes      <chr> "Updated DateExpiry to 20200512", "Added SchoolR…
#> $ rereviewed            <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ updatingsource1       <chr> "https://covid19.alaska.gov/health-mandates/", "…
#> $ updatingsource2       <chr> "https://www.alabamapublichealth.gov/legal/order…

summary(res)

#>      pid             location_id     statefips            iso2c          
#>  Length:14388       Min.   :523.0   Length:14388       Length:14388      
#>  Class :character   1st Qu.:530.0   Class :character   Class :character  
#>  Mode  :character   Median :546.0   Mode  :character   Mode  :character  
#>                     Mean   :546.5                                        
#>                     3rd Qu.:560.0                                        
#>                     Max.   :573.0                                        
#>                                                                          
#>     state           statepolicy         mandate        statewide      
#>  Length:14388       Length:14388       Mode :logical   Mode :logical  
#>  Class :character   Class :character   FALSE:930       FALSE:5664     
#>  Mode  :character   Mode  :character   TRUE :13458     TRUE :8724     
#>                                                                       
#>                                                                       
#>                                                                       
#>                                                                       
#>      swgeo            swpop            curfew        curfewstart       
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.00000   Length:14388      
#>  1st Qu.:0.0000   1st Qu.:1.0000   1st Qu.:0.00000   Class :character  
#>  Median :1.0000   Median :1.0000   Median :0.00000   Mode  :character  
#>  Mean   :0.6228   Mean   :0.9824   Mean   :0.03837                     
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:0.00000                     
#>  Max.   :1.0000   Max.   :1.0000   Max.   :1.00000                     
#>                                                                        
#>   curfewend         vaccineexempt     ingathlim          outgathlim     
#>  Length:14388       Min.   :1       Min.   :    0.00   Min.   :    0.0  
#>  Class :character   1st Qu.:1       1st Qu.:   10.00   1st Qu.:   10.0  
#>  Mode  :character   Median :1       Median :   10.00   Median :   15.0  
#>                     Mean   :1       Mean   :   96.22   Mean   :  114.8  
#>                     3rd Qu.:1       3rd Qu.:   25.00   3rd Qu.:   50.0  
#>                     Max.   :1       Max.   :10000.00   Max.   :10000.0  
#>                     NA's   :14315   NA's   :12858      NA's   :12869    
#>  ingathlimreligious outgathlimreligious businessrestrictlevel
#>  Min.   :   0.00    Min.   :   0.00     Length:14388         
#>  1st Qu.:  10.00    1st Qu.:  10.00     Class :character     
#>  Median :  50.00    Median :  25.00     Mode  :character     
#>  Mean   :  71.29    Mean   :  70.65                          
#>  3rd Qu.: 100.00    3rd Qu.: 100.00                          
#>  Max.   :1000.00    Max.   :1000.00                          
#>  NA's   :13751      NA's   :13958                            
#>  schoolrestrictlevel publicmasklevel      dateissued        
#>  Length:14388        Length:14388       Min.   :2020-02-29  
#>  Class :character    Class :character   1st Qu.:2020-06-10  
#>  Mode  :character    Mode  :character   Median :2020-10-06  
#>                                         Mean   :2020-09-29  
#>                                         3rd Qu.:2021-01-13  
#>                                         Max.   :2021-04-20  
#>                                                             
#>   dateenacted           dateexpiry           dateended         
#>  Min.   :2020-02-29   Min.   :2020-03-21   Min.   :2020-04-24  
#>  1st Qu.:2020-06-11   1st Qu.:2020-07-02   1st Qu.:2020-09-25  
#>  Median :2020-10-06   Median :2020-11-05   Median :2021-03-03  
#>  Mean   :2020-09-30   Mean   :2020-10-23   Mean   :2020-12-26  
#>  3rd Qu.:2021-01-15   3rd Qu.:2021-02-07   3rd Qu.:2021-03-31  
#>  Max.   :2021-07-15   Max.   :2021-08-31   Max.   :2021-04-20  
#>                       NA's   :3626         NA's   :11792       
#>    extends            expands             eases               ends          
#>  Length:14388       Length:14388       Length:14388       Length:14388      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     joins              leaves           appliesto         policycodingnotes 
#>  Length:14388       Length:14388       Length:14388       Length:14388      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  policysource       policysourceid      lastupdated         lastupdatednotes  
#>  Length:14388       Length:14388       Min.   :2020-03-24   Length:14388      
#>  Class :character   Class :character   1st Qu.:2020-11-25   Class :character  
#>  Mode  :character   Mode  :character   Median :2021-01-30   Mode  :character  
#>                                        Mean   :2021-01-18                     
#>                                        3rd Qu.:2021-03-14                     
#>                                        Max.   :2021-04-21                     
#>                                                                               
#>    rereviewed     updatingsource1    updatingsource2   
#>  Min.   :0.0000   Length:14388       Length:14388      
#>  1st Qu.:1.0000   Class :character   Class :character  
#>  Median :1.0000   Mode  :character   Mode  :character  
#>  Mean   :0.9267                                        
#>  3rd Qu.:1.0000                                        
#>  Max.   :1.0000                                        
#>