United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Examples

res = us_state_distancing_policy()

#> Warning: Problem with `mutate()` input `DateEnacted`.
#> ℹ  8 failed to parse.
#> ℹ Input `DateEnacted` is `lubridate::ymd(DateEnacted)`.

#> Warning:  8 failed to parse.

colnames(res)

#>  [1] "PID"                                                    
#>  [2] "Counties on the County Monitoring List must close malls"
#>  [3] "StateFIPS"                                              
#>  [4] "iso2c"                                                  
#>  [5] "state"                                                  
#>  [6] "StatePolicy"                                            
#>  [7] "Mandate"                                                
#>  [8] "StateWide"                                              
#>  [9] "SWGeo"                                                  
#> [10] "SWPop"                                                  
#> [11] "Encourages"                                             
#> [12] "InGathLim"                                              
#> [13] "OutGathLim"                                             
#> [14] "InGathLimReligious"                                     
#> [15] "OutGathLimReligious"                                    
#> [16] "PublicMaskLevel"                                        
#> [17] "DateIssued"                                             
#> [18] "DateEnacted"                                            
#> [19] "DateExpiry"                                             
#> [20] "DateEnded"                                              
#> [21] "Extends"                                                
#> [22] "Expands"                                                
#> [23] "Eases"                                                  
#> [24] "Ends"                                                   
#> [25] "Depends"                                                
#> [26] "Joins"                                                  
#> [27] "PolicyCodingNotes"                                      
#> [28] "PolicySource"                                           
#> [29] "PolicySourceID"                                         
#> [30] "LastUpdated"                                            
#> [31] "LastUpdatedNotes"                                       
#> [32] "UpdatingSource1"                                        
#> [33] "UpdatingSource2"                                        

library(dplyr)
dplyr::glimpse(res)

#> Rows: 5,682
#> Columns: 33
#> $ PID                                                       <chr> "AL0001", "…
#> $ `Counties on the County Monitoring List must close malls` <dbl> 523, 523, 5…
#> $ StateFIPS                                                 <chr> "00001", "0…
#> $ iso2c                                                     <chr> "AL", "AL",…
#> $ state                                                     <chr> "Alabama", …
#> $ StatePolicy                                               <chr> "EmergDec",…
#> $ Mandate                                                   <lgl> TRUE, TRUE,…
#> $ StateWide                                                 <lgl> TRUE, TRUE,…
#> $ SWGeo                                                     <dbl> 1, 1, 0, 0,…
#> $ SWPop                                                     <dbl> 1, 1, 1, 1,…
#> $ Encourages                                                <dbl> NA, NA, NA,…
#> $ InGathLim                                                 <dbl> NA, NA, NA,…
#> $ OutGathLim                                                <dbl> NA, NA, NA,…
#> $ InGathLimReligious                                        <dbl> NA, NA, NA,…
#> $ OutGathLimReligious                                       <dbl> NA, NA, NA,…
#> $ PublicMaskLevel                                           <chr> NA, NA, NA,…
#> $ DateIssued                                                <date> 2020-03-13…
#> $ DateEnacted                                               <date> 2020-03-13…
#> $ DateExpiry                                                <date> NA, 2020-0…
#> $ DateEnded                                                 <date> NA, NA, NA…
#> $ Extends                                                   <chr> NA, NA, NA,…
#> $ Expands                                                   <chr> NA, NA, NA,…
#> $ Eases                                                     <chr> NA, NA, NA,…
#> $ Ends                                                      <chr> NA, NA, NA,…
#> $ Depends                                                   <chr> NA, NA, NA,…
#> $ Joins                                                     <chr> NA, NA, NA,…
#> $ PolicyCodingNotes                                         <chr> "Public Hea…
#> $ PolicySource                                              <chr> "https://go…
#> $ PolicySourceID                                            <chr> "Alabama000…
#> $ LastUpdated                                               <date> 2020-03-24…
#> $ LastUpdatedNotes                                          <chr> NA, "Update…
#> $ UpdatingSource1                                           <chr> "https://co…
#> $ UpdatingSource2                                           <chr> "https://ww…

summary(res)

#>      PID            Counties on the County Monitoring List must close malls
#>  Length:5682        Min.   :523.0                                          
#>  Class :character   1st Qu.:534.0                                          
#>  Mode  :character   Median :550.0                                          
#>                     Mean   :548.5                                          
#>                     3rd Qu.:561.0                                          
#>                     Max.   :575.0                                          
#>                                                                            
#>   StateFIPS            iso2c              state           StatePolicy       
#>  Length:5682        Length:5682        Length:5682        Length:5682       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>   Mandate        StateWide           SWGeo            SWPop       
#>  Mode :logical   Mode :logical   Min.   :0.0000   Min.   :0.0000  
#>  FALSE:571       FALSE:1630      1st Qu.:0.0000   1st Qu.:1.0000  
#>  TRUE :5111      TRUE :4052      Median :1.0000   Median :1.0000  
#>                                  Mean   :0.7341   Mean   :0.9776  
#>                                  3rd Qu.:1.0000   3rd Qu.:1.0000  
#>                                  Max.   :1.0000   Max.   :1.0000  
#>                                                                   
#>    Encourages     InGathLim         OutGathLim    InGathLimReligious
#>  Min.   :1      Min.   :    0.0   Min.   :    0   Min.   :   0.00   
#>  1st Qu.:1      1st Qu.:   10.0   1st Qu.:   10   1st Qu.:  10.00   
#>  Median :1      Median :   10.0   Median :   15   Median :  20.00   
#>  Mean   :1      Mean   :  127.6   Mean   :  137   Mean   :  63.28   
#>  3rd Qu.:1      3rd Qu.:   50.0   3rd Qu.:   50   3rd Qu.:  50.00   
#>  Max.   :1      Max.   :10000.0   Max.   :10000   Max.   :1000.00   
#>  NA's   :5621   NA's   :5092      NA's   :5093    NA's   :5261      
#>  OutGathLimReligious PublicMaskLevel      DateIssued        
#>  Min.   :   0.00     Length:5682        Min.   :2020-02-29  
#>  1st Qu.:  10.00     Class :character   1st Qu.:2020-04-30  
#>  Median :  25.00     Mode  :character   Median :2020-06-03  
#>  Mean   :  58.05                        Mean   :2020-06-03  
#>  3rd Qu.:  50.00                        3rd Qu.:2020-07-09  
#>  Max.   :1000.00                        Max.   :2020-09-01  
#>  NA's   :5291                                               
#>   DateEnacted           DateExpiry           DateEnded         
#>  Min.   :202-08-31    Min.   :2020-03-21   Min.   :2020-04-26  
#>  1st Qu.:2020-05-04   1st Qu.:2020-05-16   1st Qu.:2020-06-01  
#>  Median :2020-06-03   Median :2020-06-27   Median :2020-06-12  
#>  Mean   :2020-02-08   Mean   :2020-06-25   Mean   :2020-06-21  
#>  3rd Qu.:2020-07-08   3rd Qu.:2020-08-01   3rd Qu.:2020-07-17  
#>  Max.   :2020-09-14   Max.   :2021-06-30   Max.   :2020-09-14  
#>  NA's   :8            NA's   :2069         NA's   :5157        
#>    Extends            Expands             Eases               Ends          
#>  Length:5682        Length:5682        Length:5682        Length:5682       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>    Depends             Joins           PolicyCodingNotes  PolicySource      
#>  Length:5682        Length:5682        Length:5682        Length:5682       
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  PolicySourceID      LastUpdated         LastUpdatedNotes   UpdatingSource1   
#>  Length:5682        Min.   :2020-03-24   Length:5682        Length:5682       
#>  Class :character   1st Qu.:2020-06-11   Class :character   Class :character  
#>  Mode  :character   Median :2020-07-13   Mode  :character   Mode  :character  
#>                     Mean   :2020-07-07                                        
#>                     3rd Qu.:2020-08-02                                        
#>                     Max.   :2020-09-02                                        
#>                                                                               
#>  UpdatingSource2   
#>  Length:5682       
#>  Class :character  
#>  Mode  :character  
#>                    
#>                    
#>                    
#>