United States social distancing policies for COVID-19

An updated, curated set of social distancing and non-pharmaceutical interventions for states in the United States and District of Columbia.

us_state_distancing_policy()

Source

https://github.com/COVID19StatePolicy/SocialDistancing/

Kudos

Joe Wasserman: http://www.joewasserman.com/

See also

Examples

res = us_state_distancing_policy()

#> Warning: download failed
#>   web resource path: ‘https://raw.githubusercontent.com/COVID19StatePolicy/SocialDistancing/master/data/USstatesCov19distancingpolicyBETA.csv’
#>   local file path: ‘/Users/sdavis2/Library/Caches/sars2pack/file8f0a525bf2a8’
#>   reason: Not Found (HTTP 404).

#> Warning: bfcdownload() failed
#>   rid: BFC43
#>   file: ‘https://raw.githubusercontent.com/COVID19StatePolicy/SocialDistancing/master/data/USstatesCov19distancingpolicyBETA.csv’
#>   reason: download failed

#> Error: download failed; see warnings()

colnames(res)

#> [1] "location_name"    "date"             "metric"           "mean"            
#> [5] "infections_mean"  "upper"            "infections_upper" "lower"           
#> [9] "infections_lower"

library(dplyr)
dplyr::glimpse(res)

#> Rows: 439,690
#> Columns: 9
#> $ location_name    <chr> "Abruzzo", "Abruzzo", "Abruzzo", "Abruzzo", "Abruzzo…
#> $ date             <date> 2019-12-18, 2019-12-18, 2019-12-18, 2019-12-18, 201…
#> $ metric           <chr> "allbed", "ICUbed", "InvVen", "deaths", "admis", "ne…
#> $ mean             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ infections_mean  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ upper            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ infections_upper <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ lower            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ infections_lower <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…

summary(res)

#>  location_name           date               metric               mean         
#>  Length:439690      Min.   :2019-12-18   Length:439690      Min.   :   -25.0  
#>  Class :character   1st Qu.:2020-02-19   Class :character   1st Qu.:     0.0  
#>  Mode  :character   Median :2020-04-22   Mode  :character   Median :     0.0  
#>                     Mean   :2020-04-21                      Mean   :   302.6  
#>                     3rd Qu.:2020-06-23                      3rd Qu.:    11.0  
#>                     Max.   :2020-08-24                      Max.   :147040.1  
#>                                                                               
#>  infections_mean        upper           infections_upper       lower         
#>  Min.   :     0.0   Min.   :   -25.00   Min.   :     0.0   Min.   :   -25.0  
#>  1st Qu.:     0.0   1st Qu.:     0.00   1st Qu.:     0.0   1st Qu.:     0.0  
#>  Median :     0.0   Median :     0.00   Median :     0.0   Median :     0.0  
#>  Mean   :   263.6   Mean   :   477.84   Mean   :   607.1   Mean   :   228.7  
#>  3rd Qu.:     0.0   3rd Qu.:    20.55   3rd Qu.:     0.0   3rd Qu.:     3.5  
#>  Max.   :256964.9   Max.   :226971.33   Max.   :618152.5   Max.   :113181.8  
#>  NA's   :3751                           NA's   :3751                         
#>  infections_lower
#>  Min.   :     0  
#>  1st Qu.:     0  
#>  Median :     0  
#>  Mean   :   124  
#>  3rd Qu.:     0  
#>  Max.   :223565  
#>  NA's   :3751