Global case tracking from the World Health Organization (WHO)

Each row of data reports cumulative counts based on best reporting up to the moment published. Cases and deaths are reported for all affected countries.

who_cases()

Source

https://covid19.who.int/

https://covid19.who.int/WHO-COVID-19-global-data.csv

Value

a data.frame

See also

Author

Sean Davis seandavi@gmail.com

Examples

res = who_cases()
colnames(res)

#> [1] "date"       "iso2c"      "country"    "who_region" "subset"    
#> [6] "count"     

dplyr::glimpse(res)

#> Rows: 241,266
#> Columns: 6
#> $ date       <date> 2020-01-03, 2020-01-03, 2020-01-04, 2020-01-04, 2020-01-05…
#> $ iso2c      <chr> "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF", "AF",…
#> $ country    <chr> "Afghanistan", "Afghanistan", "Afghanistan", "Afghanistan",…
#> $ who_region <chr> "EMRO", "EMRO", "EMRO", "EMRO", "EMRO", "EMRO", "EMRO", "EM…
#> $ subset     <chr> "confirmed", "deaths", "confirmed", "deaths", "confirmed", …
#> $ count      <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…